### 2026-01-29 #9916 @editorialbot
[PRE REVIEW]: Inference Perf: A Benchmarking Tool for GenAI Inference
**Submitting author:** @achandrasekar (<a href="http://orcid.org/0009-0009-3481-9173">Ashok Chandrasekar</a>)
**Repository:** https://github.com/kubernetes-sigs/inference-perf
**Branch with paper.md** (empty if default branch):
**Version:** v0.3.0
