### 2026-01-30 f9001ac
Fix docs to reference std, actual field is std_dev (#335)
https://github.com/kubernetes-sigs/inference-perf/blob/main/inference_perf/config.py#L62
