### 2026-01-30 d2be18c
Update pdm dependencies
Updates pdm dependencies, infernece-perf version, and adds
tiktoken for kimi k2.5 support.
