### 2026-01-31 02d3c1a
Add Sachin to authors in JOSS paper (#334)
