### 2026-01-31 c4026e8
make slo headers configurable
