### 2026-01-31 a6bbd5b
make slo headers configurable
