### 2026-01-31 331f440
make slo headers configurable 2
