+++ inference_perf/config.py
@@ -32,7 +32,7 @@ class APIConfig(BaseModel):
     type: APIType = APIType.Completion
     streaming: bool = False
     headers: Optional[dict[str, str]] = None
-    slo_unit: str = None
+    slo_unit: Optional[str] = None
     slo_tpot_header: Optional[str] = None
     slo_ttft_header: Optional[str] = None
 
