### 2026-01-31 ffc0836
make slo headers configurable 3
