### 2026-01-31 cc74692
make slo headers configurable 7
