+++ inference_perf/reportgen/base.py
@@ -248,14 +248,14 @@ def summarize_requests(
                 ),
                 "time_per_output_token": summarize(
                     [
-                        (x.info.output_token_times[-1] - x.info.output_token_times[0]) / (len(x.info.output_token_times) - 1)
+                       x.tpot
                         for x in streamable
                         if len(x.info.output_token_times) > 1
                     ],
                     percentiles,
                 ),
                 "time_to_first_token": summarize(
-                    [x.info.output_token_times[0] - x.start_time for x in streamable], percentiles
+                    [x.ttft for x in streamable], percentiles
                 ),
                 "inter_token_latency": summarize(
                     [
