### 2026-01-31 8e5e23d
make slo headers configurable 9
