### 2026-01-31 e9e42df
precalculate ntpot in RequestLifecycleMetric
