### 2026-01-31 c1ec975
precalculate ntpot in RequestLifecycleMetric
