### 2026-02-01 b944533
precalculate ntpot
