+++ docs/config.md
@@ -22,15 +22,20 @@ This document provides complete documentation for all configuration options avai
 
 ### API Configuration
 
-Controls the API interaction behavior:
+Controls the API interaction behavior. If SLO headers are present, each request is evaluated for SLO compliance and SLO-related metrics are reported:
 
 ```yaml
 api:
-  type: completion  # API type (completion|chat) (default: completion), completion is the default since the chat API is not typically enabled on model servers such as vLLM by default without additional configuration.
-  streaming: false  # Enable/disable streaming (default: false), needs to be enabled for metrics like TTFT, ITL and TPOT to be measured
-  headers:          # Add custom http headers to the request sent to the inference server
+  type: completion             # API type (completion|chat). completion is default since chat may require extra server config
+  streaming: false             # Enable streaming for TTFT, ITL, and TPOT metrics
+  headers:                     # Optional custom HTTP headers
     x-inference-model: llama
     x-routing-strategy: round-robin
+    x-tpot-slo-ms: "2"
+    x-ttft-slo-ms: "1000"
+  slo_unit: "s"               # Optional SLO unit (e.g., ms, s) default is ms
+  slo_tpot_header: "x-tpot-slo-s"        # Optional header name for TPOT SLO Header default is x-tpot-slo-ms
+  slo_ttft_header: "x-ttft-slo-s"        # Optional header name for TTFT SLO Header default is x-ttft-slo-ms
 ```  
 
 ### Data Generation
@@ -53,12 +58,20 @@ data:
     mean: 50
     std_dev: 10
     total_count: 100
-  shared_prefix:
-    num_unique_system_prompts: 10     # Number of distinct shared prefixes (formerly num_groups)
-    num_users_per_system_prompt: 10   # Number of unique questions per shared prefix (formerly num_prompts_per_group)
-    system_prompt_len: 100            # Length of the shared prefix (in tokens)
-    question_len: 50                  # Length of the unique question part (in tokens)
-    output_len: 50                    # Target length for the model's generated output (in tokens)
+  shared_prefix:              # For shared_prefix type
+    num_groups: 10            # Number of shared prefix groups
+    num_prompts_per_group: 10 # Unique questions per group
+    system_prompt_len: 100    # Shared prefix length (tokens)
+    
+    question_len: 50          # Question length (tokens)
+    question_len_std: 5       # optional defaults to 0
+    question_len_min: 10      # optional defaults to 1
+    question_len_max: 10      # optional defaults to None
+
+    output_len: 50            # Target output length (tokens) 
+    output_len_std: 5       # optional defaults to 0
+    output_len_min: 10      # optional defaults to 1
+    output_len_max: 10      # optional defaults to None     
 ```
 
 ### Load Configuration
