### 2026-02-02 9267a56
test(loadgen): add unit tests for worker concurrency distribution
Add a new test suite to validate how LoadGenerator splits concurrency across workers.
The tests cover evenly divisible cases, remainder handling, and scenarios where concurrency is lower than the number of workers.
This also fixes Python 3.9 compatibility by replacing Type | None with Optional[Type], and adds lightweight test-time mocks for asyncio.TaskGroup and typing.TypeAlias so the suite runs cleanly on Python versions below 3.11.

Signed-off-by: Sathvik <Sathvik.S@ibm.com>
