### 2026-02-03 3404c0f
feat: add structured output support for vLLM backend
Add --guided-json flag to enable JSON schema enforcement during benchmarking.
This allows realistic benchmarking of production workloads that use
response_format constraints.

- Add ResponseFormat config struct with json_schema and json_object types
- Pass response_format to vLLM/OpenAI compatible endpoints
- Add --guided-json CLI flag to specify schema file
- Add examples/structured-output/ with sample config and schema

See vLLM docs: https://docs.vllm.ai/en/latest/serving/openai_compatible_server.html
