+++ inference_perf/main.py
@@ -11,6 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+import logging
 from argparse import ArgumentParser
 from inference_perf.analysis.analyze import analyze_reports
 from typing import List, Optional
@@ -22,6 +23,8 @@ from inference_perf.config import (
     MetricsClientType,
     ModelServerType,
     ReportConfig,
+    ResponseFormat,
+    ResponseFormatType,
     StandardLoadStage,
     ConcurrentLoadStage,
     read_config,
@@ -106,6 +109,11 @@ def main_cli() -> None:
     parser.add_argument(
         "--log-level", help="Logging level", default="INFO", choices=["DEBUG", "INFO", "WARNING", "ERROR", "CRITICAL"]
     )
+    parser.add_argument(
+        "--guided-json",
+        help="Path to JSON schema file for structured output (response_format)",
+        required=False,
+    )
     args = parser.parse_args()
 
     setup_logging(args.log_level)
@@ -118,6 +126,18 @@ def main_cli() -> None:
         parser.error("argument -c/--config_file is required when not using --analyze")
 
     config = read_config(args.config_file)
+
+    # Apply --guided-json CLI override if provided
+    if args.guided_json:
+        import json
+
+        with open(args.guided_json, "r") as f:
+            json_schema = json.load(f)
+        config.api.response_format = ResponseFormat(
+            type=ResponseFormatType.JSON_SCHEMA,
+            json_schema=json_schema,
+        )
+        logging.getLogger(__name__).info(f"Using structured output with schema from: {args.guided_json}")
 
     # Set stage rates to high values if using concurrent load type
     if config.load.type == LoadType.CONCURRENT:
