+++ examples/structured-output/config.yml
@@ -1,0 +1,50 @@
+# Example config for structured output benchmarking
+# Uses response_format to enforce JSON schema compliance
+
+load:
+  type: constant
+  interval: 15
+  stages:
+  - rate: 1
+    duration: 30
+
+api:
+  type: chat
+  streaming: false
+  # Structured output can also be configured here instead of CLI:
+  # response_format:
+  #   type: json_schema
+  #   name: my_schema  # Optional custom name (default: "structured_output")
+  #   json_schema:
+  #     type: object
+  #     properties:
+  #       name: {type: string}
+  #       age: {type: integer}
+  #     required: [name, age]
+
+server:
+  type: vllm
+  model_name: HuggingFaceTB/SmolLM2-135M-Instruct
+  base_url: http://0.0.0.0:8000
+
+tokenizer:
+  pretrained_model_name_or_path: HuggingFaceTB/SmolLM2-135M-Instruct
+
+data:
+  type: synthetic
+  input_distribution:
+    min: 10
+    max: 100
+    mean: 50
+    std_dev: 10
+  output_distribution:
+    min: 10
+    max: 100
+    mean: 50
+    std_dev: 10
+
+report:
+  request_lifecycle:
+    summary: true
+    per_stage: true
+    per_request: false
