### 2026-02-04 faed0a7
refactor: Remove multi-turn Chat API support per reviewer request
