+++ inference_perf/datagen/shared_prefix_datagen.py
@@ -4,7 +4,7 @@ import numpy as np
 
 from inference_perf.apis.base import InferenceAPIData, LazyLoadInferenceAPIData
 from inference_perf.apis.completion import CompletionAPIData
-from inference_perf.apis.user_session import LocalUserSession, UserSessionCompletionAPIData, UserSessionChatAPIData
+from inference_perf.apis.user_session import LocalUserSession, UserSessionCompletionAPIData
 from inference_perf.apis.chat import ChatCompletionAPIData, ChatMessage
 from inference_perf.config import APIConfig, APIType, DataConfig
 from inference_perf.utils.custom_tokenizer import CustomTokenizer
@@ -50,9 +50,6 @@ class SharedPrefixDataGenerator(DataGenerator, LazyLoadDataMixin):
         self.prompts: List[str] = []  # For completion API
         self.prompt_pairs: List[tuple[str, str]] = []  # (shared_prefix, question) pairs for chat API
         self.user_sessions: List[LocalUserSession] = []
-        # For multi-turn: store questions and shared_prefixes by group for correct selection
-        self.questions_by_group: List[List[str]] = []
-        self.shared_prefixes_by_group: List[str] = []
         self._generate_prompts()
 
     def get_supported_apis(self) -> List[APIType]:
@@ -68,53 +65,17 @@ class SharedPrefixDataGenerator(DataGenerator, LazyLoadDataMixin):
         return True
 
     def load_lazy_data(self, data: LazyLoadInferenceAPIData) -> InferenceAPIData:
+        i = data.data_index % len(self.prompts)
         if self.enable_multi_turn_chat:
             user_id = data.data_index % len(self.user_sessions)
             round = data.data_index // len(self.user_sessions)
-
-            # Get user session and its group_id (stored during creation, not calculated)
-            user_session = self.user_sessions[user_id]
-            group_id = user_session.group_id
-
-            # For each round, use a different question from the same group
-            # Cycle through questions in the same group
-            question_in_group = round % self.num_prompts_per_group
-            question_text = self.questions_by_group[group_id][question_in_group]
-            shared_prefix = self.shared_prefixes_by_group[group_id]
-
-            if self.api_config.type == APIType.Chat:
-                # Chat API + Multi-turn: Use UserSessionChatAPIData
-                # Note: system message is included but will be filtered in to_payload()
-                # since the actual system message comes from user_session.context (history)
-                messages = [ChatMessage(role="system", content=shared_prefix), ChatMessage(role="user", content=question_text)]
-                return UserSessionChatAPIData(
-                    messages=messages,
-                    max_tokens=self.output_len,
-                    user_session=user_session,
-                    target_round=round,
-                )
-            else:
-                # Completion API + Multi-turn: Use UserSessionCompletionAPIData
-                # Only pass the question, shared_prefix is already in user_session.context
-                return UserSessionCompletionAPIData(
-                    prompt=question_text,
-                    max_tokens=self.output_len,
-                    user_session=user_session,
-                    target_round=round,
-                )
-        else:
-            # Single-turn: use data_index directly
-            i = data.data_index % len(self.prompts)
-            if self.api_config.type == APIType.Chat:
-                shared_prefix, question = self.prompt_pairs[i]
-                messages = [ChatMessage(role="system", content=shared_prefix), ChatMessage(role="user", content=question)]
-                return ChatCompletionAPIData(messages=messages, max_tokens=self.output_len)
-            else:
-                return CompletionAPIData(prompt=self.prompts[i], max_tokens=self.output_len)
-
-    def get_request_by_index(self, n: int) -> InferenceAPIData:
-        i = n % len(self.prompts)
-        if self.api_config.type == APIType.Chat:
+            return UserSessionCompletionAPIData(
+                prompt=self.prompts[i],
[…]
