+++ inference_perf/apis/user_session.py
@@ -7,25 +7,22 @@ from aiohttp import ClientResponse
 from inference_perf.apis import CompletionAPIData, InferenceInfo
 from inference_perf.utils.custom_tokenizer import CustomTokenizer
 from inference_perf.config import APIConfig
-from inference_perf.apis.chat import ChatCompletionAPIData, ChatMessage
 
 logger = logging.getLogger(__name__)
 
 
 class LocalUserSession:
     user_session_id: str
-    context: list[ChatMessage] | str
-    group_id: int
+    context: str
 
-    def __init__(self, user_session_id: str, context: list[ChatMessage] | str = "", group_id: int = 0):
+    def __init__(self, user_session_id: str, context: str = ""):
         self.user_session_id = user_session_id
-        self.context = context if context else ""
-        self.group_id = group_id
+        self.contexts = context if context else ""
         self._current_round = 0
         self._in_flight: asyncio.Lock = asyncio.Lock()
         self._waiting_rounds: asyncio.Queue[asyncio.Future[bool]] = asyncio.Queue()
 
-    async def get_context(self, round: int) -> list[ChatMessage] | str:
+    async def get_context(self, round: int) -> str:
         if not self._waiting_rounds.empty() or self._in_flight.locked():
             # entering waiting queue
             future: asyncio.Future[bool] = asyncio.Future()
@@ -33,10 +30,10 @@ class LocalUserSession:
             await future
         await self._in_flight.acquire()
         self._current_round += 1
-        return self.context
+        return self.contexts
 
-    def update_context(self, response: list[ChatMessage] | str) -> None:
-        self.context = response
+    def update_context(self, response: str) -> None:
+        self.contexts = response
 
         if not self._waiting_rounds.empty():
             future = self._waiting_rounds.get_nowait()
@@ -53,8 +50,6 @@ class UserSessionCompletionAPIData(CompletionAPIData):
     async def to_payload(self, model_name: str, max_tokens: int, ignore_eos: bool, streaming: bool) -> dict[str, Any]:
         self._session_context = await self.user_session.get_context(self.target_round)
         # TODO: Currently, only prompt style (concat messages) support. Adding support for messages style payload.
-        if not isinstance(self._session_context, str):
-            raise TypeError("UserSessionCompletionAPIData expects string context")
         self.prompt = self._session_context + " " + self.prompt
         # TODO: The combined prompt (session context + current prompt) might exceed the model's
         #       maximum sequence length. Implement truncation logic/strategy to prevent
@@ -65,14 +60,21 @@ class UserSessionCompletionAPIData(CompletionAPIData):
         inference_info.extra_info["user_session"] = self.user_session.user_session_id
         inference_info.extra_info["chat_round"] = self.user_session._current_round
 
-    async def process_response(self, response: ClientResponse, config: APIConfig, tokenizer: CustomTokenizer) -> InferenceInfo:
+    async def process_response(
+        self, response: ClientResponse, config: APIConfig, tokenizer: CustomTokenizer, lora_adapter: Optional[str] = None
+    ) -> InferenceInfo:
         inference_info = await super().process_response(response, config, tokenizer)
         self.update_inference_info(inference_info)
         self.user_session.update_context(self.prompt + " " + self.model_response)
         return inference_info
 
     async def process_failure(
-        self, response: Optional[ClientResponse], config: APIConfig, tokenizer: CustomTokenizer, exception: Exception
+        self,
+        response: Optional[ClientResponse],
+        config: APIConfig,
+        tokenizer: CustomTokenizer,
+        exception: Exception,
+        lora_adapter: Optional[str] = None,
     ) -> Optional[InferenceInfo]:
[…]
