### 2026-02-04 022411c
refactor: Remove multi-turn Chat API support per reviewer request
