+++ inference_perf/datagen/shared_prefix_datagen.py
@@ -5,7 +5,7 @@ import numpy as np
 
 from inference_perf.apis.base import InferenceAPIData, LazyLoadInferenceAPIData
 from inference_perf.apis.completion import CompletionAPIData
-from inference_perf.apis.user_session import LocalUserSession, UserSessionCompletionAPIData, UserSessionChatAPIData
+from inference_perf.apis.user_session import LocalUserSession, UserSessionCompletionAPIData
 from inference_perf.apis.chat import ChatCompletionAPIData, ChatMessage
 from inference_perf.config import APIConfig, APIType, DataConfig, Distribution
 from inference_perf.utils.custom_tokenizer import CustomTokenizer
@@ -75,8 +75,8 @@ class SharedPrefixDataGenerator(DataGenerator, LazyLoadDataMixin):
             )
             self.output_len_list_per_group.append(output_lens.tolist())
 
-        self.prompts: List[str] = []  # For completion API
-        self.prompt_pairs: List[tuple[str, str]] = []  # (shared_prefix, question) pairs for chat API
+        self.prompts: List[str] = []
+        self.prompt_pairs: List[tuple[str, str]] = []  # (shared_prefix, question) pairs for Chat API
         self.user_sessions: List[LocalUserSession] = []
         self.flat_output_lens: List[int] = []
         self._generate_prompts()
@@ -108,27 +108,8 @@ class SharedPrefixDataGenerator(DataGenerator, LazyLoadDataMixin):
             )
         elif self.api_config.type == APIType.Chat:
             shared_prefix, question = self.prompt_pairs[i]
-            messages = [
-                ChatMessage(role="system", content=shared_prefix),
-                ChatMessage(role="user", content=question)
-            ]
-            return ChatCompletionAPIData(messages=messages, max_tokens=output_len)
-        else:
-            # Single-turn: use data_index directly
-            i = data.data_index % len(self.prompts)
-            if self.api_config.type == APIType.Chat:
-                shared_prefix, question = self.prompt_pairs[i]
-                messages = [ChatMessage(role="system", content=shared_prefix), ChatMessage(role="user", content=question)]
-                return ChatCompletionAPIData(messages=messages, max_tokens=self.output_len)
-            else:
-                return CompletionAPIData(prompt=self.prompts[i], max_tokens=self.output_len)
-
-    def get_request_by_index(self, n: int) -> InferenceAPIData:
-        i = n % len(self.prompts)
-        if self.api_config.type == APIType.Chat:
-            shared_prefix, question = self.prompt_pairs[i]
             messages = [ChatMessage(role="system", content=shared_prefix), ChatMessage(role="user", content=question)]
-            return ChatCompletionAPIData(messages=messages, max_tokens=self.output_len)
+            return ChatCompletionAPIData(messages=messages, max_tokens=output_len)
         else:
             return CompletionAPIData(prompt=self.prompts[i], max_tokens=output_len)
 
@@ -138,18 +119,9 @@ class SharedPrefixDataGenerator(DataGenerator, LazyLoadDataMixin):
 
         i = 0
         while True:
-            if self.enable_multi_turn_chat:
-                prefered_worker_id = i % self.num_groups
-                yield LazyLoadInferenceAPIData(data_index=i, prefered_worker_id=prefered_worker_id)
-                i += 1
-            elif self.api_config.type == APIType.Chat:
-                shared_prefix, question = self.prompt_pairs[i]
-                messages = [ChatMessage(role="system", content=shared_prefix), ChatMessage(role="user", content=question)]
-                yield ChatCompletionAPIData(messages=messages, max_tokens=self.output_len)
-                i = (i + 1) % len(self.prompts)
-            else:
[…]
