### 2026-02-04 cf804b7
refactor: Remove multi-turn Chat API support per reviewer request
