+++ inference_perf/apis/user_session.py
@@ -7,25 +7,22 @@ from aiohttp import ClientResponse
 from inference_perf.apis import CompletionAPIData, InferenceInfo
 from inference_perf.utils.custom_tokenizer import CustomTokenizer
 from inference_perf.config import APIConfig
-from inference_perf.apis.chat import ChatCompletionAPIData, ChatMessage
 
 logger = logging.getLogger(__name__)
 
 
 class LocalUserSession:
     user_session_id: str
-    context: list[ChatMessage] | str
-    group_id: int
+    context: str
 
-    def __init__(self, user_session_id: str, context: list[ChatMessage] | str = "", group_id: int = 0):
+    def __init__(self, user_session_id: str, context: str = ""):
         self.user_session_id = user_session_id
-        self.context = context if context else ""
-        self.group_id = group_id
+        self.contexts = context if context else ""
         self._current_round = 0
         self._in_flight: asyncio.Lock = asyncio.Lock()
         self._waiting_rounds: asyncio.Queue[asyncio.Future[bool]] = asyncio.Queue()
 
-    async def get_context(self, round: int) -> list[ChatMessage] | str:
+    async def get_context(self, round: int) -> str:
         if not self._waiting_rounds.empty() or self._in_flight.locked():
             # entering waiting queue
             future: asyncio.Future[bool] = asyncio.Future()
@@ -33,10 +30,10 @@ class LocalUserSession:
             await future
         await self._in_flight.acquire()
         self._current_round += 1
-        return self.context
+        return self.contexts
 
-    def update_context(self, response: list[ChatMessage] | str) -> None:
-        self.context = response
+    def update_context(self, response: str) -> None:
+        self.contexts = response
 
         if not self._waiting_rounds.empty():
             future = self._waiting_rounds.get_nowait()
@@ -53,8 +50,6 @@ class UserSessionCompletionAPIData(CompletionAPIData):
     async def to_payload(self, effective_model_name: str, max_tokens: int, ignore_eos: bool, streaming: bool) -> dict[str, Any]:
         self._session_context = await self.user_session.get_context(self.target_round)
         # TODO: Currently, only prompt style (concat messages) support. Adding support for messages style payload.
-        if not isinstance(self._session_context, str):
-            raise TypeError("UserSessionCompletionAPIData expects string context")
         self.prompt = self._session_context + " " + self.prompt
         # TODO: The combined prompt (session context + current prompt) might exceed the model's
         #       maximum sequence length. Implement truncation logic/strategy to prevent
@@ -88,78 +83,6 @@ class UserSessionCompletionAPIData(CompletionAPIData):
         return inference_info
 
 
-class UserSessionChatAPIData(ChatCompletionAPIData):
-    model_config = ConfigDict(arbitrary_types_allowed=True)
-    user_session: LocalUserSession = Field(exclude=True)
-    target_round: int
-
-    async def to_payload(self, model_name: str, max_tokens: int, ignore_eos: bool, streaming: bool) -> dict[str, Any]:
-        self._session_context = await self.user_session.get_context(self.target_round)
-        # Append current messages to the session context (history)
-        # self.messages contains the new user message for this turn (may include system)
-        # self._session_context contains the history (system prompt + previous turns)
-        if isinstance(self._session_context, list):
-            # History already exists, append only the new user message(s)
-            # Remove system from current messages if it exists (already in history)
-            new_messages = [msg for msg in self.messages if msg.role != "system"]
-            full_messages = self._session_context + new_messages
-        else:
-            # First turn: context is not a list yet, use all messages (including system)
-            full_messages = self.messages
-
-        # We temporarily override self.messages to generate payload, then restore?
-        # Or just construct payload manually.
-        # ChatCompletionAPIData.to_payload uses self.messages.
[…]
