+++ inference_perf/datagen/shared_prefix_datagen.py
@@ -5,7 +5,7 @@ import numpy as np
 
 from inference_perf.apis.base import InferenceAPIData, LazyLoadInferenceAPIData
 from inference_perf.apis.completion import CompletionAPIData
-from inference_perf.apis.user_session import LocalUserSession, UserSessionCompletionAPIData, UserSessionChatAPIData
+from inference_perf.apis.user_session import LocalUserSession, UserSessionCompletionAPIData
 from inference_perf.apis.chat import ChatCompletionAPIData, ChatMessage
 from inference_perf.config import APIConfig, APIType, DataConfig, Distribution
 from inference_perf.utils.custom_tokenizer import CustomTokenizer
@@ -45,7 +45,7 @@ class SharedPrefixDataGenerator(DataGenerator, LazyLoadDataMixin):
         self.num_prompts_per_group: int = self.shared_prefix.num_prompts_per_group
         self.system_prompt_len: int = self.shared_prefix.system_prompt_len
         self.enable_multi_turn_chat: bool = self.shared_prefix.enable_multi_turn_chat
-        
+
         # Use distribution configs, or fall back to question_len/output_len with std_dev=0
         q_len = self.shared_prefix.question_len
         o_len = self.shared_prefix.output_len
@@ -74,12 +74,9 @@ class SharedPrefixDataGenerator(DataGenerator, LazyLoadDataMixin):
                 self.shared_prefix.num_prompts_per_group,
             )
             self.output_len_list_per_group.append(output_lens.tolist())
-        
-        
 
-
-        self.prompts: List[str] = []  # For completion API
-        self.prompt_pairs: List[tuple[str, str]] = []  # (shared_prefix, question) pairs for chat API
+        self.prompts: List[str] = []
+        self.prompt_pairs: List[tuple[str, str]] = []  # (shared_prefix, question) pairs for Chat API
         self.user_sessions: List[LocalUserSession] = []
         self.flat_output_lens: List[int] = []
         self._generate_prompts()
@@ -111,27 +108,8 @@ class SharedPrefixDataGenerator(DataGenerator, LazyLoadDataMixin):
             )
         elif self.api_config.type == APIType.Chat:
             shared_prefix, question = self.prompt_pairs[i]
-            messages = [
-                ChatMessage(role="system", content=shared_prefix),
-                ChatMessage(role="user", content=question)
-            ]
-            return ChatCompletionAPIData(messages=messages, max_tokens=output_len)
-        else:
-            # Single-turn: use data_index directly
-            i = data.data_index % len(self.prompts)
-            if self.api_config.type == APIType.Chat:
-                shared_prefix, question = self.prompt_pairs[i]
-                messages = [ChatMessage(role="system", content=shared_prefix), ChatMessage(role="user", content=question)]
-                return ChatCompletionAPIData(messages=messages, max_tokens=self.output_len)
-            else:
-                return CompletionAPIData(prompt=self.prompts[i], max_tokens=self.output_len)
-
-    def get_request_by_index(self, n: int) -> InferenceAPIData:
-        i = n % len(self.prompts)
-        if self.api_config.type == APIType.Chat:
-            shared_prefix, question = self.prompt_pairs[i]
             messages = [ChatMessage(role="system", content=shared_prefix), ChatMessage(role="user", content=question)]
-            return ChatCompletionAPIData(messages=messages, max_tokens=self.output_len)
+            return ChatCompletionAPIData(messages=messages, max_tokens=output_len)
         else:
             return CompletionAPIData(prompt=self.prompts[i], max_tokens=output_len)
 
@@ -141,18 +119,9 @@ class SharedPrefixDataGenerator(DataGenerator, LazyLoadDataMixin):
 
         i = 0
[…]
