+++ tests/datagen/test_shared_prefix.py
@@ -1,0 +1,239 @@
+import pytest
+from unittest.mock import Mock
+
+from inference_perf.datagen.shared_prefix_datagen import SharedPrefixDataGenerator
+from inference_perf.apis.completion import CompletionAPIData
+from inference_perf.apis.chat import ChatCompletionAPIData
+from inference_perf.apis.user_session import UserSessionCompletionAPIData
+from inference_perf.apis.base import LazyLoadInferenceAPIData
+from inference_perf.config import APIConfig, APIType, DataConfig
+
+
+def create_mock_tokenizer() -> Mock:
+    """Create a mock tokenizer for testing."""
+    mock_tokenizer = Mock()
+    mock_hf_tokenizer = Mock()
+    mock_hf_tokenizer.vocab_size = 32000
+    mock_hf_tokenizer.decode = Mock(side_effect=lambda ids, **kwargs: f"text_{len(ids)}")
+    mock_tokenizer.get_tokenizer.return_value = mock_hf_tokenizer
+    return mock_tokenizer
+
+
+def create_api_config(api_type: APIType) -> APIConfig:
+    """Create an APIConfig for testing."""
+    return APIConfig(type=api_type)
+
+
+def create_data_config(
+    num_groups: int = 2,
+    num_prompts_per_group: int = 3,
+    system_prompt_len: int = 10,
+    question_len: int = 5,
+    output_len: int = 20,
+    enable_multi_turn_chat: bool = False,
+) -> DataConfig:
+    """Create a DataConfig with shared_prefix settings for testing."""
+    config = DataConfig()
+    config.shared_prefix = Mock()
+    config.shared_prefix.num_groups = num_groups
+    config.shared_prefix.num_prompts_per_group = num_prompts_per_group
+    config.shared_prefix.system_prompt_len = system_prompt_len
+    config.shared_prefix.question_len = question_len
+    config.shared_prefix.output_len = output_len
+    config.shared_prefix.enable_multi_turn_chat = enable_multi_turn_chat
+    return config
+
+
+class TestSharedPrefixDataGeneratorBasic:
+    """Basic tests for SharedPrefixDataGenerator."""
+
+    def test_get_supported_apis(self) -> None:
+        """Test that both Completion and Chat APIs are supported."""
+        api_config = create_api_config(APIType.Completion)
+        data_config = create_data_config()
+        tokenizer = create_mock_tokenizer()
+
+        generator = SharedPrefixDataGenerator(api_config, data_config, tokenizer)
+
+        supported = generator.get_supported_apis()
+        assert APIType.Completion in supported
+        assert APIType.Chat in supported
+
+    def test_prompts_count(self) -> None:
+        """Test that correct number of prompts are generated."""
+        api_config = create_api_config(APIType.Completion)
+        data_config = create_data_config(num_groups=2, num_prompts_per_group=3)
+        tokenizer = create_mock_tokenizer()
+
+        generator = SharedPrefixDataGenerator(api_config, data_config, tokenizer)
+
+        assert len(generator.prompts) == 6  # 2 groups * 3 prompts
+        assert len(generator.prompt_pairs) == 6
+
+    def test_prompt_pairs_structure(self) -> None:
+        """Test that prompt_pairs contain (shared_prefix, question) tuples."""
+        api_config = create_api_config(APIType.Chat)
+        data_config = create_data_config()
+        tokenizer = create_mock_tokenizer()
+
+        generator = SharedPrefixDataGenerator(api_config, data_config, tokenizer)
+
+        for shared_prefix, question in generator.prompt_pairs:
+            assert isinstance(shared_prefix, str)
+            assert isinstance(question, str)
+
+
+class TestSharedPrefixCompletionAPI:
+    """Tests for Completion API support."""
+
[…]
