### 2026-02-04 6100992
test: Add unit tests for SharedPrefixDataGenerator Also revert shuffle synchronization to match origin/main - only shuffle prompts list. Shuffle synchronization for prompt_pairs and user_sessions will be handled separately in PR #331.
