+++ paper/paper.bib
@@ -15,18 +15,28 @@
   url = {https://github.com/kubernetes-sigs/wg-serving}
 }
 
-@article{vllm,
-  title={vLLM: Easy, Fast, and Cheap LLM Serving with PagedAttention},
-  author={Kwon, Woosuk and Li, Zhuohan and Zhuang, Siyuan and Sheng, Ying and Zheng, Lianmin and Yu, Cody Hao and Rostaing, Joseph and Zhang, Hao and Song, Ion Stoica},
-  journal={arXiv preprint arXiv:2309.06180},
-  year={2023}
+@inproceedings{vllm,
+  author = {Kwon, Woosuk and Li, Zhuohan and Zhuang, Siyuan and Sheng, Ying and Zheng, Lianmin and Yu, Cody Hao and Rostaing, Joseph E. and Zhang, Hao and Hendry, Gilbert and Stoica, Ion},
+  title = {vLLM: Easy, Fast, and Cheap LLM Serving with PagedAttention},
+  year = {2023},
+  isbn = {9798400702233},
+  publisher = {Association for Computing Machinery},
+  address = {New York, NY, USA},
+  url = {https://doi.org/10.1145/3600006.3613165},
+  doi = {10.1145/3600006.3613165},
+  booktitle = {Proceedings of the 29th ACM Symposium on Operating Systems Principles},
+  pages = {611–626}
 }
 
-@article{sglang,
-  title={SGLang: Efficient Execution of Structured Language Model Programs},
-  author={Zheng, Lianmin and Li, Li and Zhang, Hao and Zhuang, Yonghao and Chen, Zhijie and Huang, Yanping and Morris, Meredith Ringel and Gonzalez, Joseph E and Stoica, Ion},
-  journal={arXiv preprint arXiv:2312.07104},
-  year={2023}
+@inproceedings{sglang,
+  title = {SGLang: Efficient Execution of Structured Language Model Programs},
+  author = {Zheng, Lianmin and Yin, Liangsheng and Xie, Zhiqiang and Sun, Chuyue and Huang, Jeff and Yu, Cody Hao and Cao, Shiyi and Kozyrakis, Christos and Stoica, Ion and Gonzalez, Joseph E. and Barrett, Clark and Sheng, Ying},
+  booktitle = {Advances in Neural Information Processing Systems 37},
+  pages = {62557--62583},
+  year = {2024},
+  publisher = {Curran Associates, Inc.},
+  doi = {10.52202/079017-2000},
+  url = {https://proceedings.neurips.cc/paper_files/paper/2024/file/724be4472168f31ba1c9ac630f15dec8-Paper-Conference.pdf}
 }
 
 @misc{tgi,
