+++ tests/test_config.py
@@ -1,4 +1,12 @@
-from inference_perf.config import read_config, deep_merge, Config, APIType, DataGenType, LoadType, MetricsClientType
+from inference_perf.config import (
+    APIType,
+    Config,
+    DataGenType,
+    LoadType,
+    MetricsClientType,
+    deep_merge,
+    read_config,
+)
 import os
 
 
@@ -36,12 +44,11 @@ def test_deep_merge() -> None:
 
 def test_shared_prefix_aliases() -> None:
     # Test using the short names (field names)
-    config_short = Config(
-        data={
-            "type": DataGenType.SharedPrefix,
-            "shared_prefix": {
-                "num_groups": 5,
-                "num_prompts_per_group": 20
+    config_short = Config.model_validate(
+        {
+            "data": {
+                "type": DataGenType.SharedPrefix,
+                "shared_prefix": {"num_groups": 5, "num_prompts_per_group": 20},
             }
         }
     )
@@ -50,12 +57,11 @@ def test_shared_prefix_aliases() -> None:
     assert config_short.data.shared_prefix.num_prompts_per_group == 20
 
     # Test using the long names (aliases)
-    config_long = Config(
-        data={
-            "type": DataGenType.SharedPrefix,
-            "shared_prefix": {
-                "num_unique_system_prompts": 7,
-                "num_users_per_system_prompt": 15
+    config_long = Config.model_validate(
+        {
+            "data": {
+                "type": DataGenType.SharedPrefix,
+                "shared_prefix": {"num_unique_system_prompts": 7, "num_users_per_system_prompt": 15},
             }
         }
     )
