### 2026-02-05 4f57a0b
Add aliases for shared_prefix config fields (#311)
Adds aliases for num_groups and num_prompts_per_group to
num_unique_system_prmpts and num_users_per_system_prompt, respectively.

Both original name and alias are supported, but the output config will
use the new aliases even if the input was the old name.

Also updates documentation to use new aliases and fixes #304 for more
consistent documentation.
