### 2026-02-06 #9916 @achandrasekar
Passes all the checks now. For reviewer recommendation, searching for Kubernetes + AI/ML experience from the [list](https://reviewers.joss.theoj.org/lookup) these folks might be a good fit: summitsingh, neerazz, varshaprasad96. PTAL.
