### 2026-02-06 e3e690b
Update transformers (#336)
Needed for newer model tokenizers (kimi 2.5)
