+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "inference-perf"
-version = "0.0.1"
+version = "0.4.0"
 description = "A GenAI inference performance benchmarking tool."
 authors = []
 dependencies = [
@@ -8,13 +8,14 @@ dependencies = [
     "pydantic>=2.10.6",
     "numpy>=2.2.2",
     "datasets>=3.3.2",
-    "transformers>=4.50.2",
+    "transformers==4.57.6",
     "google-cloud-storage>=3.1.0",
     "pyyaml>=6.0.2",
     "boto3>=1.39.0",
     "uvloop>=0.21.0",
     "tqdm>=4.67.1",
     "aiofiles>=25.1.0",
+    "tiktoken>=0.12.0",
 ]
 requires-python = ">=3.12"
 readme = "README.md"
