+++ inference_perf/reportgen/base.py
@@ -26,10 +26,16 @@ from inference_perf.client.metricsclient.prometheus_client import PrometheusMetr
 from inference_perf.client.requestdatacollector import RequestDataCollector
 from inference_perf.config import Config, PrometheusMetricsReportConfig, ReportConfig
 from inference_perf.utils import ReportFile
-from inference_perf.client.modelserver.openai_client import safe_float
 
 logger = logging.getLogger(__name__)
 
+def safe_float(value: Any) -> float:
+    """NOTE: Only for use in summarize_requests after validating safe access"""
+    try:
+        return float(value)
+    except (TypeError, ValueError):
+        return 0
+    
 def summarize(items: List[float], percentiles: List[float]) -> Optional[dict[str, float]]:
     if len(items) == 0:
         return None
@@ -49,82 +55,113 @@ class ResponsesSummary(BaseModel):
     successes: dict[str, Any]
     failures: dict[str, Any]
 
-def calculate_slo_metrics(metrics: List[RequestLifecycleMetric]) -> Optional[dict[str, Any]]:
+def calculate_slo_metrics(
+    metrics: List[RequestLifecycleMetric],
+    ttft_values: List[Optional[float]],
+    tpot_values: List[Optional[float]]
+) -> Optional[dict[str, Any]]:
     """
-    Calculate SLO attainment statistics from request metrics.
+    Calculate SLO attainment statistics from request metrics and pre-calculated latencies.
+    
+    Args:
+        metrics: List of successful RequestLifecycleMetric objects.
+        ttft_values: List of calculated TTFT values corresponding to metrics (same index). None if not streamable.
+        tpot_values: List of calculated TPOT values corresponding to metrics (same index). None if not streamable.
     
     Returns None if no SLO thresholds were set, otherwise returns dict with SLO metrics.
     """
-    # Filter out errors
-    valid_metrics = [m for m in metrics if m.error is None]
-    
-    if not valid_metrics:
+    if not metrics:
         return None
     
     # Check if any metrics have SLO fields populated
-    has_ttft_slo = any(m.ttft_slo_met is not None for m in valid_metrics)
-    has_tpot_slo = any(m.tpot_slo_met is not None for m in valid_metrics)
+    has_ttft_slo = any(m.ttft_slo_sec is not None for m in metrics)
+    has_tpot_slo = any(m.tpot_slo_sec is not None for m in metrics)
 
-    ttft_slo = max([m.ttft_slo for m in valid_metrics if m.ttft_slo is not None], default=None)
-    tpot_slo = max([m.tpot_slo for m in valid_metrics if m.tpot_slo is not None], default=None)
-    
-    # Calculate the sum of input and output tokens for all valid metrics that met SLO
+    ttft_slo_limit = max([m.ttft_slo_sec for m in metrics if m.ttft_slo_sec is not None], default=None)
+    tpot_slo_limit = max([m.tpot_slo_sec for m in metrics if m.tpot_slo_sec is not None], default=None)
    
     if not has_ttft_slo and not has_tpot_slo:
         return None  # No SLO tracking enabled
     
     slo_metrics: dict[str, Any] = {}
-    total = len(valid_metrics)
+    total = len(metrics)
     
-    total_benchmark_time = max(m.end_time for m in valid_metrics) - min(m.start_time for m in valid_metrics)
+    total_benchmark_time = max(m.end_time for m in metrics) - min(m.start_time for m in metrics)
     goodput_token_counts = []
-    for m in valid_metrics:
-        # Check presence of info and tokens
-        if not (m.info and m.info.input_tokens is not None and m.info.output_tokens is not None):
-            continue
-            
[…]

[…]
