+++ inference_perf/apis/chat.py
@@ -43,40 +43,15 @@ class ChatCompletionAPIData(InferenceAPIData):
     ) -> dict[str, Any]:
         if self.max_tokens == 0:
             self.max_tokens = max_tokens
-<<<<<<< HEAD
-<<<<<<< HEAD
-=======
->>>>>>> 9b789b2 (fix test errors)
-        return {
-            "model": effective_model_name,
-            "messages": [{"role": m.role, "content": m.content} for m in self.messages],
-            "max_tokens": self.max_tokens,
-            "ignore_eos": ignore_eos,
-            "stream": streaming,
-<<<<<<< HEAD
-        }
-=======
-        if streaming:
-                return {
+            return {
                     "model": effective_model_name,
                     "messages": [{"role": m.role, "content": m.content} for m in self.messages],
                     "max_tokens": self.max_tokens,
                     "ignore_eos": ignore_eos,
                     "stream": streaming,
-                    "stream_options": {"include_usage": "true"},
+                    **({"stream_options": {"include_usage": "true"}} if streaming else {}),
                 }
-        else:
-                return {
-                    "model": effective_model_name,
-                    "messages": [{"role": m.role, "content": m.content} for m in self.messages],
-                    "max_tokens": self.max_tokens,
-                    "ignore_eos": ignore_eos,
-                }
->>>>>>> 41bb6d3 (allow shared prefix question and system prompt variance and calculate slo metrics)
-=======
-            **({"stream_options": {"include_usage": "true"}} if streaming else {}),
-        }
->>>>>>> 9b789b2 (fix test errors)
+
 
     async def process_response(
         self, response: ClientResponse, config: APIConfig, tokenizer: CustomTokenizer, lora_adapter: Optional[str] = None
