+++ inference_perf/datagen/shared_prefix_datagen.py
@@ -153,4 +153,4 @@ class SharedPrefixDataGenerator(DataGenerator, LazyLoadDataMixin):
             random.shuffle(self.user_sessions)        
         else:
             random.shuffle(self.prompts)
-ß
+
