+++ inference_perf/config.py
@@ -86,12 +86,12 @@ class SharedPrefix(BaseModel):
     question_len: int = 50
     question_len_std: float = 0  # Variation in question length within a group
     question_len_min: int = 1
-    question_len_max: int = 2048
+    question_len_max: int = 32768
 
     output_len: int = 50
     output_len_std: float = 0  # Variation in output length within a group
     output_len_min: int = 1
-    output_len_max: int = 2048
+    output_len_max: int = 32768
     enable_multi_turn_chat: bool = False
 
 
