### 2026-02-10 b006b55
Add raw time series metric output.
Adds new report.prometheus configuration that outputs a raw_metrics
directory in the report direct. One prom file for each metric
that is available. Example:

reports-20260225-165238/
├── config.yaml
├── raw_metrics
│   ├── vllm:e2e_request_latency_seconds_count.prom
│   ├── vllm:generation_tokens_total.prom
│   ├── vllm:kv_cache_usage_perc.prom
│   ├── vllm:num_preemptions_total.prom
│   ├── vllm:num_requests_waiting.prom
│   └── vllm:prompt_tokens_total.prom
├── stage_0_lifecycle_metrics.json
├── stage_0_prometheus_metrics.json
├── summary_lifecycle_metrics
[…]
