### 2026-02-11 e7023e8
restore shared prefix defaults
