### 2026-02-12 1ab94a1
fix: use worker_id in request queue for ocncurrent load generation
