### 2026-02-12 89e4cb3
use dist class for shared prefix
