### 2026-02-13 fc877d3
allow shared prefix question and system prompt variance and calculate… (#301)
This pull request introduces several enhancements and new features to
the inference performance benchmarking and reporting framework. The main
focus is on supporting Service Level Objective (SLO) tracking for
latency metrics (TTFT and TPOT), making prompt and output length
distributions more flexible, and improving metric calculation and
reporting. The changes touch data models, configuration, data
generation, metric collection, and reporting.

**Key changes include:**

### SLO Tracking and Metric Enhancements

* Add
[…]
