+++ tests/reportgen/test_calculate_slo_metrics.py
@@ -1,0 +1,192 @@
+import pytest
+from unittest.mock import Mock
+from inference_perf.apis import RequestLifecycleMetric
+from inference_perf.reportgen.base import calculate_slo_metrics
+
+@pytest.fixture(scope="module", autouse=True)
+def test_calculate_slo_metrics_empty_metrics() -> None:
+    """Test with empty metrics list returns None."""
+    result = calculate_slo_metrics([], [], [])
+    assert result is None
+
+
+
+def test_calculate_slo_metrics_no_slo_thresholds() -> None:
+    """Test with metrics but no SLO thresholds returns None."""
+    metric = Mock(spec=RequestLifecycleMetric)
+    metric.ttft_slo_sec = None
+    metric.tpot_slo_sec = None
+    metric.start_time = 0.0
+    metric.end_time = 1.0
+    
+    result = calculate_slo_metrics([metric], [None], [None])
+    assert result is None
+
+
+
+def test_calculate_slo_metrics_ttft_slo_met() -> None:
+    """Test TTFT SLO calculation when SLO is met."""
+    metric = Mock(spec=RequestLifecycleMetric)
+    metric.ttft_slo_sec = 0.5
+    metric.tpot_slo_sec = None
+    metric.start_time = 0.0
+    metric.end_time = 2.0
+    metric.info = Mock()
+    metric.info.input_tokens = 10
+    metric.info.output_tokens = 5
+    
+    ttft_value = 0.3  # Below SLO
+    result = calculate_slo_metrics([metric], [ttft_value], [None])
+    
+    assert result is not None
+    assert "ttft_slo" in result
+    assert result["ttft_slo"]["attainment_pct"] == 100.0
+    assert result["ttft_slo"]["requests_met"] == 1
+    assert result["ttft_slo"]["requests_failed"] == 0
+    assert result["ttft_slo"]["slo"] == 0.5
+
+
+
+def test_calculate_slo_metrics_ttft_slo_failed() -> None:
+    """Test TTFT SLO calculation when SLO is not met."""
+    metric = Mock(spec=RequestLifecycleMetric)
+    metric.ttft_slo_sec = 0.2
+    metric.tpot_slo_sec = None
+    metric.start_time = 0.0
+    metric.end_time = 2.0
+    metric.info = Mock()
+    metric.info.input_tokens = 10
+    metric.info.output_tokens = 5
+    
+    ttft_value = 0.5  # Above SLO
+    result = calculate_slo_metrics([metric], [ttft_value], [None])
+    
+    assert result is not None
+    assert result["ttft_slo"]["attainment_pct"] == 0.0
+    assert result["ttft_slo"]["requests_met"] == 0
+    assert result["ttft_slo"]["requests_failed"] == 1
+
+
+
+def test_calculate_slo_metrics_ttft_slo_missing_value() -> None:
+    """Test TTFT SLO when ttft value is None (non-streamable)."""
+    metric = Mock(spec=RequestLifecycleMetric)
+    metric.ttft_slo_sec = 0.5
+    metric.tpot_slo_sec = None
+    metric.start_time = 0.0
+    metric.end_time = 2.0
+    metric.info = Mock()
+    metric.info.input_tokens = 10
+    metric.info.output_tokens = 5
+    
+    result = calculate_slo_metrics([metric], [None], [None])
+    
+    assert result is not None
+    assert result["ttft_slo"]["attainment_pct"] == 100.0
+    assert result["ttft_slo"]["requests_failed"] == 0
+
+
+
+def test_calculate_slo_metrics_tpot_slo_met() -> None:
+    """Test TPOT SLO calculation when SLO is met."""
+    metric = Mock(spec=RequestLifecycleMetric)
[…]
