+++ inference_perf/apis/chat.py
@@ -44,12 +44,14 @@ class ChatCompletionAPIData(InferenceAPIData):
         if self.max_tokens == 0:
             self.max_tokens = max_tokens
         return {
-            "model": effective_model_name,
-            "messages": [{"role": m.role, "content": m.content} for m in self.messages],
-            "max_tokens": self.max_tokens,
-            "ignore_eos": ignore_eos,
-            "stream": streaming,
-        }
+                    "model": effective_model_name,
+                    "messages": [{"role": m.role, "content": m.content} for m in self.messages],
+                    "max_tokens": self.max_tokens,
+                    "ignore_eos": ignore_eos,
+                    "stream": streaming,
+                    **({"stream_options": {"include_usage": "true"}} if streaming else {}),
+                }
+
 
     async def process_response(
         self, response: ClientResponse, config: APIConfig, tokenizer: CustomTokenizer, lora_adapter: Optional[str] = None
