+++ inference_perf/client/modelserver/sglang_client.py
@@ -211,4 +211,69 @@ class SGlangModelServerClient(openAIModelServerClient):
             num_requests_swapped=None,
             prefix_cache_hits=None,
             prefix_cache_queries=None,
+            avg_num_requests_running=None,
+            avg_request_queue_time=None,
+            median_request_queue_time=None,
+            p90_request_queue_time=None,
+            p99_request_queue_time=None,
+            avg_request_inference_time=None,
+            median_request_inference_time=None,
+            p90_request_inference_time=None,
+            p99_request_inference_time=None,
+            avg_request_prefill_time=None,
+            median_request_prefill_time=None,
+            p90_request_prefill_time=None,
+            p99_request_prefill_time=None,
+            avg_request_decode_time=None,
+            median_request_decode_time=None,
+            p90_request_decode_time=None,
+            p99_request_decode_time=None,
+            avg_request_prompt_tokens=None,
+            median_request_prompt_tokens=None,
+            p90_request_prompt_tokens=None,
+            p99_request_prompt_tokens=None,
+            avg_request_generation_tokens=None,
+            median_request_generation_tokens=None,
+            p90_request_generation_tokens=None,
+            p99_request_generation_tokens=None,
+            avg_request_max_num_generation_tokens=None,
+            median_request_max_num_generation_tokens=None,
+            p90_request_max_num_generation_tokens=None,
+            p99_request_max_num_generation_tokens=None,
+            avg_request_params_n=None,
+            median_request_params_n=None,
+            p90_request_params_n=None,
+            p99_request_params_n=None,
+            avg_request_params_max_tokens=None,
+            median_request_params_max_tokens=None,
+            p90_request_params_max_tokens=None,
+            p99_request_params_max_tokens=None,
+            request_success_count=None,
+            avg_iteration_tokens=None,
+            median_iteration_tokens=None,
+            p90_iteration_tokens=None,
+            p99_iteration_tokens=None,
+            prompt_tokens_cached=None,
+            prompt_tokens_recomputed=None,
+            external_prefix_cache_hits=None,
+            external_prefix_cache_queries=None,
+            mm_cache_hits=None,
+            mm_cache_queries=None,
+            corrupted_requests=None,
+            avg_request_prefill_kv_computed_tokens=None,
+            median_request_prefill_kv_computed_tokens=None,
+            p90_request_prefill_kv_computed_tokens=None,
+            p99_request_prefill_kv_computed_tokens=None,
+            avg_kv_block_idle_before_evict=None,
+            median_kv_block_idle_before_evict=None,
+            p90_kv_block_idle_before_evict=None,
+            p99_kv_block_idle_before_evict=None,
+            avg_kv_block_lifetime=None,
+            median_kv_block_lifetime=None,
+            p90_kv_block_lifetime=None,
+            p99_kv_block_lifetime=None,
+            avg_kv_block_reuse_gap=None,
+            median_kv_block_reuse_gap=None,
+            p90_kv_block_reuse_gap=None,
+            p99_kv_block_reuse_gap=None,
         )
