+++ inference_perf/client/modelserver/vllm_client.py
@@ -61,12 +61,21 @@ class vLLMModelServerClient(openAIModelServerClient):
 
     def get_prometheus_metric_metadata(self) -> PrometheusMetricMetadata:
         return PrometheusMetricMetadata(
+            # Queue Length
             avg_queue_length=ModelServerPrometheusMetric(
                 "vllm:num_requests_waiting",
                 "mean",
                 "gauge",
                 self.metric_filters,
             ),
+            # Running Requests
+            avg_num_requests_running=ModelServerPrometheusMetric(
+                "vllm:num_requests_running",
+                "mean",
+                "gauge",
+                self.metric_filters,
+            ),
+            # Time to First Token
             avg_time_to_first_token=ModelServerPrometheusMetric(
                 "vllm:time_to_first_token_seconds",
                 "mean",
@@ -91,66 +100,102 @@ class vLLMModelServerClient(openAIModelServerClient):
                 "histogram",
                 self.metric_filters,
             ),
+            # Time per Output Token (renamed in v1)
             avg_time_per_output_token=ModelServerPrometheusMetric(
-                "vllm:time_per_output_token_seconds",
+                "vllm:request_time_per_output_token_seconds",
                 "mean",
                 "histogram",
                 self.metric_filters,
             ),
             median_time_per_output_token=ModelServerPrometheusMetric(
-                "vllm:time_per_output_token_seconds",
+                "vllm:request_time_per_output_token_seconds",
                 "median",
                 "histogram",
                 self.metric_filters,
             ),
             p90_time_per_output_token=ModelServerPrometheusMetric(
-                "vllm:time_per_output_token_seconds",
+                "vllm:request_time_per_output_token_seconds",
                 "p90",
                 "histogram",
                 self.metric_filters,
             ),
             p99_time_per_output_token=ModelServerPrometheusMetric(
-                "vllm:time_per_output_token_seconds",
+                "vllm:request_time_per_output_token_seconds",
                 "p99",
                 "histogram",
                 self.metric_filters,
             ),
+            # Inter-Token Latency (new in v1)
+            avg_inter_token_latency=ModelServerPrometheusMetric(
+                "vllm:inter_token_latency_seconds",
+                "mean",
+                "histogram",
+                self.metric_filters,
+            ),
+            median_inter_token_latency=ModelServerPrometheusMetric(
+                "vllm:inter_token_latency_seconds",
+                "median",
+                "histogram",
+                self.metric_filters,
+            ),
+            p90_inter_token_latency=ModelServerPrometheusMetric(
+                "vllm:inter_token_latency_seconds",
+                "p90",
+                "histogram",
+                self.metric_filters,
+            ),
+            p99_inter_token_latency=ModelServerPrometheusMetric(
+                "vllm:inter_token_latency_seconds",
+                "p99",
+                "histogram",
+                self.metric_filters,
+            ),
+            # Prompt Tokens (renamed in v1: prompt_tokens_total -> prompt_tokens)
             avg_prompt_tokens=ModelServerPrometheusMetric(
-                "vllm:prompt_tokens_total",
+                "vllm:prompt_tokens",
                 "mean",
                 "counter",
                 self.metric_filters,
             ),
             prompt_tokens_per_second=ModelServerPrometheusMetric(
-                "vllm:prompt_tokens_total",
+                "vllm:prompt_tokens",
                 "rate",
                 "counter",
                 self.metric_filters,
             ),
+            # Generation Tokens (renamed in v1: generation_tokens_total -> generation_tokens)
             avg_output_tokens=ModelServerPrometheusMetric(
-                "vllm:generation_tokens_total",
+                "vllm:generation_tokens",
                 "mean",
                 "counter",
                 self.metric_filters,
             ),
             output_tokens_per_second=ModelServerPrometheusMetric(
[…]

[…]
