+++ inference_perf/client/modelserver/openai_client.py
@@ -30,7 +30,7 @@ import ssl
 logger = logging.getLogger(__name__)
     
 class openAIModelServerClient(ModelServerClient):
-    _session: "openAIModelServerClientSession | None" = None
+    _session: "Optional[openAIModelServerClientSession]" = None
     _session_lock = asyncio.Lock()
 
     def __init__(
