### 2026-02-15 6daca23
feat: added configurable base seed for loadgen workers
