### 2026-02-16 c6a3f63
Update title and URL for wg-serving reference
