### 2026-02-23 931d5ee
fix: restart ShareGPT iterator when exhausted
