### 2026-02-24 fccf5e4
feat: vLLM latest (0.15.0)  production metrics (#348)
<!--  Thanks for sending a pull request!  Here are some tips for you:

1. If this is your first time, please read our contributor guidelines:
https://git.k8s.io/community/contributors/guide#your-first-contribution
and developer guide
https://git.k8s.io/community/contributors/devel/development.md#development-guide
2. Please label this pull request according to what type of issue you
are addressing, especially if this is a release targeted pull request.
For reference on required PR/issue labels, read here:

https://git.k8s.io/community/contri
[…]
