### 2026-02-24 599d379
feat: added configurable base seed for loadgen workers (#349)
<!--  Thanks for sending a pull request!  Here are some tips for you:

1. If this is your first time, please read our contributor guidelines:
https://git.k8s.io/community/contributors/guide#your-first-contribution
and developer guide
https://git.k8s.io/community/contributors/devel/development.md#development-guide
2. Please label this pull request according to what type of issue you
are addressing, especially if this is a release targeted pull request.
For reference on required PR/issue labels, read here:

https://git.k8s.io/communit
[…]
