+++ inference_perf/reportgen/base.py
@@ -29,13 +29,15 @@ from inference_perf.utils import ReportFile
 
 logger = logging.getLogger(__name__)
 
+
 def safe_float(value: Any) -> float:
     """NOTE: Only for use in summarize_requests after validating safe access"""
     try:
         return float(value)
     except (TypeError, ValueError):
         return 0.0
-    
+
+
 def summarize(items: List[float], percentiles: List[float]) -> Optional[dict[str, float]]:
     if len(items) == 0:
         return None
@@ -55,111 +57,107 @@ class ResponsesSummary(BaseModel):
     successes: dict[str, Any]
     failures: dict[str, Any]
 
+
 def calculate_slo_metrics(
-    metrics: List[RequestLifecycleMetric],
-    ttft_values: List[Optional[float]],
-    tpot_values: List[Optional[float]]
+    metrics: List[RequestLifecycleMetric], ttft_values: List[Optional[float]], tpot_values: List[Optional[float]]
 ) -> Optional[dict[str, Any]]:
     """
     Calculate SLO attainment statistics from request metrics and pre-calculated latencies.
-    
+
     Args:
         metrics: List of successful RequestLifecycleMetric objects.
         ttft_values: List of calculated TTFT values corresponding to metrics (same index). None if not streamable.
         tpot_values: List of calculated TPOT values corresponding to metrics (same index). None if not streamable.
-    
+
     Returns None if no SLO thresholds were set, otherwise returns dict with SLO metrics.
     """
     if not metrics:
         return None
-    
+
     # Check if any metrics have SLO fields populated
     has_ttft_slo = any(m.ttft_slo_sec is not None for m in metrics)
     has_tpot_slo = any(m.tpot_slo_sec is not None for m in metrics)
 
     ttft_slo_limit = max([m.ttft_slo_sec for m in metrics if m.ttft_slo_sec is not None], default=None)
     tpot_slo_limit = max([m.tpot_slo_sec for m in metrics if m.tpot_slo_sec is not None], default=None)
-   
+
     if not has_ttft_slo and not has_tpot_slo:
         return None  # No SLO tracking enabled
-    
+
     slo_metrics: dict[str, Any] = {}
     total = len(metrics)
-    
+
     total_benchmark_time = max(m.end_time for m in metrics) - min(m.start_time for m in metrics)
     goodput_token_counts = []
-    
+
     ttft_results = []
     tpot_results = []
-    
+
     # Iterate through metrics and their corresponding calculated values.
     # Using zip(..., strict=True) ensures that all sequences have the same length
     # and prevents misalignment or IndexError if they do not.
     for m, ttft, tpot in zip(metrics, ttft_values, tpot_values, strict=True):
-        
         # Check TTFT SLO (Only if streamable / ttft exists)
         ttft_met = None
         if m.ttft_slo_sec is not None:
             if ttft is not None:
-                ttft_met = (ttft <= m.ttft_slo_sec)
+                ttft_met = ttft <= m.ttft_slo_sec
             else:
-                ttft_met = True # If TTFT SLO is set but we don't have a TTFT value (non-streamable), we can choose to consider it as met or not. Here we consider it met to avoid penalizing non-streamable requests.
+                ttft_met = True  # If TTFT SLO is set but we don't have a TTFT value (non-streamable), we can choose to consider it as met or not. Here we consider it met to avoid penalizing non-streamable requests.
         ttft_results.append(ttft_met)
-        
+
         # Check TPOT SLO (Only if streamable / tpot exists)
         tpot_met = None
         if m.tpot_slo_sec is not None:
             if tpot is not None:
-                tpot_met = (tpot <= m.tpot_slo_sec)
+                tpot_met = tpot <= m.tpot_slo_sec
             else:
-                tpot_met = True # Similar logic as TTFT for non-streamable requests
[…]
