+++ inference_perf/apis/user_session.py
@@ -47,7 +47,9 @@ class UserSessionCompletionAPIData(CompletionAPIData):
     user_session: LocalUserSession = Field(exclude=True)
     target_round: int
 
-    async def to_payload(self, effective_model_name: str, max_tokens: int, ignore_eos: bool, streaming: bool) -> dict[str, Any]:
+    async def to_payload(
+        self, effective_model_name: str, max_tokens: int, ignore_eos: bool, streaming: bool
+    ) -> dict[str, Any]:
         self._session_context = await self.user_session.get_context(self.target_round)
         # TODO: Currently, only prompt style (concat messages) support. Adding support for messages style payload.
         self.prompt = self._session_context + " " + self.prompt
