### 2026-02-25 f0bb44b
enforce monotonically increasing code coverage
