### 2026-02-25 dd2df75
simplify code coverage check script
