### 2026-02-26 3d69ebf
fix: handle StopIteration when ShareGPT dataset is exhausted
When load generation rate * duration exceeds the ShareGPT dataset size,
the streaming iterator raises StopIteration. itertools.cycle() doesn't
work with streaming datasets because they can only be consumed once.

This fix implements manual cycling by:
- Extracting dataset loading to _load_dataset() method
- Adding _next_data() helper that catches StopIteration and reloads
- Removing itertools.cycle() in favor of explicit reloading

Fixes kubernetes-sigs#100
