### 2026-02-26 ab4b101
test: add unit test for StopIteration handling in HFShareGPTDataGenerator
- Test that _next_data() reloads dataset when StopIteration is raised
- Test normal iteration behavior when dataset is not exhausted
- Verifies the fix for dataset exhaustion cycling
