### 2026-02-26 9d428aa
add experimental shared_prefix trace reading capability
