+++ inference_perf/config.py
@@ -40,6 +40,7 @@ class APIConfig(BaseModel):
 
 class TraceFormat(Enum):
     AZURE_PUBLIC_DATASET = "AzurePublicDataset"
+    SHARED_PREFIX = "SharedPrefix"
 
 
 class TraceConfig(BaseModel):
