### 2026-02-27 2747298
Assert monotonically increasing code coverage (#357)
Fixes: #345

To validate locally, run `pdm check:cov`

Example output:
```
Run pdm run check:cov
--- Generating baseline coverage from 'main' ---
Setting up fresh virtualenv for baseline...
Installing dependencies in baseline environment...
Running tests on main branch...
✅ Baseline generated: coverage_main.json
--- Generating current coverage ---
✅ Current report generated: coverage.json

--- Coverage Results ---
Main Branch:    43.66%
Current Branch: 43.66%
✅ PASS: Coverage is maintained or improved.
```
