### 2026-02-27 16a3c77
Fix saturation detection and harden load generator (Clean)
- Fix 1-second misalignment in saturation detection logic

- Add robust test suite for saturation detection
