+++ inference_perf/loadgen/load_generator.py
@@ -59,6 +59,7 @@ else:
     TypeAlias = Any
 from types import FrameType
 import time
+import math
 import multiprocessing as mp
 from multiprocessing.synchronize import Event as SyncEvent
 from multiprocessing.sharedctypes import Synchronized
@@ -303,7 +304,7 @@ class LoadGenerator:
         self,
         stage_id: int,
         rate: float,
-        duration: int,
+        duration: float,
         request_queue: RequestQueue[RequestQueueData],
         active_requests_counter: "Synchronized[int]",
         finished_requests_counter: "Synchronized[int]",
@@ -327,10 +328,10 @@ class LoadGenerator:
         start_time_epoch = time.time()
         start_time = time.perf_counter() + 1
 
-        if self.datagen.trace is not None:
+        if self.datagen.trace is not None and stage_id >= 0:
             num_requests = self.datagen.get_request_count()
         else:
-            num_requests = int(rate * duration)
+            num_requests = math.ceil(rate * duration)
 
         stage_status = StageStatus.RUNNING
 
@@ -413,43 +414,173 @@ class LoadGenerator:
         cancel_signal: SyncEvent,
     ) -> None:
         """
-        Runs a preliminary load test to automatically determine the server's saturation point
-        and generate a suitable series of load stages for the main benchmark.
+        Runs a Stepped Load (Ramping) test to determine the server's saturation point.
 
-        An aggregator task samples the active requests and then the burn down rate is
-        calculated from the samples. Saturation is derived from a percentile of the
-        sampled burn down rates.
+        Iterates through increasing request rates (steps). For each step:
+        1. Runs for a specific duration.
+        2. Measures the actual throughput (finished requests / time).
+        3. Checks if the system is saturated (Actual Throughput < Target Rate * tolerance).
+
+        If saturation is detected, the previous successful rate is used as the saturation point.
         """
-        logger.info("Running preprocessing stage")
-        results: List[Tuple[float, int]] = []
+        logger.info("Running preprocessing stage (Stepped Load)")
 
         if self.sweep_config is None:
             raise Exception("sweep_config cannot be none")
 
-        # Aggregator collects timestamped value of active_requests throughout the preprocessing
-        async def aggregator() -> None:
+        # Adaptive Saturation Detection (Exponential Search + Binary Refinement)
+        # Phase 1: Probe at moderate rate
+        step_duration = 10
+        if self.sweep_config.stage_duration > 10:
+            step_duration = self.sweep_config.stage_duration
+
+        throughput_tolerance = 0.90
+        max_rate_cap = 50000.0
+
+        samples: list[tuple[float, int]] = []
+
+        # Start aggregator for this step
+        async def aggregator(results: List[Tuple[float, int]] = samples) -> None:
             while True:
-                results.append((time.perf_counter(), active_requests_counter.value))
+                results.append((time.perf_counter(), finished_requests_counter.value))
                 await sleep(0.5)
 
         aggregator_task = create_task(aggregator())
 
-        stage_id = -1
-        duration = 5
-        rate = self.sweep_config.num_requests / duration
-        timeout = self.sweep_config.timeout
-        start_time = time.perf_counter()
-        await self.run_stage(
-            stage_id,
-            rate,
-            duration,
-            request_queue,
-            active_requests_counter,
-            finished_requests_counter,
-            request_phase,
-            timeout=timeout,
-            cancel_signal=cancel_signal,
-        )
+        lower_bound = 0.0
+        upper_bound = 0.0
+        saturation_point = 0.0
+        found_saturation = False
+
+        async def measure_rate(target_rate: float) -> tuple[bool, float]:
+            nonlocal samples
+            samples.clear()
+            with finished_requests_counter.get_lock():
[…]
