### 2026-03-01 c9fd630
fix: Ensure aiohttp response bodies are always consumed in `openai_client` to prevent connection leaks
