+++ inference_perf/client/modelserver/openai_client.py
@@ -107,6 +107,7 @@ class openAIModelServerClient(ModelServerClient):
                 self._session = openAIModelServerClientSession(self)
             session = self._session
         await session.process_request(data, stage_id, scheduled_time, lora_adapter)
+
     async def close(self) -> None:
         """Close the internal session created by process_request, if any."""
         if self._session is not None:
@@ -170,14 +171,15 @@ class openAIModelServerClientSession(ModelServerClientSession):
         request_data = json.dumps(payload)
 
         start = time.perf_counter()
+        response: Optional[aiohttp.ClientResponse] = None
         response_info = None
         error = None
         response_content = ""
-        end_time = None
         caught_exception = None
 
         try:
-            async with self.session.post(self.client.uri + data.get_route(), headers=headers, data=request_data) as response:
+            async with self.session.post(self.client.uri + data.get_route(), headers=headers, data=request_data) as resp:
+                response = resp
                 try:
                     if response.status == 200:
                         response_info = await data.process_response(
@@ -186,47 +188,37 @@ class openAIModelServerClientSession(ModelServerClientSession):
                             tokenizer=self.client.tokenizer,
                             lora_adapter=lora_adapter,
                         )
-                        response_content = await response.text()
                     else:
-                        response_content = await response.text()
                         error = ErrorResponseInfo(
-                            error_msg=response_content,
+                            error_msg=await response.text(),
                             error_type=f"HTTP Error {response.status}",
                         )
-                except Exception as e:
-                    caught_exception = e
-                    logger.error("Error processing response:", exc_info=True)
-                    if not response_content:
-                        try:
-                            response_content = await response.text()
-                        except Exception:
+                finally:
+                    # Always try to consume the payload to return the connection to the pool
+                    try:
+                        response_content = await response.text()
+                    except Exception:
+                        if not response_content:
                             response_content = "Failed to read response text"
-
-                    error = ErrorResponseInfo(
-                        error_msg=f"Error processing response: {str(e)}",
-                        error_type=type(e).__name__,
-                    )
-            end_time = time.perf_counter()
 
         except aiohttp.ClientError as e:
             caught_exception = e
             logger.error("Client error during request:", exc_info=True)
             error = ErrorResponseInfo(error_msg=str(e), error_type=type(e).__name__)
-            end_time = time.perf_counter()
         except asyncio.TimeoutError as e:
             caught_exception = e
             logger.error("Request timed out:", exc_info=True)
             error = ErrorResponseInfo(error_msg="Request timed out", error_type="TimeoutError")
-            end_time = time.perf_counter()
         except Exception as e:
             caught_exception = e
             logger.error("Unexpected error during request processing:", exc_info=True)
             error = ErrorResponseInfo(error_msg=str(e), error_type=type(e).__name__)
-            end_time = time.perf_counter()
+
+        end_time = time.perf_counter()
 
         if caught_exception is not None and not response_info:
             response_info = await data.process_failure(
-                response=response if "response" in locals() else None,
+                response=response,
                 config=self.client.api_config,
                 tokenizer=self.client.tokenizer,
                 exception=caught_exception,
