### 2026-03-01 1dc5e22
refactor: consolidate exception handling in `_send_request` using `isinstance` checks and add type hint for `caught_exception`.
