+++ inference_perf/client/modelserver/openai_client.py
@@ -175,7 +175,7 @@ class openAIModelServerClientSession(ModelServerClientSession):
         response_info = None
         error = None
         response_content = ""
-        caught_exception = None
+        caught_exception: Optional[Exception] = None
 
         try:
             async with self.session.post(self.client.uri + data.get_route(), headers=headers, data=request_data) as resp:
@@ -201,18 +201,17 @@ class openAIModelServerClientSession(ModelServerClientSession):
                         if not response_content:
                             response_content = "Failed to read response text"
 
-        except aiohttp.ClientError as e:
-            caught_exception = e
-            logger.error("Client error during request:", exc_info=True)
-            error = ErrorResponseInfo(error_msg=str(e), error_type=type(e).__name__)
-        except asyncio.TimeoutError as e:
-            caught_exception = e
-            logger.error("Request timed out:", exc_info=True)
-            error = ErrorResponseInfo(error_msg="Request timed out", error_type="TimeoutError")
         except Exception as e:
             caught_exception = e
-            logger.error("Unexpected error during request processing:", exc_info=True)
-            error = ErrorResponseInfo(error_msg=str(e), error_type=type(e).__name__)
+            if isinstance(e, asyncio.TimeoutError):
+                logger.error("Request timed out:", exc_info=True)
+                error = ErrorResponseInfo(error_msg="Request timed out", error_type="TimeoutError")
+            elif isinstance(e, aiohttp.ClientError):
+                logger.error("Client error during request:", exc_info=True)
+                error = ErrorResponseInfo(error_msg=str(e), error_type=type(e).__name__)
+            else:
+                logger.error("Unexpected error during request processing:", exc_info=True)
+                error = ErrorResponseInfo(error_msg=str(e), error_type=type(e).__name__)
 
         end_time = time.perf_counter()
 
