### 2026-03-01 b9a256e
refactor: remove unused `patch` import from test file.
