### 2026-03-01 68fdd33
finalize generic openai ModelServerType
