+++ inference_perf/main.py
@@ -41,6 +41,7 @@ from inference_perf.client.modelserver import (
     ModelServerClient,
     vLLMModelServerClient,
     SGlangModelServerClient,
+    openAIModelServerClient,
     MockModelServerClient,
 )
 from inference_perf.client.metricsclient.base import MetricsClient, PerfRuntimeParameters
@@ -222,6 +223,24 @@ def main_cli() -> None:
                 lora_config=config.load.lora_traffic_split,
             )
             # tgi_client supports inferring the tokenizer
+            tokenizer = model_server_client.tokenizer
+        if config.server.type == ModelServerType.OPENAI:
+            model_server_client = openAIModelServerClient(
+                reportgen.get_metrics_collector(),
+                api_config=config.api,
+                uri=config.server.base_url,
+                model_name=config.server.model_name,
+                tokenizer_config=config.tokenizer,
+                ignore_eos=config.server.ignore_eos,
+                max_tcp_connections=config.load.worker_max_tcp_connections,
+                additional_filters=config.metrics.prometheus.filters if config.metrics and config.metrics.prometheus else [],
+                api_key=config.server.api_key,
+                timeout=config.load.request_timeout,
+                cert_path=config.server.cert_path,
+                key_path=config.server.key_path,
+                lora_config=config.load.lora_traffic_split,
+            )
+            # OpenAI requires a tokenizer configured externally if not strictly string-based or inferred from the model_name
             tokenizer = model_server_client.tokenizer
         if config.server.type == ModelServerType.MOCK:
             model_server_client = MockModelServerClient(
