+++ inference_perf/client/modelserver/openai_client.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from abc import abstractmethod
+
 from inference_perf.client.requestdatacollector import RequestDataCollector
 from inference_perf.config import APIConfig, APIType, CustomTokenizerConfig, MultiLoRAConfig
 from inference_perf.apis import InferenceAPIData, InferenceInfo, RequestLifecycleMetric, ErrorResponseInfo
@@ -117,9 +117,8 @@ class openAIModelServerClient(ModelServerClient):
     def get_supported_apis(self) -> List[APIType]:
         return [APIType.Completion, APIType.Chat]
 
-    @abstractmethod
     def get_prometheus_metric_metadata(self) -> PrometheusMetricMetadata:
-        raise NotImplementedError
+        return PrometheusMetricMetadata()  # type: ignore[typeddict-item]
 
     def get_supported_models(self) -> List[dict[str, Any]]:
         try:
