### 2026-03-01 8dbb66d
test: Add  unit tests for the OpenAI client's API and metadata retrieval methods.
