### 2026-03-06 7fb4b5f
Remove --guided-json CLI flag, keep config-based response_format
Drop the CLI flag in favor of configuring response_format directly
in the YAML config file under api.response_format. This is cleaner
and consistent with how all other options are configured.

Update example config to use inline response_format instead of
commented-out section.
