### 2026-03-12 1a31b2c
feat(testing): increase test coverage for load_generator
Adds comprehensive unit tests for the LoadGenerator class, covering:
- LoRA adapter selection
- Timer retrieval logic
- Stage timeout handling
- Single-worker execution mode
- SIGINT signal handling
- Queue draining

These changes increase the test coverage for load_generator.py from 19% to over 42%.
