+++ tests/loadgen/test_load_generator.py
@@ -12,12 +12,19 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 import unittest
-from unittest.mock import MagicMock
+from unittest.mock import MagicMock, AsyncMock, patch
 import multiprocessing as mp
 import asyncio
 import typing
 import sys
+import numpy as np
 from typing import Any
+
+from inference_perf.loadgen.load_generator import LoadGenerator, RequestQueueData
+from inference_perf.config import LoadConfig, LoadType, TraceConfig, TraceFormat, StandardLoadStage
+from inference_perf.client.modelserver import ModelServerClient
+from inference_perf.apis import InferenceAPIData
+from inference_perf.utils.request_queue import RequestQueue
 
 # Patch asyncio.TaskGroup for Python < 3.11
 if sys.version_info < (3, 11):
@@ -103,6 +110,152 @@ class TestLoadGeneratorConcurrency(unittest.TestCase):
         self.assertEqual(self.load_generator.workers[1].shared_max_concurrency.value, 1)  # type: ignore
         self.assertEqual(self.load_generator.workers[2].shared_max_concurrency.value, 1)  # type: ignore
         self.assertEqual(self.load_generator.workers[3].shared_max_concurrency.value, 0)  # type: ignore
+
+
+class TestLoadGenerator(unittest.IsolatedAsyncioTestCase):
+    def setUp(self) -> None:
+        self.mock_datagen = MagicMock(spec=DataGenerator)
+        self.mock_datagen.get_data.return_value = iter([MagicMock(preferred_worker_id=-1) for _ in range(100)])
+        self.mock_datagen.trace = None
+
+        self.mock_client = AsyncMock(spec=ModelServerClient)
+
+        self.load_config = LoadConfig(
+            type=LoadType.CONSTANT,
+            num_workers=2,
+            worker_max_concurrency=10,
+            interval=1,
+            stages=[StandardLoadStage(rate=10, duration=1)],
+            circuit_breakers=[],
+            base_seed=42,
+        )
+
+        with patch("inference_perf.loadgen.load_generator.get_circuit_breaker"):
+            self.load_generator = LoadGenerator(self.mock_datagen, self.load_config)
+
+    def test_get_lora_adapter(self) -> None:
+        # No config
+        self.assertIsNone(self.load_generator._get_lora_adapter())
+
+        # With config
+        self.load_generator.lora_adapters = ["adapter1", "adapter2"]
+        self.load_generator.lora_weights = [1.0, 0.0]
+
+        np.random.seed(42)  # For reproducibility
+        self.assertEqual(self.load_generator._get_lora_adapter(), "adapter1")
+
+        self.load_generator.lora_weights = [0.0, 1.0]
+        self.assertEqual(self.load_generator._get_lora_adapter(), "adapter2")
+
+    def test_get_timer(self) -> None:
+        from inference_perf.loadgen.load_timer import ConstantLoadTimer, PoissonLoadTimer, TraceReplayLoadTimer
+
+        # Constant
+        self.load_generator.load_type = LoadType.CONSTANT
+        timer = self.load_generator.get_timer(10, 5)
+        self.assertIsInstance(timer, ConstantLoadTimer)
+
+        # Poisson
+        self.load_generator.load_type = LoadType.POISSON
+        timer = self.load_generator.get_timer(10, 5)
+        self.assertIsInstance(timer, PoissonLoadTimer)
+
+        # Trace (Needs trace config setup)
+        self.load_generator.load_type = LoadType.TRACE_REPLAY
+        self.load_generator.trace = TraceConfig(format=TraceFormat.AZURE_PUBLIC_DATASET, file="dummy.csv")
+        self.load_generator.trace_reader = MagicMock()
+        timer = self.load_generator.get_timer(10, 5)
[…]
