+++ inference_perf/client/modelserver/openai_client.py
@@ -160,6 +160,9 @@ class openAIModelServerClientSession(ModelServerClientSession):
             ignore_eos=self.client.ignore_eos,
             streaming=self.client.api_config.streaming,
         )
+        if payload.get("stream"):
+            payload["stream_options"] = {"include_usage": True}
+
         headers = {"Content-Type": "application/json"}
 
         if self.client.api_key:
