### 2026-03-18 #9916 @logological
@achandrasekar wrote:
> For reviewer recommendation, searching for Kubernetes + AI/ML experience from the [list](https://reviewers.joss.theoj.org/lookup) these folks might be a good fit: summitsingh, neerazz, varshaprasad96. PTAL.

I've sent out a half dozen invitations to potential reviewers and am waiting to hear back.  I did reach out to varshaprasad96 but belatedly noticed that she shares an affiliation with one of your co-authors, which may raise a [conflict of interest](https://github.com/openjournals/joss/blob/main/COI.md).
