### 2026-03-23 0949afc
add Tree-of-Thought synthetic tracegen
This introduces the ability to generate advanced "Shared Prefix" traces
specifically targeting Tree-of-Thought (ToT) workloads.
