+++ inference_perf/tracegen/__init__.py
@@ -1,0 +1,1 @@
+# Trace generation modules
