+++ inference_perf/loadgen/load_timer.py
@@ -13,7 +13,7 @@
 # limitations under the License.
 import time
 from abc import ABC, abstractmethod
-from typing import Generator, Optional, Tuple
+from typing import Generator, Optional, Tuple, Any
 import numpy as np
 from pathlib import Path
 from inference_perf.utils.trace_reader import NewTraceReader, LegacyTraceReader, NewTraceEntry
@@ -104,9 +104,15 @@ class TraceReplayLoadTimer(LoadTimer):
 
     def start_timer(self, initial: Optional[float] = None) -> Generator[float, None, None]:
         start_time = time.monotonic() if initial is None else initial
-        if isinstance(self._trace_reader, NewTraceReader):
-            for entry in self._trace_reader.load_entries(self._trace_file):
+        if hasattr(self._trace_reader, "load_entries"):
+            # SharedPrefixTraceReader logic
+            # Cast to Any to avoid "Item "LegacyTraceReader" of "NewTraceReader[Any] | LegacyTraceReader" has no attribute "load_entries""
+            reader: Any = self._trace_reader
+            for entry in reader.load_entries(self._trace_file):
                 yield start_time + entry.timestamp
         else:
-            for timestamp, _, _ in self._trace_reader.load_traces(self._trace_file):
-                yield start_time + timestamp
+            # AzurePublicDatasetReader logic
+            if isinstance(self._trace_reader, LegacyTraceReader):
+                reader_legacy = self._trace_reader
+                for timestamp, _, _ in reader_legacy.load_traces(self._trace_file):
+                    yield start_time + timestamp
