### 2026-03-24 a1517f1
Add wg-sreving serving catalog
Following https://docs.google.com/document/d/1DbhKWRKnBNQanWJZ1saBhkA6wwNP8gk0y4rzBWZYl48/edit?tab=t.0#heading=h.22pdcrnq88rz

catalog moves to inference-perf
