+++ serving-catalog/core/deployment/vllm/gemma-2b/gke/README.md
@@ -1,0 +1,14 @@
+# Gemma-2b
+
+## Configuration
+| Kind | Model Server | Model | Provider | Accelerator |
+| --- | --- | --- | --- | --- |
+| Deployment | vLLM | gemma-2b | GKE | GPU L4 |
+
+## Usage
+
+The template can be deployed with the following commands:
+
+```
+kustomize build core/deployment/vllm/gemma-2b/gke | kubectl apply -f -
+```
