+++ serving-catalog/core/deployment/vllm/llama3-70b/gke/README.md
@@ -1,0 +1,14 @@
+# Llama-70b
+
+## Configuration
+| Kind | Model Server | Model | Provider | Accelerator |
+| --- | --- | --- | --- | --- |
+| Deployment | vLLM | llama3-70b | GKE | GPU L4 |
+
+## Usage
+
+The template can be deployed with the following commands:
+
+```
+kustomize build core/deployment/vllm/llama3-70b/gke | kubectl apply -f -
+```
