+++ serving-catalog/core/lws/vllm/base/service.yaml
@@ -1,0 +1,14 @@
+apiVersion: v1
+kind: Service
+metadata:
+  name: vllm-leader
+spec:
+  ports:
+    - name: http
+      port: 8080
+      protocol: TCP
+      targetPort: 8080
+  selector:
+    role: leader
+    deployment: vllm-multihost-base
+  type: ClusterIP
