+++ serving-catalog/core/deployment/jetstream/base/deployment.patch.yaml
@@ -1,0 +1,25 @@
+apiVersion: apps/v1
+kind: Deployment
+metadata:
+  name: maxengine-server
+spec:
+  replicas: 2
+  selector:
+    matchLabels:
+      app: maxengine-server
+      deployment: maxengine-server
+  template:
+    metadata:
+      labels:
+        app: maxengine-server
+        deployment: maxengine-server
+    spec:
+      containers:
+      - name: inference-server
+        image: us-docker.pkg.dev/cloud-tpu-images/inference/maxengine-server:v0.2.2
+        ports:
+        - containerPort: 9000
+      - name: jetstream-http
+        image: us-docker.pkg.dev/cloud-tpu-images/inference/jetstream-http:v0.2.2
+        ports:
+        - containerPort: 8000
