+++ serving-catalog/core/deployment/components/gke/resources/gpu/8-H100-80GB/kustomization.yaml
@@ -1,0 +1,18 @@
+# kustomization.yaml
+apiVersion: kustomize.config.k8s.io/v1alpha1
+kind: Component
+
+patches:
+  - path: h100-80gb.yaml
+    target:
+      group: apps
+      version: v1
+      kind: Deployment
+  - patch: |-
+      - op: add
+        path: /spec/template/spec/containers/0/env/-
+        value:
+          name: TENSOR_PARALLEL_SIZE
+          value: "8"
+    target:
+      kind: Deployment
