+++ serving-catalog/core/lws/vllm/base/configmap.yaml
@@ -1,0 +1,4 @@
+apiVersion: v1
+kind: ConfigMap
+metadata:
+  name: vllm-multihost-config
