### 2026-03-25 e20c392
fix: resolve CI failures - remove unused import, fix test fixture
- Remove unused `import logging` in main.py (ruff F401)
- Set api_config.response_format = None in mock_client fixture to
  prevent MagicMock from being passed to json.dumps
