+++ inference_perf/datagen/otel_trace_utils.py
@@ -1,0 +1,595 @@
+"""
+Reconstruct raw LLM output from parsed API responses for token estimation.
+
+This module helps estimate the actual tokens generated by an LLM by reconstructing
+what the raw output likely looked like before being parsed by LiteLLM or other API wrappers.
+
+Supports:
+- Text responses
+- Tool/function calls
+- Mixed responses (text + tool calls)
+- Multiple tool calls
+"""
+
+import json
+from typing import Dict, List, Any, Union, Optional
+
+
+def reconstruct_llm_output(response_data: Union[str, Dict, List]) -> str:
+    """
+    Reconstruct the raw LLM output from any parsed response structure.
+
+    This is the main entry point that handles all types of LLM outputs:
+    - Pure text responses
+    - Tool/function calls
+    - Mixed responses (text + tool calls)
+
+    Args:
+        response_data: Can be:
+            - A JSON string from OTEL trace (gen_ai.output.messages)
+            - A parsed dict/list structure
+            - OpenAI API response format
+            - LiteLLM parsed format
+
+    Returns:
+        String representation of the raw LLM output (for token estimation)
+
+    Example:
+        >>> otel_output = '[{"role": "assistant", "content": "Hello!"}]'
+        >>> raw_output = reconstruct_llm_output(otel_output)
+        >>> print(raw_output)
+        Hello!
+    """
+
+    # Parse if string
+    if isinstance(response_data, str):
+        try:
+            response_data = json.loads(response_data)
+        except json.JSONDecodeError:
+            # If it's not JSON, assume it's already raw text
+            return response_data
+
+    # Extract the message content
+    message = _extract_message(response_data)
+
+    if not message:
+        return ""
+
+    # Reconstruct based on what's in the message
+    output_parts = []
+
+    # 1. Handle text content
+    text_content = _extract_text_content(message)
+    if text_content:
+        output_parts.append(text_content)
+
+    # 2. Handle tool calls
+    tool_calls = _extract_tool_calls(message)
+    if tool_calls:
+        for tool_call in tool_calls:
+            tool_output = _format_tool_call(tool_call)
+            output_parts.append(tool_output)
+
+    return "\n".join(output_parts) if output_parts else ""
+
+
+def _extract_message(data: Any) -> Optional[Dict]:
+    """
+    Extract the assistant message from various response formats.
+
+    Handles:
+    - List format: [{"role": "assistant", ...}]
+    - Dict with message: {"message": {...}}
+    - Dict with choices: {"choices": [{"message": {...}}]}
+    - Direct message dict: {"role": "assistant", ...}
+    """
+    if isinstance(data, list):
+        # Find the assistant message
+        for item in data:
+            if isinstance(item, dict) and item.get("role") == "assistant":
+                return item
+        # If no role specified, return first item
+        if data and isinstance(data[0], dict):
+            return data[0]
+
+    elif isinstance(data, dict):
+        # Check for OpenAI choices format
+        if "choices" in data and data["choices"]:
+            choice = data["choices"][0]
+            if "message" in choice:
+                return choice["message"]
+            return choice
+
+        # Check for direct message key
+        if "message" in data:
+            return data["message"]
+
+        # Check if data itself is a message
+        if "role" in data or "content" in data or "parts" in data or "tool_calls" in data:
+            return data
+
+    return None
+
+
+def _extract_text_content(message: Dict) -> str:
+    """
+    Extract text content from a message.
+
+    Handles:
+    - content field (string)
+    - content field (list of content parts)
+    - parts field (OTEL format)
+    """
+    text_parts = []
+
+    # Check for direct content field
[…]
