+++ inference_perf/datagen/otel_trace_replay_datagen.py
@@ -1,0 +1,1282 @@
+# Copyright 2025 The Kubernetes Authors.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""
+OTel Trace Replay Data Generator
+
+This module provides a data generator that replays LLM requests from OpenTelemetry trace files.
+It processes OTel JSON traces, extracts LLM calls, infers dependencies, and replays them
+with proper timing and dependency enforcement.
+
+Output-aware replay
+-------------------
+When a call has "output" segments (i.e., a predecessor's output was injected as an assistant
+message in the original trace), the replayer substitutes the recorded assistant message content
+with the *actual* output generated by the predecessor call at replay time.
+
+This ensures that growing-context patterns (multi-turn conversations, agent chains) are replayed
+faithfully: the KV cache sees the real generated text, not the static recorded text.
+
+The substitution is performed in `load_lazy_data` using the `NodeOutputRegistry`, which maps
+node_id → actual output text. All registry data is stored in per-worker plain dicts; session-to-
+worker affinity ensures that the node writing an output and the nodes reading it always run on
+the same worker process.
+
+The actual output is captured in `OTelChatCompletionAPIData.process_response`, which overrides
+the base `ChatCompletionAPIData.process_response` to register the output after each LLM call.
+
+Cross-process state
+-------------------
+Two orthogonal coordination problems exist in this system:
+
+  Problem A — intra-worker, between coroutines on the same event loop:
+    A node waits for its predecessor's output before building its request.
+    Handled by NodeOutputRegistry using plain dicts and asyncio.Event.
+    No IPC involved; session-to-worker affinity makes cross-process sharing unnecessary.
+
+  Problem B — cross-process, worker → main process:
+    The main process drives the session pool loop and needs to know when sessions
+    complete and whether they failed. Workers produce these facts.
+    Handled by SessionSharedState using Manager.dict() for both signals.
+
+OTelChatCompletionAPIData holds refs to both objects and uses each for its
+appropriate purpose. SessionSharedState is the single place where Manager.dict()
+instances are created and owned.
+"""
+
+import asyncio
+import glob
+import json
+import logging
+import random
+import time
+from dataclasses import dataclass, field, replace as dc_replace
+from multiprocessing.managers import SyncManager
+from pathlib import Path
+from typing import Any, Dict, List, Optional, Set
+
+from aiohttp import ClientResponse
+
+from inference_perf.apis import (
+    ChatCompletionAPIData,
+    InferenceAPIData,
+    InferenceInfo,
+    LazyLoadInferenceAPIData,
+    SessionLifecycleMetric,
+)
+from inference_perf.apis.chat import ChatMessage
+from inference_perf.config import APIConfig, APIType, DataConfig
+from inference_perf.datagen.base import SessionGenerator, LazyLoadDataMixin
+from inference_perf.datagen.otel_trace_to_replay_graph import (
+    InputSegment,
+    ReplayGraph,
+    build_raw_calls,
+    build_graph,
+)
+from inference_perf.utils.custom_tokenizer import CustomTokenizer
+
+logger = logging.getLogger(__name__)
+
+# ---------------------------------------------------------------------------
+# SessionGraphState — tracks graph traversal state for one session
[…]
