### 2026-03-26 c5201d7
feat: add OTel trace replay for agentic workload benchmarking
Add OTel trace replay support for session-based agentic workload benchmarking,
including replay graph construction, dependency handling, output substitution,
and session-level metrics.

Co-authored-by: Lena Dankin <lenad@il.ibm.com>
Co-authored-by: oritht <oritht@il.ibm.com>
