+++ inference_perf/config.py
@@ -16,7 +16,7 @@ from datetime import datetime
 from enum import Enum
 from os import cpu_count
 import time
-from typing import Any, List, Optional, Union
+from typing import Any, List, Optional, Union, Dict
 
 import yaml
 from pydantic import AliasChoices, BaseModel, ConfigDict, Field, HttpUrl, model_validator
@@ -86,6 +86,7 @@ class DataGenType(Enum):
     CNNDailyMail = "cnn_dailymail"
     InfinityInstruct = "infinity_instruct"
     BillsumConversations = "billsum_conversations"
+    OTelTraceReplay = "otel_trace_replay"
 
 
 # Represents the distribution for input prompts and output generations.
@@ -121,6 +122,47 @@ class SharedPrefix(BaseModel):
     enable_multi_turn_chat: bool = False
 
 
+class OTelTraceReplayConfig(BaseModel):
+    """Configuration for OTel trace replay data generator."""
+
+    trace_directory: Optional[str] = Field(None, description="Directory containing OTel JSON trace files")
+    trace_files: Optional[List[str]] = Field(None, description="List of paths to specific OTel JSON trace files")
+
+    # Model configuration
+    use_static_model: bool = Field(False, description="Use a single static model for all requests")
+    static_model_name: str = Field("", description="Static model name (required if use_static_model=True)")
+    model_mapping: Optional[Dict[str, str]] = Field(None, description="Map recorded model names to target models")
+
+    # Request configuration
+    default_max_tokens: int = Field(1000, gt=0, description="Default max_tokens if not specified in trace")
+
+    # Error handling
+    include_errors: bool = Field(True, description="Include spans with error status")
+    skip_invalid_files: bool = Field(False, description="Skip invalid trace files instead of failing")
+
+    @model_validator(mode="after")
+    def validate_static_model(self) -> "OTelTraceReplayConfig":
+        # Validate that exactly one of trace_directory or trace_files is provided
+        sources_provided = sum(
+            [
+                self.trace_directory is not None,
+                self.trace_files is not None,
+            ]
+        )
+
+        if sources_provided == 0:
+            raise ValueError("Either trace_directory or trace_files must be provided")
+        if sources_provided > 1:
+            raise ValueError("Cannot specify both trace_directory and trace_files; choose one")
+
+        # Validate static model configuration
+        if self.use_static_model and not self.static_model_name:
+            raise ValueError("static_model_name is required when use_static_model=True")
+        if not self.use_static_model and self.static_model_name and not self.model_mapping:
+            raise ValueError("Either use_static_model must be True or model_mapping must be provided")
+        return self
+
+
 class DataConfig(BaseModel):
     type: DataGenType = DataGenType.Mock
 
@@ -135,6 +177,9 @@ class DataConfig(BaseModel):
     # Trace file is only supported for random dataset at this moment
     trace: Optional[TraceConfig] = None
 
+    # OTel trace replay configuration
+    otel_trace_replay: Optional[OTelTraceReplayConfig] = None
+
 
 class ModelServerType(Enum):
     VLLM = "vllm"
@@ -148,6 +193,7 @@ class LoadType(Enum):
     POISSON = "poisson"
     TRACE_REPLAY = "trace_replay"
     CONCURRENT = "concurrent"
+    TRACE_SESSION_REPLAY = "trace_session_replay"
 
 
 class MetricsClientType(Enum):
@@ -197,6 +243,73 @@ class ConcurrentLoadStage(LoadStage):
         return self
 
 
+logger = logging.getLogger(__name__)
+
+
+class TraceSessionReplayLoadStage(LoadStage):
+    """Load stage for TRACE_SESSION_REPLAY load type.
+
[…]
