### 2026-03-27 1b7e697
Add rich progress bar and a table with results
This change adds a rich progress bar for per stage and overall progress
of the benchmark. It also prints selected columns from the report on to
stdout at the end of the run. This includes latency, throughput, request
rates and prompt and output token breakdown.
