+++ inference_perf/utils/cli_summary.py
@@ -1,0 +1,179 @@
+# Copyright 2026 The Kubernetes Authors.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+import re
+from typing import List, Dict, Any, Optional
+from rich.console import Console
+from rich.table import Table
+from rich import print as rprint
+from inference_perf.utils.report_file import ReportFile
+
+
+def extract_stage_id(report_name: str) -> Optional[int]:
+    """Extract stage ID from report name (e.g. 'stage_0_lifecycle_metrics')."""
+    match = re.match(r"stage_(\d+)_lifecycle_metrics", report_name)
+    if match:
+        return int(match.group(1))
+    return None
+
+
+def print_summary_table(reports: List[ReportFile]) -> None:
+    """Print a summary table of all stages to stdout using rich."""
+    stage_reports: Dict[int, Dict[str, Any]] = {}
+
+    for report in reports:
+        stage_id = extract_stage_id(report.name)
+        if stage_id is not None:
+            stage_reports[stage_id] = report.contents
+
+    if not stage_reports:
+        rprint("[yellow]No per-stage lifecycle metrics found to display summary table.[/yellow]")
+        return
+
+    # Sort stages by ID
+    sorted_stages = sorted(stage_reports.keys())
+    
+    console = Console()
+    table = Table(title="[bold magenta]Inference Performance Summary[/bold magenta]", show_header=True, header_style="bold cyan")
+    
+    table.add_column("Stage", justify="right")
+    table.add_column("Req Rate", justify="right")
+    table.add_column("Achieved Rate", justify="right")
+    table.add_column("Error Rate", justify="right")
+    
+    # TTFT Columns
+    table.add_column("TTFT Mean\n(ms)", justify="right")
+    table.add_column("TTFT Med\n(ms)", justify="right")
+    table.add_column("TTFT P90\n(ms)", justify="right")
+    
+    # ITL Columns
+    table.add_column("ITL Mean\n(ms)", justify="right")
+    table.add_column("ITL Med\n(ms)", justify="right")
+    table.add_column("ITL P90\n(ms)", justify="right")
+    
+    # Throughput Columns
+    table.add_column("Req/s", justify="right")
+    table.add_column("In Tokens/s", justify="right")
+    table.add_column("Out Tokens/s", justify="right")
+    table.add_column("Tot Tokens/s", justify="right")
+
+    for stage_id in sorted_stages:
+        contents = stage_reports[stage_id]
+        load_summary = contents.get("load_summary", {})
+        successes = contents.get("successes", {})
+        failures = contents.get("failures", {})
+        
+        req_rate = load_summary.get("requested_rate", 0.0)
+        ach_rate = load_summary.get("achieved_rate", 0.0)
+
+        # Error Rate calculation
+        success_count = successes.get("count", 0)
+        failed_count = failures.get("count", 0)
+        total_count = success_count + failed_count
+        error_rate = failed_count / total_count if total_count > 0 else 0.0
+        error_rate_pct = error_rate * 100.0
+        
[…]
