+++ inference_perf/loadgen/load_generator.py
@@ -59,7 +59,7 @@ from functools import partial
 import logging
 import uvloop
 import numpy as np
-from tqdm import tqdm
+from rich.progress import Progress, TextColumn, BarColumn, TaskProgressColumn, TimeElapsedColumn, TimeRemainingColumn, MofNCompleteColumn
 import signal
 
 logger = logging.getLogger(__name__)
@@ -308,8 +308,9 @@ class LoadGenerator:
         cancel_signal: Optional[SyncEvent] = None,
         timeout: Optional[float] = None,
         concurrency_level: Optional[int] = None,
+        progress_ctx: Optional[Progress] = None,
     ) -> None:
-        logger.info("Stage %d - run started", stage_id)
+        logger.debug("Stage %d - run started", stage_id)
 
         if timeout is not None and cancel_signal is None:
             raise Exception("run_stage timeout requires cancel_signal to be not None!")
@@ -350,31 +351,33 @@ class LoadGenerator:
             )
 
         # Wait until all requests are finished processing
-        with tqdm(total=1.0, desc=f"Stage {stage_id} progress") as pbar:
-            timed_out = False
-            timeout_progress = (time.perf_counter() - start_time) / timeout if timeout else 0
-            prog, last = min(1.0, max(timeout_progress, finished_requests_counter.value / num_requests)), 0.0
-            pbar.update(prog)
-            last = prog
-            while finished_requests_counter.value < num_requests:
-                if timeout and start_time + timeout < time.perf_counter():
-                    pbar.close()
-                    logger.info(f"Loadgen timed out after {timeout:0.2f}s")
-                    timed_out = True
-                    break
-                if self.interrupt_sig:
-                    pbar.close()
-                    logger.info("Loadgen encountered SIGINT")
-                    break
-                if cb := next((cb for cb in self.circuit_breakers if cb.is_open()), None):
-                    logger.warning(f'Loadgen detects circuit breakers "{cb.name}" open, exit the stage.')
-                    timed_out = True
-                    break
-                await sleep(1)
-                timeout_progress = (time.perf_counter() - start_time) / timeout if timeout else 0
-                prog = min(1.0, max(timeout_progress, finished_requests_counter.value / num_requests))
-                pbar.update(prog - last)
-                last = prog
+        stage_task = None
+        if progress_ctx:
+            stage_task = progress_ctx.add_task(description=f"Stage {stage_id} Requests", total=num_requests)
+
+        timed_out = False
+        while finished_requests_counter.value < num_requests:
+            if timeout and start_time + timeout < time.perf_counter():
+                if progress_ctx and stage_task:
+                    progress_ctx.remove_task(stage_task)
+                logger.info(f"Loadgen timed out after {timeout:0.2f}s")
+                timed_out = True
+                break
+            if self.interrupt_sig:
+                if progress_ctx and stage_task:
+                    progress_ctx.remove_task(stage_task)
+                logger.info("Loadgen encountered SIGINT")
+                break
+            if cb := next((cb for cb in self.circuit_breakers if cb.is_open()), None):
+                logger.warning(f'Loadgen detects circuit breakers "{cb.name}" open, exit the stage.')
+                timed_out = True
+                break
+            await sleep(1)
+            if progress_ctx and stage_task:
+                progress_ctx.update(stage_task, completed=finished_requests_counter.value)
+
+        if progress_ctx and stage_task:
+            progress_ctx.remove_task(stage_task)
 
         # Trigger cleanup if timed out or received SIGINT
         if (timed_out or self.interrupt_sig) and cancel_signal:
@@ -401,7 +404,7 @@ class LoadGenerator:
             status=stage_status,
             concurrency_level=concurrency_level,
         )
[…]
