### 2026-03-27 #10314 @editorialbot
[REVIEW]: Inference Perf: A Benchmarking Tool for GenAI Inference
**Submitting author:** @achandrasekar (<a href="http://orcid.org/0009-0009-3481-9173">Ashok Chandrasekar</a>)
**Repository:** https://github.com/kubernetes-sigs/inference-perf
**Branch with paper.md** (empty if default branch):
**Version:** 0.6.1
