### 2026-03-29 71e4b14
add visualization + improve matching of outputs that are composed of multiple tool calls.
