+++ inference_perf/datagen/otel_trace_to_replay_graph.py
@@ -44,6 +44,7 @@ from datetime import datetime, timezone
 from pathlib import Path
 from typing import Any, Dict, List, Optional, Set
 
+from inference_perf.datagen.export_replay_graph_to_dot import export_to_dot
 from inference_perf.datagen.otel_trace_utils import reconstruct_llm_output, reconstruct_llm_input,  \
     reconstruct_each_part_in_message_info
 
@@ -384,30 +385,134 @@ def is_causal_dep(a: RawCall, b: RawCall) -> bool:
             return True
     #try matching parts
     if isinstance(a.out_message, ComplexOtelMessage) and "parts" in a.out_message.message_info and len(a.out_message.message_info["parts"]) > 1:
-        #this means this output message contains several part, and will be interpreted as more than one message in the calls history
-        first_part_match_candidates = []
-        first_part_text = a.out_message.message_info["parts_text"][0]
-        for i in range(len(b.messages)):
-            if output_matches_message(first_part_text, b.messages[i]):
-                first_part_match_candidates.append(i)
-        for candidate_i in first_part_match_candidates:
-            candidate_ok = True
-            #in case of a tool call, we'll see first the tool call, and then the result of the execution.
-            offset = 1 if a.out_message.message_info["parts"][0]["type"] != "tool_call" else 2
-            for part, part_text in zip(a.out_message.message_info["parts"][1:], a.out_message.message_info["parts_text"][1:]):
-                #part_text is the text that would appear in the input messages list. we also need part to determine the type, to set the offset
-                next_index_to_check = candidate_i + offset
-                if next_index_to_check >= len(b.messages):
-                    candidate_ok = False
-                    break
-                if not output_matches_message(part_text, b.messages[next_index_to_check]):
-                    candidate_ok = False
-                    break
-                else:
-                    offset = 1 if part["type"] != "tool_call" else 2
-            if candidate_ok:
+        #this means this output message contains several parts, and will be interpreted as more than one message in the calls history
+        parts = a.out_message.message_info["parts"]
+        parts_text = a.out_message.message_info["parts_text"]
+        
+        # Determine structure: check if first part is content (text) or tool_call
+        first_part_is_content = parts[0]["type"] != "tool_call"
+        
+        # Case 1: Only tool calls [tool_call_1, tool_call_2, ..., tool_call_n]
+        # Case 2: Content + tool calls [content, tool_call_1, tool_call_2, ..., tool_call_n]
+        
+        if not first_part_is_content:
+            # Case 1: Only tool calls - each tool call appears separately in input messages
+            return _try_match_parts(parts, parts_text, b.messages, combine_content_with_tools=False)
+        else:
+            # Case 2: Content + tool calls - try two matching strategies:
+            # Strategy A: Each part appears separately (content has offset 1, tools have offset 2)
+            if _try_match_parts(parts, parts_text, b.messages, combine_content_with_tools=False):
                 return True
+            # Strategy B: Content is combined with each tool call
+            # [content + tool_call_1, content + tool_call_2, ..., content + tool_call_n]
+            # Combined messages are treated as tool calls (offset 2)
+            if _try_match_parts(parts, parts_text, b.messages, combine_content_with_tools=True):
+                return True
+        
         return False
+    return False
+
+
+def _try_match_parts(parts: list, parts_text: list, b_messages: list, combine_content_with_tools: bool) -> bool:
+    """
+    Unified function to match parts in b_messages.
+    
+    Args:
+        parts: List of part dictionaries with 'type' field
+        parts_text: List of text representations for each part
+        b_messages: List of messages to search in
+        combine_content_with_tools: If True and first part is content, expect content combined with each tool call
+    
[…]
