### 2026-03-30 35e0e74
small fix in config
