+++ examples/otel/configs/per_case_config/fan_in_dependencies.yml
@@ -28,7 +28,7 @@ load:
 server:
   type: vllm
   base_url: "http://localhost:8000"
-  model_name: ""HuggingFaceTB/SmolLM2-135M-Instruct"
+  model_name: "HuggingFaceTB/SmolLM2-135M-Instruct"
 
 storage:
   local_storage:
