### 2026-03-30 dad555d
add copyright comment to the top for all code files
