+++ inference_perf/loadgen/load_generator.py
@@ -19,6 +19,7 @@ from .load_timer import LoadTimer, ConstantLoadTimer, PoissonLoadTimer, TraceRep
 from inference_perf.datagen import DataGenerator, SessionGenerator, LazyLoadDataMixin
 from inference_perf.apis import InferenceAPIData
 from inference_perf.client.modelserver import ModelServerClient
+from inference_perf.client.modelserver.otel_instrumentation import get_otel_instrumentation
 from inference_perf.circuit_breaker import get_circuit_breaker
 from inference_perf.metrics import SessionMetricsCollector
 from inference_perf.config import (
@@ -50,7 +51,7 @@ else:
     # Runtime usage will still require Python 3.11+.
     TaskGroup = object
 
-from typing import List, Tuple, Optional, NamedTuple, Union, Set, Dict
+from typing import List, Tuple, Optional, NamedTuple, Union, Set, Dict, Any
 from types import FrameType
 import time
 import multiprocessing as mp
@@ -407,6 +408,26 @@ class LoadGenerator:
         )  # Sessions waiting to start
         completed_session_ids: Set[str] = set()  # Session IDs that have completed
         session_dispatch_times: Dict[str, float] = {}  # session_id → wall-clock dispatch time
+        
+        # Cache OTEL instrumentation to avoid redundant calls
+        otel_instr = get_otel_instrumentation()
+        session_spans: Dict[str, object] = {}  # session_id → OTEL span object
+        stage_span = None  # Stage-level span (if trace_per_stage is enabled)
+        stage_context_dict = None  # Stage-level context for propagation
+
+        # Start stage-level span if trace_per_stage is enabled
+        if otel_instr.trace_per_stage:
+            stage_info = {
+                "num_sessions": effective_num_sessions,
+                "concurrent_sessions": concurrent_sessions,
+            }
+            if session_rate is not None:
+                stage_info["session_rate"] = session_rate
+            if timeout is not None:
+                stage_info["timeout"] = timeout
+            
+            stage_span, stage_context_dict = otel_instr.start_stage_span(stage_id, stage_info)
+            logger.info(f"Started stage-level OTEL span for stage {stage_id}")
 
         # Track dispatch timing
         sessions_dispatched = 0
@@ -451,6 +472,16 @@ class LoadGenerator:
                 f"Starting session {session_idx}: {session_id} "
                 f"({len(active_session_indices)} active, {len(pending_session_indices)} pending)"
             )
+
+            # Start OTEL session span (as child of stage span if trace_per_stage is enabled)
+            parent_ctx = stage_context_dict if otel_instr.trace_per_stage else None
+            span, context_dict = otel_instr.start_session_span(session_id, session_info, parent_ctx)
+            if span is not None:
+                session_spans[session_id] = span
+            
+            # Store OTEL context in shared state for workers to access
+            if context_dict and hasattr(self.datagen, 'shared_state'):
+                self.datagen.shared_state.store_session_otel_context(session_id, context_dict)
 
             # Activate session in DataGen (marks root nodes as ready)
             self.datagen.activate_session(session_id)
@@ -498,17 +529,35 @@ class LoadGenerator:
                     pbar.close()
                     logger.info("Loadgen encountered SIGINT")
                     stage_status = StageStatus.FAILED
+                    # Clean up any active session spans (using cached otel_instr)
+                    for sid in list(session_spans.keys()):
+                        otel_instr.end_session_span(session_spans[sid], "Session interrupted by SIGINT")
+                        if hasattr(self.datagen, 'shared_state'):
+                            self.datagen.shared_state.cleanup_session_otel_context(sid)
+                        del session_spans[sid]
                     break
 
[…]
