### 2026-03-30 aaa50c9
feat: Add session-level OTEL spans with nested LLM request spans (#4)
* Add standard GenAI OTEL instrumentation to LLM API calls

- Added OpenTelemetry dependencies (opentelemetry-api, opentelemetry-sdk, opentelemetry-exporter-otlp-proto-grpc, opentelemetry-semantic-conventions-ai)
- Created OTelInstrumentation module with support for Console and OTLP exporters
- Instrumented all model server clients (OpenAI, vLLM, SGLang, TGI) with OTEL tracing
- Added GenAI semantic convention attributes:
  - gen_ai.input.messages: Full conversation history as JSON
  - gen_ai.output.text: Model response text
[…]
