### 2026-03-30 c94a96e
Eliminate SessionSharedState and replace with event-driven architecture
Major architectural improvement to OTEL trace replay system:

REMOVED:
- SessionSharedState class (70+ lines) using Manager.dict() for cross-process communication
- All polling-based completion checking via shared state
- OTEL context storage in shared state

ADDED:
- WorkerSessionTracker: Per-worker session state tracking using plain dicts (no IPC overhead)
- session_completion_queue: Event-driven mp.Queue for worker→main completion notifications
- otel_context field in InferenceAPIData base class for context propagation

[…]
