+++ inference_perf/loadgen/load_generator.py
@@ -495,10 +495,6 @@ class LoadGenerator:
             span, context_dict = otel_instr.start_session_span(session_id, session_info, parent_ctx)
             if span is not None:
                 session_spans[session_id] = span
-            
-            # Store OTEL context in shared state for workers to access
-            if context_dict and hasattr(self.datagen, 'shared_state'):
-                self.datagen.shared_state.store_session_otel_context(session_id, context_dict)
 
             # Activate session in DataGen (marks root nodes as ready)
             self.datagen.activate_session(session_id)
@@ -517,8 +513,9 @@ class LoadGenerator:
                 if worker_id >= 0:
                     worker_id = worker_id % active_workers
 
-                # Stamp session_id so the client can attach it to RequestLifecycleMetric
+                # Stamp session_id and OTEL context so workers can use them
                 lazy_data.session_id = session_id
+                lazy_data.otel_context = context_dict  # Embed OTEL context in data
 
                 event_time = time.perf_counter()
                 request_queue.put((stage_id, lazy_data, event_time, lora_adapter), worker_id)
@@ -549,8 +546,6 @@ class LoadGenerator:
                     # Clean up any active session spans (using cached otel_instr)
                     for sid in list(session_spans.keys()):
                         otel_instr.end_session_span(session_spans[sid], "Session interrupted by SIGINT")
-                        if hasattr(self.datagen, 'shared_state'):
-                            self.datagen.shared_state.cleanup_session_otel_context(sid)
                         del session_spans[sid]
                     break
 
@@ -560,8 +555,6 @@ class LoadGenerator:
                     # Clean up any active session spans (using cached otel_instr)
                     for sid in list(session_spans.keys()):
                         otel_instr.end_session_span(session_spans[sid], f"Session failed due to circuit breaker: {cb.name}")
-                        if hasattr(self.datagen, 'shared_state'):
-                            self.datagen.shared_state.cleanup_session_otel_context(sid)
                         del session_spans[sid]
                     break
 
@@ -572,8 +565,6 @@ class LoadGenerator:
                     # Clean up any active session spans (using cached otel_instr)
                     for sid in list(session_spans.keys()):
                         otel_instr.end_session_span(session_spans[sid], "Session timed out")
-                        if hasattr(self.datagen, 'shared_state'):
-                            self.datagen.shared_state.cleanup_session_otel_context(sid)
                         del session_spans[sid]
                     break
 
@@ -589,18 +580,13 @@ class LoadGenerator:
                             completed_session_ids.add(session_id)
                             newly_completed.append(session_idx)
                             
-                            # End OTEL session span and cleanup context (using cached otel_instr)
+                            # End OTEL session span (using cached otel_instr)
                             if session_id in session_spans:
                                 # Check if session failed
                                 session_failed = (hasattr(self.datagen, 'shared_state') and
                                                 self.datagen.shared_state.is_session_failed(session_id))
                                 error_msg = "Session failed" if session_failed else None
                                 otel_instr.end_session_span(session_spans[session_id], error_msg)
-                                
-                                # Cleanup OTEL context to prevent memory leak
-                                if hasattr(self.datagen, 'shared_state'):
-                                    self.datagen.shared_state.cleanup_session_otel_context(session_id)
-                                
                                 del session_spans[session_id]
                             
                             logger.info(
