+++ inference_perf/datagen/otel_trace_replay_datagen.py
@@ -36,8 +36,8 @@ the same worker process.
 The actual output is captured in `OTelChatCompletionAPIData.process_response`, which overrides
 the base `ChatCompletionAPIData.process_response` to register the output after each LLM call.
 
-Cross-process state
--------------------
+Cross-process coordination
+---------------------------
 Two orthogonal coordination problems exist in this system:
 
   Problem A — intra-worker, between coroutines on the same event loop:
@@ -46,25 +46,28 @@ Two orthogonal coordination problems exist in this system:
     No IPC involved; session-to-worker affinity makes cross-process sharing unnecessary.
 
   Problem B — cross-process, worker → main process:
-    The main process drives the session pool loop and needs to know when sessions
-    complete and whether they failed. Workers produce these facts.
-    Handled by SessionSharedState using Manager.dict() for both signals.
+    The main process needs to know when sessions complete for OTEL span closure and reporting.
+    Workers detect session completion and push notifications via mp.Queue.
+    
+    - WorkerSessionTracker: Per-worker tracking of node completions and session failures
+    - session_completion_queue: Workers push completion data when last node finishes
+    - Main process consumes from queue in check_session_completed()
 
-OTelChatCompletionAPIData holds refs to both objects and uses each for its
-appropriate purpose. SessionSharedState is the single place where Manager.dict()
-instances are created and owned.
+OTelChatCompletionAPIData holds refs to WorkerSessionTracker and completion_queue.
+No Manager.dict() instances are used; all cross-process communication is via Queue.
 """
 
 import asyncio
 import glob
 import json
 import logging
+import multiprocessing as mp
 import random
 import time
 from dataclasses import dataclass, field, replace as dc_replace
 from multiprocessing.managers import SyncManager
 from pathlib import Path
-from typing import Any, Dict, List, Optional, Set
+from typing import Any, Dict, List, Optional, Set, Tuple
 
 from aiohttp import ClientResponse
 
@@ -128,80 +131,113 @@ class OTelInferenceInfo(InferenceInfo):
 
 
 # ---------------------------------------------------------------------------
-# SessionSharedState — all cross-process state in one place
+# WorkerSessionTracker — tracks session state within a worker process
 # ---------------------------------------------------------------------------
 
 
-class SessionSharedState:
-    """Owns all state that must cross the worker → main-process boundary.
 
-    Two facts need to be visible to the main process while workers are running:
-      1. Which nodes have completed (so the main process can detect session completion).
-      2. Which sessions have failed (so workers can skip dependent nodes and the
-         main process can retire failed sessions).
+# ---------------------------------------------------------------------------
+# WorkerSessionTracker — tracks session state within a worker process
+# ---------------------------------------------------------------------------
 
-    Both are stored in Manager.dict() instances so writes on any worker are
-    immediately visible to every other process. This is the *only* place in the
-    OTel replay path where Manager proxies are created.
 
-    Usage:
-        shared = SessionSharedState(mp_manager)   # main process, before fork
-        # workers call shared.record_node_completed() and shared.mark_session_failed()
-        # main process calls shared.is_node_completed() and shared.is_session_failed()
-    """
-
-    def __init__(self, manager: Optional[SyncManager] = None) -> None:
-        if manager is not None:
-            self._node_completions: Dict[str, float] = manager.dict()  # type: ignore[assignment]
-            self._failed_sessions: Dict[str, bool] = manager.dict()  # type: ignore[assignment]
-            self._session_otel_contexts: Dict[str, Dict[str, str]] = manager.dict()  # type: ignore[assignment]
-        else:
-            self._node_completions = {}
-            self._failed_sessions = {}
[…]
