+++ examples/otel/README.md
@@ -36,7 +36,7 @@ with the *actual* text generated by the predecessor during the current run. This
 growing-context patterns (agent chains, multi-turn conversations) produce realistic KV-cache
 behaviour rather than replaying stale recorded text.
 
-Two orthogonal coordination objects handle this:
+Three coordination mechanisms handle this:
 
 - **`NodeOutputRegistry`** — intra-worker only. Holds plain dicts (`node_id → output text`,
   `node_id → input messages`) and one `asyncio.Event` per node. When a node completes,
@@ -44,18 +44,23 @@ Two orthogonal coordination objects handle this:
   on the same worker. No IPC — session-to-worker affinity guarantees all nodes of a session
   run on the same worker.
 
-- **`SessionSharedState`** — cross-process. Holds two `Manager.dict()` instances: one mapping
-  `node_id → completion_time` (so the main process can detect session completion) and one
-  mapping `session_id → failed` (so workers can skip dependents after a failure). This is the
-  only place in the OTel replay path where multiprocessing proxies are created.
+- **`WorkerSessionTracker`** — per-worker session state tracking. Each worker independently
+  tracks which nodes have completed and which sessions have failed within its assigned sessions.
+  No cross-process communication needed due to session-to-worker affinity.
 
-Each `OTelChatCompletionAPIData` holds references to both. Before dispatching an HTTP request,
-the worker calls `wait_for_predecessors_and_substitute()`, which awaits predecessors via
-`registry.require_async()` (zero threads — pure `asyncio.Event` suspension), checks
-`shared_state` for session failure before and after waiting, then substitutes output segments
-with actual predecessor text. After the response returns, `on_completion()` writes to both
-`registry` (unblocking dependents on the same worker) and `shared_state` (signalling the main
-process) in a single call.
+- **`session_completion_queue`** — event-driven worker→main communication. When the last node
+  of a session completes, the worker pushes a completion notification (with node completion
+  times and failure status) to an `mp.Queue`. The main process consumes from this queue in
+  `check_session_completed()` instead of polling shared state.
+
+Each `OTelChatCompletionAPIData` holds references to `registry`, `worker_tracker`, and
+`completion_queue`. Before dispatching an HTTP request, the worker calls
+`wait_for_predecessors_and_substitute()`, which awaits predecessors via `registry.require_async()`
+(zero threads — pure `asyncio.Event` suspension), checks `worker_tracker` for session failure
+before and after waiting, then substitutes output segments with actual predecessor text. After
+the response returns, `on_completion()` writes to `registry` (unblocking dependents) and
+`worker_tracker` (recording completion), and pushes to `completion_queue` if this was the last
+node in the session.
 
 ### 3. Load shape
 

[…]
