+++ inference_perf/datagen/otel_trace_replay_datagen.py
@@ -48,7 +48,7 @@ Two orthogonal coordination problems exist in this system:
   Problem B — cross-process, worker → main process:
     The main process needs to know when sessions complete for OTEL span closure and reporting.
     Workers detect session completion and push notifications via mp.Queue.
-    
+
     - WorkerSessionTracker: Per-worker tracking of node completions and session failures
     - session_completion_queue: Workers push completion data when last node finishes
     - Main process consumes from queue in check_session_completed()
@@ -61,13 +61,12 @@ import asyncio
 import glob
 import json
 import logging
-import multiprocessing as mp
 import random
 import time
 from dataclasses import dataclass, field, replace as dc_replace
 from multiprocessing.managers import SyncManager
 from pathlib import Path
-from typing import Any, Dict, List, Optional, Set, Tuple
+from typing import Any, Dict, List, Optional, Set
 
 from aiohttp import ClientResponse
 
@@ -135,7 +134,6 @@ class OTelInferenceInfo(InferenceInfo):
 # ---------------------------------------------------------------------------
 
 
-
 # ---------------------------------------------------------------------------
 # WorkerSessionTracker — tracks session state within a worker process
 # ---------------------------------------------------------------------------
@@ -143,101 +141,97 @@ class OTelInferenceInfo(InferenceInfo):
 
 class WorkerSessionTracker:
     """Tracks session state within a single worker process.
-    
+
     Since all nodes of a session run on the same worker (session-to-worker affinity),
     each worker can independently track:
     - Which nodes have completed
     - Whether the session has failed
     - Node completion times
-    
+
     This eliminates the need for Manager.dict() and polling for node completion tracking.
     Session completion is communicated to main process via mp.Queue.
     """
-    
+
     def __init__(self):
         """Initialize empty tracking state."""
         # Map: qualified_node_id -> completion_time
         self._node_completions: Dict[str, float] = {}
-        
+
         # Set of session_ids that have failed
         self._failed_sessions: Set[str] = set()
-    
+
     def record_node_completed(self, qualified_node_id: str, completion_time: float) -> None:
         """Record that a node has completed.
-        
+
         Args:
             qualified_node_id: Full node ID (session_id:node_xxx)
             completion_time: Wall-clock time when node completed
         """
         self._node_completions[qualified_node_id] = completion_time
-    
+
     def is_node_completed(self, qualified_node_id: str) -> bool:
         """Check if a node has completed.
-        
+
         Args:
             qualified_node_id: Full node ID (session_id:node_xxx)
-            
+
         Returns:
             True if the node has completed
         """
         return qualified_node_id in self._node_completions
-    
+
     def get_node_completion_time(self, qualified_node_id: str) -> Optional[float]:
         """Get the completion time for a node.
-        
+
         Args:
             qualified_node_id: Full node ID (session_id:node_xxx)
-            
+
         Returns:
             Completion time, or None if not completed
         """
         return self._node_completions.get(qualified_node_id)
-    
+
     def mark_session_failed(self, session_id: str) -> None:
         """Mark a session as failed.
-        
+
         Args:
             session_id: The session ID
         """
         self._failed_sessions.add(session_id)
-    
+
     def is_session_failed(self, session_id: str) -> bool:
         """Check if a session has failed.
-        
+
         Args:
             session_id: The session ID
-            
+
         Returns:
             True if the session has failed
         """
         return session_id in self._failed_sessions
-    
+
     def get_session_node_count(self, session_id: str) -> int:
[…]
