+++ inference_perf/loadgen/load_generator.py
@@ -51,7 +51,7 @@ else:
     # Runtime usage will still require Python 3.11+.
     TaskGroup = object
 
-from typing import List, Tuple, Optional, NamedTuple, Union, Set, Dict, Any
+from typing import List, Tuple, Optional, NamedTuple, Union, Set, Dict
 from types import FrameType
 import time
 import multiprocessing as mp
@@ -425,7 +425,7 @@ class LoadGenerator:
         )  # Sessions waiting to start
         completed_session_ids: Set[str] = set()  # Session IDs that have completed
         session_dispatch_times: Dict[str, float] = {}  # session_id → wall-clock dispatch time
-        
+
         # Cache OTEL instrumentation to avoid redundant calls
         otel_instr = get_otel_instrumentation()
         session_spans: Dict[str, object] = {}  # session_id → OTEL span object
@@ -442,7 +442,7 @@ class LoadGenerator:
                 stage_info["session_rate"] = session_rate
             if timeout is not None:
                 stage_info["timeout"] = timeout
-            
+
             stage_span, stage_context_dict = otel_instr.start_stage_span(stage_id, stage_info)
             logger.info(f"Started stage-level OTEL span for stage {stage_id}")
 
@@ -579,16 +579,17 @@ class LoadGenerator:
                         if session_id not in completed_session_ids:
                             completed_session_ids.add(session_id)
                             newly_completed.append(session_idx)
-                            
+
                             # End OTEL session span (using cached otel_instr)
                             if session_id in session_spans:
                                 # Check if session failed
-                                session_failed = (hasattr(self.datagen, 'shared_state') and
-                                                self.datagen.shared_state.is_session_failed(session_id))
+                                session_failed = hasattr(
+                                    self.datagen, "shared_state"
+                                ) and self.datagen.shared_state.is_session_failed(session_id)
                                 error_msg = "Session failed" if session_failed else None
                                 otel_instr.end_session_span(session_spans[session_id], error_msg)
                                 del session_spans[session_id]
-                            
+
                             logger.info(
                                 f"Session {session_idx} ({session_id}) completed "
                                 f"({len(completed_session_ids)}/{effective_num_sessions} total)"
