### 2026-03-30 15311c6
Remove historical references to Manager.dict() and SessionSharedState
Clean up documentation and comments by removing references to the old
implementation. Focus on describing the current architecture without
explaining what was changed.

Changes:
- inference_perf/datagen/otel_trace_replay_datagen.py: Removed 2 references
- examples/otel/REVIEW.md: Removed comparison to old approach
- examples/otel/README.md: Removed reference to SessionSharedState
