### 2026-03-30 5a2a5e3
Refactor check_session_completed to separate queue processing
Extract queue draining logic into _process_completion_queue() helper method.
This improves code clarity by separating concerns:
- _process_completion_queue(): Drains queue and updates all session states
- check_session_completed(): Simple state lookup after processing queue

Benefits:
- Clearer separation of concerns
- More maintainable code
- Easier to understand what each method does
- No performance change (queue was already drained on every call)
