### 2026-03-30 9c73d8b
fix: Use cryptographic ID generator for unique OTEL trace/span IDs
Implement CryptographicIdGenerator using os.urandom() to generate
unique trace and span IDs, bypassing Python's random module. This
ensures unique IDs across benchmark runs even when random seed is
fixed for reproducibility.

The custom generator:
- Uses os.urandom() for cryptographically secure random bytes
- Generates 64-bit span IDs and 128-bit trace IDs
- Ensures IDs are never 0 per OpenTelemetry spec requirements
- Prevents span ID collisions when running benchmarks multiple times

Fixes issue where fixed random seed cause
[…]
