### 2026-03-30 d7f9643
Use nested dict in WorkerSessionTracker to eliminate string operations
Changed _node_completions from Dict[str, float] to Dict[str, Dict[str, float]]
(session_id -> node_id -> time). Eliminates string splitting and prefix matching.
Updated method signatures and call sites accordingly.
