### 2026-03-30 a9235b4
fix: correct various typos in code and documentation
- Fix typos in README.md (kuberentes-sigs, inferece-perf, previouse, betweem)
- Fix typo in docs/loadgen.md (concurreny)
- Fix typo in docstring in prometheus_client/base.py (relevent)
- Refactor 'prefered' to 'preferred' across the codebase for consistency

Ref: https://github.com/llm-d/llm-d-benchmark/pull/872
