### 2026-03-31 0815658
Add CLI flags as a way to specify benchmark config
