+++ docs/cli_flags.md
@@ -1,0 +1,95 @@
+# Inference-Perf CLI Flags
+
+These command line flags are automatically generated from the internal `Config` schema. You can override any configuration directly from the CLI without using a yaml configuration file.
+
+| Flag | Type | Description |
+| --- | --- | --- |
+| `--api.type` | Enum (completion, chat) | Matches api.type in config |
+| `--api.streaming` | boolean | Matches api.streaming in config |
+| `--api.headers` | JSON | Matches api.headers in config |
+| `--api.slo_unit` | str | Matches api.slo_unit in config |
+| `--api.slo_tpot_header` | str | Matches api.slo_tpot_header in config |
+| `--api.slo_ttft_header` | str | Matches api.slo_ttft_header in config |
+| `--api.response_format.type` | Enum (json_schema, json_object) | Matches api.response_format.type in config |
+| `--api.response_format.name` | str | Matches api.response_format.name in config |
+| `--api.response_format.json_schema` | JSON | Matches api.response_format.json_schema in config |
+| `--data.type` | Enum (mock, shareGPT, synthetic, random, shared_prefix, cnn_dailymail, infinity_instruct, billsum_conversations) | Matches data.type in config |
+| `--data.path` | str | Matches data.path in config |
+| `--data.input_distribution.min` | int | Matches data.input_distribution.min in config |
+| `--data.input_distribution.max` | int | Matches data.input_distribution.max in config |
+| `--data.input_distribution.mean` | float | Matches data.input_distribution.mean in config |
+| `--data.input_distribution.std_dev` | float | Matches data.input_distribution.std_dev in config |
+| `--data.input_distribution.total_count` | int | Matches data.input_distribution.total_count in config |
+| `--data.output_distribution.min` | int | Matches data.output_distribution.min in config |
+| `--data.output_distribution.max` | int | Matches data.output_distribution.max in config |
+| `--data.output_distribution.mean` | float | Matches data.output_distribution.mean in config |
+| `--data.output_distribution.std_dev` | float | Matches data.output_distribution.std_dev in config |
+| `--data.output_distribution.total_count` | int | Matches data.output_distribution.total_count in config |
+| `--data.shared_prefix.num_groups` | int | Matches data.shared_prefix.num_groups in config |
+| `--data.shared_prefix.num_prompts_per_group` | int | Matches data.shared_prefix.num_prompts_per_group in config |
+| `--data.shared_prefix.system_prompt_len` | int | Matches data.shared_prefix.system_prompt_len in config |
+| `--data.shared_prefix.question_len` | int | Matches data.shared_prefix.question_len in config |
+| `--data.shared_prefix.output_len` | int | Matches data.shared_prefix.output_len in config |
+| `--data.shared_prefix.question_distribution.min` | int | Matches data.shared_prefix.question_distribution.min in config |
+| `--data.shared_prefix.question_distribution.max` | int | Matches data.shared_prefix.question_distribution.max in config |
+| `--data.shared_prefix.question_distribution.mean` | float | Matches data.shared_prefix.question_distribution.mean in config |
+| `--data.shared_prefix.question_distribution.std_dev` | float | Matches data.shared_prefix.question_distribution.std_dev in config |
+| `--data.shared_prefix.question_distribution.total_count` | int | Matches data.shared_prefix.question_distribution.total_count in config |
+| `--data.shared_prefix.output_distribution.min` | int | Matches data.shared_prefix.output_distribution.min in config |
+| `--data.shared_prefix.output_distribution.max` | int | Matches data.shared_prefix.output_distribution.max in config |
[…]
