### 2026-03-31 6317dce
add dependency types + expand the mechanism of the detection of causal dependencies.
