+++ inference_perf/datagen/otel_trace_to_replay_graph.py
@@ -124,13 +124,20 @@ def messages_equal(a: Dict[str, Any], b: Dict[str, Any]) -> bool:
     return a.role == b.role and norm_text(message_content_text(a)) == norm_text(message_content_text(b))
 
 
-def output_matches_message(output_text: str, msg: Dict[str, Any]) -> bool:
+def output_matches_message(output_text: str, msg: Dict[str, Any], allow_partial_match=False) -> bool:
     """Return True if msg is an assistant message whose content matches output_text."""
     if msg.role != "assistant":
         return False
     msg_text = norm_text(message_content_text(msg))
     out_text = norm_text(output_text)
-    return msg_text == out_text
+    if msg_text == out_text:
+        return True
+    if not allow_partial_match:
+        return False
+    else: #try partial match
+        if out_text in msg_text: #out_text is entirely contained in msg_text
+            return True
+    return False
 
 
 # ---------------------------------------------------------------------------
@@ -381,27 +388,93 @@ def build_raw_calls(spans: List[Dict[str, Any]], include_errors: bool = False) -
 
 # ---------------------------------------------------------------------------
 # Causal dependency detection (message-level)
-# ---------------------------------------------------------------------------
+from enum import Enum
 
-def is_causal_dep(a: RawCall, b: RawCall) -> bool:
-    """Return True if call B causally depends on call A.
+
+class DEPENDENCY_TYPE(Enum):
+    """Types of dependencies between different llm call nodes."""
+    CAUSAL_FULL_MATCH = "full_match" #full text of output a matches the full text of b
+    CAUSAL_TOOL_CALL_IDS_MATCHED = "tool_call_ids_matched" #all tool call IDs from output a appear in b's messages
+    CAUSAL_SPLIT_PARTS_MATCHED = "split_parts_matched" #the output of a is split into parts, each part appears in b's list of messages.
+    CAUSAL_CONTENT_AND_SPLIT_TOOLS_MATCH = "content_and_split_tools_match" #the output of a contains [content, tool_call_1, tool_call_2, etc], b's list of messages contains [content+tool_call_1, content+tool_call_2, etc).
+    CAUSAL_DROP_CONTENT_SPLIT_PARTS = "drop_content_split_parts" #the output of a contains [content, tool_call_1, tool_call_2, etc], b's list of message contains [tool_call_1, tool_call_2]
+    TEMPORAL = "temporal" #outputs are not matches, this is a temporal dependency only
+
+# ---------------------------------------------------------------------------
+def _try_match_tool_call_ids(a_parts: list, b_messages: list) -> bool:
+    """
+    Check if all tool call IDs from a_parts appear in b_messages.
+    Args:
+        a_parts: List of part dictionaries from output message A
+        b_messages: List of messages from call B
+    
+    Returns:
+        True if all tool calls in a_parts have IDs and all those IDs appear in b_messages
+    """
+    # Extract tool call parts that have IDs
+    tool_call_parts = [p for p in a_parts if p["type"] == "tool_call"]
+    if not tool_call_parts:
+        return False
+    
+    # Get all tool call IDs from a_parts
+    tool_call_ids = [p.get("id") for p in tool_call_parts]
+    
+    # Check if all tool calls have IDs
+    if not all(tc_id is not None for tc_id in tool_call_ids):
+        return False
+    
+    # Extract all tool call IDs from b_messages
+    b_tool_call_ids = set()
+    for msg in b_messages:
+        #we message_info may contain parts to tool_calls.
+        if isinstance(msg, ComplexOtelMessage):
+            if "tool_calls" in msg.message_info:
+                for tool_call in msg.message_info["tool_calls"]:
+                    if tool_call.get("id"):
+                        b_tool_call_ids.add(tool_call["id"])
+            elif "parts" in msg.message_info:
[…]
