### 2026-03-31 b60da91
simplify trace files loading
