### 2026-03-31 5163870
Add comprehensive tests for OTEL trace replay architecture
Tests cover NodeOutputRegistry, WorkerSessionTracker, OTelChatCompletionAPIData,
and OTelTraceReplayDataGenerator. Validates event-driven completion, asyncio.Event
waiting, output substitution, and session lifecycle management. 26 tests passing.
