+++ tests/test_otel_replay_datagen.py
@@ -1,0 +1,774 @@
+#!/usr/bin/env python3
+# Copyright 2025 The Kubernetes Authors.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""
+Tests for OTelTraceReplayDataGenerator output-aware replay.
+
+Tests the OTEL trace replay architecture:
+- NodeOutputRegistry: intra-worker output coordination via asyncio.Event
+- WorkerSessionTracker: per-worker session state tracking
+- OTelChatCompletionAPIData: output substitution and completion notification
+- Session completion via mp.Queue (event-driven, not polling)
+
+Key architectural principles tested:
+1. Session-to-worker affinity: all nodes of a session run on same worker
+2. Zero-thread waiting: asyncio.Event suspension, no OS threads
+3. Event-driven completion: queue notifications, not polling
+4. Nested dict structure: efficient O(1) lookups without string operations
+"""
+import asyncio
+import json
+import multiprocessing as mp
+import sys
+import time
+from pathlib import Path
+from typing import Any, Dict, List, Optional
+from unittest.mock import AsyncMock, MagicMock, patch
+
+import pytest
+
+# Ensure project root is on path
+sys.path.insert(0, str(Path(__file__).parent.parent))
+
+from inference_perf.datagen.otel_trace_replay_datagen import (
+    NodeOutputRegistry,
+    OTelChatCompletionAPIData,
+    OTelInferenceInfo,
+    OTelTraceReplayDataGenerator,
+    OTelTraceReplayEvent,
+    WorkerSessionTracker,
+)
+from inference_perf.datagen.otel_trace_to_replay_graph import (
+    InputSegment,
+    build_graph,
+    build_raw_calls,
+)
+from inference_perf.apis.chat import ChatMessage
+from inference_perf.config import APIConfig, APIType
+
+
+# ---------------------------------------------------------------------------
+# Helpers
+# ---------------------------------------------------------------------------
+
+SIMPLE_CHAIN_JSON = Path(__file__).parent.parent / "examples/otel/test_traces/simple/simple_chain.json"
+
+
+def make_api_config() -> APIConfig:
+    return APIConfig(type=APIType.Chat, streaming=False)
+
+
+def make_mock_tokenizer():
+    """Return a mock tokenizer that counts words as tokens."""
+    tok = MagicMock()
+    tok.count_tokens = lambda text: max(1, len((text or "").split()))
+    return tok
+
+
+def make_mock_response(content: str) -> MagicMock:
+    """Return a mock aiohttp ClientResponse for non-streaming mode."""
+    response = MagicMock()
+    response.json = AsyncMock(return_value={
+        "choices": [{"message": {"role": "assistant", "content": content}}]
+    })
+    return response
+
+
+def make_mock_api_config_streaming(streaming: bool = False) -> MagicMock:
+    cfg = MagicMock()
+    cfg.streaming = streaming
+    return cfg
+
+
+# ---------------------------------------------------------------------------
+# NodeOutputRegistry tests
+# ---------------------------------------------------------------------------
+
+class TestNodeOutputRegistry:
+    """Test intra-worker output coordination via asyncio.Event."""
+
+    def test_record_and_get(self):
[…]
