### 2026-03-31 2ad7076
Remove REVIEW.md, merge relevant sections into the readme
