### 2026-03-31 ad6ce6f
Remove REVIEW.md, merge relevant sections into the readme
