+++ examples/otel/REVIEW.md
@@ -1,213 +1,0 @@
-# OTel Trace Replay: Agentic Workload Benchmarking
-
-This document describes a proposed addition to `inference-perf` that enables benchmarking of **agentic workloads** by replaying LLM call graphs extracted from OpenTelemetry traces. The change adds a new data-generation path alongside the existing one, without modifying any existing behavior.
-
----
-
-## Motivation
-
-Existing load types (`CONSTANT`, `POISSON`, `CONCURRENT`) generate independent, single-turn requests. Agentic applications — tool-calling agents, multi-turn conversations, RAG pipelines — produce **chains of dependent LLM calls** where each call's input includes the output of one or more predecessors. Replaying these workloads requires:
-
-1. A way to express **dependency graphs** between requests within a session.
-2. A mechanism to **substitute real generated outputs** into dependent requests at replay time, so KV cache patterns faithfully reflect production traffic.
-3. **Session-level concurrency control** separate from per-request concurrency.
-4. **Session-level metrics** (success rate, duration, total tokens per session) alongside the existing per-request metrics.
-
----
-
-## API Changes: Generator Class Hierarchy
-
-### Before
-
-```
-DataGenerator (ABC)
-```
-
-`DataGenerator` combined common API validation logic with request-level generation. All generators had to inherit from it even when they didn't stream individual requests.
-
-### After
-
-```
-BaseGenerator (ABC)
-├── DataGenerator       — unchanged request-based generation
-└── SessionGenerator      — new session-based generation for agentic workloads
-```
-
-**`BaseGenerator`** (`inference_perf/datagen/base.py`) is a new abstract base that holds only the shared initialization — API config validation, tokenizer storage, and the `is_prefered_worker_requested()` hook. Both `DataGenerator` and `SessionGenerator` extend it.
-
-**`DataGenerator`** is unchanged from the outside. It retains all its existing abstract methods (`get_data`, `is_io_distribution_supported`, `is_shared_prefix_supported`) and is used by all existing load types.
-
-**`SessionGenerator`** is a new ABC for session-based replay. It defines the contract LoadGen uses to drive agentic workloads:
-
-| Method | Purpose |
-|---|---|
-| `get_session_count()` | Total sessions in the corpus |
-| `get_session_info(index)` | Metadata (session_id, file_path, num_events) |
-| `get_session_event_indices(index)` | Event indices belonging to a session |
-| `get_session_events(index)` | Session events as `LazyLoadInferenceAPIData` |
-| `activate_session(session_id)` | Called by LoadGen when starting a session; marks root nodes ready |
-| `check_session_completed(session_id)` | Polled by LoadGen to detect session completion |
-| `build_session_metric(...)` | Constructs a `SessionLifecycleMetric` after completion |
-| `cleanup_session(session_id)` | Releases per-session state to prevent memory growth |
-
----
-
-## New Implementation: `OTelTraceReplayDataGenerator`
-
-**File:** `inference_perf/datagen/otel_trace_replay_datagen.py`
-
-`OTelTraceReplayDataGenerator` implements `SessionGenerator`. It parses OTel JSON trace files, builds a dependency graph per session, and serves events to LoadGen in the correct order.
-
-### Trace Parsing and Graph Construction
-
-OTel traces contain LLM spans with timestamps and message content. The generator (using helpers in `otel_trace_to_replay_graph.py`) extracts all LLM calls from a trace file and infers dependencies between them by time-window analysis: if a call's input messages overlap with a predecessor's output, a dependency edge is added to the graph. The result is a `ReplayGraph` — a DAG of nodes, where each node is one LLM call and edges represent "must complete before" relationships.
-
-### Output-Aware Replay
-
[…]
