+++ inference_perf/datagen/otel_trace_to_replay_graph.py
@@ -59,8 +59,11 @@ from pathlib import Path
 from typing import Any, Dict, List, Optional, Set
 
 from inference_perf.datagen.export_replay_graph_to_dot import export_to_dot
-from inference_perf.datagen.otel_trace_utils import reconstruct_llm_output, reconstruct_llm_input,  \
-    reconstruct_each_part_in_message_info
+from inference_perf.datagen.otel_trace_utils import (
+    reconstruct_llm_output,
+    reconstruct_llm_input,
+    reconstruct_each_part_in_message_info,
+)
 
 logger = logging.getLogger(__name__)
 
@@ -134,8 +137,8 @@ def output_matches_message(output_text: str, msg: Dict[str, Any], allow_partial_
         return True
     if not allow_partial_match:
         return False
-    else: #try partial match
-        if out_text in msg_text: #out_text is entirely contained in msg_text
+    else:  # try partial match
+        if out_text in msg_text:  # out_text is entirely contained in msg_text
             return True
     return False
 
@@ -157,7 +160,7 @@ def _convert_content_and_tool_calls_to_parts(message: Dict[str, Any]) -> Dict[st
     parts = []
 
     # Add content as a text part if present
-    content = message.get('content')
+    content = message.get("content")
     if content:
         if isinstance(content, str):
             parts.append({"type": "text", "content": content})
@@ -166,28 +169,27 @@ def _convert_content_and_tool_calls_to_parts(message: Dict[str, Any]) -> Dict[st
             parts.extend(content)
 
     # Add tool_calls as tool_call parts
-    tool_calls = message.get('tool_calls', [])
+    tool_calls = message.get("tool_calls", [])
     for tc in tool_calls:
         if isinstance(tc, dict):
             # OpenAI format: {"id": "...", "type": "function", "function": {"name": "...", "arguments": "..."}}
-            if 'function' in tc:
-                parts.append({
-                    "type": "tool_call",
-                    "id": tc.get('id'),
-                    "name": tc['function'].get('name'),
-                    "arguments": tc['function'].get('arguments')
-                })
+            if "function" in tc:
+                parts.append(
+                    {
+                        "type": "tool_call",
+                        "id": tc.get("id"),
+                        "name": tc["function"].get("name"),
+                        "arguments": tc["function"].get("arguments"),
+                    }
+                )
             # Direct format: {"name": "...", "arguments": "..."}
-            elif 'name' in tc:
-                parts.append({
-                    "type": "tool_call",
-                    "id": tc.get('id'),
-                    "name": tc.get('name'),
-                    "arguments": tc.get('arguments')
-                })
+            elif "name" in tc:
+                parts.append(
+                    {"type": "tool_call", "id": tc.get("id"), "name": tc.get("name"), "arguments": tc.get("arguments")}
+                )
 
     # Create new message with parts
-    result = {"role": message.get('role', 'assistant'), "parts": parts}
+    result = {"role": message.get("role", "assistant"), "parts": parts}
     return result
 
 
@@ -214,7 +216,13 @@ def extract_messages(span: Dict[str, Any]) -> List[Dict[str, Any]]:
                 if "tool_calls" in x:
                     # Transform message with content + tool_calls into parts format
                     message_with_parts = _convert_content_and_tool_calls_to_parts(x)
-                    res.append(ComplexOtelMessage(role=role, message_info=message_with_parts, raw_reconstructed_text=reconstruct_llm_input(message_with_parts)))
+                    res.append(
+                        ComplexOtelMessage(
+                            role=role,
[…]
