+++ tests/test_otel_replay_datagen.py
@@ -28,6 +28,7 @@ Key architectural principles tested:
 3. Event-driven completion: queue notifications, not polling
 4. Nested dict structure: efficient O(1) lookups without string operations
 """
+
 import asyncio
 import json
 import multiprocessing as mp
@@ -80,9 +81,7 @@ def make_mock_tokenizer():
 def make_mock_response(content: str) -> MagicMock:
     """Return a mock aiohttp ClientResponse for non-streaming mode."""
     response = MagicMock()
-    response.json = AsyncMock(return_value={
-        "choices": [{"message": {"role": "assistant", "content": content}}]
-    })
+    response.json = AsyncMock(return_value={"choices": [{"message": {"role": "assistant", "content": content}}]})
     return response
 
 
@@ -95,6 +94,7 @@ def make_mock_api_config_streaming(streaming: bool = False) -> MagicMock:
 # ---------------------------------------------------------------------------
 # NodeOutputRegistry tests
 # ---------------------------------------------------------------------------
+
 
 class TestNodeOutputRegistry:
     """Test intra-worker output coordination via asyncio.Event."""
@@ -142,11 +142,10 @@ class TestNodeOutputRegistry:
         """Recording the same node twice should raise ValueError."""
         reg = NodeOutputRegistry()
         reg.record("node_001", "first output", [])
-        
+
         # Attempting to record the same node again should raise ValueError
         with pytest.raises(ValueError, match="Node node_001 has already been recorded"):
             reg.record("node_001", "second output", [])
-
 
     def test_get_messages_by_node_id(self):
         """get_messages_by_node_id returns input messages."""
@@ -163,7 +162,7 @@ class TestNodeOutputRegistry:
         """Recording with empty messages list should still store the messages (as empty list)."""
         reg = NodeOutputRegistry()
         reg.record("node_001", "output", [])
-        
+
         # Should return empty list, not None
         retrieved = reg.get_messages_by_node_id("node_001")
         assert retrieved is not None
@@ -173,6 +172,7 @@ class TestNodeOutputRegistry:
 # ---------------------------------------------------------------------------
 # WorkerSessionTracker tests
 # ---------------------------------------------------------------------------
+
 
 class TestWorkerSessionTracker:
     """Test per-worker session state tracking with nested dict structure."""
@@ -221,6 +221,7 @@ class TestWorkerSessionTracker:
 # OTelChatCompletionAPIData tests
 # ---------------------------------------------------------------------------
 
+
 class TestOTelChatCompletionAPIData:
     """Test output capture, substitution, and completion notification."""
 
@@ -268,15 +269,15 @@ class TestOTelChatCompletionAPIData:
         registry = NodeOutputRegistry()
         tracker = WorkerSessionTracker()
         api_data = self._make_api_data("session_1:node_0", registry, tracker, total_nodes=2)
-        
+
         info = OTelInferenceInfo(
             output_text="test output",
             input_tokens=10,
             output_tokens=5,
         )
-        
+
         api_data.on_completion(info)
-        
+
         # Should be recorded in tracker with unqualified node_id
         assert tracker.is_node_completed("session_1", "node_0")
         assert tracker.get_node_completion_time("session_1", "node_0") is not None
@@ -286,26 +287,26 @@ class TestOTelChatCompletionAPIData:
         """on_completion pushes to queue when last node completes."""
         registry = NodeOutputRegistry()
         tracker = WorkerSessionTracker()
-        
+
         # Use a mock queue that captures put_nowait calls
         completion_notifications = []
         mock_queue = MagicMock()
         mock_queue.put_nowait = lambda data: completion_notifications.append(data)
-        
+
         # Session has 2 nodes total
         # First complete node_0
[…]
