+++ inference_perf/client/modelserver/otel_instrumentation.py
@@ -38,7 +38,7 @@ try:
     from opentelemetry.sdk.trace.id_generator import IdGenerator
     from opentelemetry.semconv_ai import SpanAttributes
     from opentelemetry.trace.propagation.tracecontext import TraceContextTextMapPropagator
-    
+
     OTEL_AVAILABLE = True
 except ImportError:
     OTEL_AVAILABLE = False
@@ -54,43 +54,43 @@ logger = logging.getLogger(__name__)
 class CryptographicIdGenerator(IdGenerator):
     """
     Custom ID generator that uses os.urandom() for cryptographically secure random IDs.
-    
+
     This bypasses Python's random module completely, ensuring unique Trace and Span IDs
     even when the random seed is fixed for reproducibility in other parts of the application.
-    
+
     OpenTelemetry spec requires that IDs cannot be 0, so we regenerate if we get 0.
     """
-    
+
     def generate_span_id(self) -> int:
         """Generate a random 64-bit span ID using os.urandom()."""
         # Generate 8 random bytes and convert to a 64-bit integer
         # Ensure the ID is never 0 (OpenTelemetry spec requirement)
         span_id = 0
         while span_id == 0:
-            span_id = int.from_bytes(os.urandom(8), byteorder='big')
+            span_id = int.from_bytes(os.urandom(8), byteorder="big")
         return span_id
-    
+
     def generate_trace_id(self) -> int:
         """Generate a random 128-bit trace ID using os.urandom()."""
         # Generate 16 random bytes and convert to a 128-bit integer
         # Ensure the ID is never 0 (OpenTelemetry spec requirement)
         trace_id = 0
         while trace_id == 0:
-            trace_id = int.from_bytes(os.urandom(16), byteorder='big')
+            trace_id = int.from_bytes(os.urandom(16), byteorder="big")
         return trace_id
 
 
 class OTelInstrumentation:
     """
     OpenTelemetry instrumentation for LLM API calls.
-    
+
     Provides tracing capabilities following GenAI semantic conventions.
     """
-    
+
     def __init__(self):
         """
         Initialize OTEL instrumentation.
-        
+
         Reads configuration from environment variables:
         - OTEL_TRACES_ENABLED: Set to "true" to enable tracing (default: false)
         - OTEL_EXPORTER_OTLP_ENDPOINT: OTLP endpoint (e.g., "http://localhost:4317")
@@ -102,10 +102,10 @@ class OTelInstrumentation:
         self.service_name = os.getenv("OTEL_SERVICE_NAME", "inference-perf")
         self.otlp_endpoint = os.getenv("OTEL_EXPORTER_OTLP_ENDPOINT")
         self.trace_per_stage = os.getenv("OTEL_TRACE_PER_STAGE", "false").lower() == "true"
-        
+
         self.enabled = enabled and OTEL_AVAILABLE
         self.tracer: Optional[Any] = None
-        
+
         if not OTEL_AVAILABLE and enabled:
             logger.warning(
                 "OpenTelemetry packages not installed. "
@@ -113,40 +113,40 @@ class OTelInstrumentation:
                 "opentelemetry-exporter-otlp-proto-grpc opentelemetry-semantic-conventions-ai"
             )
             self.enabled = False
-        
+
         if self.enabled:
             self._setup_tracer()
             logger.info(f"OTEL tracing enabled for service: {self.service_name}")
-    
+
     def _setup_tracer(self) -> None:
         """Set up the OpenTelemetry tracer."""
         if not OTEL_AVAILABLE:
             return
-            
+
         # Get or create tracer provider
         provider = trace.get_tracer_provider()
-        
+
         # If no provider is set, create a default one
-        if not hasattr(provider, 'add_span_processor'):
[…]
