### 2026-03-31 16d56ac
Add missing configure_otel method to allow unit testing
