### 2026-03-31 38c7a7d
Define input config via Protobufs, autogenerate python types
