+++ api/generated/python/buf/validate/validate_pb2.py
@@ -1,0 +1,469 @@
+# -*- coding: utf-8 -*-
+# Generated by the protocol buffer compiler.  DO NOT EDIT!
+# NO CHECKED-IN PROTOBUF GENCODE
+# source: buf/validate/validate.proto
+# Protobuf Python Version: 6.30.1
+"""Generated protocol buffer code."""
+from google.protobuf import descriptor as _descriptor
+from google.protobuf import descriptor_pool as _descriptor_pool
+from google.protobuf import runtime_version as _runtime_version
+from google.protobuf import symbol_database as _symbol_database
+from google.protobuf.internal import builder as _builder
+_runtime_version.ValidateProtobufRuntimeVersion(
+    _runtime_version.Domain.PUBLIC,
+    6,
+    30,
+    1,
+    '',
+    'buf/validate/validate.proto'
+)
+# @@protoc_insertion_point(imports)
+
+_sym_db = _symbol_database.Default()
+
+
+from google.protobuf import descriptor_pb2 as google_dot_protobuf_dot_descriptor__pb2
+from google.protobuf import duration_pb2 as google_dot_protobuf_dot_duration__pb2
+from google.protobuf import field_mask_pb2 as google_dot_protobuf_dot_field__mask__pb2
+from google.protobuf import timestamp_pb2 as google_dot_protobuf_dot_timestamp__pb2
+
+
+DESCRIPTOR = _descriptor_pool.Default().AddSerializedFile(b'\n\x1b\x62uf/validate/validate.proto\x12\x0c\x62uf.validate\x1a google/protobuf/descriptor.proto\x1a\x1egoogle/protobuf/duration.proto\x1a google/protobuf/field_mask.proto\x1a\x1fgoogle/protobuf/timestamp.proto\"7\n\x04Rule\x12\n\n\x02id\x18\x01 \x01(\t\x12\x0f\n\x07message\x18\x02 \x01(\t\x12\x12\n\nexpression\x18\x03 \x01(\t\"\x86\x01\n\x0cMessageRules\x12\x16\n\x0e\x63\x65l_expression\x18\x05 \x03(\t\x12\x1f\n\x03\x63\x65l\x18\x03 \x03(\x0b\x32\x12.buf.validate.Rule\x12-\n\x05oneof\x18\x04 \x03(\x0b\x32\x1e.buf.validate.MessageOneofRuleJ\x04\x08\x01\x10\x02R\x08\x64isabled\"4\n\x10MessageOneofRule\x12\x0e\n\x06\x66ields\x18\x01 \x03(\t\x12\x10\n\x08required\x18\x02 \x01(\x08\"\x1e\n\nOneofRules\x12\x10\n\x08required\x18\x01 \x01(\x08\"\x8b\t\n\nFieldRules\x12\x16\n\x0e\x63\x65l_expression\x18\x1d \x03(\t\x12\x1f\n\x03\x63\x65l\x18\x17 \x03(\x0b\x32\x12.buf.validate.Rule\x12\x10\n\x08required\x18\x19 \x01(\x08\x12$\n\x06ignore\x18\x1b \x01(\x0e\x32\x14.buf.validate.Ignore\x12)\n\x05\x66loat\x18\x01 \x01(\x0b\x32\x18.buf.validate.FloatRulesH\x00\x12+\n\x06\x64ouble\x18\x02 \x01(\x0b\x32\x19.buf.validate.DoubleRulesH\x00\x12)\n\x05int32\x18\x03 \x01(\x0b\x32\x18.buf.validate.Int32RulesH\x00\x12)\n\x05int64\x18\x04 \x01(\
[…]
