### 2026-03-31 292ad20
Fix all mypy strict type errors and consolidate error handling
- Fixed 164 mypy --strict type errors across 11 files
- Added missing type annotations and proper generic types
- Consolidated error handling in openai_client to fix duplicate process_failure() calls
- All CI checks passing: format, lint, type-check, unit tests, e2e tests
