+++ inference_perf/metrics/session_collector.py
@@ -88,6 +88,3 @@ class SessionMetricsCollector:
             sm.total_output_tokens = out
             sm.error = error_by_session.get(sm.session_id)
             sm.success = (sm.num_nodes_completed == sm.num_nodes) and (sm.error is None)
-
-
-# Made with Bob
