+++ inference_perf/datagen/otel_trace_utils.py
@@ -29,7 +29,7 @@ import json
 from typing import Dict, List, Any, Union, Optional
 
 
-def reconstruct_llm_output(response_data: Union[str, Dict, List]) -> str:
+def reconstruct_llm_output(response_data: Union[str, Dict[str, Any], List[Any]]) -> str:
     """
     Reconstruct the raw LLM output from any parsed response structure.
 
@@ -56,15 +56,17 @@ def reconstruct_llm_output(response_data: Union[str, Dict, List]) -> str:
     """
 
     # Parse if string
+    data_to_process: Union[str, Dict[str, Any], List[Any]] = response_data
     if isinstance(response_data, str):
         try:
-            response_data = json.loads(response_data)
+            parsed_data: Union[Dict[str, Any], List[Any]] = json.loads(response_data)
+            data_to_process = parsed_data
         except json.JSONDecodeError:
             # If it's not JSON, assume it's already raw text
             return response_data
 
     # Extract the message content
-    message = _extract_message(response_data)
+    message = _extract_message(data_to_process)
 
     if not message:
         return ""
@@ -87,7 +89,7 @@ def reconstruct_llm_output(response_data: Union[str, Dict, List]) -> str:
     return "\n".join(output_parts) if output_parts else ""
 
 
-def _extract_message(data: Any) -> Optional[Dict]:
+def _extract_message(data: Any) -> Optional[Dict[str, Any]]:
     """
     Extract the assistant message from various response formats.
 
@@ -109,14 +111,18 @@ def _extract_message(data: Any) -> Optional[Dict]:
     elif isinstance(data, dict):
         # Check for OpenAI choices format
         if "choices" in data and data["choices"]:
-            choice = data["choices"][0]
+            choice: Any = data["choices"][0]
             if "message" in choice:
-                return choice["message"]
-            return choice
+                result: Dict[str, Any] = choice["message"]
+                return result
+            # Cast choice to Dict[str, Any] since we know it's a dict at this point
+            choice_dict: Dict[str, Any] = choice
+            return choice_dict
 
         # Check for direct message key
         if "message" in data:
-            return data["message"]
+            msg: Dict[str, Any] = data["message"]
+            return msg
 
         # Check if data itself is a message
         if "role" in data or "content" in data or "parts" in data or "tool_calls" in data:
@@ -125,7 +131,7 @@ def _extract_message(data: Any) -> Optional[Dict]:
     return None
 
 
-def _extract_text_content(message: Dict) -> str:
+def _extract_text_content(message: Dict[str, Any]) -> str:
     """
     Extract text content from a message.
 
@@ -167,7 +173,7 @@ def _extract_text_content(message: Dict) -> str:
     return "\n".join(text_parts)
 
 
-def _extract_tool_calls(message: Dict) -> List[Dict]:
+def _extract_tool_calls(message: Dict[str, Any]) -> List[Dict[str, Any]]:
     """
     Extract tool calls from a message.
 
@@ -206,7 +212,7 @@ def _extract_tool_calls(message: Dict) -> List[Dict]:
     return tool_calls
 
 
-def _format_tool_call(tool_call: Dict) -> str:
+def _format_tool_call(tool_call: Dict[str, Any]) -> str:
     """
     Format a tool call as the LLM would have generated it.
 
@@ -223,7 +229,7 @@ def _format_tool_call(tool_call: Dict) -> str:
     return f"<|tool_call|>{function_name}<|tool_args|>{arguments}<|end|>"
 
 
-def reconstruct_each_part_in_message_info(message_info):
+def reconstruct_each_part_in_message_info(message_info: Dict[str, Any]) -> Dict[str, Any]:
[…]
