+++ examples/otel/run_with_jaeger.sh
@@ -82,4 +82,4 @@ echo "✓ Benchmark complete!"
 echo "View traces at: http://localhost:16686"
 echo "=================================================="
 
-# Made with Bob
+

[…]
