+++ examples/otel/configs/per_case_config/fan_in_dependencies.yml
@@ -40,5 +40,5 @@ report:
     per_stage: false
     per_request: false
 
-# Made with Bob
+
 
