+++ examples/otel/configs/per_case_config/customer_support_escalation.yml
@@ -47,4 +47,4 @@ report:
     per_stage: false
     per_request: false
 
-# Made with Bob
+
