+++ examples/otel/configs/advanced/graph-replay.yml
@@ -44,4 +44,3 @@ report:
     per_stage: true
     per_request: false
 
-# Made with Bob
