+++ examples/otel/configs/external/synthetic_conversations.yml
@@ -46,4 +46,4 @@ report:
 # financial_planning, pet_care, learning_language, mental_health)
 # Each trace has 5-15 conversation turns with varied content
 
-# Made with Bob
+

[…]
