### 2026-03-31 92e9b85
remove unused configure_otel func
