+++ inference_perf/client/modelserver/otel_instrumentation.py
@@ -494,48 +494,3 @@ def get_otel_instrumentation(service_name: Optional[str] = None, enabled: Option
         _global_instrumentation = OTelInstrumentation(service_name=service_name, enabled=enabled)
 
     return _global_instrumentation
-
-
-def configure_otel(service_name: Optional[str] = None, enabled: Optional[bool] = None) -> None:
-    """
-    Configure the global OTEL instrumentation instance.
-
-    This function allows programmatic configuration of OTEL settings,
-    overriding environment variables. It will recreate the global instance
-    with the new settings.
-
-    Args:
-        service_name: Service name for tracing (overrides OTEL_SERVICE_NAME env var)
-        enabled: Whether to enable tracing (overrides OTEL_TRACES_ENABLED env var)
-    """
-    global _global_instrumentation
-
-    # Shutdown existing instance if it exists
-    if _global_instrumentation is not None:
-        _global_instrumentation.shutdown()
-
-    # Temporarily override environment variables
-    original_service_name = os.getenv("OTEL_SERVICE_NAME")
-    original_enabled = os.getenv("OTEL_TRACES_ENABLED")
-
-    try:
-        if service_name is not None:
-            os.environ["OTEL_SERVICE_NAME"] = service_name
-        if enabled is not None:
-            os.environ["OTEL_TRACES_ENABLED"] = "true" if enabled else "false"
-
-        # Create new instance with updated configuration
-        _global_instrumentation = OTelInstrumentation()
-
-    finally:
-        # Restore original environment variables
-        if service_name is not None:
-            if original_service_name is not None:
-                os.environ["OTEL_SERVICE_NAME"] = original_service_name
-            else:
-                os.environ.pop("OTEL_SERVICE_NAME", None)
-        if enabled is not None:
-            if original_enabled is not None:
-                os.environ["OTEL_TRACES_ENABLED"] = original_enabled
-            else:
-                os.environ.pop("OTEL_TRACES_ENABLED", None)
