### 2026-03-31 f3962c5
remove unused configure_otel func
