+++ tests/utils/test_cli_summary.py
@@ -1,23 +1,23 @@
 import unittest
-from unittest.mock import patch
+from unittest.mock import patch, MagicMock
 from inference_perf.utils.cli_summary import extract_stage_id, print_summary_table
 from inference_perf.utils.report_file import ReportFile
 
 
 class TestCliSummary(unittest.TestCase):
-    def test_extract_stage_id(self):
+    def test_extract_stage_id(self) -> None:
         self.assertEqual(extract_stage_id("stage_0_lifecycle_metrics"), 0)
         self.assertEqual(extract_stage_id("stage_12_lifecycle_metrics"), 12)
         self.assertIsNone(extract_stage_id("invalid_name"))
         self.assertIsNone(extract_stage_id("stage_abc_lifecycle_metrics"))
 
     @patch("inference_perf.utils.cli_summary.rprint")
-    def test_print_summary_table_empty(self, mock_rprint):
+    def test_print_summary_table_empty(self, mock_rprint: MagicMock) -> None:
         print_summary_table([])
         mock_rprint.assert_called_once_with("[yellow]No per-stage lifecycle metrics found to display summary table.[/yellow]")
 
     @patch("inference_perf.utils.cli_summary.Console.print")
-    def test_print_summary_table_with_reports(self, mock_console_print):
+    def test_print_summary_table_with_reports(self, mock_console_print: MagicMock) -> None:
         # Create a mock report file
         contents = {
             "load_summary": {"requested_rate": 10.0, "achieved_rate": 9.5},
