### 2026-04-01 117d169
refactor: extract SSE streaming parser to reduce code duplication
- Created new streaming_parser.py module with parse_sse_stream function
- Refactored chat.py, completion.py, and otel_trace_replay_datagen.py to use shared parser
- Reduced duplicate SSE parsing logic across multiple API implementations
- Maintained existing functionality while improving code maintainability
- Used callback pattern for API-specific content extraction

Signed-off-by: Yoav Katz <katz@il.ibm.com>
