### 2026-04-01 73b3567
fix: regenerate pdm.lock with analysis group included
Fix PDM lockfile and mypy type checking for OpenTelemetry integration

The lockfile was missing the 'analysis' optional dependency group, causing GitHub workflow failures when running 'pdm install -G analysis'.

Changes:

Regenerated lockfile using 'pdm lock -G :all' to include all groups (default, analysis, dev, lint, otel, test, types)
Updated GitHub workflows to install both analysis and otel groups together: pdm sync -d -G analysis -G otel
Fixed mypy type checking for OpenTelemetry's IdGenerator by using a local stub class under TYPE_CH
[…]
