+++ pyproject.toml
@@ -26,14 +26,6 @@ license = {text = "Apache-2.0"}
 inference-perf = "inference_perf:main_cli"
 
 [project.optional-dependencies]
-otel = [
-    "opentelemetry-api>=1.30.0",
-    "opentelemetry-sdk>=1.30.0",
-    "opentelemetry-instrumentation>=0.51b0",
-    "opentelemetry-instrumentation-aiohttp-client>=0.51b0",
-    "opentelemetry-semantic-conventions-ai>=0.5.0",
-    "opentelemetry-exporter-otlp-proto-grpc>=1.30.0",
-]
 analysis = [
     "matplotlib>=3.9.0",
 ]
@@ -47,7 +39,14 @@ dev = [
     "hypothesis>=6.146.0",
     "pytest-asyncio>=1.3.0",
 ]
-
+otel = [
+    "opentelemetry-api>=1.30.0",
+    "opentelemetry-sdk>=1.30.0",
+    "opentelemetry-instrumentation>=0.51b0",
+    "opentelemetry-instrumentation-aiohttp-client>=0.51b0",
+    "opentelemetry-semantic-conventions-ai>=0.5.0",
+    "opentelemetry-exporter-otlp-proto-grpc>=1.30.0",
+]
 [dependency-groups]
 lint = [
     "mypy>=1.14.1",
@@ -156,6 +155,7 @@ module = [
     "datasets.*",
     "matplotlib.*",
     "boto3.*",
+    "opentelemetry.*",
 ]
 ignore_missing_imports = true
 
