### 2026-04-01 523868d
test: Add comprehensive tests for OTEL trace replay features (#9)
- Add tests for otel_trace_utils.py (5% -> 63% coverage)
- Add error path tests for otel_trace_replay_datagen.py (55% -> 56%)
- Fix coverage regression: 47.45% -> 51.0% overall
- Total: 12 new tests (10 utils + 2 error paths)

Tests cover:
- LLM output/input reconstruction from various formats
- Message extraction (OpenAI, LiteLLM, OTEL)
- Tool call handling and formatting
- Token estimation
- Session failure handling
- Predecessor wait failure checks
