### 2026-04-01 ccea0aa
Create proto type for output reports
