+++ tests/reportgen/test_report_serialization.py
@@ -1,0 +1,204 @@
+from typing import Any
+import pytest
+from google.protobuf import json_format
+from api.generated.python.api.outputs.reports.native.v1 import native_pb2
+from inference_perf.reportgen.base import ResponsesSummary
+
+
+@pytest.fixture
+def mock_report_data() -> ResponsesSummary:
+    return ResponsesSummary(
+        load_summary={
+            "count": 60,
+            "schedule_delay": {
+                "mean": 0.00044983585054675737,
+                "min": -0.0006079088198021054,
+                "p0.1": -0.0005960080651566386,
+                "p1": -0.0004889012733474374,
+                "p5": -0.0002833204809576273,
+                "p10": -0.0002156400354579091,
+                "p25": 0.00015595677541568875,
+                "median": 0.0005210512899793684,
+                "p75": 0.0007308190106414258,
+                "p90": 0.0009630696615204217,
+                "p95": 0.0011399090639315544,
+                "p99": 0.0015012771042529482,
+                "p99.9": 0.0016855146751040626,
+                "max": 0.0017059855163097382,
+            },
+            "send_duration": 58.51001774193719,
+            "requested_rate": 1.0,
+            "achieved_rate": 1.0254654214024423,
+        },
+        successes={
+            "count": 60,
+            "latency": {
+                "request_latency": {
+                    "mean": 0.48048226445680486,
+                    "min": 0.47253478108905256,
+                    "p0.1": 0.4726674221145222,
+                    "p1": 0.47386119134374893,
+                    "p5": 0.4752197924011853,
+                    "p10": 0.475656164821703,
+                    "p25": 0.4778720944887027,
+                    "median": 0.4796094944467768,
+                    "p75": 0.48374426225200295,
+                    "p90": 0.48580294536659496,
+                    "p95": 0.48651613232214,
+                    "p99": 0.489441333723953,
+                    "p99.9": 0.4899891945802374,
+                    "max": 0.4900500680087134,
+                },
+                "normalized_time_per_output_token": {
+                    "mean": 0.0076659232914650915,
+                    "min": 0.007025085322846494,
+                    "p0.1": 0.007034749209689163,
+                    "p1": 0.007121724191273178,
+                    "p5": 0.007270589464601609,
+                    "p10": 0.00731774064945057,
+                    "p25": 0.007364180857817141,
+                    "median": 0.007457680346515889,
+                    "p75": 0.007612397971411516,
+                    "p90": 0.00841689943131713,
+                    "p95": 0.008489072611285209,
+                    "p99": 0.01064582721341003,
+                    "p99.9": 0.013100320472625663,
+                    "max": 0.013373041945871793,
+                },
+                "time_per_output_token": {
+                    "mean": 0.0034734172405264745,
+                    "min": 0.0034534218531578433,
+                    "p0.1": 0.003453474750541974,
+                    "p1": 0.0034539508269991466,
+                    "p5": 0.003455240081665431,
+                    "p10": 0.0034566809881709576,
+                    "p25": 0.003459879160405664,
+                    "median": 0.003468046240533825,
+                    "p75": 0.0034801688969366314,
+                    "p90": 0.0034946888832324873,
+                    "p95": 0.003515250179709058,
+                    "p99": 0.003521602816623455,
+                    "p99.9": 0.0035254566003550423,
+                    "max": 0.003525884798547441,
+                },
+                "time_to_first_token": {
+                    "mean": 0.03112733216257766,
+                    "min": 0.02462312299758196,
+                    "p0.1": 0.024646742470911705,
+                    "p1": 0.02485931773087941,
[…]
