+++ scripts/check_coverage_regression.py
@@ -97,7 +97,7 @@ def generate_current(output_path: Path):
     print("--- Generating current coverage ---")
     # Use the current python executable to run pytest directly
     # This avoids issues with nested 'pdm run' calls and ensures we use the same environment
-    cmd = f"{sys.executable} -m pytest --cov=inference_perf --cov-report=json:{output_path.absolute()} tests/"
+    cmd = f"PYTHONPATH=.:api/generated/python {sys.executable} -m pytest --cov=inference_perf --cov-report=json:{output_path.absolute()} tests/"
     run_command(cmd)
     print(f"✅ Current report generated: {output_path.name}")
 
