+++ api/generated/python/api/outputs/reports/native/v1/native_pb2.py
@@ -1,0 +1,48 @@
+# -*- coding: utf-8 -*-
+# Generated by the protocol buffer compiler.  DO NOT EDIT!
+# source: api/outputs/reports/native/v1/native.proto
+# Protobuf Python Version: 4.25.6
+"""Generated protocol buffer code."""
+from google.protobuf import descriptor as _descriptor
+from google.protobuf import descriptor_pool as _descriptor_pool
+from google.protobuf import symbol_database as _symbol_database
+from google.protobuf.internal import builder as _builder
+# @@protoc_insertion_point(imports)
+
+_sym_db = _symbol_database.Default()
+
+
+
+
+DESCRIPTOR = _descriptor_pool.Default().AddSerializedFile(b'\n*api/outputs/reports/native/v1/native.proto\x12\x1d\x61pi.outputs.reports.native.v1\"\xfa\x01\n\x0cSummaryStats\x12\x12\n\x04mean\x18\x01 \x01(\x01R\x04mean\x12\x10\n\x03min\x18\x02 \x01(\x01R\x03min\x12\x10\n\x03max\x18\x03 \x01(\x01R\x03max\x12^\n\x0bpercentiles\x18\x04 \x03(\x0b\x32<.api.outputs.reports.native.v1.SummaryStats.PercentilesEntryR\x0bpercentiles\x12\x12\n\x04rate\x18\x05 \x01(\x01R\x04rate\x1a>\n\x10PercentilesEntry\x12\x10\n\x03key\x18\x01 \x01(\tR\x03key\x12\x14\n\x05value\x18\x02 \x01(\x01R\x05value:\x02\x38\x01\"\xf4\x03\n\x0eLatencyMetrics\x12T\n\x0frequest_latency\x18\x01 \x01(\x0b\x32+.api.outputs.reports.native.v1.SummaryStatsR\x0erequestLatency\x12s\n normalized_time_per_output_token\x18\x02 \x01(\x0b\x32+.api.outputs.reports.native.v1.SummaryStatsR\x1cnormalizedTimePerOutputToken\x12^\n\x15time_per_output_token\x18\x03 \x01(\x0b\x32+.api.outputs.reports.native.v1.SummaryStatsR\x12timePerOutputToken\x12Z\n\x13time_to_first_token\x18\x04 \x01(\x0b\x32+.api.outputs.reports.native.v1.SummaryStatsR\x10timeToFirstToken\x12[\n\x13inter_token_latency\x18\x05 \x01(\x0b\x32+.api.outputs.reports.native.v1.SummaryStatsR\x11interTokenLatency\"\xd2\x01\n\x11ThroughputMetrics\x12/\n\x14input_tokens_per_sec\x18\x01 \x01(\x01R\x11inputTokensPerSec\x12\x31\n\x15output_tokens_per_sec\x18\x02 \x01(\x01R\x12outputTokensPerSec\x12/\n\x14total_tokens_per_sec\x18\x03 \x01(\x01R\x11totalTokensPerSec\x12(\n\x10requests_per_sec\x18\x04 \x01(\x01R\x0erequestsPerSec\"\xbb\x01\n\rSloAttainment\x12%\n\x0e\x61ttainment_pct\x18\x01 \x01(\x01R\rattainmentPct\x12!\n\x0crequests_met\x18\x02 \x01(\x03R\x0brequestsMet\x12\'\n\x0frequests_failed\x18\x03 \x01(\x03R\x0erequestsFailed\x12%\n\x0etotal_requests\x18\x04 \x01(\x03R\rtotalRequests\x12\x10\n\x03slo\x18\x05 \x01(\x01R
[…]
