+++ api/generated/python/api/config/v1/config_p2p.py
@@ -1,6 +1,6 @@
 # This is an automatically generated file, please do not change
 # gen by protobuf_to_pydantic[v0.3.3.1](https://github.com/so1n/protobuf_to_pydantic)
-# Protobuf Version: 6.33.6
+# Protobuf Version: 6.33.2
 # Pydantic Version: 2.12.5
 from enum import IntEnum
 from google.protobuf.message import Message  # type: ignore
