+++ buf.gen.yaml
@@ -6,4 +6,4 @@ plugins:
     out: api/generated/python
   - remote: buf.build/community/pseudomuto-doc:v1.5.1
     out: docs/api/config
-    opt: markdown,index.md
+    opt: markdown,index.md,source_relative
