### 2026-04-01 93ab0d6
cleanup prometheus metric querying
