### 2026-04-01 0f3b9a6
cleanup prometheus metric querying
