+++ inference_perf/client/modelserver/sglang_client.py
@@ -12,10 +12,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from inference_perf.client.modelserver.openai_client import openAIModelServerClient
+from inference_perf.client.modelserver.openai_client import openAIModelServerClient, OpenAIMetrics
 from inference_perf.metrics.request_collector import RequestMetricCollector
 from inference_perf.config import APIConfig, APIType, CustomTokenizerConfig, MultiLoRAConfig
-from .base import PrometheusMetricMetadata, ModelServerPrometheusMetric
+from .base import GaugeMetric, HistogramMetric, CounterMetric, RequestsMetric
 from typing import List, Optional
 import logging
 
@@ -55,225 +55,19 @@ class SGlangModelServerClient(openAIModelServerClient):
     def get_supported_apis(self) -> List[APIType]:
         return [APIType.Completion, APIType.Chat]
 
-    def get_prometheus_metric_metadata(self) -> PrometheusMetricMetadata:
-        return PrometheusMetricMetadata(
-            avg_queue_length=ModelServerPrometheusMetric(
-                "sglang:num_queue_reqs",
-                "mean",
-                "gauge",
-                self.metric_filters,
+    def get_prometheus_metric_metadata(self) -> OpenAIMetrics:
+        return OpenAIMetrics(
+            prompt_tokens=CounterMetric("prompt_tokens", "sglang:prompt_tokens_total", self.metric_filters),
+            output_tokens=CounterMetric("output_tokens", "sglang:generation_tokens_total", self.metric_filters),
+            requests=RequestsMetric("sglang:e2e_request_latency_seconds_count", self.metric_filters),
+            request_latency=HistogramMetric("request_latency", "sglang:e2e_request_latency_seconds", self.metric_filters),
+            queue_length=GaugeMetric("queue_length", "sglang:num_queue_reqs", self.metric_filters),
+            time_per_output_token=HistogramMetric(
+                "time_per_output_token", "sglang:time_per_output_token_seconds", self.metric_filters
             ),
-            avg_time_to_first_token=ModelServerPrometheusMetric(
-                "sglang:time_to_first_token_seconds",
-                "mean",
-                "histogram",
-                self.metric_filters,
-            ),
-            median_time_to_first_token=ModelServerPrometheusMetric(
-                "sglang:time_to_first_token_seconds",
-                "median",
-                "histogram",
-                self.metric_filters,
-            ),
-            p90_time_to_first_token=ModelServerPrometheusMetric(
-                "sglang:time_to_first_token_seconds",
-                "p90",
-                "histogram",
-                self.metric_filters,
-            ),
-            p99_time_to_first_token=ModelServerPrometheusMetric(
-                "sglang:time_to_first_token_seconds",
-                "p99",
-                "histogram",
-                self.metric_filters,
-            ),
-            avg_inter_token_latency=ModelServerPrometheusMetric(
-                "sglang:inter_token_latency_seconds",
-                "mean",
-                "histogram",
-                self.metric_filters,
-            ),
-            median_inter_token_latency=ModelServerPrometheusMetric(
-                "sglang:inter_token_latency_seconds",
-                "median",
-                "histogram",
-                self.metric_filters,
-            ),
-            p90_inter_token_latency=ModelServerPrometheusMetric(
-                "sglang:inter_token_latency_seconds",
-                "p90",
-                "histogram",
-                self.metric_filters,
-            ),
-            p99_inter_token_latency=ModelServerPrometheusMetric(
-                "sglang:inter_token_latency_seconds",
-                "p99",
-                "histogram",
-                self.metric_filters,
-            ),
-            avg_prompt_tokens=ModelServerPrometheusMetric(
[…]
