### 2026-04-01 48c2189
cleanup prometheus metric querying
