### 2026-04-01 44383de
cleanup prometheus metric querying
