+++ inference_perf/client/modelserver/openai_client.py
@@ -26,7 +26,7 @@ from inference_perf.apis.anthropic_messages import ANTHROPIC_VERSION, parse_anth
 from inference_perf.apis.streaming_parser import StreamInterruptedError
 from inference_perf.payloads import RequestMetrics, Text
 from inference_perf.utils import CustomTokenizer
-from .base import ModelServerClient, ModelServerClientSession, PrometheusMetricMetadata
+from .base import ModelServerClient, ModelServerClientSession, Metric, BaseMetrics
 from .otel_instrumentation import get_otel_instrumentation
 from typing import List, Optional, Any, Dict
 import aiohttp
@@ -39,6 +39,45 @@ import ssl
 
 
 logger = logging.getLogger(__name__)
+
+
+class OpenAIMetrics(BaseMetrics):
+    def __init__(
+        self,
+        prompt_tokens: Metric[Any],
+        output_tokens: Metric[Any],
+        requests: Metric[Any],
+        request_latency: Metric[Any],
+        queue_length: Metric[Any],
+        time_per_output_token: Metric[Any],
+        custom_metrics: Optional[List[Metric[Any]]] = None,
+    ) -> None:
+        super().__init__(custom_metrics)
+        self.prompt_tokens = prompt_tokens
+        self.output_tokens = output_tokens
+        self.requests = requests
+        self.request_latency = request_latency
+        self.queue_length = queue_length
+        self.time_per_output_token = time_per_output_token
+
+    def get_all_metrics(self) -> List[Metric[Any]]:
+        seen = set()
+        metrics: List[Metric[Any]] = []
+
+        def add_metric(m: Metric[Any]) -> None:
+            if id(m) not in seen:
+                seen.add(id(m))
+                metrics.append(m)
+
+        for attr_value in self.__dict__.values():
+            if isinstance(attr_value, Metric):
+                add_metric(attr_value)
+            elif isinstance(attr_value, list):
+                for item in attr_value:
+                    if isinstance(item, Metric):
+                        add_metric(item)
+
+        return metrics
 
 
 class openAIModelServerClient(ModelServerClient):
@@ -139,7 +178,7 @@ class openAIModelServerClient(ModelServerClient):
         return []
 
     @abstractmethod
-    def get_prometheus_metric_metadata(self) -> PrometheusMetricMetadata:
+    def get_prometheus_metric_metadata(self) -> OpenAIMetrics:
         raise NotImplementedError
 
     def get_supported_models(self) -> List[dict[str, Any]]:
