### 2026-04-01 5c53e7f
cleanup prometheus metric querying
