+++ inference_perf/client/modelserver/mock_client.py
@@ -23,7 +23,7 @@ from inference_perf.apis import (
     UnaryResponseMetrics,
 )
 from inference_perf.payloads import RequestMetrics, Text
-from .base import ModelServerClient, ModelServerPrometheusMetric, PrometheusMetricMetadata
+from .base import ModelServerClient, BaseMetrics
 import asyncio
 import time
 import logging
@@ -102,112 +102,5 @@ class MockModelServerClient(ModelServerClient):
     def get_supported_apis(self) -> List[APIType]:
         return [APIType.Completion, APIType.Chat, APIType.AnthropicMessages]
 
-    def get_prometheus_metric_metadata(self) -> PrometheusMetricMetadata:
-        mock_prometheus_metric = ModelServerPrometheusMetric(
-            name="mock_metric",
-            op="mean",
-            type="counter",
-            filters=[],
-        )
-        return PrometheusMetricMetadata(
-            # Throughput
-            prompt_tokens_per_second=mock_prometheus_metric,
-            output_tokens_per_second=mock_prometheus_metric,
-            requests_per_second=mock_prometheus_metric,
-            # Latency
-            avg_request_latency=mock_prometheus_metric,
-            median_request_latency=mock_prometheus_metric,
-            p90_request_latency=mock_prometheus_metric,
-            p99_request_latency=mock_prometheus_metric,
-            # Request
-            total_requests=mock_prometheus_metric,
-            avg_prompt_tokens=mock_prometheus_metric,
-            avg_output_tokens=mock_prometheus_metric,
-            avg_queue_length=mock_prometheus_metric,
-            # Others
-            avg_time_to_first_token=None,
-            median_time_to_first_token=None,
-            p90_time_to_first_token=None,
-            p99_time_to_first_token=None,
-            avg_time_per_output_token=None,
-            median_time_per_output_token=None,
-            p90_time_per_output_token=None,
-            p99_time_per_output_token=None,
-            avg_inter_token_latency=None,
-            median_inter_token_latency=None,
-            p90_inter_token_latency=None,
-            p99_inter_token_latency=None,
-            avg_kv_cache_usage=None,
-            median_kv_cache_usage=None,
-            p90_kv_cache_usage=None,
-            p99_kv_cache_usage=None,
-            num_preemptions_total=None,
-            num_requests_swapped=None,
-            prefix_cache_hits=None,
-            prefix_cache_queries=None,
-            avg_num_requests_running=None,
-            avg_request_queue_time=None,
-            median_request_queue_time=None,
-            p90_request_queue_time=None,
-            p99_request_queue_time=None,
-            avg_request_inference_time=None,
-            median_request_inference_time=None,
-            p90_request_inference_time=None,
-            p99_request_inference_time=None,
-            avg_request_prefill_time=None,
-            median_request_prefill_time=None,
-            p90_request_prefill_time=None,
-            p99_request_prefill_time=None,
-            avg_request_decode_time=None,
-            median_request_decode_time=None,
-            p90_request_decode_time=None,
-            p99_request_decode_time=None,
-            avg_request_prompt_tokens=None,
-            median_request_prompt_tokens=None,
-            p90_request_prompt_tokens=None,
-            p99_request_prompt_tokens=None,
-            avg_request_generation_tokens=None,
-            median_request_generation_tokens=None,
-            p90_request_generation_tokens=None,
-            p99_request_generation_tokens=None,
-            avg_request_max_num_generation_tokens=None,
-            median_request_max_num_generation_tokens=None,
-            p90_request_max_num_generation_tokens=None,
-            p99_request_max_num_generation_tokens=None,
[…]
