+++ inference_perf/client/server_metrics/base.py
@@ -13,13 +13,16 @@
 # limitations under the License.
 from abc import ABC, abstractmethod
 from enum import Enum, auto
-from typing import Optional, TypedDict
-from pydantic import BaseModel
+from typing import Optional
+from pydantic import BaseModel, Field
 
-
-# Base class for accumulating metrics objects on
-class MetricsMetadata(TypedDict):
-    pass
+from inference_perf.client.modelserver.base import (
+    BaseMetrics,
+    HistogramResult,
+    GaugeResult,
+    RequestsResult,
+    CounterResult,
+)
 
 
 class StageStatus(Enum):
@@ -41,7 +44,11 @@ class StageRuntimeInfo(BaseModel):
 
 class PerfRuntimeParameters:
     def __init__(
-        self, start_time: float, duration: float, model_server_metrics: MetricsMetadata, stages: dict[int, StageRuntimeInfo]
+        self,
+        start_time: float,
+        duration: float,
+        model_server_metrics: BaseMetrics,
+        stages: dict[int, StageRuntimeInfo],
     ) -> None:
         self.start_time = start_time
         self.duration = duration
@@ -51,122 +58,52 @@ class PerfRuntimeParameters:
 
 class ModelServerMetrics(BaseModel):
     # Throughput
-    prompt_tokens_per_second: float = 0.0
-    output_tokens_per_second: float = 0.0
-    requests_per_second: float = 0.0
+    prompt_tokens: HistogramResult = Field(default_factory=HistogramResult)
+    output_tokens: HistogramResult = Field(default_factory=HistogramResult)
+    requests: RequestsResult = Field(default_factory=RequestsResult)
 
     # Latency
-    avg_request_latency: float = 0.0
-    median_request_latency: float = 0.0
-    p90_request_latency: float = 0.0
-    p99_request_latency: float = 0.0
-    avg_time_to_first_token: float = 0.0
-    median_time_to_first_token: float = 0.0
-    p90_time_to_first_token: float = 0.0
-    p99_time_to_first_token: float = 0.0
-    avg_time_per_output_token: float = 0.0
-    median_time_per_output_token: float = 0.0
-    p90_time_per_output_token: float = 0.0
-    p99_time_per_output_token: float = 0.0
-    avg_inter_token_latency: float = 0.0
-    median_inter_token_latency: float = 0.0
-    p90_inter_token_latency: float = 0.0
-    p99_inter_token_latency: float = 0.0
+    request_latency: HistogramResult = Field(default_factory=HistogramResult)
+    time_to_first_token: HistogramResult = Field(default_factory=HistogramResult)
+    time_per_output_token: HistogramResult = Field(default_factory=HistogramResult)
+    inter_token_latency: HistogramResult = Field(default_factory=HistogramResult)
 
-    # Request
-    total_requests: int = 0
-    avg_prompt_tokens: int = 0
-    avg_output_tokens: int = 0
-    avg_queue_length: int = 0
-    num_preemptions_total: int = 0
-    num_requests_swapped: int = 0
+    # Gauges
+    queue_length: GaugeResult = Field(default_factory=GaugeResult)
+    num_requests_running: GaugeResult = Field(default_factory=GaugeResult)
+    kv_cache_usage: GaugeResult = Field(default_factory=GaugeResult)
 
-    # Usage
-    avg_kv_cache_usage: float = 0.0
-    median_kv_cache_usage: float = 0.0
-    p90_kv_cache_usage: float = 0.0
-    p99_kv_cache_usage: float = 0.0
+    # Single-value customs
+    num_requests_swapped: CounterResult = Field(default_factory=CounterResult)
[…]

[…]
