### 2026-04-01 35f831a
cleanup prometheus metric querying
