+++ inference_perf/client/modelserver/base.py
@@ -13,140 +13,121 @@
 # limitations under the License.
 from abc import ABC, abstractmethod
 from typing import List, Optional, Tuple
-from inference_perf.client.server_metrics.base import MetricsMetadata
 from inference_perf.config import APIConfig, APIType
 from inference_perf.apis import InferenceAPIData
 
 
-class ModelServerPrometheusMetric:
-    def __init__(self, name: str, op: str, type: str, filters: List[str]) -> None:
-        self.name = name
+class Metric(ABC):
+    metric_name: str
+
+    @abstractmethod
+    def get_queries(self, duration: float) -> List[Tuple[str, str]]:
+        """Returns a list of tuples containing (target_attr, query)."""
+        pass
+
+
+class GaugeMetric(Metric):
+    def __init__(self, base_name: str, metric_name: str, filters: List[str]) -> None:
+        self.base_name = base_name
+        self.metric_name = metric_name
+        self.filters = ",".join(filters)
+
+    def get_queries(self, duration: float) -> List[Tuple[str, str]]:
+        f = self.filters
+        m = self.metric_name
+        return [
+            (f"avg_{self.base_name}", f"avg_over_time({m}{{{f}}}[{duration:.0f}s])"),
+            (f"median_{self.base_name}", f"quantile_over_time(0.5, {m}{{{f}}}[{duration:.0f}s])"),
+            (f"p90_{self.base_name}", f"quantile_over_time(0.9, {m}{{{f}}}[{duration:.0f}s])"),
+            (f"p99_{self.base_name}", f"quantile_over_time(0.99, {m}{{{f}}}[{duration:.0f}s])"),
+        ]
+
+
+class CounterMetric(Metric):
+    def __init__(self, base_name: str, metric_name: str, filters: List[str]) -> None:
+        self.base_name = base_name
+        self.metric_name = metric_name
+        self.filters = ",".join(filters)
+
+    def get_queries(self, duration: float) -> List[Tuple[str, str]]:
+        f = self.filters
+        m = self.metric_name
+        return [
+            (f"avg_{self.base_name}", f"sum(increase({m}{{{f}}}[{duration:.0f}s]))"),
+            (f"{self.base_name}_per_second", f"sum(rate({m}{{{f}}}[{duration:.0f}s]))"),
+        ]
+
+
+class RequestsMetric(Metric):
+    def __init__(self, metric_name: str, filters: List[str]) -> None:
+        self.metric_name = metric_name
+        self.filters = ",".join(filters)
+
+    def get_queries(self, duration: float) -> List[Tuple[str, str]]:
+        m = self.metric_name
+        if m.startswith("{") and m.endswith("}"):
+            selector = f"{{{m[1:-1]},{self.filters}}}" if self.filters else m
+        else:
+            selector = f"{m}{{{self.filters}}}"
+        return [
+            ("total_requests", f"sum(increase({selector}[{duration:.0f}s]))"),
+            ("requests_per_second", f"sum(rate({selector}[{duration:.0f}s]))"),
+        ]
+
+
+class HistogramMetric(Metric):
+    def __init__(self, base_name: str, metric_name: str, filters: List[str]) -> None:
+        self.base_name = base_name
+        self.metric_name = metric_name
+        self.filters = ",".join(filters)
+
+    def get_queries(self, duration: float) -> List[Tuple[str, str]]:
+        f = self.filters
+        m = self.metric_name
+        return [
+            (
+                f"avg_{self.base_name}",
[…]

[…]
