### 2026-04-01 ff22022
cleanup prometheus metric querying
