### 2026-04-01 29a69d3
Add --url flag and config autofilling logic
