+++ e2e/utils/benchmark.py
@@ -88,6 +88,7 @@ async def run_benchmark_minimal(
     executable: Union[str, List[str]] = "inference-perf",
     timeout_sec: Optional[int] = 300,
     extra_env: Optional[Dict[str, str]] = None,
+    url: Optional[str] = None,
 ) -> BenchmarkResult:
     """
     Minimal wrapper:
@@ -110,6 +111,8 @@ async def run_benchmark_minimal(
     else:
         args = list(executable)
     args.extend(["--config_file", str(cfg_path), "--log-level", "DEBUG"])
+    if url:
+        args.extend(["--url", url])
     logger.debug(f"starting inference-perf, {args=}")
 
     proc = await asyncio.create_subprocess_exec(
