+++ inference_perf/detector/test_detector.py
@@ -1,0 +1,260 @@
+# Copyright 2025 The Kubernetes Authors.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from unittest.mock import MagicMock, patch
+
+from inference_perf.config import ModelServerType
+from inference_perf.detector.detector import EnvironmentType, detect_server_type
+
+
+def test_detect_vllm_via_metrics() -> None:
+    with patch("requests.get") as mock_get:
+        mock_response = MagicMock()
+        mock_response.status_code = 200
+        mock_response.text = "vllm:num_requests_waiting 0"
+        mock_get.return_value = mock_response
+
+        assert detect_server_type("http://dummy") == ModelServerType.VLLM
+
+
+def test_detect_sglang_via_metrics() -> None:
+    with patch("requests.get") as mock_get:
+        mock_response = MagicMock()
+        mock_response.status_code = 200
+        mock_response.text = "sglang:num_queue_reqs 0"
+        mock_get.return_value = mock_response
+
+        assert detect_server_type("http://dummy") == ModelServerType.SGLANG
+
+
+def test_detect_tgi_via_metrics() -> None:
+    with patch("requests.get") as mock_get:
+        mock_response = MagicMock()
+        mock_response.status_code = 200
+        mock_response.text = "tgi_queue_size 0"
+        mock_get.return_value = mock_response
+
+        assert detect_server_type("http://dummy") == ModelServerType.TGI
+
+
+def test_detect_vllm_via_v1_models() -> None:
+    with patch("requests.get") as mock_get:
+        # /metrics returns empty text
+        mock_metrics = MagicMock()
+        mock_metrics.status_code = 200
+        mock_metrics.text = ""
+
+        # /v1/models returns 200
+        mock_models = MagicMock()
+        mock_models.status_code = 200
+
+        mock_get.side_effect = [mock_metrics, mock_models]
+
+        assert detect_server_type("http://dummy") == ModelServerType.VLLM
+
+
+def test_detect_tgi_via_generate() -> None:
+    with patch("requests.get") as mock_get:
+        # /metrics fails
+        mock_metrics = MagicMock()
+        mock_metrics.status_code = 404
+
+        # /v1/models fails
+        mock_models = MagicMock()
+        mock_models.status_code = 404
+
+        # /generate returns 405 Method Not Allowed
+        mock_generate = MagicMock()
+        mock_generate.status_code = 405
+
+        mock_get.side_effect = [mock_metrics, mock_models, mock_generate]
+
+        assert detect_server_type("http://dummy") == ModelServerType.TGI
+
+
+def test_detect_fallback_to_vllm() -> None:
+    with patch("requests.get") as mock_get:
+        # All probes raise exception
+        mock_get.side_effect = Exception("Connection error")
+
+        assert detect_server_type("http://dummy") == ModelServerType.VLLM
+
+
+def test_detect_environment_gke() -> None:
+    with patch("requests.get") as mock_get, patch.dict("os.environ", {"KUBERNETES_SERVICE_HOST": "10.0.0.1"}):
+        mock_response = MagicMock()
+        mock_response.status_code = 200
+        mock_get.return_value = mock_response
+
+        from inference_perf.detector.detector import detect_environment
+
[…]

[…]
