### 2026-04-01 e7b55e6
Add --url flag and config autofilling logic
