+++ inference_perf/detector/detector.py
@@ -1,0 +1,220 @@
+# Copyright 2025 The Kubernetes Authors.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+import logging
+import os
+from typing import Any, Optional
+
+import requests
+from inference_perf.config import (
+    APIType,
+    Config,
+    MetricsClientConfig,
+    MetricsClientType,
+    ModelServerType,
+    PrometheusClientConfig,
+    deep_merge,
+)
+
+from enum import Enum, auto
+
+
+class EnvironmentType(Enum):
+    GKE = auto()
+    GCP = auto()
+    EKS = auto()
+    AWS = auto()
+    K8S = auto()
+    LOCAL = auto()
+
+
+logger = logging.getLogger(__name__)
+
+
+def autofill_config(url: str, base_config: Optional[Config] = None) -> Config:
+    """Generate or patch a Config object for a given URL by auto-detecting the server type and environment."""
+    detected_type = detect_server_type(url)
+    detected_api_type = detect_api_type(url)
+    detected_model_name = detect_model_name(url)
+    env = detect_environment()
+    metrics_cfg = autofill_metrics_config(env)
+
+    if base_config is None:
+        base_config = Config()
+
+    base_dict = base_config.model_dump(mode="json")
+
+    overrides: dict[str, Any] = {
+        "server": {
+            "type": detected_type.value,
+            "base_url": url,
+        },
+        "api": {
+            "type": detected_api_type.value,
+        },
+    }
+    if detected_model_name:
+        overrides["server"]["model_name"] = detected_model_name
+
+    if metrics_cfg:
+        overrides["metrics"] = metrics_cfg.model_dump(mode="json")
+
+    # Apply default load stages if none are defined
+    if "load" not in base_dict or not base_dict["load"].get("stages"):
+        overrides["load"] = {"stages": [{"rate": 1.0, "duration": 10}]}
+
+    merged = deep_merge(base_dict, overrides)
+    return Config(**merged)
+
+
+def detect_server_type(url: str) -> ModelServerType:
+    """Detect the type of model server by probing endpoints.
+
+    Heuristics:
+    1. Check /metrics for server-specific prefixes (vllm:, sglang:, tgi_).
+    2. Check /v1/models to see if it's OpenAI compatible (defaults to vLLM).
+    3. Check /generate (TGI specific).
+    """
+    # 1. Try /metrics
+    try:
+        response = requests.get(f"{url}/metrics", timeout=5)
+        if response.status_code == 200:
+            content = response.text
+            if "vllm:" in content:
+                logger.info("Auto-detected vLLM server via /metrics")
+                return ModelServerType.VLLM
+            if "sglang:" in content:
+                logger.info("Auto-detected SGLang server via /metrics")
+                return ModelServerType.SGLANG
+            if "tgi_" in content:
+                logger.info("Auto-detected TGI server via /metrics")
+                return ModelServerType.TGI
+    except Exception as e:
+        logger.debug("/metrics probe failed: %s", e)
+
+    # 2. Try /v1/models (OpenAI compatibility)
+    try:
+        response = requests.get(f"{url}/v1/models", timeout=5)
+        if response.status_code == 200:
+            logger.info("Auto-detected OpenAI compatible server via /v1/models (defaulting to vLLM)")
[…]
