### 2026-04-02 558d215
Harden loadgen saturation detection
Addresses: #379

Minor cleanup, makes sure all python files have a licence at the top.
