+++ inference_perf/loadgen/load_generator.py
@@ -56,6 +56,7 @@ else:
 from typing import List, Tuple, Optional, NamedTuple, Union, Set, Dict
 from types import FrameType
 import time
+import math
 import multiprocessing as mp
 from queue import Empty
 from multiprocessing.synchronize import Barrier as SyncBarrier, Event as SyncEvent
@@ -180,6 +181,7 @@ class Worker(mp.Process):
                     lora_adapter: Optional[str],
                 ) -> None:
                     inflight = False
+                    cancelled = False
                     try:
                         current_time = time.perf_counter()
                         sleep_time = request_time - current_time
@@ -203,7 +205,7 @@ class Worker(mp.Process):
 
                         await self.client.process_request(request_data, stage_id, request_time, lora_adapter)
                     except CancelledError:
-                        pass
+                        cancelled = True
                     except Exception as e:
                         logger.error(f"[DEBUG] Exception in task: {type(e).__name__}: {e}", exc_info=True)
                         raise
@@ -211,8 +213,9 @@ class Worker(mp.Process):
                         with self.active_requests_counter.get_lock():
                             if inflight:
                                 self.active_requests_counter.value -= 1
-                        with self.finished_requests_counter.get_lock():
-                            self.finished_requests_counter.value += 1
+                        if not cancelled:
+                            with self.finished_requests_counter.get_lock():
+                                self.finished_requests_counter.value += 1
                         queue.task_done()
                         semaphore.release()
 
@@ -703,7 +706,7 @@ class LoadGenerator:
         self,
         stage_id: int,
         rate: float,
-        duration: int,
+        duration: float,
         request_queue: RequestQueue[RequestQueueData],
         active_requests_counter: "Synchronized[int]",
         finished_requests_counter: "Synchronized[int]",
@@ -728,10 +731,10 @@ class LoadGenerator:
         start_time_epoch = time.time()
         start_time = time.perf_counter() + 1
 
-        if isinstance(self.datagen, DataGenerator) and self.datagen.trace is not None:
+        if isinstance(self.datagen, DataGenerator) and self.datagen.trace is not None and stage_id >= 0:
             num_requests = self.datagen.get_request_count()
         else:
-            num_requests = int(rate * duration)
+            num_requests = math.ceil(rate * duration)
 
         stage_status = StageStatus.RUNNING
 
@@ -824,43 +827,197 @@ class LoadGenerator:
         cancel_signal: SyncEvent,
     ) -> None:
         """
-        Runs a preliminary load test to automatically determine the server's saturation point
-        and generate a suitable series of load stages for the main benchmark.
+        Runs a Stepped Load (Ramping) test to determine the server's saturation point.
 
-        An aggregator task samples the active requests and then the burn down rate is
-        calculated from the samples. Saturation is derived from a percentile of the
-        sampled burn down rates.
+        Iterates through increasing request rates (steps). For each step:
+        1. Runs for a specific duration.
+        2. Measures the actual throughput (finished requests / time).
+        3. Checks if the system is saturated (Actual Throughput < Target Rate * tolerance).
+
+        If saturation is detected, the previous successful rate is used as the saturation point.
         """
-        logger.info("Running preprocessing stage")
-        results: List[Tuple[float, int]] = []
+        logger.info("Running preprocessing stage (Stepped Load)")
 
         if self.sweep_config is None:
             raise Exception("sweep_config cannot be none")
 
-        # Aggregator collects timestamped value of active_requests throughout the preprocessing
-        async def aggregator() -> None:
+        # Adaptive Saturation Detection (Exponential Search + Binary Refinement)
[…]
