+++ tests/required/loadgen/test_preprocess_saturation.py
@@ -1,0 +1,143 @@
+import unittest
+from unittest.mock import MagicMock, AsyncMock, patch
+from typing import Any
+import asyncio
+import sys
+from inference_perf.loadgen.load_generator import LoadGenerator
+from inference_perf.config import LoadConfig, LoadType, SweepConfig, StageGenType, StandardLoadStage, ConcurrentLoadStage
+from inference_perf.datagen.random_datagen import RandomDataGenerator
+import multiprocessing as mp
+
+# Patch asyncio.TaskGroup/TypeAlias if needed (copied from existing tests)
+if sys.version_info < (3, 11):
+
+    class MockTaskGroup:
+        async def __aenter__(self):
+            return self
+
+        async def __aexit__(self, *args):
+            pass
+
+        def create_task(self, coro):
+            return asyncio.create_task(coro)
+
+    asyncio.TaskGroup = MockTaskGroup
+
+
+class TestPreprocessSaturation(unittest.IsolatedAsyncioTestCase):
+    async def test_preprocess_generates_1_75x_saturation(self) -> None:
+        # Setup
+        mock_datagen = MagicMock(spec=RandomDataGenerator)
+        mock_datagen.get_request_count.return_value = 100
+
+        # Target 200 QPS max, 5 steps -> 40, 80, 120, 160, 200
+        sweep_config = SweepConfig(type=StageGenType.LINEAR, num_stages=5, num_requests=200, stage_duration=10, timeout=15)
+
+        load_config = LoadConfig(type=LoadType.CONSTANT, sweep=sweep_config, num_workers=1)
+
+        with patch("inference_perf.loadgen.load_generator.get_circuit_breaker"):
+            generator = LoadGenerator(mock_datagen, load_config)
+
+        # Mock counters
+        active_counter = mp.Value("i", 0)
+        finished_counter = mp.Value("i", 0)
+        request_phase = mp.Event()
+        cancel_signal = mp.Event()
+        client = AsyncMock()
+        queue = MagicMock()
+
+        # Mock time control
+        self.current_time = 1000.0
+
+        def mock_time() -> float:
+            return self.current_time
+
+        async def mock_sleep(delay: float) -> None:
+            # Just yield control, do NOT advance time here
+            # Time is driven by mock_run_stage
+            await asyncio.sleep(0)
+
+        # Mock run_stage to simulate execution
+        async def mock_run_stage(
+            stage_id: int, rate: float, duration: float, *args: list[Any], **kwargs: dict[str, Any]
+        ) -> None:
+            # Simulate stage running
+            start_t = self.current_time
+
+            # Simulate throughput: Cap at 100 QPS
+            # If rate <= 100, we get rate * duration requests
+            # If rate > 100, we get 100 * duration requests
+            effective_rate = min(rate, 100.0)
+            target_requests = int(effective_rate * duration)
+
+            # Increment time in small steps to allow aggregator to sample
+            step = 0.5
+            current = 0.0
+            while current < duration:
+                # Update counter
+                elapsed = current
+                # Interpolate finished requests
+                with finished_counter.get_lock():
+                    finished_counter.value = int(effective_rate * elapsed)
+
+                # Advance time explicitly
+                self.current_time += step
+                await asyncio.sleep(0)  # Yield to aggregator
+                current += step
+
+            # Final update
+            with finished_counter.get_lock():
+                finished_counter.value = target_requests
+
+            # Ensure time matches end
+            self.current_time = start_t + duration
+
+        generator.run_stage = AsyncMock(side_effect=mock_run_stage)  # type: ignore[method-assign]
+
+        with (
[…]
