+++ examples/vllm/config-sweep.yml
@@ -6,7 +6,6 @@ load:
     timeout: 60                 # Length of time to run load to determine saturation
     num_stages: 10              # Number of stages to generate
     stage_duration: 60          # Duration of each generated stage
-    saturation_percentile: 95   # Percentile of sampled rates to select as saturation point
 api:
   type: completion
 server:

[…]
