### 2026-04-02 c337e5c
Harden loadgen saturation detection
Addresses: #379

Minor cleanup, makes sure all python files have a licence at the top.
