### 2026-04-02 df003db
Add distribution sampling for shared_prefix
Add support for multiple statistical distribution types
(normal, skew_normal, lognormal, uniform, poisson) for
shared_prefix benchmark parameters, enabling realistic
production traffic modeling for capacity planning.

- Add DistributionType enum and DistributionConfig model
- Allow question_len, output_len, system_prompt_len to be
  either a fixed int or an inline distribution config
- Add num_tool_calls_per_turn for agentic workload modeling
- Add seed parameter for deterministic benchmark runs
- Implement sample_from_distribution() with all 5 types
[…]
