+++ tests/utils/test_distribution.py
@@ -1,0 +1,143 @@
+# Copyright 2026 The Kubernetes Authors.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+import numpy as np
+import pytest
+
+from inference_perf.config import DistributionConfig, DistributionType
+from inference_perf.utils.distribution import generate_distribution, sample_from_distribution
+
+
+class TestSampleFromDistribution:
+    def test_normal_within_bounds(self) -> None:
+        config = DistributionConfig(distribution=DistributionType.NORMAL, mean=500.0, min=100, max=900, std_dev=100.0)
+        result = sample_from_distribution(config, 1000, rng=np.random.default_rng(42))
+        assert result.min() >= 100
+        assert result.max() <= 900
+        assert len(result) == 1000
+
+    def test_normal_approximate_mean(self) -> None:
+        config = DistributionConfig(distribution=DistributionType.NORMAL, mean=200.0, min=10, max=2000, std_dev=50.0)
+        result = sample_from_distribution(config, 10000, rng=np.random.default_rng(42))
+        assert abs(result.mean() - 200.0) < 5.0
+
+    def test_skew_normal_within_bounds(self) -> None:
+        config = DistributionConfig(
+            distribution=DistributionType.SKEW_NORMAL, mean=200.0, min=10, max=2000, std_dev=80.0, skew=2.5
+        )
+        result = sample_from_distribution(config, 1000, rng=np.random.default_rng(42))
+        assert result.min() >= 10
+        assert result.max() <= 2000
+
+    def test_skew_normal_positive_skew_has_right_tail(self) -> None:
+        config = DistributionConfig(
+            distribution=DistributionType.SKEW_NORMAL, mean=100.0, min=0, max=1000, std_dev=50.0, skew=5.0
+        )
+        result = sample_from_distribution(config, 10000, rng=np.random.default_rng(42))
+        # With positive skew, the median should be less than the mean of the samples
+        median = float(np.median(result))
+        mean = float(result.mean())
+        assert median <= mean
+
+    def test_skew_normal_zero_skew_approximates_normal(self) -> None:
+        config_skew = DistributionConfig(
+            distribution=DistributionType.SKEW_NORMAL, mean=500.0, min=0, max=1000, std_dev=100.0, skew=0.0
+        )
+        config_normal = DistributionConfig(distribution=DistributionType.NORMAL, mean=500.0, min=0, max=1000, std_dev=100.0)
+        rng1 = np.random.default_rng(42)
+        rng2 = np.random.default_rng(42)
+        result_skew = sample_from_distribution(config_skew, 10000, rng=rng1)
+        result_normal = sample_from_distribution(config_normal, 10000, rng=rng2)
+        # Means should be close (both target 500)
+        assert abs(float(result_skew.mean()) - float(result_normal.mean())) < 20.0
+
+    def test_lognormal_within_bounds(self) -> None:
+        config = DistributionConfig(distribution=DistributionType.LOGNORMAL, mean=150.0, min=1, max=4096, std_dev=60.0)
+        result = sample_from_distribution(config, 1000, rng=np.random.default_rng(42))
+        assert result.min() >= 1
+        assert result.max() <= 4096
+
[…]
