+++ examples/vllm/config-shared-prefix-tool-calls.yml
@@ -1,0 +1,41 @@
+load:
+  type: constant
+  num_workers: 2
+  worker_max_concurrency: 10
+  stages:
+  - rate: 20
+    duration: 5
+api:
+  type: completion
+server:
+  type: vllm
+  model_name: HuggingFaceTB/SmolLM2-135M-Instruct
+  base_url: http://0.0.0.0:8000
+  ignore_eos: true
+tokenizer:
+  pretrained_model_name_or_path: HuggingFaceTB/SmolLM2-135M-Instruct
+data:
+  type: shared_prefix
+  shared_prefix:
+    seed: 42
+    num_unique_system_prompts: 2
+    num_users_per_system_prompt: 10
+    system_prompt_len: 100
+    question_len: 50
+    output_len: 50
+    num_tool_calls_per_turn:              # Agentic workload: skewed tool call distribution
+      distribution: poisson
+      mean: 4.2
+      min: 0
+      max: 50
+    enable_multi_turn_chat: true
+metrics:
+  type: prometheus
+  prometheus:
+    url: http://localhost:9090
+    scrape_interval: 15
+report:
+  request_lifecycle:
+    summary: true
+    per_stage: true
+    per_request: true

[…]
