### 2026-04-03 #10314 @neerazz
I went through the paper, the codebase, and the actual Python source in detail — cloned the repo and verified claims against the v0.3.0 tag. My background: I work on Kubernetes infrastructure and AI/ML serving pipelines at scale. This tool sits right in my day-to-day.

The short version: inference-perf fills a real gap and the multi-process load generation architecture is solid engineering. But I found places where the paper claims things the code doesn't do yet, and a measurement correctness issue in the streaming response handling. I've filed issues on the repo for each.

## What I can check off now

The tool installs cleanly, CI is healthy (4 workflows), documentation covers the main use cases. Development history is solid — first commit Jan 2025, 14 months of sustained work, 37 contributors from Google, Capital One, IBM Research, and Red Hat. The trace replay feature with Azure public dataset support is well implemented. And the llm-d-inference-sim approach for GPU-free E2E testing is a nice engineering choice.

I've checked off the development history, general checks, documentation, and installation items I can verify.

## Issues that need addressing

### Paper claims features the code doesn't have

The Software Design section (paper.md line 69) says the Client supports "protocols (HTTP, gRPC, streaming)." I searched the v0.3.0 codebase:

- No `.proto` files anywhere in the repository
- No `grpcio` or `grpcio-tools` in `pyproject.toml`
- The internal design doc (`docs/design.md` line 29) uses future tense: "will also support different protocols like http and grpc" — indicating this is planned, not implemented

HTTP with OpenAI-compatible APIs is what's actually implemented. The design doc acknowledges this is aspirational, but the paper presents it as current.

Similarly, the paper lists three price-performance metrics (paper.md lines 103-107): price per million output tokens, price per million input tokens, throughput per dollar. The docs (`docs/metrics.md` lines 27-35) define formulas. But searching the Python source under `inference_perf/` for "price", "cost", "dollar", or "accelerator" returns zero results. These metrics are documented but not implemented in code.

→ Issues filed: (links to follow in next comment)

### Measurement issue in streaming response timestamps

In `inference_perf/apis/chat.py` lines 59-61, every incoming SSE event gets timestamped before the code checks whether the event contains content:

```python
message, buffer = buffer.split(b"\n\n", 1)
output_token_times.append(time.perf_counter())   # line 61 — timestamp HERE
for line in message.split(sep=b"\n"):             # line 62 — THEN check content
    if line.startswith(b"data:"):
        data_str = line.removeprefix(b"data: ").strip()
        if data_str == b"[DONE]":                 # line 65 — [DONE] already timestamped
```

The same pattern exists in `inference_perf/apis/completion.py` lines 55-57. The `[DONE]` event, role-only first chunks, and usage statistics all get timestamps. Then in `reportgen/base.py`:

- **TPOT** (line 165): `(output_token_times[-1] - output_token_times[0]) / (len(output_token_times) - 1)` — denominator includes non-content events
- **TTFT** (line 170): `output_token_times[0] - x.start_time` — could measure time to a role-only chunk, not first content
- **ITL** (line 175): consecutive timestamp diffs include intervals to/from non-content events

The test file `tests/apis/test_chat.py` only tests payload construction (lines 7-17) — it doesn't test `process_response()` or the SSE parsing at all.

### Load pattern reproducibility

The `ConstantLoadTimer` (`loadgen/load_timer.py` line 42-46) and `PoissonLoadTimer` (line 73-76) both create unseeded random number generators:

```python
# TODO: Make random state a global seed
self._rand = np.random.default_rng()
```

The TODO on line 45 explicitly acknowledges this gap. Non-trace load patterns (constant, Poisson) will produce different results across runs. Trace replay IS deterministic (it reads timestamps from a file), but for the other load types, the same configuration won't produce the same traffic pattern.

### State-of-the-field gaps

The paper (paper.md lines 53-55) splits existing tools into "web benchmarks" (k6, Locust) and "model server benchmarks" (vLLM, TGI, GenAI-Perf). This misses tools in the same niche as inference-perf: SGLang's `bench_serving` does async TTFT/TPOT/ITL with multi-backend support, MLPerf Inference from MLCommons is the industry standard, and LLMPerf from Anyscale targets the same problem.

GenAI-Perf is described as "tightly coupled to specific frameworks" (paper.md line 49) but GenAI-Perf supports the OpenAI API and can benchmark vLLM and TGI directly.

### wg-serving framing

paper.md line 45 says inference-perf is "part of the inference benchmarking and metrics standardization effort in the Kubernetes wg-serving" in present tense. The paper.bib entry (line 15) links to `https://github.com/kubernetes/community/tree/master/archive/wg-serving`. The `/archive/` path means wg-serving has been dissolved. Yuan Tang, a paper co-author, authored the git commit that moved wg-serving to the archive. The project originated from wg-serving work — that's worth acknowledging — but the current framing implies active institutional backing that no longer exists.

### Research Impact overstated

paper.md line 119 says: "inference-perf is the primary benchmarking tool used by state of the art open source distributed inference frameworks like llm-d." The paper.bib cites `llm-d-kv-cache/benchmarking/37-capacity/README.md` as evidence. The llm-d repository (2,853 stars) doesn't mention inference-perf. llm-d has its own tool (llm-d-inference-sim). The cited README file appears empty. inference-perf can benchmark llm-d deployments like any OpenAI-compatible endpoint, but that's different from being llm-d's "primary benchmarking tool."

### Software Design section

Lists six components with descriptions but doesn't discuss trade-offs. Worth noting: `docs/design.md` is more candid — line 29 says gRPC "will also" be supported (future), while the paper presents it as current. JOSS asks for "meaningful design thinking": why Python for 10K+ QPS? How does multi-process address GIL? What are the bottlenecks?

### Prometheus integration direction

paper.md line 70 says the tool "supports exporting metrics to Prometheus." The actual code (`client/metricsclient/prometheus_client/base.py`) is a Prometheus query client — it READS FROM an existing Prometheus server to collect model server metrics. The direction in the paper is backwards. inference-perf queries Prometheus; it doesn't export to it.

### Minor items

- paper.md line 61: typo "infernce-perf"
- CONTRIBUTING.md is the generic Kubernetes template with no project-specific content
- `summarize_requests()` in `reportgen/base.py` (the function calculating ALL reported metrics) has zero test coverage

## Suggestions (not blocking)

- A comparison table in State of the field
- One reproducible benchmark experiment with methodology
- A mention of coordinated omission (Gil Tene) and how `schedule_delay` tracking relates

## Bottom line

**Major revision.** The tool has real merit — trace replay works well, the multi-process load generator is sound, the modular architecture is clean. But the paper has accuracy issues: claimed features that aren't implemented (gRPC, price metrics), overstated impact evidence, and an incomplete competitive landscape. Most fixes are paper edits plus the SSE timestamp code fix. Happy to re-review once addressed.

*AI assistance disclosure: I used AI for grammar checking and contextualizing public materials. All evaluative decisions — the SSE timestamp bug identification, the wg-serving framing concern, the state-of-field gap analysis, the Prometheus direction reversal — are my own judgments from reading the code and paper.*
