### 2026-04-03 kubernetes-sigs/inference-perf#390 @neerazz
Paper claims gRPC support that doesn't exist in the codebase
The JOSS paper (openjournals/joss-reviews#10314) states in `paper/paper.md` line 69:

> Abstractions for different model servers and protocols (HTTP, gRPC, streaming)

Verified against the v0.3.0 tag:

**What I searched:**
- `rg "grpc" inference_perf/` → zero results in Python source
- `rg "grpc" pyproject.toml` → no grpcio dependency
- `find . -name "*.proto"` → no proto files

**What I found:**
- `docs/design.md` line 29 uses future tense: "will also support different protocols like http and grpc" — indicating this is planned, not implemented
- All model server clients (`client/modelserver/vllm_client.py`, `sglang_client.py`, `tgi_client.py`) use HTTP via aiohttp

The tool currently supports HTTP with OpenAI-compatible APIs only. The internal design doc correctly frames gRPC as planned, but the paper presents it as a current capability.

**Source references:**
- Paper claim: `paper/paper.md:69`
- Design doc (future tense): `docs/design.md:29`
- Dependencies: `pyproject.toml` (no grpcio)

Filed as part of JOSS review: openjournals/joss-reviews#10314
