### 2026-04-03 kubernetes-sigs/inference-perf#391 @neerazz
Price-performance metrics defined in docs but not implemented in code
The JOSS paper (`paper/paper.md` lines 103-107) lists three price-performance metrics:
- Price per million output tokens
- Price per million input tokens
- Throughput per dollar

`docs/metrics.md` (lines 27-35) defines formulas using accelerator $ / second.

**What I searched:**
- `rg -i "price|cost|dollar|accelerator" inference_perf/` → zero results

**What I found:**
- `docs/metrics.md` has the Price/Performance section with formulas
- `inference_perf/reportgen/base.py` `summarize_requests()` function (lines 122-193) calculates throughput and latency metrics but has no price-related code
- No config field for accelerator cost anywhere in `inference_perf/config.py`

These metrics are documented but have no implementation.

**Source references:**
- Paper claim: `paper/paper.md:103-107`
- Doc formulas: `docs/metrics.md:27-35`
- Implementation (absent): `inference_perf/reportgen/base.py:122-193`

Filed as part of JOSS review: openjournals/joss-reviews#10314
