### 2026-04-03 kubernetes-sigs/inference-perf#392 @neerazz
SSE event timestamping includes non-content events, affecting TPOT/ITL/TTFT accuracy
In `inference_perf/apis/chat.py` lines 59-61, every SSE event is timestamped before the code checks whether it contains content:

```python
message, buffer = buffer.split(b"\n\n", 1)
output_token_times.append(time.perf_counter())  # timestamp ALL events
for line in message.split(sep=b"\n"):            # THEN check content
```

The same pattern in `inference_perf/apis/completion.py` lines 55-57.

This means `[DONE]` signals (line 65), role-only first chunks, and usage statistics all get timestamps in `output_token_times`. The downstream metrics use these timestamps:

- **TPOT** (`reportgen/base.py:165`): `(output_token_times[-1] - output_token_times[0]) / (len(output_token_times) - 1)` — denominator inflated by non-content events
- **TTFT** (`reportgen/base.py:170`): `output_token_times[0] - x.start_time` — may measure time to a role-only chunk, not first generated content
- **ITL** (`reportgen/base.py:175`): consecutive diffs include intervals between content and non-content events

The test file `tests/apis/test_chat.py` only tests `to_payload()` (lines 7-17), not `process_response()` or the SSE parsing.

**Source references:**
- Bug location: `inference_perf/apis/chat.py:59-61`
- Same pattern: `inference_perf/apis/completion.py:55-57`
- Impact on TPOT: `inference_perf/reportgen/base.py:165`
- Impact on TTFT: `inference_perf/reportgen/base.py:170`
- Impact on ITL: `inference_perf/reportgen/base.py:175`
- Test gap: `tests/apis/test_chat.py:7-17`

Filed as part of JOSS review: openjournals/joss-reviews#10314
