### 2026-04-03 kubernetes-sigs/inference-perf#393 @neerazz
Load timer RNGs are not seeded — non-trace load patterns are not reproducible
`inference_perf/loadgen/load_timer.py` creates unseeded random number generators for both load types:

```python
# ConstantLoadTimer, line 45-46
# TODO: Make random state a global seed
self._rand = np.random.default_rng()

# PoissonLoadTimer, line 76
self._rand = np.random.default_rng()
```

The TODO on line 45 explicitly acknowledges this gap. Running the same configuration twice with constant or poisson load types will produce different traffic patterns because `default_rng()` without a seed uses system entropy.

Trace replay (`TraceReplayLoadTimer`, line 100) IS deterministic since it reads timestamps from a file. But for constant and Poisson load patterns, there's no way to reproduce the same run.

Either propagate a configurable seed to all RNGs, or document that exact reproducibility is only available with trace replay.

**Source references:**
- `ConstantLoadTimer` unseeded: `inference_perf/loadgen/load_timer.py:45-46`
- `PoissonLoadTimer` unseeded: `inference_perf/loadgen/load_timer.py:76`
- `TraceReplayLoadTimer` (deterministic): `inference_perf/loadgen/load_timer.py:100-108`

Filed as part of JOSS review: openjournals/joss-reviews#10314
