### 2026-04-03 kubernetes-sigs/inference-perf#394 @neerazz
Paper references wg-serving as active but it has been archived
`paper/paper.md` line 45:

> As part of the inference benchmarking and metrics standardization effort in the Kubernetes wg-serving

Present tense, implying an active working group. The bibliography entry in `paper/paper.bib` (line 15) links to:

https://github.com/kubernetes/community/tree/master/archive/wg-serving

The `/archive/` directory path indicates wg-serving has been dissolved. The git commit that moved wg-serving to the archive was authored by terrytangyuan (Yuan Tang), who is listed as a co-author on this paper.

The project originated from wg-serving work — that's worth acknowledging — but the present-tense framing implies institutional backing from an active Kubernetes working group that no longer exists. Please either:
- Use past tense and clarify where the standardization effort now lives
- Point to the active successor (SIG Scalability sponsorship?)

**Source references:**
- Paper claim: `paper/paper.md:45`
- Bibliography URL: `paper/paper.bib:15` → archive/wg-serving
- Acknowledgements: `paper/paper.md:127`

Filed as part of JOSS review: openjournals/joss-reviews#10314
