### 2026-04-03 kubernetes-sigs/inference-perf#395 @neerazz
State of the field should acknowledge SGLang bench_serving, MLPerf Inference, and LLMPerf
`paper/paper.md` lines 53-55 divides existing tools into "web benchmarks" (k6, Locust) and "model server benchmarks" (vLLM, TGI, GenAI-Perf). This two-category taxonomy omits tools in the same niche as inference-perf:

1. **SGLang bench_serving**: Async HTTP benchmarking client with TTFT/TPOT/ITL metrics, multi-backend support (sglang, vllm, trt). Part of the SGLang project already cited in this paper.

2. **MLPerf Inference LoadGen (MLCommons)**: The industry standard for inference benchmarking. Has LLM-specific scenarios (Server, Offline). A paper about benchmarking "standardization" should at minimum reference it.

3. **LLMPerf (Anyscale/Ray)**: Dedicated LLM inference benchmarking tool.

Also, `paper.md` line 49 characterizes GenAI-Perf tools as "tightly coupled to specific frameworks." GenAI-Perf supports the OpenAI API and can benchmark vLLM and TGI directly. This characterization is not accurate.

**Source references:**
- Paper claim: `paper/paper.md:53-55`
- GenAI-Perf characterization: `paper/paper.md:49`
- `paper.bib` already cites SGLang (line 31) but bench_serving is not discussed

Filed as part of JOSS review: openjournals/joss-reviews#10314
