### 2026-04-03 kubernetes-sigs/inference-perf#396 @neerazz
Research Impact claim about llm-d usage lacks supporting evidence
`paper/paper.md` line 119:

> inference-perf is the primary benchmarking tool used by state of the art open source distributed inference frameworks like llm-d

`paper/paper.bib` (line 125) cites `llm-d-kv-cache/benchmarking/37-capacity/README.md` as evidence.

**What I checked:**
- The llm-d repository (2,853 stars) does not mention inference-perf
- llm-d has its own performance testing project (llm-d-inference-sim, 103 stars)
- The cited file (`llm-d-kv-cache/benchmarking/37-capacity/README.md`) returns empty content

inference-perf can benchmark llm-d deployments (as it can benchmark any OpenAI-compatible endpoint), but calling it the "primary benchmarking tool used by" llm-d overstates the relationship.

**Source references:**
- Paper claim: `paper/paper.md:119`
- Citation: `paper/paper.bib:119-127` (llm-d-kv-cache)
- llm-d repo: https://github.com/llm-d/llm-d (no mention of inference-perf)

Filed as part of JOSS review: openjournals/joss-reviews#10314
