### 2026-04-03 kubernetes-sigs/inference-perf#397 @neerazz
CONTRIBUTING.md needs project-specific contribution guidelines
`CONTRIBUTING.md` is the verbatim Kubernetes community template with no project-specific content. For JOSS compliance ("clear guidelines for third parties wishing to contribute"), it should include:

1. **Development environment setup** (the README mentions `nix develop` and `pdm` but `CONTRIBUTING.md` doesn't)
2. **How to run tests locally** (the `tests/` directory has 5 test files; `pdm run test` is how to run them)
3. **How to add a new backend or dataset type**
4. **Code style expectations** (project uses `ruff` + `mypy` strict, but this isn't documented for contributors)

**Source references:**
- Current `CONTRIBUTING.md`: verbatim Kubernetes template
- README dev setup: `README.md` Getting Started section
- Test structure: `tests/test_config.py`, `tests/test_trace_replay.py`, `tests/test_logger.py`, `tests/apis/test_chat.py`, `tests/apis/test_completion.py`

Filed as part of JOSS review: openjournals/joss-reviews#10314
