### 2026-04-03 kubernetes-sigs/inference-perf#398 @neerazz
Typo in paper: "it's" → "its" (possessive)
In `paper.md`, the phrase "it's architecture" uses the contraction "it's" (it is) where the possessive "its" is needed.

**Location:** `paper.md` — search for "it's architecture"

**Fix:** Replace `it's` with `its`.

Filed as part of JOSS review: openjournals/joss-reviews#10314
