### 2026-04-03 #10314 @neerazz
Filed the following issues as part of my review:

1. kubernetes-sigs/inference-perf#390 — gRPC support is absent and not documented as out-of-scope
2. kubernetes-sigs/inference-perf#391 — No price/cost-per-token metrics
3. kubernetes-sigs/inference-perf#392 — SSE stream parser silently drops malformed chunks
4. kubernetes-sigs/inference-perf#393 — No way to seed or replay a fixed request load
5. kubernetes-sigs/inference-perf#394 — No connection to wg-serving benchmarking standards
6. kubernetes-sigs/inference-perf#395 — Paper omits comparison to vLLM Benchmark / Paladin / llmperf
7. kubernetes-sigs/inference-perf#396 — llm-d is listed as a backend but has no implementation
8. kubernetes-sigs/inference-perf#397 — CONTRIBUTING.md needs project-specific contribution guidelines
9. kubernetes-sigs/inference-perf#398 — Typo in paper: "it's" → "its" (possessive)
