### 2026-04-03 5b657f5
Add automatic doc generation for cli flags
