+++ tests/reportgen/test_summarize_requests.py
@@ -1,0 +1,106 @@
+import pytest
+from inference_perf.reportgen.base import summarize_requests
+from inference_perf.apis.base import RequestLifecycleMetric, InferenceInfo, StreamedInferenceResponseInfo
+
+
+def test_summarize_requests_tpot_calculation() -> None:
+    info = InferenceInfo(
+        input_tokens=5, response_info=StreamedInferenceResponseInfo(output_tokens=10, output_token_times=[1.0, 2.0, 3.0])
+    )
+
+    metric = RequestLifecycleMetric(
+        scheduled_time=0.0, start_time=0.0, end_time=10.0, request_data="test_request", info=info, error=None
+    )
+
+    # Duration = 3.0 - 1.0 = 2.0
+    # Actual tokens = 10
+    # Expected TPOT = 2.0 / (10 - 1) = 2.0 / 9 = 0.222...
+
+    result = summarize_requests([metric], [50])
+
+    assert result is not None
+    successes = result.successes
+    assert "latency" in successes
+    assert "time_per_output_token" in successes["latency"]
+    tpot_summary = successes["latency"]["time_per_output_token"]
+    assert tpot_summary["mean"] == pytest.approx(2.0 / 9.0)
+
+
+def test_summarize_requests_tpot_fallback() -> None:
+    # Test fallback when output_tokens is not available or <= 1
+    info = InferenceInfo(
+        input_tokens=5, response_info=StreamedInferenceResponseInfo(output_tokens=0, output_token_times=[1.0, 2.0, 3.0])
+    )
+
+    metric = RequestLifecycleMetric(
+        scheduled_time=0.0, start_time=0.0, end_time=10.0, request_data="test_request", info=info, error=None
+    )
+
+    # Duration = 3.0 - 1.0 = 2.0
+    # Count = 3
+
+    result = summarize_requests([metric], [50])
+
+    assert result is not None
+    successes = result.successes
+    tpot_summary = successes["latency"]["time_per_output_token"]
+    assert tpot_summary is None
+
+
+def test_summarize_requests_with_chunks() -> None:
+    from unittest.mock import Mock
+
+    mock_tokenizer = Mock()
+    # Assume 1 chunk has 2 tokens, another has 3 tokens
+    mock_tokenizer.count_tokens.side_effect = lambda text: 2 if "hello" in text else (3 if "world" in text else 0)
+
+    info = InferenceInfo(
+        input_tokens=5,
+        response_info=StreamedInferenceResponseInfo(
+            response_chunks=['{"choices": [{"text": "hello"}]}', '{"choices": [{"text": "world"}]}'], chunk_times=[1.0, 2.0]
+        ),
+    )
+
+    metric = RequestLifecycleMetric(
+        scheduled_time=0.0, start_time=0.0, end_time=10.0, request_data="test_request", info=info, error=None
+    )
+
+    result = summarize_requests([metric], [50], tokenizer=mock_tokenizer)
+
+    assert result is not None
+    successes = result.successes
+
+    tpot_summary = successes["latency"]["time_per_output_token"]
+    assert tpot_summary["mean"] == pytest.approx(0.25)
+
+    ttft_summary = successes["latency"]["time_to_first_token"]
+    assert ttft_summary["mean"] == pytest.approx(1.0)
+
+
+def test_summarize_requests_token_mismatch() -> None:
+    from unittest.mock import Mock
+
+    mock_tokenizer = Mock()
+    mock_tokenizer.count_tokens.side_effect = lambda text: 2 if "hello" in text else (3 if "world" in text else 0)
+
+    info = InferenceInfo(
+        input_tokens=5,
+        response_info=StreamedInferenceResponseInfo(
+            response_chunks=[
[…]
