+++ inference_perf/reportgen/base.py
@@ -12,8 +12,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 import logging
+import json
 from collections import defaultdict
 from typing import Any, Dict, List, Optional, Union, TYPE_CHECKING
+from inference_perf.utils.custom_tokenizer import CustomTokenizer
 
 if TYPE_CHECKING:
     from inference_perf.datagen import DataGenerator, SessionGenerator
@@ -22,7 +24,7 @@ if TYPE_CHECKING:
 import numpy as np
 from pydantic import BaseModel
 
-from inference_perf.apis import RequestLifecycleMetric, SessionLifecycleMetric
+from inference_perf.apis import RequestLifecycleMetric, SessionLifecycleMetric, StreamedInferenceResponseInfo
 from inference_perf.client.metricsclient import MetricsClient, PerfRuntimeParameters
 from inference_perf.client.metricsclient.base import ModelServerMetrics
 from inference_perf.client.metricsclient.prometheus_client import PrometheusMetricsClient
@@ -99,6 +101,7 @@ def calculate_goodput_metrics(
     for i, m in enumerate(metrics):
         is_good = True
 
+<<<<<<< HEAD
         effective_ttft_slo = (
             m.ttft_slo_sec
             if m.ttft_slo_sec is not None
@@ -108,6 +111,49 @@ def calculate_goodput_metrics(
             m.tpot_slo_sec
             if m.tpot_slo_sec is not None
             else (goodput_config.constraints.get("tpot") if goodput_config else None)
+=======
+        # Goodput Calculation Logic
+        ttft_ok = (ttft_met is True) if has_ttft_slo else True
+        tpot_ok = (tpot_met is True) if has_tpot_slo else True
+
+        if ttft_ok and tpot_ok:
+            if m.info.input_tokens is not None and m.info.response_info and m.info.response_info.output_tokens is not None:
+                goodput_token_counts.append(m.info.input_tokens + m.info.response_info.output_tokens)
+
+    total_goodput_tokens = sum(goodput_token_counts)
+    goodput_rate = total_goodput_tokens / total_benchmark_time if total_benchmark_time > 0 else 0.0
+
+    # TTFT SLO metrics
+    if has_ttft_slo:
+        ttft_met_count = sum(1 for res in ttft_results if res is True)
+        ttft_failed_count = sum(1 for res in ttft_results if res is False)
+
+        slo_metrics["ttft_slo"] = {
+            "attainment_pct": (ttft_met_count / total * 100) if total > 0 else 0,
+            "requests_met": ttft_met_count,
+            "requests_failed": ttft_failed_count,
+            "total_requests": total,
+            "slo": ttft_slo_limit,
+        }
+
+    # TPOT SLO metrics
+    if has_tpot_slo:
+        tpot_met_count = sum(1 for res in tpot_results if res is True)
+        tpot_failed_count = sum(1 for res in tpot_results if res is False)
+
+        slo_metrics["tpot_slo"] = {
+            "attainment_pct": (tpot_met_count / total * 100) if total > 0 else 0,
+            "requests_met": tpot_met_count,
+            "requests_failed": tpot_failed_count,
+            "total_requests": total,
+            "slo": tpot_slo_limit,
+        }
+
+    # Combined SLO (both must be met)
+    if has_ttft_slo and has_tpot_slo:
+        combined_met = sum(
+            1 for t_res, p_res in zip(ttft_results, tpot_results, strict=True) if t_res is True and p_res is True
+>>>>>>> 4fbc041 (TPOT, ITL from repsonse tokens, not chunks)
         )
 
[…]

[…]
