### 2026-04-04 #10314 @neerazz-parafin
@logological Sure, happy to walk you through it.

I actually spent a few days on this before posting anything. Started around March 30 — cloned the repo at v0.3.0, read through the paper, then started poking around the Python source. My usual process is pretty low-tech: grep for whatever the paper claims, see if the code backs it up. That's how I caught the gRPC thing — paper says it's supported, searched for .proto files and grpcio in pyproject.toml, nothing there. Then I noticed `docs/design.md` uses future tense ("will also support") for the same feature the paper presents as current. The SSE timestamp issue came from tracing the streaming response path in chat.py line by line — I was trying to understand how they compute TPOT and realized the timestamps include non-content events.

The reason it all showed up at once: I wrote everything locally in markdown files first. The review, the issue bodies, all of it. I like to get the full picture before posting anything so I'm not going back and editing things after the fact. Once I was happy with the drafts, I posted the review comment manually, then used a GitHub MCP integration to create the nine issues from my pre-written drafts. That's why the timing looks compressed — the actual work happened over days, the posting was just the last step.

For the AI piece: I used Cursor (IDE with Claude) to help clean up my rough notes into more readable prose and fix grammar. The substantive analysis — finding the SSE bug, the gap between paper claims and code, the wg-serving archive, the missing competitors — that's all from reading the source. You can't really get that from an AI summary; I had to follow the actual control flow in chat.py and cross-reference specific line numbers in design.md versus paper.md to spot the inconsistencies.

Happy to clarify anything else.
