### 2026-04-05 8d088db
refactor(loadgen): replace tqdm with progress_ctx in run_session_stage
Migrate run_session_stage to use Rich Progress context instead of tqdm,
matching the pattern in run_stage. Track progress by completed sessions
and update log verbosity (INFO -> DEBUG for some messages).
