+++ inference_perf/loadgen/load_generator.py
@@ -352,6 +352,7 @@ class LoadGenerator:
         finished_requests_counter: "Synchronized[int]",
         request_phase: SyncEvent,
         cancel_signal: Optional[SyncEvent] = None,
+        progress_ctx: Optional[Progress] = None,
     ) -> None:
         """Run a session-based trace replay stage.
 
@@ -476,7 +477,7 @@ class LoadGenerator:
             session_info = self.datagen.get_session_info(session_idx)
             session_id = session_info["session_id"]
 
-            logger.info(
+            logger.debug(
                 f"Starting session {session_idx}: {session_id} "
                 f"({len(active_session_indices)} active, {len(pending_session_indices)} pending)"
             )
@@ -524,111 +525,122 @@ class LoadGenerator:
                 session_interval = 1.0 / session_rate
                 next_dispatch_time = max(next_dispatch_time + session_interval, now)
 
-            logger.info(f"Dispatched {dispatched_count} events for session {session_idx}")
+            logger.debug(f"Dispatched {dispatched_count} events for session {session_idx}")
             return dispatched_count
 
         # Main dispatch and wait loop
-        with tqdm(total=1.0, desc=f"Stage {stage_id} progress ({effective_num_sessions} sessions)") as pbar:
-            while True:
-                # Check for interrupts
-                if self.interrupt_sig:
-                    pbar.close()
-                    logger.info("Loadgen encountered SIGINT")
-                    stage_status = StageStatus.FAILED
-                    # Clean up any active session spans (using cached otel_instr)
-                    for sid in list(session_spans.keys()):
-                        otel_instr.end_session_span(session_spans[sid], "Session interrupted by SIGINT")
-                        del session_spans[sid]
-                    break
+        stage_task = None
+        if progress_ctx:
+            stage_task = progress_ctx.add_task(description=f"Stage {stage_id} Sessions", total=effective_num_sessions)
 
-                if cb := next((cb for cb in self.circuit_breakers if cb.is_open()), None):
-                    logger.warning(f'Loadgen detects circuit breakers "{cb.name}" open, exit the stage.')
-                    stage_status = StageStatus.FAILED
-                    # Clean up any active session spans (using cached otel_instr)
-                    for sid in list(session_spans.keys()):
-                        otel_instr.end_session_span(session_spans[sid], f"Session failed due to circuit breaker: {cb.name}")
-                        del session_spans[sid]
-                    break
+        while True:
+            # Check for interrupts
+            if self.interrupt_sig:
+                if progress_ctx and stage_task:
+                    progress_ctx.remove_task(stage_task)
+                logger.info("Loadgen encountered SIGINT")
+                stage_status = StageStatus.FAILED
+                # Clean up any active session spans (using cached otel_instr)
+                for sid in list(session_spans.keys()):
+                    otel_instr.end_session_span(session_spans[sid], "Session interrupted by SIGINT")
+                    del session_spans[sid]
+                break
 
-                if timeout is not None and time.perf_counter() - start_time >= timeout:
-                    pbar.close()
-                    logger.warning(f"Stage {stage_id}: timeout after {timeout:.1f}s")
-                    stage_status = StageStatus.FAILED
-                    # Clean up any active session spans (using cached otel_instr)
-                    for sid in list(session_spans.keys()):
-                        otel_instr.end_session_span(session_spans[sid], "Session timed out")
-                        del session_spans[sid]
-                    break
+            if cb := next((cb for cb in self.circuit_breakers if cb.is_open()), None):
+                if progress_ctx and stage_task:
+                    progress_ctx.remove_task(stage_task)
+                logger.warning(f'Loadgen detects circuit breakers "{cb.name}" open, exit the stage.')
[…]
