### 2026-04-05 816ec2d
test: fix MagicMock spec for DataGenerator in progress test
Use spec=DataGenerator when creating mock_datagen to pass isinstance()
check in LoadGenerator.run() method. This prevents TypeError being
raised during non-multiprocessing test execution.
