### 2026-04-05 9378046
add http error handling for OTEL replay
