+++ inference_perf/client/modelserver/openai_client.py
@@ -28,6 +28,8 @@ import logging
 import requests
 import ssl
 
+from ...datagen.otel_trace_replay_datagen import OTelChatCompletionAPIData
+
 logger = logging.getLogger(__name__)
 
 
@@ -299,10 +301,34 @@ class openAIModelServerClientSession(ModelServerClientSession):
                                 lora_adapter=lora_adapter,
                             )
                         else:
-                            error = ErrorResponseInfo(
-                                error_msg=response_content,
-                                error_type=f"HTTP Error {response.status}",
-                            )
+                            """
+                            Handle HTTP error responses (status != 200).
+                            
+                            For OTEL trace replay, we need to call process_failure() to:
+                            1. Mark the session as failed in shared state
+                            2. Register empty output via on_completion() to unblock dependent nodes
+                            3. Maintain dependency graph integrity when nodes fail
+                            
+                            This ensures that if request X fails and request Y depends on X's output,
+                            Y can still proceed (with empty input) rather than hanging indefinitely.
+                            
+                            Note: The original code only logged errors for non-200 responses without
+                            calling process_failure(). This special handling for OTelChatCompletionAPIData
+                            ensures proper failure propagation in trace replay scenarios.
+                            """
+                            if isinstance(data, OTelChatCompletionAPIData) and response is not None:
+                                error = ErrorResponseInfo(
+                                    error_msg=response_content,
+                                    error_type=f"HTTP Error {response.status}",
+                                )
+                                exception = Exception(f"{error.error_type}: {error.error_msg}")
+                                response_info = await data.process_failure(
+                                    response=response,
+                                    config=self.client.api_config,
+                                    tokenizer=self.client.tokenizer,
+                                    exception=exception,
+                                    lora_adapter=lora_adapter,
+                                )
                     except Exception as read_error:
                         # Handle errors reading response body
                         if not response_content:
