+++ inference_perf/datagen/otel_trace_replay_datagen.py
@@ -195,13 +195,17 @@ class WorkerSessionTracker:
         """
         return self._node_completions.get(session_id, {}).get(node_id)
 
-    def mark_session_failed(self, session_id: str) -> None:
+    def mark_session_failed(self, session_id: str, shared_failed_sessions: Any = None) -> None:
         """Mark a session as failed.
 
         Args:
             session_id: The session ID
+            shared_failed_sessions: Optional shared list for cross-process communication
         """
         self._failed_sessions.add(session_id)
+        # Also add to shared list if available (for cross-process visibility)
+        if shared_failed_sessions is not None and session_id not in shared_failed_sessions:
+            shared_failed_sessions.append(session_id)
 
     def is_session_failed(self, session_id: str) -> bool:
         """Check if a session has failed.
@@ -394,6 +398,9 @@ class OTelChatCompletionAPIData(ChatCompletionAPIData):
     # Session completion notification
     completion_queue: Any  # mp.Queue for notifying main process
     total_nodes_in_session: int  # Total number of nodes in this session
+
+    # Shared failed sessions list for cross-process failure tracking
+    shared_failed_sessions: Any = None  # mp.Manager().list() or None
 
     # Dependency information for async waiting
     predecessor_node_ids: List[str] = field(default_factory=list)
@@ -665,7 +672,8 @@ class OTelChatCompletionAPIData(ChatCompletionAPIData):
         session_id = self._extract_session_id()
 
         # Mark the entire session as failed to prevent other nodes from processing
-        self.worker_tracker.mark_session_failed(session_id)
+        # Pass shared_failed_sessions if available for cross-process visibility
+        self.worker_tracker.mark_session_failed(session_id, self.shared_failed_sessions)
 
         # Create empty inference info
         empty_info = OTelInferenceInfo(
@@ -841,8 +849,12 @@ class OTelTraceReplayDataGenerator(SessionGenerator, LazyLoadDataMixin):
         # Main process consumes from this queue to update state and close OTEL spans.
         if mp_manager is not None:
             self.session_completion_queue: Any = mp_manager.Queue()  # Queue for (session_id, completion_data)
+            # Shared set for tracking failed sessions across all worker processes
+            # This allows the main process to detect when sessions have failed
+            self.shared_failed_sessions: Any = mp_manager.list()  # Using list as a shared set
         else:
             self.session_completion_queue = None
+            self.shared_failed_sessions = None
 
         # Load and process all OTel trace files
         self.sessions: List[OTelTraceSession] = []
@@ -1240,7 +1252,24 @@ class OTelTraceReplayDataGenerator(SessionGenerator, LazyLoadDataMixin):
             logger.warning(f"Attempted to check unknown session: {session_id}")
             return False
 
-        return state.is_complete
+        # A session is considered completed if either:
+        # 1. It has completed normally (is_complete = True), OR
+        # 2. It has failed (marked in shared_failed_sessions for multiprocess mode)
+        if state.is_complete:
+            return True
+
+        # Check if session has failed in any worker process
+        # In multiprocess mode, workers add failed sessions to shared_failed_sessions
+        shared_failed = getattr(self, 'shared_failed_sessions', None)
+        if shared_failed is not None:
+            is_failed = session_id in shared_failed
+            if is_failed:
+                # Mark it as complete so we don't check again
+                state.is_complete = True
+                logger.info(f"Session {session_id} marked as complete due to failure")
+                return True
+
+        return False
 
     def load_lazy_data(self, data: LazyLoadInferenceAPIData) -> InferenceAPIData:
         """Load the actual request data for a given index.
[…]
