+++ inference_perf/loadgen/load_generator.py
@@ -582,10 +582,13 @@ class LoadGenerator:
 
                         # End OTEL session span (using cached otel_instr)
                         if session_id in session_spans:
-                            # Check if session failed
-                            session_failed = hasattr(
-                                self.datagen, "shared_state"
-                            ) and self.datagen.shared_state.is_session_failed(session_id)
+                            # Check if session failed using shared_failed_sessions list
+                            session_failed = False
+                            if (
+                                hasattr(self.datagen, "shared_failed_sessions")
+                                and self.datagen.shared_failed_sessions is not None
+                            ):
+                                session_failed = session_id in self.datagen.shared_failed_sessions
                             error_msg = "Session failed" if session_failed else None
                             otel_instr.end_session_span(session_spans[session_id], error_msg)
                             del session_spans[session_id]
