+++ examples/otel/README.md
@@ -48,7 +48,7 @@ with the *actual* text generated by the predecessor during the current run. This
 growing-context patterns (agent chains, multi-turn conversations) produce realistic KV-cache
 behaviour rather than replaying stale recorded text.
 
-Three coordination mechanisms handle this:
+Four coordination mechanisms handle this:
 
 - **`NodeOutputRegistry`** — intra-worker only. Holds plain dicts (`node_id → output text`,
   `node_id → input messages`) and one `asyncio.Event` per node. When a node completes,
@@ -65,14 +65,58 @@ Three coordination mechanisms handle this:
   times and failure status) to an `mp.Queue`. The main process consumes from this queue in
   `check_session_completed()` instead of polling shared state.
 
-Each `OTelChatCompletionAPIData` holds references to `registry`, `worker_tracker`, and
-`completion_queue`. Before dispatching an HTTP request, the worker calls
-`wait_for_predecessors_and_substitute()`, which awaits predecessors via `registry.require_async()`
-(zero threads — pure `asyncio.Event` suspension), checks `worker_tracker` for session failure
-before and after waiting, then substitutes output segments with actual predecessor text. After
-the response returns, `on_completion()` writes to `registry` (unblocking dependents) and
-`worker_tracker` (recording completion), and pushes to `completion_queue` if this was the last
-node in the session.
+- **`shared_failed_sessions`** — worker→main failure notification. An `mp.Manager().list()`
+  that allows workers to notify the main process when a session fails. When `process_failure()`
+  is called, the worker adds the session_id to this shared list. The main process checks this
+  list when ending OTEL session spans to mark failed sessions with error messages.
+
+Each `OTelChatCompletionAPIData` holds references to `registry`, `worker_tracker`,
+`completion_queue`, and `shared_failed_sessions`. Before dispatching an HTTP request, the worker
+calls `wait_for_predecessors_and_substitute()`, which awaits predecessors via
+`registry.require_async()` (zero threads — pure `asyncio.Event` suspension), checks
+`worker_tracker` for session failure before and after waiting, then substitutes output segments
+with actual predecessor text. After the response returns, `on_completion()` writes to `registry`
+(unblocking dependents) and `worker_tracker` (recording completion), and pushes to
+`completion_queue` if this was the last node in the session. If a request fails,
+`process_failure()` marks the session as failed in both `worker_tracker` and
+`shared_failed_sessions`.
+
+### Failure handling
+
+When a node fails (network error, timeout, HTTP error), the system ensures dependent nodes don't
+hang and the session completes gracefully:
+
+**Worker-level failure handling:**
+1. `process_failure()` is called on the failed node's `OTelChatCompletionAPIData`
+2. The worker marks the entire session as failed in `WorkerSessionTracker` (local to that worker)
+3. Empty output (`""`) is registered in `NodeOutputRegistry` via `on_completion()`
+4. Dependent nodes are unblocked and can proceed with empty input
+
+**Session-level failure propagation (within a worker):**
+- **Pre-wait check**: Before waiting for predecessors, each node checks if its session has failed
+- **Post-wait check**: After predecessors complete, each node checks again (a predecessor may have failed during the wait)
+- **Skip mechanism**: If either check detects failure, the node sets `skip_request = True`, registers empty output, and returns without making an HTTP request
+- **Completion tracking**: Failed nodes still count as "completed" for session lifecycle tracking
+- **Session-to-worker affinity**: All nodes of a session run on the same worker, so `WorkerSessionTracker` (local to each worker) is sufficient for intra-session failure propagation
+
+**Worker-to-main-process communication:**
+- `shared_failed_sessions` (an `mp.Manager().list()`) allows workers to notify the main process of failures
[…]
