### 2026-04-06 e88acf8
improve tracking failed sessions for main process.
