+++ examples/otel/README.md
@@ -48,7 +48,7 @@ with the *actual* text generated by the predecessor during the current run. This
 growing-context patterns (agent chains, multi-turn conversations) produce realistic KV-cache
 behaviour rather than replaying stale recorded text.
 
-Four coordination mechanisms handle this:
+Three coordination mechanisms handle this:
 
 - **`NodeOutputRegistry`** — intra-worker only. Holds plain dicts (`node_id → output text`,
   `node_id → input messages`) and one `asyncio.Event` per node. When a node completes,
@@ -63,23 +63,18 @@ Four coordination mechanisms handle this:
 - **`session_completion_queue`** — event-driven worker→main communication. When the last node
   of a session completes, the worker pushes a completion notification (with node completion
   times and failure status) to an `mp.Queue`. The main process consumes from this queue in
-  `check_session_completed()` instead of polling shared state.
+  `check_session_completed()` instead of polling shared state. The completion notification
+  includes a `"failed"` field indicating whether the session failed.
 
-- **`shared_failed_sessions`** — worker→main failure notification. An `mp.Manager().list()`
-  that allows workers to notify the main process when a session fails. When `process_failure()`
-  is called, the worker adds the session_id to this shared list. The main process checks this
-  list when ending OTEL session spans to mark failed sessions with error messages.
-
-Each `OTelChatCompletionAPIData` holds references to `registry`, `worker_tracker`,
-`completion_queue`, and `shared_failed_sessions`. Before dispatching an HTTP request, the worker
-calls `wait_for_predecessors_and_substitute()`, which awaits predecessors via
-`registry.require_async()` (zero threads — pure `asyncio.Event` suspension), checks
-`worker_tracker` for session failure before and after waiting, then substitutes output segments
-with actual predecessor text. After the response returns, `on_completion()` writes to `registry`
-(unblocking dependents) and `worker_tracker` (recording completion), and pushes to
-`completion_queue` if this was the last node in the session. If a request fails,
-`process_failure()` marks the session as failed in both `worker_tracker` and
-`shared_failed_sessions`.
+Each `OTelChatCompletionAPIData` holds references to `registry`, `worker_tracker`, and
+`completion_queue`. Before dispatching an HTTP request, the worker calls
+`wait_for_predecessors_and_substitute()`, which awaits predecessors via `registry.require_async()`
+(zero threads — pure `asyncio.Event` suspension), checks `worker_tracker` for session failure
+before and after waiting, then substitutes output segments with actual predecessor text. After
+the response returns, `on_completion()` writes to `registry` (unblocking dependents) and
+`worker_tracker` (recording completion), and pushes to `completion_queue` if this was the last
+node in the session. The completion notification includes the failure status from
+`worker_tracker`.
 
 ### Failure handling
 
@@ -100,10 +95,10 @@ hang and the session completes gracefully:
 - **Session-to-worker affinity**: All nodes of a session run on the same worker, so `WorkerSessionTracker` (local to each worker) is sufficient for intra-session failure propagation
 
 **Worker-to-main-process communication:**
-- `shared_failed_sessions` (an `mp.Manager().list()`) allows workers to notify the main process of failures
-- When a worker marks a session as failed, it also adds the session_id to this shared list
-- The main process (load generator) checks this list when ending OTEL session spans
-- Failed sessions are marked with an error message in their OTEL span: `"Session failed"`
+- The `session_completion_queue` includes a `"failed"` field in the completion notification
+- When the last node completes, `on_completion()` checks `worker_tracker.is_session_failed()` and includes this in the queue notification
+- The main process calls `_process_completion_queue()` which stores the failure status in `SessionGraphState.failed`
[…]
