### 2026-04-06 4991a7e
refactor(otel-replay): propagate node failure via NodeFailedError instead of empty registry entries
  Add NodeFailedError and record_failure() to NodeOutputRegistry so failed
  nodes signal successors by firing the existing asyncio.Event without
  writing to _node_output_text. require_async() raises NodeFailedError on
  wake, which wait_for_predecessors_and_substitute() catches to set
  skip_request and propagate failure to its own successors.

  Skipped nodes no longer call on_completion() or record_node_completed().
  process_failure() sends the immediate session-failure notification with
  
[…]
