+++ tests/test_otel_replay_datagen.py
@@ -46,6 +46,7 @@ import pytest
 sys.path.insert(0, str(Path(__file__).parent.parent))
 
 from inference_perf.datagen.otel_trace_replay_datagen import (
+    NodeFailedError,
     NodeOutputRegistry,
     OTelChatCompletionAPIData,
     OTelInferenceInfo,
@@ -169,6 +170,38 @@ class TestNodeOutputRegistry:
         retrieved = reg.get_messages_by_node_id("node_001")
         assert retrieved is not None
         assert retrieved == []
+
+    def test_record_failure_fast_path(self) -> None:
+        """require_async raises NodeFailedError immediately for an already-failed node."""
+        reg = NodeOutputRegistry()
+        reg.record_failure("node_001")
+        assert reg.is_node_failed("node_001")
+
+        with pytest.raises(NodeFailedError) as exc_info:
+            asyncio.run(reg.require_async("node_001"))
+        assert exc_info.value.node_id == "node_001"
+
+    def test_record_failure_wakes_waiter_with_error(self) -> None:
+        """record_failure wakes a coroutine blocked in require_async with NodeFailedError."""
+        reg = NodeOutputRegistry()
+
+        async def _run() -> None:
+            async def fail_producer() -> None:
+                await asyncio.sleep(0.05)
+                reg.record_failure("node_001")
+
+            asyncio.create_task(fail_producer())
+            await reg.require_async("node_001", timeout_sec=2.0)
+
+        with pytest.raises(NodeFailedError):
+            asyncio.run(_run())
+
+    def test_record_failure_idempotent(self) -> None:
+        """Calling record_failure multiple times for the same node is safe."""
+        reg = NodeOutputRegistry()
+        reg.record_failure("node_001")
+        reg.record_failure("node_001")  # should not raise
+        assert reg.is_node_failed("node_001")
 
 
 # ---------------------------------------------------------------------------
@@ -335,8 +368,9 @@ class TestOTelChatCompletionAPIData:
 
         # Should have set skip_request flag
         assert api_data.skip_request is True
-        # Should have recorded empty output
-        assert registry.get_output_by_node_id("session_1:node_1") == ""
+        # Registry should record failure (not empty output)
+        assert registry.is_node_failed("session_1:node_1")
+        assert registry.get_output_by_node_id("session_1:node_1") is None
 
     @pytest.mark.asyncio
     async def test_wait_for_predecessors_waits_and_substitutes(self) -> None:
@@ -814,10 +848,11 @@ class TestOTelChatCompletionAPIDataErrorPaths:
         # Session should be marked as failed
         assert tracker.is_session_failed("session_1")
 
-        # Empty output should be registered
-        assert registry.get_output_by_node_id("session_1:node_0") == ""
+        # Registry should record failure for the node (not empty output)
+        assert registry.is_node_failed("session_1:node_0")
+        assert registry.get_output_by_node_id("session_1:node_0") is None
 
-        # Should return empty inference info (base InferenceInfo, not OTelInferenceInfo)
+        # Should return empty inference info
         assert isinstance(result, OTelInferenceInfo)
         assert result.output_text == ""
         assert result.input_tokens == 0
@@ -825,7 +860,7 @@ class TestOTelChatCompletionAPIDataErrorPaths:
 
     @pytest.mark.asyncio
     async def test_wait_for_predecessors_post_wait_failure_check(self) -> None:
-        """Test post-wait failure check skips node if session failed during wait."""
+        """Test failure propagation: when predecessor fails, successor skips via NodeFailedError."""
         registry = NodeOutputRegistry()
         tracker = WorkerSessionTracker()
 
@@ -841,22 +876,107 @@ class TestOTelChatCompletionAPIDataErrorPaths:
             predecessor_node_ids=["session_1:node_0"],
         )
 
[…]
