+++ inference_perf/datagen/otel_trace_replay_datagen.py
@@ -91,6 +91,19 @@ from inference_perf.utils.custom_tokenizer import CustomTokenizer
 logger = logging.getLogger(__name__)
 
 # ---------------------------------------------------------------------------
+# NodeFailedError — raised by require_async when a predecessor node failed
+# ---------------------------------------------------------------------------
+
+
+class NodeFailedError(Exception):
+    """Raised by NodeOutputRegistry.require_async when the awaited node failed."""
+
+    def __init__(self, node_id: str) -> None:
+        super().__init__(f"Predecessor node {node_id!r} failed")
+        self.node_id = node_id
+
+
+# ---------------------------------------------------------------------------
 # SessionGraphState — tracks graph traversal state for one session
 # ---------------------------------------------------------------------------
 
@@ -271,6 +284,10 @@ class NodeOutputRegistry:
         # worker's single event loop; no lock needed.
         self._node_events: Dict[str, asyncio.Event] = {}
 
+        # Set of node_ids that failed. record_failure() adds here and sets the
+        # node's event so waiting coroutines wake and raise NodeFailedError.
+        self._failed_node_ids: Set[str] = set()
+
     def record(self, node_id: str, output_text: str, messages: List[Any]) -> None:
         """Register output for a completed node and wake any async waiters.
 
@@ -315,6 +332,22 @@ class NodeOutputRegistry:
         """
         return list(self._node_output_text.keys())
 
+    def record_failure(self, node_id: str) -> None:
+        """Mark a node as failed and wake any async waiters with a failure signal.
+
+        Idempotent: calling multiple times for the same node is safe.
+        Only sets the event — does not write to _node_output_text.
+        """
+        self._failed_node_ids.add(node_id)
+        if node_id not in self._node_events:
+            self._node_events[node_id] = asyncio.Event()
+        self._node_events[node_id].set()
+        logger.debug(f"Recorded failure for node {node_id}")
+
+    def is_node_failed(self, node_id: str) -> bool:
+        """Return True if this node has been marked as failed."""
+        return node_id in self._failed_node_ids
+
     async def require_async(self, node_id: str, timeout_sec: float = 3600.0) -> str:
         """Wait asynchronously for a node's output.
 
@@ -322,8 +355,13 @@ class NodeOutputRegistry:
         the calling coroutine until record() is called for this node_id, then resumes.
 
         Raises:
+            NodeFailedError: if the node failed (via record_failure)
             TimeoutError: if output not available within timeout_sec
         """
+        # Fast path: predecessor already failed.
+        if node_id in self._failed_node_ids:
+            raise NodeFailedError(node_id)
+
         # Fast path: predecessor already completed.
         output = self._node_output_text.get(node_id)
         if output is not None:
@@ -336,8 +374,10 @@ class NodeOutputRegistry:
             self._node_events[node_id] = asyncio.Event()
         event = self._node_events[node_id]
 
-        # Second fast-path check: no await has occurred since the first check,
+        # Second fast-path checks: no await has occurred since the first checks,
         # but be defensive in case this method's call site changes in the future.
+        if node_id in self._failed_node_ids:
+            raise NodeFailedError(node_id)
         output = self._node_output_text.get(node_id)
         if output is not None:
             return output
@@ -351,6 +391,10 @@ class NodeOutputRegistry:
                 f"NodeOutputRegistry: output for '{node_id}' not available after "
                 f"{timeout_sec:.1f}s. Check that the predecessor is not blocked or failed."
             ) from e
+
+        # Check if the event was fired due to failure.
+        if node_id in self._failed_node_ids:
+            raise NodeFailedError(node_id)
 
         # record() writes _node_output_text before calling event.set(). Both run on
[…]

[…]
