### 2026-04-06 d06fe87
feat(otel-replay): surface cancelled LLM calls in session-level reports
  Add num_nodes_cancelled to SessionLifecycleMetric and SessionGraphState,
  populated from the cancelled_nodes count in the failure completion
  notification. Summary reports now include total_nodes, total_nodes_completed,
  and total_nodes_cancelled for a clear at-a-glance breakdown.
