### 2026-04-06 4ebb6dd
Fix mypy errors
