### 2026-04-06 4446b8a
add e2e testing for prometheus querying
