+++ e2e/tests/test_llm_d_inference_sim.py
@@ -133,7 +133,8 @@ async def test_completion_successful_run(data: dict, load: dict):
                         "per_request": True,
                     },
                 },
-            }
+            },
+            extra_env={"PYTHONPATH": "/workspace"}
         )
 
     assert result.success, "Benchmark failed"
