### 2026-04-06 a12bb8e
Address review comments from jjk-g
- Change skip request log from info to debug level to reduce noise
- Add validation to ensure datagen type matches load_type in LoadGenerator
  - TRACE_SESSION_REPLAY requires SessionGenerator
  - Other load types require DataGenerator
  - Raises TypeError with clear message if mismatch detected
- Improve error message clarity: "corpus" → "trace files"
- Remove unnecessary sleep(1) in progress update loop (change to sleep(0))
