### 2026-04-06 904aacc
Fix Prometheus metrics E2E test and metric names in vllm_client.py
