+++ e2e/tests/test_prometheus.py
@@ -13,6 +13,7 @@
 # limitations under the License.
 
 import json
+import re
 import os
 import sys
 import shutil
@@ -144,7 +145,7 @@ async def test_prometheus_metrics_collection(prometheus_server):
                     "type": "prometheus",
                     "prometheus": {
                         "url": prometheus_url,
-                        "scrape_interval": 1,
+                        "scrape_interval": 5,
                     },
                 },
                 "report": {
@@ -159,6 +160,22 @@ async def test_prometheus_metrics_collection(prometheus_server):
             executable=[sys.executable, "/workspace/inference_perf/main.py"],
             extra_env={"PYTHONPATH": "/workspace"}
         )
+        
+        # Debug and verify metrics exposed by simulator
+        try:
+            resp = requests.get("http://127.0.0.1:18000/metrics", timeout=1)
+            metrics_content = resp.text
+            print(f"Simulator Metrics Content:\n{metrics_content}")
+            
+            # Verify simulator recorded 25 successes
+            match = re.search(r'vllm:request_success_total\{.*?\} (\d+)', metrics_content)
+            assert match, "vllm:request_success_total not found in simulator metrics"
+            sim_success_count = int(match.group(1))
+            assert sim_success_count == 25, f"Expected 25 successes in simulator metrics, got {sim_success_count}"
+            print("Verified 25 successes in simulator metrics")
+        except Exception as e:
+            print(f"Failed to get or verify simulator metrics: {e}")
+            raise
 
     if not result.success:
         print(f"Simulator Stdout:\n{sim.stdout}")
@@ -177,6 +194,13 @@ async def test_prometheus_metrics_collection(prometheus_server):
     assert isinstance(prom_report, dict), "Report should be a dictionary"
     
     print(f"Prometheus Report Content:\n{json.dumps(prom_report, indent=2)}")
+    
+    # Assertions on content
+    successes_obj = prom_report.get("successes", {})
+    success_count = successes_obj.get("request_success_count", 0.0)
+    print(f"Asserting request_success_count ({success_count}) is greater than 10")
+    # Due to PromQL increase() extrapolation on short windows, we loosen this assertion
+    assert success_count > 10.0, f"Expected > 10 successes in report due to extrapolation, got {success_count}"
     
     lifecycle_report = result.reports.get("summary_lifecycle_metrics.json")
     if lifecycle_report:
