### 2026-04-06 4f0f34e
Fix test mocks to use DataGenerator spec
Update TestLoadGeneratorConcurrency to use MagicMock(spec=DataGenerator)
to satisfy isinstance validation added in LoadGenerator.__init__.
