### 2026-04-06 ba92d7b
Fix mypy error in google_managed_prometheus_client.py
