+++ e2e/tests/test_prometheus.py
@@ -13,19 +13,20 @@
 # limitations under the License.
 
 import json
+import pathlib
 import re
-import os
-import sys
 import shutil
-import subprocess
-import time
-from pathlib import Path
+import sys
+
 import pytest
 import requests
 
-from utils.llm_d_inference_sim import LLMDInferenceSimRunner
 from utils.benchmark import run_benchmark_minimal
+from utils.llm_d_inference_sim import LLMDInferenceSimRunner
 from utils.testdata import extract_tarball
+
+PROJECT_ROOT = pathlib.Path(__file__).parent.parent.parent.resolve()
+MAIN_PY_PATH = PROJECT_ROOT / "inference_perf" / "main.py"
 
 TEST_MODEL_NAME = "google/gemma-3-270m"
 TEST_MODEL_TARBALL = "e2e/testdata/models/google_gemma-3-270m.tar.gz"
@@ -33,70 +34,6 @@ TEST_MODEL_TARBALL = "e2e/testdata/models/google_gemma-3-270m.tar.gz"
 
 def is_prometheus_available() -> bool:
     return shutil.which("prometheus") is not None
-
-
-@pytest.fixture(scope="module")
-def prometheus_server():
-    """Starts a lightweight ephemeral Prometheus instance."""
-    import tempfile
-
-    with tempfile.TemporaryDirectory() as tmpdir:
-        tmp_path = Path(tmpdir)
-        config_path = tmp_path / "prometheus.yml"
-
-        # Write minimal config pointing to the simulator
-        config_path.write_text(
-            """
-global:
-  scrape_interval: 1s
-scrape_configs:
-  - job_name: 'llm-d-inference-sim'
-    static_configs:
-      - targets: ['127.0.0.1:18000']
-""",
-            encoding="utf-8",
-        )
-
-        data_dir = tmp_path / "data"
-        data_dir.mkdir()
-
-        # Start prometheus
-        proc = subprocess.Popen(
-            [
-                "prometheus",
-                f"--config.file={config_path}",
-                f"--storage.tsdb.path={data_dir}",
-                "--web.listen-address=127.0.0.1:9090",
-            ],
-            stdout=subprocess.PIPE,
-            stderr=subprocess.STDOUT,
-        )
-
-        # Wait for ready
-        ready = False
-        for _ in range(30):
-            try:
-                resp = requests.get("http://127.0.0.1:9090/-/ready", timeout=1)
-                if resp.status_code == 200:
-                    ready = True
-                    break
-            except Exception:
-                pass
-            time.sleep(1)
-
-        if not ready:
-            proc.terminate()
-            stdout, _ = proc.communicate()
-            raise Exception(f"Prometheus failed to become ready. Output:\n{stdout.decode()}")
-
-        yield "http://127.0.0.1:9090"
-
-        # Teardown
-        proc.terminate()
-        try:
-            proc.wait(timeout=5)
-        except subprocess.TimeoutExpired:
-            proc.kill()
 
 
 @pytest.mark.asyncio
@@ -125,7 +62,7 @@ async def test_prometheus_metrics_collection(prometheus_server):
                 },
                 "load": {
                     "type": "constant",
-                    "stages": [{"rate": 5, "duration": 5}],
+                    "stages": [{"rate": 5, "duration": 30}],
                     "num_workers": 1,
                 },
                 "api": {
@@ -157,53 +94,162 @@ async def test_prometheus_metrics_collection(prometheus_server):
                     },
                 },
             },
-            executable=[sys.executable, "/workspace/inference_perf/main.py"],
-            extra_env={"PYTHONPATH": "/workspace"}
+            executable=[sys.executable, str(MAIN_PY_PATH)],
[…]
