### 2026-04-06 37314e1
feat: Implement distributed Redis-based load generator for high-scale benchmarking
This feature introduces a distributed architecture for inference-perf, enabling it to scale beyond a single host and handle >10,000 QPS. By using Redis as a message broker and task queue, load is distributed across an Orchestrator and multiple Worker instances, overcoming Python GIL limitations and single-host resource constraints.

Specifics:
- Added distributed architecture components: Orchestrator, Worker, and Submitter (CLI).
- Implemented Redis-based job streaming, task distribution, and real-time telemetry
[…]
