### 2026-04-06 a5a7b7d
rename: use event instead of node for OTel trace replay graph elements
