+++ tests/test_otel_replay_datagen.py
@@ -18,13 +18,13 @@
 Tests for OTelTraceReplayDataGenerator output-aware replay.
 
 Tests the OTEL trace replay architecture:
-- NodeOutputRegistry: intra-worker output coordination via asyncio.Event
+- EventOutputRegistry: intra-worker output coordination via asyncio.Event
 - WorkerSessionTracker: per-worker session state tracking
 - OTelChatCompletionAPIData: output substitution and completion notification
 - Session completion via mp.Queue (event-driven, not polling)
 
 Key architectural principles tested:
-1. Session-to-worker affinity: all nodes of a session run on same worker
+1. Session-to-worker affinity: all events of a session run on same worker
 2. Zero-thread waiting: asyncio.Event suspension, no OS threads
 3. Event-driven completion: queue notifications, not polling
 4. Nested dict structure: efficient O(1) lookups without string operations
@@ -46,8 +46,8 @@ import pytest
 sys.path.insert(0, str(Path(__file__).parent.parent))
 
 from inference_perf.datagen.otel_trace_replay_datagen import (
-    NodeFailedError,
-    NodeOutputRegistry,
+    EventFailedError,
+    EventOutputRegistry,
     OTelChatCompletionAPIData,
     OTelInferenceInfo,
     OTelTraceReplayDataGenerator,
@@ -95,67 +95,67 @@ def make_mock_api_config_streaming(streaming: bool = False) -> MagicMock:
 
 
 # ---------------------------------------------------------------------------
-# NodeOutputRegistry tests
+# EventOutputRegistry tests
 # ---------------------------------------------------------------------------
 
 
-class TestNodeOutputRegistry:
+class TestEventOutputRegistry:
     """Test intra-worker output coordination via asyncio.Event."""
 
     def test_record_and_get(self) -> None:
         """Basic record/retrieve functionality."""
-        reg = NodeOutputRegistry()
-        reg.record("node_001", "Hello world", [])
-        assert reg.get_output_by_node_id("node_001") == "Hello world"
+        reg = EventOutputRegistry()
+        reg.record("event_001", "Hello world", [])
+        assert reg.get_output_by_event_id("event_001") == "Hello world"
 
     def test_get_missing_returns_none(self) -> None:
-        """Missing nodes return None."""
-        reg = NodeOutputRegistry()
-        assert reg.get_output_by_node_id("nonexistent") is None
+        """Missing events return None."""
+        reg = EventOutputRegistry()
+        assert reg.get_output_by_event_id("nonexistent") is None
 
     def test_require_async_already_registered(self) -> None:
         """Fast path: output already present before require_async is called."""
-        reg = NodeOutputRegistry()
-        reg.record("node_001", "Some output", [])
-        result = asyncio.run(reg.require_async("node_001"))
+        reg = EventOutputRegistry()
+        reg.record("event_001", "Some output", [])
+        result = asyncio.run(reg.require_async("event_001"))
         assert result == "Some output"
 
     def test_require_async_waits_for_record(self) -> None:
         """require_async suspends via asyncio.Event until record() is called."""
-        reg = NodeOutputRegistry()
+        reg = EventOutputRegistry()
 
         async def _run() -> str:
             async def producer() -> None:
                 await asyncio.sleep(0.05)
-                reg.record("node_001", "delayed output", [])
+                reg.record("event_001", "delayed output", [])
 
             asyncio.create_task(producer())
-            return await reg.require_async("node_001", timeout_sec=2.0)
+            return await reg.require_async("event_001", timeout_sec=2.0)
 
         result = asyncio.run(_run())
         assert result == "delayed output"
 
     def test_require_async_timeout(self) -> None:
         """require_async raises TimeoutError when output never arrives."""
-        reg = NodeOutputRegistry()
[…]
