### 2026-04-07 063a238
Address review feedback
- Merge DistributionConfig into existing Distribution class
  by adding type, variance, and skew fields with backward-
  compatible defaults (type defaults to normal)
- Remove num_tool_calls_per_turn (not wired into benchmark;
  to be handled by separate agentic tool-calling PRs)
- Remove dual-path _resolve_param; use unified
  _resolve_distribution with single Distribution model
- Remove tool-calls example config

Issue: #385
