### 2026-04-07 kubernetes-sigs/inference-perf#394 @achandrasekar
WG Serving archival happened after we submitted the paper. Updated now to clarify in the paper as a part of #412.
