### 2026-04-07 kubernetes-sigs/inference-perf#391 @achandrasekar
The metrics doc in the repo is meant to provide the definition of various metrics that are important in measuring inference performance and not just the one the tool reports. The paper did not specifically call out that the tool doesn't directly report the price-perf metrics. Updated the metrics section in the paper to clarify this as a part of #412.
