### 2026-04-07 8ce1625
Address review feedback: add warning for ambiguous model selection and skip autoconfig on failure
