### 2026-04-07 f67e800
Remove detector from protos branch (added by mistake)
