### 2026-04-07 6ffd772
- Simplify type annotations from Union to base class - Remove unused event index mappings and lookup method - Clean up unused .gitignore
