### 2026-04-09 ede7a1b
Generate flags for Otel feature
