### 2026-04-09 8bc0853
Fix streaming metrics calculation by avoiding stream consumption before parsing
