### 2026-04-09 e1f0307
Fix type checks
