### 2026-04-09 51ad5fa
fix: streaming response body consumed before SSE parser
When streaming is enabled, response.text() on line 294 consumed the
entire response body before process_response() could iterate it via
parse_sse_stream(). This resulted in:
- output_tokens always 0
- TTFT, ITL, TPOT all null
- output_token_times empty

Fix: call process_response() first for 200 responses so the SSE parser
can iterate response.content, then read the body only for error responses.

For OTel span recording, streaming responses will have empty
response_content — the actual output is captured by the SSE parser
and stored in I
[…]
