### 2026-04-09 dba6a71
feat: add synthetic OTel trace generator for agentic workloads
Generates deterministic OTel JSON trace files from configurable
distributions, compatible with the otel_trace_replay pipeline.

Features:
- Two-part system prompts: shared prefix for KV cache + dynamic suffix
- Multiple distribution types: normal, lognormal, uniform, fixed
- Right-skewed turn counts matching production agentic workloads
- Deterministic seeding for reproducible A/B comparisons

Usage:
  python tools/generate_synthetic_traces.py \
    --config examples/otel/configs/synthetic/synthetic_trace_config.yaml

  python tool
[…]
