+++ tools/generate_synthetic_traces.py
@@ -1,0 +1,300 @@
+#!/usr/bin/env python3
+"""Generate synthetic OTel trace files for agentic workload benchmarking.
+
+Produces OTel-format JSON trace files compatible with inference-perf's
+otel_trace_replay data type (PR #372). Each file represents one multi-turn
+conversation with a shared system prompt prefix for prefix caching.
+
+Usage:
+    python generate_synthetic_traces.py --config synthetic_trace_config.yaml
+    python generate_synthetic_traces.py --config synthetic_trace_config.yaml --stats
+"""
+
+import argparse
+import json
+import os
+import uuid
+from dataclasses import dataclass, field
+from datetime import datetime, timedelta
+from pathlib import Path
+from typing import List, Optional
+
+import numpy as np
+import yaml
+
+
+# ---------------------------------------------------------------------------
+# Distribution sampling
+# ---------------------------------------------------------------------------
+
+@dataclass
+class DistributionConfig:
+    type: str = "normal"  # normal, lognormal, uniform, fixed
+    min: int = 10
+    max: int = 1024
+    mean: float = 512
+    std_dev: float = 200
+
+
+def sample_distribution(config: DistributionConfig, count: int, rng: np.random.RandomState) -> np.ndarray:
+    """Sample from a distribution, returning integer values clipped to [min, max]."""
+    if config.type == "fixed":
+        values = np.full(count, config.mean)
+    elif config.type == "uniform":
+        values = rng.uniform(config.min, config.max, size=count)
+    elif config.type == "lognormal":
+        # Parameterize lognormal to approximate target mean within [min, max].
+        # We shift so that min maps to 0, then use lognormal for the remainder.
+        target_mean = config.mean - config.min
+        if target_mean <= 0:
+            values = np.full(count, config.min)
+        else:
+            # sigma controls the skew; derive from std_dev relative to mean
+            sigma = np.log(1 + (config.std_dev / max(target_mean, 1)) ** 2) ** 0.5
+            mu = np.log(target_mean) - sigma**2 / 2
+            values = rng.lognormal(mean=mu, sigma=sigma, size=count) + config.min
+    elif config.type == "normal":
+        values = rng.normal(loc=config.mean, scale=config.std_dev, size=count)
+    else:
+        raise ValueError(f"Unknown distribution type: {config.type}")
+
+    return np.clip(np.round(values), config.min, config.max).astype(int)
+
+
+def parse_distribution(d) -> DistributionConfig:
+    """Parse a distribution config from YAML (dict or scalar)."""
+    if isinstance(d, (int, float)):
+        return DistributionConfig(type="fixed", min=int(d), max=int(d), mean=float(d), std_dev=0)
+    if isinstance(d, dict):
+        return DistributionConfig(**d)
+    raise ValueError(f"Invalid distribution config: {d}")
+
+
+# ---------------------------------------------------------------------------
+# Conversation blueprint
+# ---------------------------------------------------------------------------
+
+@dataclass
+class TurnBlueprint:
+    input_tokens: int
+    output_tokens: int
+
+
+@dataclass
+class ConversationBlueprint:
+    conversation_id: str
+    system_prompt: str  # shared_prefix + dynamic_suffix
+    turns: List[TurnBlueprint] = field(default_factory=list)
+
+
+# ---------------------------------------------------------------------------
+# Random text generation
+# ---------------------------------------------------------------------------
+
+def generate_random_text(rng: np.random.RandomState, num_tokens: int, vocab_size: int = 30000) -> str:
[…]
