### 2026-04-09 67baf48
Auto-generate synthetic traces at container startup
Adds an entrypoint wrapper that generates traces from a default config
before running inference-perf. No manual generation step needed — the
user just runs ml bench run with the conversation-replay bench config.

The default trace config produces 500 conversations with production-like
distributions. Override via SYNTHETIC_TRACE_CONFIG env var if needed.
