### 2026-04-09 41047ee
add expressions API
