+++ tests/loadgen/test_load_generator.py
@@ -48,6 +48,9 @@ class MockWorker:
         self.id = id
         self.shared_max_concurrency = shared_max_concurrency
 
+    def is_alive(self) -> bool:
+        return True
+
 
 class TestLoadGeneratorConcurrency(unittest.TestCase):
     def setUp(self) -> None:
@@ -103,6 +106,58 @@ class TestLoadGeneratorConcurrency(unittest.TestCase):
         self.assertEqual(self.load_generator.workers[1].shared_max_concurrency.value, 1)  # type: ignore
         self.assertEqual(self.load_generator.workers[2].shared_max_concurrency.value, 1)  # type: ignore
         self.assertEqual(self.load_generator.workers[3].shared_max_concurrency.value, 0)  # type: ignore
+
+    @unittest.mock.patch("inference_perf.loadgen.load_generator.sleep", new_callable=unittest.mock.AsyncMock)
+    def test_run_stage_dynamic_concurrency(self, mock_sleep: Any) -> None:
+        # Setup workers
+        self.load_generator.workers = []
+        for i in range(4):
+            shared_val = mp.Value("i", 0)
+            self.load_generator.workers.append(MockWorker(i, shared_val))  # type: ignore
+
+        # Mock _set_worker_concurrency to record calls
+        patcher1 = unittest.mock.patch.object(self.load_generator, "_set_worker_concurrency")
+        mock_set_concurrency = patcher1.start()
+        self.addCleanup(patcher1.stop)
+
+        # Mock arguments for run_stage
+        request_queue = MagicMock()
+        active_requests_counter = MagicMock()
+        finished_requests_counter = MagicMock()
+        finished_requests_counter.get_lock.return_value.__enter__.return_value = None
+
+        # Simulate finished_requests_counter.value changing to exit loop
+        # We need enough values for setter and getter calls!
+        type(finished_requests_counter).value = unittest.mock.PropertyMock(side_effect=[0, 0, 10, 10])
+
+        request_phase = MagicMock()
+
+        # Mock datagen and timer
+        self.mock_datagen.trace = None
+        self.mock_datagen.get_data.return_value = iter([MagicMock(preferred_worker_id=-1)] * 10)
+
+        mock_timer = MagicMock()
+        mock_timer.start_timer.return_value = iter([0.0] * 10)
+        patcher2 = unittest.mock.patch.object(self.load_generator, "get_timer", return_value=mock_timer)
+        patcher2.start()
+        self.addCleanup(patcher2.stop)
+
+        async def run() -> None:
+            await self.load_generator.run_stage(
+                stage_id=0,
+                interval=1.0,
+                duration=10,
+                request_queue=request_queue,
+                active_requests_counter=active_requests_counter,
+                finished_requests_counter=finished_requests_counter,
+                request_phase=request_phase,
+                concurrency_level="10 + t",
+            )
+
+        asyncio.run(run())
+
+        # Verify _set_worker_concurrency was called
+        mock_set_concurrency.assert_called()
 
 
 if __name__ == "__main__":
