### 2026-04-09 994eed4
Generate traces at build time instead of runtime
ml bench run overrides the container's ENTRYPOINT, so runtime generation
via entrypoint.sh doesn't work. Instead, generate traces at Docker build
time — they're deterministic (seeded), so same config = same files.

To change distributions, edit tools/default_trace_config.yaml and rebuild.
