### 2026-04-09 51b1de3
Remove vendor-specific references from examples
Replace Shopify-specific terminology (shop context, sidekick, ml bench,
six-bench) with generic equivalents suitable for an OSS library.
