### 2026-04-09 d1105fb
Add comprehensive tests for streaming fix and trace generator
- test_streaming_response_order: verifies SSE parser receives streaming
  content and non-streaming path reads JSON body
- test_generate_synthetic_traces: 17 tests covering distributions
  (fixed/uniform/normal/lognormal), determinism, OTel format validation,
  shared prefix consistency, and end-to-end file generation
