+++ tests/test_generate_synthetic_traces.py
@@ -1,0 +1,291 @@
+"""Tests for the synthetic OTel trace generator."""
+
+import json
+import os
+import tempfile
+from pathlib import Path
+
+import numpy as np
+import pytest
+import yaml
+
+# Import from tools/ — add to path
+import sys
+
+sys.path.insert(0, str(Path(__file__).parent.parent / "tools"))
+from generate_synthetic_traces import (
+    ConversationBlueprint,
+    DistributionConfig,
+    TurnBlueprint,
+    generate_otel_trace,
+    generate_random_text,
+    generate_traces,
+    parse_distribution,
+    sample_distribution,
+)
+
+
+class TestDistributionConfig:
+    def test_parse_scalar(self) -> None:
+        config = parse_distribution(42)
+        assert config.type == "fixed"
+        assert config.mean == 42.0
+
+    def test_parse_dict(self) -> None:
+        config = parse_distribution({"type": "lognormal", "min": 3, "max": 50, "mean": 4, "std_dev": 3})
+        assert config.type == "lognormal"
+        assert config.min == 3
+        assert config.max == 50
+
+    def test_parse_invalid(self) -> None:
+        with pytest.raises(ValueError):
+            parse_distribution("invalid")
+
+
+class TestSampleDistribution:
+    def test_fixed(self) -> None:
+        config = DistributionConfig(type="fixed", min=10, max=10, mean=10, std_dev=0)
+        rng = np.random.RandomState(42)
+        values = sample_distribution(config, 100, rng)
+        assert all(v == 10 for v in values)
+
+    def test_uniform_within_bounds(self) -> None:
+        config = DistributionConfig(type="uniform", min=5, max=15, mean=10, std_dev=0)
+        rng = np.random.RandomState(42)
+        values = sample_distribution(config, 1000, rng)
+        assert all(5 <= v <= 15 for v in values)
+
+    def test_normal_within_bounds(self) -> None:
+        config = DistributionConfig(type="normal", min=1, max=100, mean=50, std_dev=10)
+        rng = np.random.RandomState(42)
+        values = sample_distribution(config, 1000, rng)
+        assert all(1 <= v <= 100 for v in values)
+
+    def test_lognormal_within_bounds(self) -> None:
+        config = DistributionConfig(type="lognormal", min=3, max=50, mean=4, std_dev=3)
+        rng = np.random.RandomState(42)
+        values = sample_distribution(config, 1000, rng)
+        assert all(3 <= v <= 50 for v in values)
+
+    def test_lognormal_right_skewed(self) -> None:
+        config = DistributionConfig(type="lognormal", min=3, max=50, mean=4, std_dev=3)
+        rng = np.random.RandomState(42)
+        values = sample_distribution(config, 10000, rng)
+        median = float(np.median(values))
+        mean = float(np.mean(values))
+        # Right-skewed: mean > median
+        assert mean >= median, f"Expected right-skew: mean={mean} should be >= median={median}"
+
+    def test_deterministic(self) -> None:
+        config = DistributionConfig(type="normal", min=1, max=100, mean=50, std_dev=10)
+        values1 = sample_distribution(config, 100, np.random.RandomState(42))
+        values2 = sample_distribution(config, 100, np.random.RandomState(42))
+        assert list(values1) == list(values2)
+
+    def test_unknown_type_raises(self) -> None:
+        config = DistributionConfig(type="unknown", min=1, max=10, mean=5, std_dev=1)
+        with pytest.raises(ValueError, match="Unknown distribution type"):
+            sample_distribution(config, 10, np.random.RandomState(42))
+
+
+class TestGenerateRandomText:
+    def test_nonempty(self) -> None:
+        rng = np.random.RandomState(42)
+        text = generate_random_text(rng, 100)
+        assert len(text) > 0
+
[…]
