### 2026-04-09 5ecf249
Add tests for streaming response fix
Verifies that process_response() is called before response body is
consumed, so the SSE streaming parser can iterate response.content.
