+++ pyproject.toml
@@ -38,6 +38,7 @@ dev = [
     "typing-extensions>=4.15.0",
     "hypothesis>=6.146.0",
     "pytest-asyncio>=1.3.0",
+    "pytest-picked",
 ]
 otel = [
     "opentelemetry-api>=1.30.0",
@@ -79,9 +80,12 @@ lint = "ruff check"
 type-check = "mypy --strict ./inference_perf ./tests"
 "check:cli-flags" = "python scripts/sync_cli_flags_doc.py --check"
 "update:cli-flags" = "python scripts/sync_cli_flags_doc.py"
+check = {composite = ["format", "lint"]}
 validate = {composite = ["format", "lint", "type-check", "check:cli-flags"]}
+setup = {shell = "pdm install && pre-commit install"}
 
 "test" = "pytest tests"
+"test:picked" = "pytest --picked"
 "test:e2e" = "pytest e2e"
 "test:e2e:docker" = "pdm run docker:e2e-test:run"
 
