+++ .pre-commit-config.yaml
@@ -4,14 +4,7 @@ repos:
   hooks:
     - id: pdm-lock-check
       name: check lock file matches pyproject
-- repo: https://github.com/astral-sh/ruff-pre-commit
-  rev: v0.9.4
-  hooks:
-    - id: ruff
-      name: run the linter
-      args: [ --fix ]
-    - id: ruff-format
-      name: run the formatter
+
 - repo: https://github.com/pre-commit/mirrors-mypy
   rev: v1.14.1
   hooks:
@@ -25,4 +18,15 @@ repos:
       entry: python scripts/check_license.py
       language: python
       files: \.py$
-
+    - id: ruff
+      name: ruff
+      entry: pdm run ruff check --fix
+      language: system
+      types: [python]
+      require_serial: true
+    - id: ruff-format
+      name: ruff-format
+      entry: pdm run ruff format
+      language: system
+      types: [python]
+      require_serial: true
