### 2026-04-09 1e1ef81
Unified single-config: inline synthetic trace generation
Add generate_synthetic field to OTelTraceReplayConfig that generates
traces at init time from distribution parameters. No separate trace
generation step or pre-built traces needed — everything in one YAML.

- Add SyntheticTraceConfig and SyntheticTraceDistribution to config.py
- Add inference_perf/datagen/synthetic_trace_generator.py module
- Integrate into OTelTraceReplayDataGenerator.__init__
- Remove build-time trace generation from Dockerfile
- Update example config to unified single-config format
