### 2026-04-09 b24dbf2
Support goodput reporting based on latency SLOs
