+++ tests/test_config.py
@@ -221,6 +221,7 @@ def test_prometheus_client_config_validation() -> None:
     with pytest.raises(ValueError, match="Exactly one of 'url' or 'google_managed' must be set"):
         PrometheusClientConfig(google_managed=False)
 
+<<<<<<< HEAD
 def test_shared_prefix_inline_distribution() -> None:
     config = Config.model_validate(
         {
@@ -334,3 +335,31 @@ def test_distribution_variance_conversion() -> None:
 def test_distribution_both_variance_and_std_dev_error() -> None:
     with pytest.raises(Exception, match="Specify either"):
         Distribution(type=DistributionType.NORMAL, mean=100.0, std_dev=10.0, variance=100.0)
+=======
+
+def test_goodput_config_parsing() -> None:
+    config_content = {
+        "report": {
+            "goodput": {
+                "constraints": {
+                    "ttft": 0.5,
+                    "tpot": 0.1,
+                },
+                "percentile": "p90",
+            }
+        }
+    }
+    with tempfile.NamedTemporaryFile(mode="w", suffix=".yaml", delete=False) as tmp:
+        yaml.dump(config_content, tmp)
+        tmp_path = tmp.name
+
+    try:
+        config = read_config(tmp_path)
+        assert config.report.goodput is not None
+        assert config.report.goodput.constraints["ttft"] == 0.5
+        assert config.report.goodput.constraints["tpot"] == 0.1
+        assert config.report.goodput.percentile == "p90"
+    finally:
+        if os.path.exists(tmp_path):
+            os.remove(tmp_path)
+>>>>>>> cfaf911 (Support goodput reporting based on latency SLOs)
