+++ inference_perf/reportgen/base.py
@@ -27,7 +27,7 @@ from inference_perf.client.metricsclient import MetricsClient, PerfRuntimeParame
 from inference_perf.client.metricsclient.base import ModelServerMetrics
 from inference_perf.client.metricsclient.prometheus_client import PrometheusMetricsClient
 from inference_perf.client.requestdatacollector import RequestDataCollector
-from inference_perf.config import Config, PrometheusMetricsReportConfig, ReportConfig, SessionLifecycleReportConfig
+from inference_perf.config import Config, PrometheusMetricsReportConfig, ReportConfig, SessionLifecycleReportConfig, GoodputConfig
 from inference_perf.metrics import SessionMetricsCollector
 from inference_perf.utils import ReportFile
 
@@ -175,6 +175,72 @@ def calculate_slo_metrics(
     return slo_metrics
 
 
+def calculate_goodput_metrics(
+    metrics: List[RequestLifecycleMetric],
+    goodput_config: Optional[GoodputConfig],
+    ttft_values: List[Optional[float]],
+    tpot_values: List[Optional[float]],
+    ntpot_values: List[float],
+    request_latency_values: List[float],
+    itl_values: List[Optional[float]],
+) -> Optional[dict[str, Any]]:
+    if not goodput_config or not goodput_config.constraints:
+        return None
+
+    total = len(metrics)
+    if total == 0:
+        return None
+
+    total_benchmark_time = max(m.end_time for m in metrics) - min(m.start_time for m in metrics)
+    
+    good_requests_count = 0
+    good_output_tokens = 0
+
+    for i, m in enumerate(metrics):
+        is_good = True
+        for metric_name, slo_value in goodput_config.constraints.items():
+            if metric_name == "ttft":
+                val = ttft_values[i]
+                if val is None or val > slo_value:
+                    is_good = False
+                    break
+            elif metric_name == "tpot":
+                val = tpot_values[i]
+                if val is None or val > slo_value:
+                    is_good = False
+                    break
+            elif metric_name == "itl":
+                val = itl_values[i]
+                if val is None or val > slo_value:
+                    is_good = False
+                    break
+            elif metric_name == "ntpot":
+                val = ntpot_values[i]
+                if val > slo_value:
+                    is_good = False
+                    break
+            elif metric_name == "request_latency":
+                val = request_latency_values[i]
+                if val > slo_value:
+                    is_good = False
+                    break
+        
+        if is_good:
+            good_requests_count += 1
+            if m.info.output_tokens is not None:
+                good_output_tokens += m.info.output_tokens
+
+    goodput_pct = (good_requests_count / total * 100) if total > 0 else 0.0
+    request_goodput_rate = good_requests_count / total_benchmark_time if total_benchmark_time > 0 else 0.0
+    token_goodput_rate = good_output_tokens / total_benchmark_time if total_benchmark_time > 0 else 0.0
+
+    return {
+        "goodput_pct": goodput_pct,
+        "request_goodput_rate": request_goodput_rate,
+        "token_goodput_rate": token_goodput_rate,
+        "good_requests": good_requests_count,
+        "total_requests": total,
+    }
 def summarize_prometheus_metrics(metrics: ModelServerMetrics) -> ResponsesSummary:
     return ResponsesSummary(
         load_summary={},  # model server doesn't report failed requests
@@ -340,6 +406,7 @@ def summarize_requests(
     percentiles: List[float],
     stage_rate: Optional[float] = None,
     stage_concurrency: Optional[int] = None,
+    goodput_config: Optional[GoodputConfig] = None,
 ) -> ResponsesSummary:
     all_successful: List[RequestLifecycleMetric] = [x for x in metrics if x.error is None]
[…]

[…]
