### 2026-04-09 a2f6d0e
feat: conversation_replay data generator for agentic workloads
Adds a new ConversationReplayDataGenerator for benchmarking agentic
tool-calling workloads where N conversations run concurrently.

Generates synthetic multi-turn conversations in-memory from configurable
distributions (normal, lognormal, uniform, fixed). Uses the existing
CONCURRENT load type for concurrency control and LocalUserSession for
sequential turn management. No file I/O, no DAG parsing — loads in
seconds regardless of conversation length.

Key additions:
- ConversationReplayDataGenerator with deterministic seeded RNG
- C
[…]
