+++ inference_perf/datagen/conversation_replay_datagen.py
@@ -1,0 +1,257 @@
+# Copyright 2026 The Kubernetes Authors.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+"""Conversation replay data generator for agentic workload benchmarking.
+
+Generates synthetic multi-turn conversations in-memory from configurable
+distributions (normal, lognormal, uniform, fixed). Uses the existing
+LocalUserSession / UserSessionCompletionAPIData / CONCURRENT load type
+infrastructure — no file I/O, no DAG parsing.
+
+Designed for benchmarking agentic tool-calling workloads where N
+conversations run concurrently as fast as possible with production-calibrated
+distributions.
+"""
+
+import logging
+from dataclasses import dataclass, field
+from typing import Generator, List, Optional
+
+import numpy as np
+
+from inference_perf.apis.base import InferenceAPIData, LazyLoadInferenceAPIData
+from inference_perf.apis.user_session import LocalUserSession, UserSessionCompletionAPIData
+from inference_perf.config import (
+    APIConfig,
+    APIType,
+    ConversationReplayConfig,
+    ConversationReplayDistribution,
+    DataConfig,
+)
+from inference_perf.utils.custom_tokenizer import CustomTokenizer
+from inference_perf.utils.distribution import generate_distribution
+
+from .base import DataGenerator, LazyLoadDataMixin
+
+logger = logging.getLogger(__name__)
+
+
+@dataclass
+class ConversationBlueprint:
+    """Pre-computed plan for a single conversation."""
+
+    conversation_id: int
+    num_turns: int
+    system_prompt: str
+    turn_prompts: List[str] = field(default_factory=list)
+    turn_output_lens: List[int] = field(default_factory=list)
+
+
+class ConversationReplayDataGenerator(DataGenerator, LazyLoadDataMixin):
+    """Generates synthetic multi-turn conversations from distribution configs.
+
+    Each conversation has:
+    - A two-part system prompt (shared prefix + dynamic per-conversation suffix)
+    - N turns with independently sampled input/output token lengths
+    - Sequential turn enforcement via LocalUserSession
+
+    Conversations are dispatched round-robin across workers using
+    preferred_worker_id for affinity. When all turns of a conversation are
+    exhausted, the session resets and replays from the beginning (recycling).
+    """
+
+    def __init__(
+        self,
+        api_config: APIConfig,
+        config: DataConfig,
+        tokenizer: Optional[CustomTokenizer],
+    ) -> None:
+        super().__init__(api_config, config, tokenizer)
+
+        if self.tokenizer is None:
+            raise ValueError("Tokenizer is required for ConversationReplayDataGenerator.")
+
+        cr_config = config.conversation_replay
+        if cr_config is None:
+            raise ValueError("conversation_replay config is required.")
+        self.cr_config: ConversationReplayConfig = cr_config
+
+        # Tokenizer vocab size for random token generation
+        hf_tokenizer = self.tokenizer.get_tokenizer()
+        if hasattr(hf_tokenizer, "vocab_size") and hf_tokenizer.vocab_size is not None:
+            self.vocab_size: int = hf_tokenizer.vocab_size
+        elif hasattr(hf_tokenizer, "get_vocab") and callable(hf_tokenizer.get_vocab):
+            self.vocab_size = len(hf_tokenizer.get_vocab())
+        else:
[…]
