### 2026-04-10 9de7243
fix: disable stop token IDs in conversation replay for accurate output token measurement
Random synthetic prompts cause chat models (e.g. Qwen3) to emit the chat
stop token (<|im_end|>) before max_tokens, severely undercounting output vs
production. Adding stop_token_ids=[] to the request payload ensures generation
runs to max_tokens, matching the sampled output_tokens_per_turn distribution.
