### 2026-04-10 a463868
Add goodput and reports documentation, update README, and clean up goodput config
