+++ inference_perf/config.py
@@ -465,7 +465,6 @@ class SessionLifecycleReportConfig(BaseModel):
 
 class GoodputConfig(BaseModel):
     constraints: Dict[str, float] = {}
-    percentile: Optional[str] = None
 
 
 class ReportConfig(BaseModel):
