+++ inference_perf/utils/cli_summary.py
@@ -47,33 +47,62 @@ def print_summary_table(reports: List[ReportFile]) -> None:
     has_goodput = any("goodput_metrics" in r.get("successes", {}) for r in stage_reports.values())
 
     console = Console()
-    table = Table(
-        title="[bold magenta]Inference Performance Summary[/bold magenta]", show_header=True, header_style="bold cyan"
+
+    # Table 1: Stage & Throughput Summary
+    summary_table = Table(
+        title="[bold magenta]Throughput and Goodput Summary[/bold magenta]", show_header=True, header_style="bold cyan"
     )
-
-    table.add_column("Stage", justify="right")
-    table.add_column("Req Rate", justify="right")
-    table.add_column("Achieved Rate", justify="right")
-    table.add_column("Error Rate", justify="right")
-
-    # TTFT Columns
-    table.add_column("TTFT Mean\n(ms)", justify="right")
-    table.add_column("TTFT Med\n(ms)", justify="right")
-    table.add_column("TTFT P90\n(ms)", justify="right")
-
-    # ITL Columns
-    table.add_column("ITL Mean\n(ms)", justify="right")
-    table.add_column("ITL Med\n(ms)", justify="right")
-    table.add_column("ITL P90\n(ms)", justify="right")
-
-    # Throughput Columns
-    table.add_column("Req/s", justify="right")
-    table.add_column("In Tokens/s", justify="right")
-    table.add_column("Out Tokens/s", justify="right")
-    table.add_column("Tot Tokens/s", justify="right")
-
+    summary_table.add_column("Stage", justify="right")
+    summary_table.add_column("Req Rate", justify="right")
+    summary_table.add_column("Achieved Rate", justify="right")
+    summary_table.add_column("Error Rate", justify="right")
+    summary_table.add_column("Req/s", justify="right")
+    summary_table.add_column("In Tokens/s", justify="right")
+    summary_table.add_column("Out Tokens/s", justify="right")
+    summary_table.add_column("Tot Tokens/s", justify="right")
     if has_goodput:
-        table.add_column("Goodput %", justify="right")
+        summary_table.add_column("Goodput %", justify="right")
+        summary_table.add_column("Req Goodput", justify="right")
+        summary_table.add_column("Token Goodput", justify="right")
+
+    # Table 2: Request & Token Latency (ms)
+    latency_table = Table(
+        title="[bold magenta]Request & Token Latency (ms)[/bold magenta]", show_header=True, header_style="bold cyan"
+    )
+    latency_table.add_column("Stage", justify="right")
+    latency_table.add_column("Req Lat Mean", justify="right")
+    latency_table.add_column("Req Lat Med", justify="right")
+    latency_table.add_column("Req Lat P90", justify="right")
+    latency_table.add_column("TTFT Mean", justify="right")
+    latency_table.add_column("TTFT Med", justify="right")
+    latency_table.add_column("TTFT P90", justify="right")
+    latency_table.add_column("ITL Mean", justify="right")
+    latency_table.add_column("ITL Med", justify="right")
+    latency_table.add_column("ITL P90", justify="right")
+
+    # Table 3: Token Generation Speed (ms)
+    speed_table = Table(
+        title="[bold magenta]Token Generation Speed (ms)[/bold magenta]", show_header=True, header_style="bold cyan"
+    )
+    speed_table.add_column("Stage", justify="right")
+    speed_table.add_column("TPOT Mean", justify="right")
+    speed_table.add_column("TPOT Med", justify="right")
[…]
