### 2026-04-10 336f57b
Fix tests and type checks
