### 2026-04-10 5428b21
Remove goodput percentile and fix associated tests
