### 2026-04-10 0ade519
Fix mypy error in google_managed_prometheus_client.py
