### 2026-04-10 010c60f
feat: closed-loop session replenishment in conversation replay
Each slot (preferred_worker_id) now resets its LocalUserSession when a
conversation completes all its turns, immediately starting a fresh
conversation from turn 0.

This enables the correct steady-state benchmark model:
- num_conversations = C (one slot per concurrent worker)
- At steady state, C slots are uniformly distributed across turns 0..N
- Mean active context ≈ (start_context + end_context) / 2 = production mean
- No drain problem: slots never go idle between conversations
- Each concurrency level is run as a separate fresh
[…]
