### 2026-04-10 afcff5c
feat: add benchmark_time_seconds to metrics report
