+++ inference_perf/reportgen/base.py
@@ -57,6 +57,7 @@ def summarize(items: List[float], percentiles: List[float]) -> Optional[dict[str
 
 
 class ResponsesSummary(BaseModel):
+    benchmark_time_seconds: float
     load_summary: dict[str, Any]
     successes: dict[str, Any]
     failures: dict[str, Any]
@@ -177,6 +178,7 @@ def calculate_slo_metrics(
 
 def summarize_prometheus_metrics(metrics: ModelServerMetrics) -> ResponsesSummary:
     return ResponsesSummary(
+        benchmark_time_seconds=0.0,
         load_summary={},  # model server doesn't report failed requests
         failures={},
         successes={
@@ -445,6 +447,7 @@ def summarize_requests(
     if slo_metrics:
         successes_dict["slo_metrics"] = slo_metrics
     return ResponsesSummary(
+        benchmark_time_seconds=total_time,
         load_summary=load_summary,
         successes=successes_dict,
         failures={
