### 2026-04-10 4956258
fix: handle streaming errors in 200 responses and add context safety reset
openai_client.py: When process_response() raises inside async with response:
(e.g. ClientPayloadError from broken SSE), call process_failure() directly
instead of re-raising. Re-raising ran ClientResponse.__aexit__ on a broken
connection which could raise a second exception masking the original and
bypassing the outer aiohttp.ClientError handler, crashing the benchmark.

conversation_replay_datagen.py: Add safety context reset when accumulated
context exceeds 900K chars (~225K tokens = max_model_len). Rare outlier
turns
[…]
