+++ inference_perf/client/modelserver/openai_client.py
@@ -332,10 +332,30 @@ class openAIModelServerClientSession(ModelServerClientSession):
                                     lora_adapter=lora_adapter,
                                 )
                     except Exception as read_error:
-                        # Handle errors reading response body
-                        if not response_content:
-                            response_content = f"Failed to read response text: {read_error}"
-                        raise
+                        # Handle errors reading response body or streaming.
+                        # For 200 responses, process_response() raised (e.g. ClientPayloadError
+                        # from a broken SSE stream). Call process_failure() here so that session
+                        # locks are released before the context manager exits. Re-raising would
+                        # run ClientResponse.__aexit__ on a broken connection, which can raise
+                        # a second exception that masks the original and bypasses the outer
+                        # aiohttp.ClientError handler.
+                        if response is not None and response.status == 200 and not response_info:
+                            caught_exception = read_error
+                            error = ErrorResponseInfo(
+                                error_msg=str(read_error),
+                                error_type=type(read_error).__name__,
+                            )
+                            response_info = await data.process_failure(
+                                response=None,
+                                config=self.client.api_config,
+                                tokenizer=self.client.tokenizer,
+                                exception=read_error,
+                                lora_adapter=lora_adapter,
+                            )
+                        else:
+                            if not response_content:
+                                response_content = f"Failed to read response text: {read_error}"
+                            raise
 
             except aiohttp.ClientError as e:
                 caught_exception = e
