+++ inference_perf/datagen/conversation_replay_datagen.py
@@ -160,6 +160,16 @@ class ConversationReplayDataGenerator(DataGenerator, LazyLoadDataMixin):
                 context=bp.system_prompt,
             )
             logger.debug("Slot %d starting conversation %d", conv_idx, convo_num)
+        elif len(self.user_sessions[conv_idx].contexts) > 900_000:
+            # Safety reset: context is approaching max_model_len (225K tokens ≈ 900K
+            # chars) due to outlier turns with very large inputs or outputs. Reset to
+            # fresh context mid-conversation rather than letting vLLM reject the request.
+            self.user_sessions[conv_idx] = LocalUserSession(
+                user_session_id=f"slot_{conv_idx}_convo_{convo_num}_reset",
+                context=bp.system_prompt,
+            )
+            turn_idx = 0
+            logger.warning("Slot %d: safety context reset at turn %d", conv_idx, turn_idx)
 
         return UserSessionCompletionAPIData(
             prompt=bp.turn_prompts[turn_idx],
