### 2026-04-10 d2983ed
fix: raise safety context reset threshold to 2.7M chars
Random Qwen3 tokens decode to ~12 chars/token on average (not 4 as
initially assumed). This means 225K token max_model_len corresponds
to ~2.7M chars, not 900K. The aggressive 900K threshold was causing
safety resets every 2-3 turns for slots that hit p90 inputs.
