### 2026-04-12 kubernetes-sigs/inference-perf#395 @achandrasekar
Closing this since it is addressed in the paper.
