### 2026-04-12 kubernetes-sigs/inference-perf#392 @achandrasekar
There are some improvements happening around this in #410 that should address any non-content tokens as well. @Bslabe123  FYI.
