### 2026-04-13 a00b541
Unify with the existing SLO mechanism
