+++ tests/reportgen/test_calculate_goodput_metrics.py
@@ -27,6 +27,8 @@ def test_calculate_goodput_metrics_empty_metrics() -> None:
 def test_calculate_goodput_metrics_no_config() -> None:
     """Test with no goodput config returns None."""
     metric = Mock(spec=RequestLifecycleMetric)
+    metric.ttft_slo_sec = None
+    metric.tpot_slo_sec = None
     result = calculate_goodput_metrics([metric], None, [0.3], [None], [0.1], [1.0], [None])
     assert result is None
 
@@ -35,6 +37,8 @@ def test_calculate_goodput_metrics_no_constraints() -> None:
     """Test with empty constraints returns None."""
     config = GoodputConfig(constraints={})
     metric = Mock(spec=RequestLifecycleMetric)
+    metric.ttft_slo_sec = None
+    metric.tpot_slo_sec = None
     result = calculate_goodput_metrics([metric], config, [0.3], [None], [0.1], [1.0], [None])
     assert result is None
 
@@ -49,6 +53,8 @@ def test_calculate_goodput_metrics_ttft_constraint() -> None:
     metric1.info = Mock()
     metric1.info.input_tokens = 0
     metric1.info.output_tokens = 10
+    metric1.ttft_slo_sec = None
+    metric1.tpot_slo_sec = None
 
     metric2 = Mock(spec=RequestLifecycleMetric)
     metric2.start_time = 2.0
@@ -56,6 +62,8 @@ def test_calculate_goodput_metrics_ttft_constraint() -> None:
     metric2.info = Mock()
     metric2.info.input_tokens = 0
     metric2.info.output_tokens = 20
+    metric2.ttft_slo_sec = None
+    metric2.tpot_slo_sec = None
 
     metrics = [metric1, metric2]
     ttft_values: list[float | None] = [0.3, 0.6]  # metric1 meets, metric2 fails
@@ -82,6 +90,8 @@ def test_calculate_goodput_metrics_total_tokens() -> None:
     metric1.info = Mock()
     metric1.info.input_tokens = 5
     metric1.info.output_tokens = 10
+    metric1.ttft_slo_sec = None
+    metric1.tpot_slo_sec = None
 
     metrics = [metric1]
     ttft_values: list[float | None] = [0.3]  # Meets constraint
@@ -106,6 +116,8 @@ def test_calculate_goodput_metrics_multiple_constraints() -> None:
     metric1.info = Mock()
     metric1.info.input_tokens = 0
     metric1.info.output_tokens = 10
+    metric1.ttft_slo_sec = None
+    metric1.tpot_slo_sec = None
 
     metric2 = Mock(spec=RequestLifecycleMetric)
     metric2.start_time = 2.0
@@ -113,6 +125,8 @@ def test_calculate_goodput_metrics_multiple_constraints() -> None:
     metric2.info = Mock()
     metric2.info.input_tokens = 0
     metric2.info.output_tokens = 20
+    metric2.ttft_slo_sec = None
+    metric2.tpot_slo_sec = None
 
     metrics = [metric1, metric2]
     ttft_values: list[float | None] = [0.3, 0.3]  # Both meet TTFT
@@ -135,6 +149,8 @@ def test_calculate_goodput_metrics_itl_constraint() -> None:
     metric1.info = Mock()
     metric1.info.input_tokens = 0
     metric1.info.output_tokens = 10
+    metric1.ttft_slo_sec = None
+    metric1.tpot_slo_sec = None
 
     metrics = [metric1]
     itl_values: list[float | None] = [0.04]  # Meets ITL
@@ -155,6 +171,8 @@ def test_calculate_goodput_metrics_ntpot_constraint() -> None:
     metric1.info = Mock()
     metric1.info.input_tokens = 0
[…]
