+++ inference_perf/reportgen/base.py
@@ -68,119 +68,6 @@ class ResponsesSummary(BaseModel):
     failures: dict[str, Any]
 
 
-def calculate_slo_metrics(
-    metrics: List[RequestLifecycleMetric], ttft_values: List[Optional[float]], tpot_values: List[Optional[float]]
-) -> Optional[dict[str, Any]]:
-    """
-    Calculate SLO attainment statistics from request metrics and pre-calculated latencies.
-
-    Args:
-        metrics: List of successful RequestLifecycleMetric objects.
-        ttft_values: List of calculated TTFT values corresponding to metrics (same index). None if not streamable.
-        tpot_values: List of calculated TPOT values corresponding to metrics (same index). None if not streamable.
-
-    Returns None if no SLO thresholds were set, otherwise returns dict with SLO metrics.
-    """
-    if not metrics:
-        return None
-
-    # Check if any metrics have SLO fields populated
-    has_ttft_slo = any(m.ttft_slo_sec is not None for m in metrics)
-    has_tpot_slo = any(m.tpot_slo_sec is not None for m in metrics)
-
-    ttft_slo_limit = max([m.ttft_slo_sec for m in metrics if m.ttft_slo_sec is not None], default=None)
-    tpot_slo_limit = max([m.tpot_slo_sec for m in metrics if m.tpot_slo_sec is not None], default=None)
-
-    if not has_ttft_slo and not has_tpot_slo:
-        return None  # No SLO tracking enabled
-
-    slo_metrics: dict[str, Any] = {}
-    total = len(metrics)
-
-    total_benchmark_time = max(m.end_time for m in metrics) - min(m.start_time for m in metrics)
-    goodput_token_counts = []
-
-    ttft_results = []
-    tpot_results = []
-
-    # Iterate through metrics and their corresponding calculated values.
-    # Using zip(..., strict=True) ensures that all sequences have the same length
-    # and prevents misalignment or IndexError if they do not.
-    for m, ttft, tpot in zip(metrics, ttft_values, tpot_values, strict=True):
-        # Check TTFT SLO (Only if streamable / ttft exists)
-        ttft_met = None
-        if m.ttft_slo_sec is not None:
-            if ttft is not None:
-                ttft_met = ttft <= m.ttft_slo_sec
-            else:
-                ttft_met = True  # If TTFT SLO is set but we don't have a TTFT value (non-streamable), we can choose to consider it as met or not. Here we consider it met to avoid penalizing non-streamable requests.
-        ttft_results.append(ttft_met)
-
-        # Check TPOT SLO (Only if streamable / tpot exists)
-        tpot_met = None
-        if m.tpot_slo_sec is not None:
-            if tpot is not None:
-                tpot_met = tpot <= m.tpot_slo_sec
-            else:
-                tpot_met = True  # Similar logic as TTFT for non-streamable requests
-        tpot_results.append(tpot_met)
-
-        # Goodput Calculation Logic
-        ttft_ok = (ttft_met is True) if has_ttft_slo else True
-        tpot_ok = (tpot_met is True) if has_tpot_slo else True
-
-        if ttft_ok and tpot_ok:
-            if m.info.input_tokens is not None and m.info.output_tokens is not None:
-                goodput_token_counts.append(m.info.input_tokens + m.info.output_tokens)
-
-    total_goodput_tokens = sum(goodput_token_counts)
-    goodput_rate = total_goodput_tokens / total_benchmark_time if total_benchmark_time > 0 else 0.0
-
-    # TTFT SLO metrics
-    if has_ttft_slo:
-        ttft_met_count = sum(1 for res in ttft_results if res is True)
[…]
