### 2026-04-13 892bb9a
Add workload catalog
