### 2026-04-13 ccd696e
Extract common graph-backed session replay runtime into ReplayGraphSessionGeneratorBase
* Move shared scheduling, predecessor waiting, output substitution, worker/session tracking, and completion handling into the shared base class
* Make OTelTraceReplayDataGenerator a concrete implementation that reads OTel traces and builds replay sessions from them
* Rename shared replay types to session-oriented names instead of OTel-specific names
* Update OpenAI client failure handling to work on the shared session replay request type
