+++ inference_perf/datagen/otel_trace_replay_datagen.py
@@ -62,45 +62,30 @@ import json
 import logging
 import random
 import time
-from dataclasses import dataclass, field, replace as dc_replace
+from dataclasses import dataclass, field
 from multiprocessing.managers import SyncManager
 from pathlib import Path
 from typing import Any, Dict, List, Optional, Set
 
 from aiohttp import ClientResponse
 
-from inference_perf.apis import (
-    ChatCompletionAPIData,
-    InferenceAPIData,
-    InferenceInfo,
-    LazyLoadInferenceAPIData,
-    SessionLifecycleMetric,
-)
-from inference_perf.apis.chat import ChatMessage
+from inference_perf.apis.chat import ChatCompletionAPIData, ChatMessage
 from inference_perf.apis.streaming_parser import parse_sse_stream
-from inference_perf.config import APIConfig, APIType, DataConfig
-from inference_perf.datagen.base import SessionGenerator, LazyLoadDataMixin
+from inference_perf.apis.user_session import InferenceInfo
+from inference_perf.config import APIConfig, DataConfig
+from inference_perf.datagen.replay_graph_session_datagen import (
+    EventFailedError,
+    ReplaySession,
+    ReplayGraphSessionGeneratorBase,
+)
+from inference_perf.datagen.replay_graph_types import InputSegment, ReplayGraph
 from inference_perf.datagen.otel_trace_to_replay_graph import (
-    InputSegment,
-    ReplayGraph,
     build_raw_calls,
     build_graph,
 )
 from inference_perf.utils.custom_tokenizer import CustomTokenizer
 
 logger = logging.getLogger(__name__)
-
-# ---------------------------------------------------------------------------
-# EventFailedError — raised by require_async when a predecessor event failed
-# ---------------------------------------------------------------------------
-
-
-class EventFailedError(Exception):
-    """Raised by EventOutputRegistry.require_async when the awaited event failed."""
-
-    def __init__(self, event_id: str) -> None:
-        super().__init__(f"Predecessor event {event_id!r} failed")
-        self.event_id = event_id
 
 
 # ---------------------------------------------------------------------------
@@ -825,7 +810,7 @@ def resolve_trace_files(trace_files: List[str]) -> List[Path]:
     return [Path(f) for f in all_files]
 
 
-class OTelTraceReplayDataGenerator(SessionGenerator, LazyLoadDataMixin):
+class OTelTraceReplayDataGenerator(ReplayGraphSessionGeneratorBase):
     """
     Data generator that replays LLM requests from OpenTelemetry trace files.
 
@@ -849,7 +834,7 @@ class OTelTraceReplayDataGenerator(SessionGenerator, LazyLoadDataMixin):
         base_seed: Optional[int] = None,
         num_workers: int = 1,
     ) -> None:
-        super().__init__(api_config, config, tokenizer)
+        super().__init__(api_config, config, tokenizer, mp_manager=mp_manager, base_seed=base_seed, num_workers=num_workers)
 
         if not hasattr(config, "otel_trace_replay") or config.otel_trace_replay is None:
             raise ValueError("otel_trace_replay configuration is required for OTelTraceReplayDataGenerator")
@@ -889,45 +874,19 @@ class OTelTraceReplayDataGenerator(SessionGenerator, LazyLoadDataMixin):
         else:
             raise ValueError("Either trace_directory or trace_files must be provided in otel_trace_replay config")
 
-        # Intra-worker output registry: plain dicts + asyncio.Event, no IPC.
-        self.output_registry = EventOutputRegistry()
+        sessions = self._load_trace_files()
+        self.initialize_sessions(sessions)
 
-        # Intra-worker session tracker: tracks event completions and session failures within worker.
-        # Each worker has its own instance due to session-to-worker affinity.
[…]
