+++ inference_perf/datagen/replay_graph_session_datagen.py
@@ -1,0 +1,767 @@
+# Copyright 2026 The Kubernetes Authors.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""Shared graph-backed SessionGenerator runtime.
+
+This module contains the session replay runtime that is agnostic to how a
+ReplayGraph was produced. Concrete generators are responsible for producing
+ReplaySession objects; this base class handles session lifecycle, worker
+affinity, lazy request materialization, and session completion tracking.
+"""
+
+from __future__ import annotations
+
+import asyncio
+import json
+import logging
+import time
+from dataclasses import dataclass, field, replace as dc_replace
+from multiprocessing.managers import SyncManager
+from typing import Any, Dict, List, Optional, Set
+
+from aiohttp import ClientResponse
+
+from inference_perf.apis import ChatCompletionAPIData, InferenceInfo, LazyLoadInferenceAPIData, SessionLifecycleMetric
+from inference_perf.apis.chat import ChatMessage
+from inference_perf.apis.streaming_parser import parse_sse_stream
+from inference_perf.config import APIConfig, APIType, DataConfig
+from inference_perf.datagen.base import LazyLoadDataMixin, SessionGenerator
+from inference_perf.datagen.replay_graph_types import InputSegment, ReplayGraph
+from inference_perf.utils.custom_tokenizer import CustomTokenizer
+
+logger = logging.getLogger(__name__)
+
+
+class EventFailedError(Exception):
+    """Raised by EventOutputRegistry.require_async when the awaited event failed."""
+
+    def __init__(self, event_id: str) -> None:
+        super().__init__(f"Predecessor event {event_id!r} failed")
+        self.event_id = event_id
+
+
+class SessionInferenceInfo(InferenceInfo):
+    """InferenceInfo subclass that also carries the raw output text."""
+
+    output_text: Optional[str] = None
+
+
+class WorkerSessionTracker:
+    """Per-worker tracking of event completions and session failures."""
+
+    def __init__(self) -> None:
+        self._event_completions: Dict[str, Dict[str, float]] = {}
+        self._failed_sessions: Set[str] = set()
+
+    def record_event_completed(self, session_id: str, event_id: str, completion_time: float) -> None:
+        if session_id not in self._event_completions:
+            self._event_completions[session_id] = {}
+        self._event_completions[session_id][event_id] = completion_time
+
+    def is_event_completed(self, session_id: str, event_id: str) -> bool:
+        return session_id in self._event_completions and event_id in self._event_completions[session_id]
+
+    def get_event_completion_time(self, session_id: str, event_id: str) -> Optional[float]:
+        return self._event_completions.get(session_id, {}).get(event_id)
+
+    def mark_session_failed(self, session_id: str) -> None:
+        self._failed_sessions.add(session_id)
+
+    def is_session_failed(self, session_id: str) -> bool:
+        return session_id in self._failed_sessions
+
+    def get_session_event_count(self, session_id: str) -> int:
+        return len(self._event_completions.get(session_id, {}))
+
+    def get_session_completion_times(self, session_id: str) -> Dict[str, float]:
+        return self._event_completions.get(session_id, {}).copy()
+
+
+class EventOutputRegistry:
[…]
