### 2026-04-13 564c05b
feat: tool_call_latency_sec distribution for agentic workload simulation
Adds per-turn tool execution latency to ConversationReplayDataGenerator.
When configured, each turn sleeps for the sampled duration after model
inference completes and before the next turn begins — simulating tool
call round-trips in agentic pipelines (e.g. 2-25s for API calls).

The sleep is injected while the session lock is held, so:
- The next turn of that conversation is blocked (correct serialisation)
- Other concurrent conversations run freely on the GPU (correct batching)
- At sufficient concurrency, GPU utilizati
[…]
