### 2026-04-13 1bb8ec5
Update documentation on the fields
