### 2026-04-13 51f956f
Fix conflict and update tests
