### 2026-04-13 4da5ced
chore: address PR review comments
- Remove workload-specific example from datagen docstring
- Enumerate supported distribution types explicitly in generate_distribution signature
- Raise ValueError for unknown dist_type instead of silently defaulting to normal
