### 2026-04-14 a91a5f5
improve runtime of build_graph. Add memoization + remove unneeded regex usage to normalize compared messages.
