+++ inference_perf/datagen/otel_trace_to_replay_graph.py
@@ -52,7 +52,6 @@ Falls back to len(text) // 4 (rough chars-per-token estimate) per message.
 import argparse
 import json
 import logging
-import re
 from enum import Enum
 from dataclasses import dataclass
 from datetime import datetime, timezone
@@ -94,11 +93,6 @@ def parse_iso(ts: str) -> float:
     return dt.timestamp()
 
 
-def norm_text(s: str) -> str:
-    """Normalize text by collapsing whitespace."""
-    return re.sub(r"\s+", " ", s or "").strip()
-
-
 def estimate_tokens(text: str) -> int:
     """Estimate token count from character length (rough: 4 chars per token)."""
     return max(1, len(text) // 4)
@@ -125,21 +119,20 @@ def message_tokens(msg: OtelMessage) -> int:
 
 def messages_equal(a: OtelMessage, b: OtelMessage) -> bool:
     """Return True if two messages have the same role and content."""
-    return a.role == b.role and norm_text(message_content_text(a)) == norm_text(message_content_text(b))
+    return a.role == b.role and message_content_text(a) == message_content_text(b)
 
 
 def output_matches_message(output_text: str, msg: OtelMessage, allow_partial_match: bool = False) -> bool:
     """Return True if msg is an assistant message whose content matches output_text."""
     if msg.role != "assistant":
         return False
-    msg_text = norm_text(message_content_text(msg))
-    out_text = norm_text(output_text)
-    if msg_text == out_text:
+    msg_text = message_content_text(msg)
+    if msg_text == output_text:
         return True
     if not allow_partial_match:
         return False
     else:  # try partial match
-        if out_text in msg_text:  # out_text is entirely contained in msg_text
+        if output_text in msg_text:  # output_text is entirely contained in msg_text
             return True
     return False
 
@@ -462,13 +455,19 @@ def _try_match_tool_call_ids(a_parts: List[Dict[str, Any]], b_messages: List[Dic
     return all(tc_id in b_tool_call_ids for tc_id in tool_call_ids)
 
 
-def get_causal_dep(a: RawCall, b: RawCall) -> Optional[DEPENDENCY_TYPE]:
+def get_causal_dep(
+    a: RawCall, b: RawCall, output_matches_for_substitutions: Optional[Dict[tuple[str, str], List[int]]] = None
+) -> Optional[DEPENDENCY_TYPE]:
     """Return the type of causal dependency if call B causally depends on call A, None otherwise.
 
     A call B depends on A if any assistant message in B's message list has content
     that matches A's output text (full content match, not a snippet).
     This means A's output was injected into B's prompt as a prior assistant turn.
     We start with trying to detect FULL_MATCH dependency. Then, if not detected, we proceed to TOOL_CALL_IDS_MATCHED. If this dependency is not detected either we proceed to the other options.
+
+    Args:
+        output_matches_for_substitutions: Optional cache storing message indices with EXACT output matches.
+                                         Key: (pred_call_id, curr_call_id), Value: list of matching indices
 
     Returns:
         FULL_MATCH: Full text of output A matches the full text in B
@@ -483,10 +482,15 @@ def get_causal_dep(a: RawCall, b: RawCall) -> Optional[DEPENDENCY_TYPE]:
     """
     if not a.out_message or not b.messages:
         return None
-    # match entire output message:
-    a_out = norm_text(a.out_message.text)
-    for msg in b.messages:
+    a_out = a.out_message.text or ""
+    for msg_idx, msg in enumerate(b.messages):
         if output_matches_message(a_out, msg, allow_partial_match=True):
+            # Cache only exact matches (for reuse in decompose_input)
[…]
