### 2026-04-14 99d4e0c
fix: Fix lint and formatting errors
