+++ inference_perf/distributed/orchestrator.py
@@ -2,13 +2,14 @@ import asyncio
 import logging
 import time
 import json
-from typing import Any, Dict, List
+from typing import Any
 from inference_perf.config import Config, LoadConfig, DataConfig, APIConfig
 from inference_perf.datagen.shared_prefix_datagen import SharedPrefixDataGenerator
 from inference_perf.distributed.redis_client import RedisClient
 from inference_perf.utils.custom_tokenizer import CustomTokenizer
 
 logger = logging.getLogger(__name__)
+
 
 class Orchestrator:
     def __init__(
@@ -17,7 +18,7 @@ class Orchestrator:
         load_config: LoadConfig = None,
         data_config: DataConfig = None,
         api_config: APIConfig = None,
-        tokenizer_config: Any = None
+        tokenizer_config: Any = None,
     ):
         self.redis = redis_client
         self.load_config = load_config
@@ -37,54 +38,57 @@ class Orchestrator:
     async def run_job_with_configs(self, load_config, data_config, api_config, tokenizer_config, job_id=None):
         # Initialize tokenizer
         tokenizer = CustomTokenizer(tokenizer_config)
-        
+
         # Initialize data generator
         datagen = SharedPrefixDataGenerator(api_config, data_config, tokenizer)
-        
+
         # Prime Redis with prompts
         await self.prime_redis(datagen)
-        
+
         # Set global start time
         global_start_time = time.time() + 5  # Give workers time to start
         await self.redis.redis.set("global_start_time", str(global_start_time))
         await self.redis.redis.set("completed_requests", "0")
         await self.redis.redis.delete(self.results_stream)  # Clear results from previous runs
         await self.redis.redis.delete("cancel_job")  # Clear cancellation flag
-        
+
         # Clear per-stage counters
         for i in range(len(load_config.stages)):
             await self.redis.redis.delete(f"completed_requests_stage_{i}")
-            
+
         # Generate and push tasks
         total_requests = await self.dispatch_tasks(datagen, load_config, global_start_time)
-        
+
         # Wait for completion
         await self.wait_for_completion(total_requests)
-        
+
         # Post-processing
         await self.post_process()
-        
+
         # Cleanup
         await self.cleanup()
-        
+
         if job_id:
             logger.info(f"Publishing completion for job {job_id}")
-            await self.redis.redis.publish("job_status", json.dumps({"job_id": job_id.decode('utf-8') if isinstance(job_id, bytes) else job_id, "status": "completed"}))
+            await self.redis.redis.publish(
+                "job_status",
+                json.dumps({"job_id": job_id.decode("utf-8") if isinstance(job_id, bytes) else job_id, "status": "completed"}),
+            )
 
     async def start_daemon(self):
         await self.redis.connect()
         logger.info("Orchestrator daemon started, waiting for jobs...")
-        
+
         # Start reading from the end of the stream
         last_id = "$"
-        
+
         while True:
             try:
                 # Read from job_stream
                 streams = await self.redis.redis.xread({"job_stream": last_id}, block=0, count=1)
-                
+
                 if streams:
-                    for stream_name, messages in streams:
+                    for _stream_name, messages in streams:
                         for message_id, message_data in messages:
                             if b"config" in message_data:
                                 config_json = message_data[b"config"].decode("utf-8")
@@ -96,16 +100,16 @@ class Orchestrator:
                                 logger.error(f"Message data missing config key: {message_data}")
                                 continue
                             logger.info(f"Received job with message_id: {message_id}")
-                            
+
                             # Parse config
                             cfg = Config.model_validate_json(config_json)
-                            
+
                             # Run the job
                             await self.run_job_with_configs(cfg.load, cfg.data, cfg.api, cfg.tokenizer, message_id)
[…]
