+++ inference_perf/reportgen/base.py
@@ -531,7 +531,9 @@ class ReportGenerator:
                 if metric.stage_id is not None:
                     stage_buckets[metric.stage_id].append(metric)
             for stage_id, metrics in stage_buckets.items():
-                logger.info(f"Processing stage_id={stage_id}, runtime_parameters.stages type={type(runtime_parameters.stages)}, length={len(runtime_parameters.stages) if isinstance(runtime_parameters.stages, list) else 'N/A'}")
+                logger.info(
+                    f"Processing stage_id={stage_id}, runtime_parameters.stages type={type(runtime_parameters.stages)}, length={len(runtime_parameters.stages) if isinstance(runtime_parameters.stages, list) else 'N/A'}"
+                )
                 try:
                     stage_rate = runtime_parameters.stages[stage_id].rate
                 except IndexError as e:
