+++ inference_perf/distributed/worker.py
@@ -1,14 +1,13 @@
 import asyncio
-import json
 import logging
 import time
-from typing import Optional
 from inference_perf.client.modelserver import ModelServerClient
 from inference_perf.distributed.redis_client import RedisClient
 from inference_perf.apis.completion import CompletionAPIData
 from inference_perf.client.requestdatacollector import LocalRequestDataCollector
 
 logger = logging.getLogger(__name__)
+
 
 class DistributedWorker:
     def __init__(
@@ -21,7 +20,7 @@ class DistributedWorker:
         results_stream: str,
         telemetry_channel: str,
         max_concurrency: int = 10,
-        publish_interval: int = 1  # Publish telemetry every N requests
+        publish_interval: int = 1,  # Publish telemetry every N requests
     ):
         self.worker_id = worker_id
         self.redis = redis_client
@@ -40,7 +39,7 @@ class DistributedWorker:
     async def run(self):
         await self.redis.connect()
         await self.redis.create_consumer_group(self.stream_name, self.group_name)
-        
+
         # Fetch global start time from Redis
         global_start_time_str = await self.redis.redis.get("global_start_time")
         if not global_start_time_str:
@@ -48,19 +47,17 @@ class DistributedWorker:
             while not global_start_time_str and not self.stop_event.is_set():
                 await asyncio.sleep(1)
                 global_start_time_str = await self.redis.redis.get("global_start_time")
-            
+
             if self.stop_event.is_set():
                 return
-                
+
         global_start_time = float(global_start_time_str)
         logger.info(f"Worker {self.worker_id} started. Global start time: {global_start_time}")
 
         while not self.stop_event.is_set():
             try:
                 # Read tasks
-                tasks = await self.redis.read_tasks(
-                    self.stream_name, self.group_name, self.worker_id, count=1
-                )
+                tasks = await self.redis.read_tasks(self.stream_name, self.group_name, self.worker_id, count=1)
                 if not tasks:
                     await asyncio.sleep(0.1)
                     continue
@@ -84,13 +81,13 @@ class DistributedWorker:
         task_id = task["_id"]
         stage_id = task.get("stage_id", 0)
         lora_adapter = task.get("lora_adapter")
-        
+
         try:
             scheduled_offset = task["scheduled_offset"]
             # Use global_start_time from task if available, fallback to cached one
             task_global_start_time = task.get("global_start_time", global_start_time)
             target_time = task_global_start_time + scheduled_offset
-            
+
             # Wait until scheduled time
             now = time.time()
             sleep_time = target_time - now
@@ -107,29 +104,26 @@ class DistributedWorker:
             # Fetch prompt from Redis Hash
             prompt_field = task["prompt_field"]
             prompt_text = await self.redis.redis.hget("test_prompts", prompt_field)
-            
+
             if not prompt_text:
                 logger.error(f"Prompt not found for field {prompt_field}")
                 await self.redis.ack_task(self.stream_name, self.group_name, task_id)
                 return
 
             # Construct API Data
-            api_data = CompletionAPIData(
-                prompt=prompt_text,
-                max_tokens=task.get("output_len", 50)
-            )
+            api_data = CompletionAPIData(prompt=prompt_text, max_tokens=task.get("output_len", 50))
 
             # Execute request
             start_perf = time.perf_counter()
             start_wall = time.time()
-            
+
             # Note: process_request logs metrics to its own collector if configured.
             # Here we want to capture results for Redis.
             # We might need to wrap it or modify it to return metrics.
             # Let's assume process_request handles it and we can access results,
             # or we implement the call here directly if process_request is too tight.
             # Looking at load_generator.py, it calls client.process_request directly.
-            
+
             try:
[…]
