### 2026-04-14 f4275b1
fix: Resolve Mypy type-checking errors across distributed modules
