### 2026-04-14 f5796fc
feat: add unit tests for distributed module
