+++ tests/distributed/test_redis_client.py
@@ -1,0 +1,134 @@
+import pytest
+import json
+from unittest.mock import AsyncMock, MagicMock, patch
+from inference_perf.distributed.redis_client import RedisClient
+from redis.exceptions import ResponseError
+
+@pytest.fixture
+def mock_redis():
+    with patch("redis.asyncio.Redis") as mock:
+        instance = mock.return_value
+        instance.xadd = AsyncMock()
+        instance.xgroup_create = AsyncMock()
+        instance.xreadgroup = AsyncMock()
+        instance.xack = AsyncMock()
+        instance.publish = AsyncMock()
+        instance.close = AsyncMock()
+        instance.pubsub = MagicMock()
+        yield instance
+
+@pytest.mark.asyncio
+async def test_connect():
+    with patch("redis.asyncio.Redis") as mock_redis_class:
+        client = RedisClient()
+        await client.connect()
+        assert client.redis is not None
+        mock_redis_class.assert_called_once_with(
+            host="localhost", port=6379, db=0, password=None, decode_responses=True
+        )
+
+@pytest.mark.asyncio
+async def test_close(mock_redis):
+    client = RedisClient()
+    await client.connect()
+    await client.close()
+    mock_redis.close.assert_called_once()
+
+@pytest.mark.asyncio
+async def test_push_task(mock_redis):
+    client = RedisClient()
+    await client.connect()
+    mock_redis.xadd.return_value = "123-0"
+    
+    task_id = await client.push_task("mystream", {"key": "value"})
+    
+    assert task_id == "123-0"
+    mock_redis.xadd.assert_called_once_with("mystream", {"data": json.dumps({"key": "value"})})
+
+@pytest.mark.asyncio
+async def test_create_consumer_group_success(mock_redis):
+    client = RedisClient()
+    await client.connect()
+    
+    await client.create_consumer_group("mystream", "mygroup")
+    
+    mock_redis.xgroup_create.assert_called_once_with("mystream", "mygroup", id="0", mkstream=True)
+
+@pytest.mark.asyncio
+async def test_create_consumer_group_busygroup(mock_redis):
+    client = RedisClient()
+    await client.connect()
+    mock_redis.xgroup_create.side_effect = ResponseError("BUSYGROUP Consumer Group name already exists")
+    
+    # Should not raise error
+    await client.create_consumer_group("mystream", "mygroup")
+
+@pytest.mark.asyncio
+async def test_create_consumer_group_other_error(mock_redis):
+    client = RedisClient()
+    await client.connect()
+    mock_redis.xgroup_create.side_effect = ResponseError("Some other error")
+    
+    with pytest.raises(ResponseError):
+        await client.create_consumer_group("mystream", "mygroup")
+
+@pytest.mark.asyncio
+async def test_read_tasks(mock_redis):
+    client = RedisClient()
+    await client.connect()
+    mock_redis.xreadgroup.return_value = [
+        ("mystream", [("123-0", {"data": json.dumps({"key": "value"})})])
+    ]
+    
+    tasks = await client.read_tasks("mystream", "mygroup", "myconsumer")
+    
+    assert len(tasks) == 1
+    assert tasks[0]["key"] == "value"
+    assert tasks[0]["_id"] == "123-0"
+    mock_redis.xreadgroup.assert_called_once_with(
+        groupname="mygroup", consumername="myconsumer", streams={"mystream": ">"}, count=1
+    )
+
+@pytest.mark.asyncio
+async def test_ack_task(mock_redis):
+    client = RedisClient()
+    await client.connect()
+    
+    await client.ack_task("mystream", "mygroup", "123-0")
+    
[…]
