+++ tests/distributed/test_orchestrator.py
@@ -4,10 +4,11 @@ import asyncio
 from unittest.mock import AsyncMock, MagicMock, patch, ANY
 from inference_perf.distributed.orchestrator import Orchestrator
 from inference_perf.distributed.redis_client import RedisClient
-from inference_perf.config import Config, LoadConfig, DataConfig, APIConfig, StandardLoadStage
+from inference_perf.config import LoadConfig, DataConfig, APIConfig, StandardLoadStage
+
 
 @pytest.fixture
-def mock_redis_client():
+def mock_redis_client() -> MagicMock:
     client = MagicMock(spec=RedisClient)
     client.redis = MagicMock()
     client.redis.set = AsyncMock()
@@ -20,97 +21,105 @@ def mock_redis_client():
     client.connect = AsyncMock()
     return client
 
+
 @pytest.fixture
-def mock_configs():
-    load_config = LoadConfig(
-        stages=[StandardLoadStage(rate=10, duration=2)]
-    )
+def mock_configs() -> tuple[LoadConfig, DataConfig, APIConfig]:
+    load_config = LoadConfig(stages=[StandardLoadStage(rate=10, duration=2)])
     data_config = DataConfig(type="mock")
     api_config = APIConfig(type="completion")
     return load_config, data_config, api_config
 
+
 @pytest.mark.asyncio
 @patch("inference_perf.distributed.orchestrator.CustomTokenizer")
 @patch("inference_perf.distributed.orchestrator.SharedPrefixDataGenerator")
-async def test_run_job_with_configs(mock_datagen_class, mock_tokenizer_class, mock_redis_client, mock_configs):
+async def test_run_job_with_configs(
+    mock_datagen_class: MagicMock,
+    mock_tokenizer_class: MagicMock,
+    mock_redis_client: MagicMock,
+    mock_configs: tuple[LoadConfig, DataConfig, APIConfig],
+) -> None:
     load_config, data_config, api_config = mock_configs
-    
+
     mock_datagen = mock_datagen_class.return_value
     mock_datagen.prompts = ["prompt1", "prompt2"]
     mock_datagen.flat_output_lens = [10, 20]
-    
+
     orchestrator = Orchestrator(mock_redis_client)
-    
+
     # Mock wait_for_completion to avoid infinite loop
-    orchestrator.wait_for_completion = AsyncMock()
-    
+    orchestrator.wait_for_completion = AsyncMock()  # type: ignore[method-assign]
+
     await orchestrator.run_job_with_configs(load_config, data_config, api_config, tokenizer_config=None)
-    
+
     # Verify Redis interactions
     mock_redis_client.redis.set.assert_any_call("global_start_time", ANY)
     mock_redis_client.redis.delete.assert_any_call("results_stream")
-    
+
     # Verify tasks were dispatched
-    assert mock_redis_client.push_task.call_count == 20 # 10 rate * 2 duration
-    
+    assert mock_redis_client.push_task.call_count == 20  # 10 rate * 2 duration
+
     # Verify prime_redis called hset
     assert mock_redis_client.redis.hset.call_count == 2
 
+
 @pytest.mark.asyncio
-async def test_wait_for_completion(mock_redis_client):
+async def test_wait_for_completion(mock_redis_client: MagicMock) -> None:
     orchestrator = Orchestrator(mock_redis_client)
-    
+
     # Mock redis.get to simulate completion
     # First call for cancel, second for completed, third for cancel, fourth for completed
     mock_redis_client.redis.get.side_effect = [None, "5", None, "20"]
-    
+
     with patch("asyncio.sleep", AsyncMock()) as mock_sleep:
         await orchestrator.wait_for_completion(total_requests=20)
-        
+
[…]
