### 2026-04-14 02a879b
feat: Implement distributed OTel trace replay with Redis backend
