### 2026-04-14 834d51e
fix: Address lint and Mypy errors in distributed OTel replay
