### 2026-04-14 f994de2
Support streaming get_data() in distributed orchestrator and add synthetic test config
