+++ inference_perf/distributed/orchestrator.py
@@ -4,8 +4,18 @@ import time
 import json
 from typing import Any, Optional
 from inference_perf.config import Config, LoadConfig, DataConfig, APIConfig, DataGenType
-from inference_perf.datagen.shared_prefix_datagen import SharedPrefixDataGenerator
-from inference_perf.datagen.otel_trace_replay_datagen import OTelTraceReplayDataGenerator
+from inference_perf.apis import LazyLoadInferenceAPIData
+from inference_perf.datagen import (
+    MockDataGenerator,
+    HFShareGPTDataGenerator,
+    SyntheticDataGenerator,
+    RandomDataGenerator,
+    SharedPrefixDataGenerator,
+    CNNDailyMailDataGenerator,
+    InfinityInstructDataGenerator,
+    BillsumConversationsDataGenerator,
+    OTelTraceReplayDataGenerator,
+)
 from inference_perf.distributed.redis_client import RedisClient
 from inference_perf.utils.custom_tokenizer import CustomTokenizer
 
@@ -53,13 +63,29 @@ class Orchestrator:
 
         # Initialize data generator
         datagen: Any
-        if data_config.type == DataGenType.OTelTraceReplay:
+        if data_config.type == DataGenType.Mock:
+            datagen = MockDataGenerator(api_config, data_config, tokenizer)
+        elif data_config.type == DataGenType.ShareGPT:
+            datagen = HFShareGPTDataGenerator(api_config, data_config, tokenizer)
+        elif data_config.type == DataGenType.Synthetic:
+            datagen = SyntheticDataGenerator(api_config, data_config, tokenizer)
+        elif data_config.type == DataGenType.Random:
+            datagen = RandomDataGenerator(api_config, data_config, tokenizer)
+        elif data_config.type == DataGenType.SharedPrefix:
+            datagen = SharedPrefixDataGenerator(api_config, data_config, tokenizer)
+        elif data_config.type == DataGenType.CNNDailyMail:
+            datagen = CNNDailyMailDataGenerator(api_config, data_config, tokenizer)
+        elif data_config.type == DataGenType.InfinityInstruct:
+            datagen = InfinityInstructDataGenerator(api_config, data_config, tokenizer)
+        elif data_config.type == DataGenType.BillsumConversations:
+            datagen = BillsumConversationsDataGenerator(api_config, data_config, tokenizer)
+        elif data_config.type == DataGenType.OTelTraceReplay:
             datagen = OTelTraceReplayDataGenerator(api_config, data_config, tokenizer)
         else:
-            datagen = SharedPrefixDataGenerator(api_config, data_config, tokenizer)
+            datagen = MockDataGenerator(api_config, data_config, tokenizer)
 
-        # Prime Redis with prompts (skip for OTel replay as tasks carry full data)
-        if data_config.type != DataGenType.OTelTraceReplay:
+        # Prime Redis with prompts (only for SharedPrefix for now, others handled in dispatch)
+        if data_config.type == DataGenType.SharedPrefix:
             await self.prime_redis(datagen)
 
         # Set global start time
@@ -190,28 +216,77 @@ class Orchestrator:
                 await self.redis.push_task(self.stream_name, task)
                 total_requests += 1
         else:
-            # Legacy mode
-            for stage_id, stage in enumerate(load_config.stages):
-                rate = stage.rate
-                duration = stage.duration
-                num_requests = int(rate * duration)
+            # Legacy mode / DataGenerator mode
+            if hasattr(datagen, "prompts") and hasattr(datagen, "flat_output_lens"):
+                # SharedPrefix optimized path (already primed in Redis)
+                for stage_id, stage in enumerate(load_config.stages):
+                    rate = stage.rate
+                    duration = stage.duration
[…]
