### 2026-04-15 42ab74e
improve the otel readme and link the top level readme to it
