+++ examples/otel/README.md
@@ -5,300 +5,190 @@ system (e.g. an agent framework instrumented with OTel); inference-perf reconstr
 original call graph — including sequential dependencies, parallel fan-outs, and shared-prefix
 patterns — and drives those calls against the target inference server under test.
 
+## Table of Contents
+
+- [Why use OTel trace replay?](#why-use-otel-trace-replay)
+- [Quick Start](#quick-start)
+- [Configuration Guide](#configuration-guide)
+- [Session-Level Metrics](#session-level-metrics)
+- [OpenTelemetry Background](#opentelemetry-background)
+- [Developer Guide](#developer-guide)
+
 ## Why use OTel trace replay?
 
-Standard load types (`constant`, `poisson`, `concurrent`) generate independent, single-turn requests.
-Agentic applications — tool-calling agents, multi-turn conversations, RAG pipelines — produce **chains
-of dependent LLM calls** where each call's input includes the output of predecessors.
+Standard load types (`constant`, `poisson`, `concurrent`) dispatch requests at pre-scheduled times determined by the load timer. While they can handle sequential multi-turn conversations (via `shared_prefix` with user sessions), they cannot model **complex dependency graphs** where:
+- Multiple LLM calls run in parallel (e.g., parallel tool calls, concurrent reasoning paths)
+- Each call's input depends on outputs from multiple predecessors
+- Timing between calls reflects real application logic (waiting for tool results, user input, etc.) rather than just clock-based scheduling
+
+Agentic applications — tool-calling agents with parallel branches, multi-step RAG pipelines, complex workflows — produce these **dependency graphs** where the structure and timing of calls is determined by the application's control flow, not a fixed schedule.
 
 OTel trace replay enables you to:
-1. Benchmark **realistic agentic workloads** with proper dependency chains
-2. Measure **KV cache effectiveness** with actual growing-context patterns
-3. Test **session-level behavior** (success rates, end-to-end latency)
-4. Replay **production traffic patterns** captured from real systems
+1. Benchmark **complex agentic workloads** with parallel execution and branching dependencies
+2. Replay **production traffic patterns** with actual timing and dependency structures from real systems
+3. Measure **KV cache effectiveness** with realistic shared-prefix and growing-context patterns
+4. Test **session-level behavior** (success rates, end-to-end latency, failure propagation) for complete workflows
 
-## How it works
+> **Note:** If you're unfamiliar with OpenTelemetry traces and spans, see the [OpenTelemetry Background](#opentelemetry-background) section.
 
-### 1. Trace → Replay Graph
+## Quick Start
 
-Each OTel trace file contains a flat list of spans. The replayer converts them into a directed
-acyclic graph (DAG):
+```bash
+# Replay a single trace against a local vLLM server
+python -m inference_perf.main \
+  --config examples/otel/configs/per_case_config/simple_chain.yml
 
-1. **Extract LLM spans** — spans with `gen_ai.input.messages` (or a `chat *` name) become session events.
-2. **Infer dependencies** — two types of edges are added:
-   - **Causal edges**: when a span's input contains an `assistant` message whose content exactly matches a predecessor's output
-   - **Temporal edges**: to the closest non-overlapping earlier span (timing fallback)
-   
-   The temporal fallback is necessary because output matching doesn't always detect all dependencies. If event X ends before event Y begins, X is considered a predecessor even if Y doesn't use X's entire output.
-3. **Transitive reduction** — redundant edges are pruned so only direct predecessors remain.
-4. **Segment decomposition** — each event's input is split into message-level segments:
-   - `shared` — leading messages identical to a predecessor (KV-cache hit opportunity)
[…]
