### 2026-04-15 f892868
Fix lint and type checking errors in tests
