### 2026-04-15 5037fe0
move otel trace replay readme to /docs
