+++ README.md
@@ -93,8 +93,8 @@ Explore detailed documentation for specific topics:
 | **Metrics** | Definitions of TTFT, TPOT, ITL, etc. | [metrics.md](./docs/metrics.md) |
 | **Goodput** | How to measure requests meeting SLOs. | [goodput.md](./docs/goodput.md) |
 | **Reports** | Understanding generated JSON reports. | [reports.md](./docs/reports.md) |
-| **OTel Instrumentation** | OpenTelemetry integration for tracing. | [otel_instrumentation.md](./docs/otel_instrumentation.md) |
-| **OTel Trace Replay** | Replay production traces with dependency graphs and session metrics. | [examples/otel/README.md](./examples/otel/README.md) |
+| **OTel Observability** | Instrument benchmark runs with OpenTelemetry tracing to export to Jaeger, Tempo, etc. | [otel_instrumentation.md](./docs/otel_instrumentation.md) |
+| **OTel Trace Replay** | Data/load type for replaying production traces with complex dependency graphs. | [otel_trace_replay.md](./docs/otel_trace_replay.md) |
 | **Analysis** | Visualizations and plots for performance metrics. | [analysis.md](./docs/analysis.md) |
 
 ---
