### 2026-04-16 55cd2a2
add random session id injection to unique messages in the session and add optional duplicate sessions in load
