+++ inference_perf/datagen/replay_graph_session_datagen.py
@@ -26,6 +26,7 @@ import asyncio
 import json
 import logging
 import time
+import uuid
 from dataclasses import dataclass, field, replace as dc_replace
 from multiprocessing.managers import SyncManager
 from typing import Any, Dict, List, Optional, Set, Tuple
@@ -43,7 +44,7 @@ from inference_perf.apis import (
 from inference_perf.apis.chat import ChatMessage
 from inference_perf.payloads import RequestMetrics, Text
 from inference_perf.apis.streaming_parser import parse_sse_stream
-from inference_perf.config import APIConfig, APIType, DataConfig
+from inference_perf.config import APIConfig, APIType, DataConfig, SessionReplayConfig
 from inference_perf.datagen.base import LazyLoadDataMixin, SessionGenerator
 from inference_perf.datagen.replay_graph_types import InputSegment, ReplayGraph
 from inference_perf.utils.custom_tokenizer import CustomTokenizer
@@ -208,6 +209,9 @@ class SessionChatCompletionAPIData(ChatCompletionAPIData):
     skip_request: bool = False
     expected_output_is_tool_call: bool = False
     expected_output_tool_names: Optional[List[str]] = None
+    # KV-cache invalidation configuration
+    inject_random_session_id: bool = False
+    session_random_string: Optional[str] = None
 
     async def to_payload(
         self, effective_model_name: str, max_tokens: int, ignore_eos: bool, streaming: bool
@@ -298,8 +302,16 @@ class SessionChatCompletionAPIData(ChatCompletionAPIData):
             logger.debug(f"Event {self.event_id} waiting {wait_sec:.3f}s (wait_ms={self.wait_ms})")
             await asyncio.sleep(wait_sec)
 
-        if any(seg.type == "output" or seg.type == "shared" for seg in self.input_segments):
-            logger.debug(f"Event {self.event_id} substituting output/shared segments")
+        # Substitute output segments with actual predecessor outputs, or inject random session ID into unique segments
+        needs_substitution = any(seg.type == "output" or seg.type == "shared" for seg in self.input_segments)
+        needs_random_injection = self.inject_random_session_id and any(seg.type == "unique" for seg in self.input_segments)
+
+        if needs_substitution or needs_random_injection:
+            if needs_substitution:
+                logger.debug(f"Event {self.event_id} substituting output/shared segments")
+            if needs_random_injection:
+                logger.debug(f"Event {self.event_id} injecting random session ID into unique segments")
+
             substituted = self._build_messages_with_substitution()
             # _build_messages_with_substitution calls record_failure and returns
             # early when substitution is not possible (e.g. tool call expected but
@@ -315,7 +327,7 @@ class SessionChatCompletionAPIData(ChatCompletionAPIData):
                 )
                 for m in substituted
             ]
-            logger.debug(f"Event {self.event_id} substitution complete, {len(self.messages)} messages")
+            logger.debug(f"Event {self.event_id} substitution/injection complete, {len(self.messages)} messages")
 
     def _build_messages_with_substitution(self) -> List[Dict[str, Any]]:
         # NOTE: when input_segments is empty, the original_messages list is returned
@@ -430,8 +442,20 @@ class SessionChatCompletionAPIData(ChatCompletionAPIData):
                         else:
                             result.append(dict(msg))
             elif seg.type == "unique":
+                # Unique message - inject random session string if enabled
                 for msg in seg_msgs:
-                    result.append(msg)
+                    if self.inject_random_session_id and self.session_random_string:
+                        # Use the session random string passed from SessionGraphState
+                        # Inject random string into message content
+                        msg_copy = dict(msg)
+                        original_content = msg_copy.get("content", "")
+
+                        # Prepend random session identifier to content
[…]
