### 2026-04-16 c6deed4
auto-inject random strings in duplicate sessions for KV-cache isolation
