+++ tests/test_otel_replay_datagen.py
@@ -1164,3 +1164,172 @@ class TestRandomSessionIDInjection:
         # Message should NOT have random string injected
         assert api_data.messages[0].content == "What is the capital of France?"
         assert "[ID:" not in api_data.messages[0].content
+
+    @pytest.mark.asyncio
+    async def test_no_injection_for_original_session_when_flag_disabled(self) -> None:
+        """Case 1: inject_random_session_id=False, original session (not duplicate) -> NO injection."""
+        registry = EventOutputRegistry()
+        tracker = WorkerSessionTracker()
+
+        messages = [
+            {"role": "user", "content": "What is the capital of France?"},
+        ]
+
+        segments = [
+            InputSegment(type="unique", message_count=1, token_count=10, source_event_id=None),
+        ]
+
+        api_data = OTelChatCompletionAPIData(
+            messages=[ChatMessage(role="user", content="What is the capital of France?")],
+            max_tokens=50,
+            event_id="session_001:event_0",  # Original session (no _dup)
+            registry=registry,
+            worker_tracker=tracker,
+            completion_queue=None,
+            total_events_in_session=1,
+            inject_random_session_id=False,  # Flag disabled
+            session_random_string="test123456789abc",  # Random string available but not used
+            input_segments=segments,
+            original_messages=messages,
+        )
+
+        # Trigger substitution
+        await api_data.wait_for_predecessors_and_substitute()
+
+        # Message should NOT have random string injected (flag off, not duplicate)
+        assert api_data.messages[0].content == "What is the capital of France?"
+        assert "[SESS:" not in api_data.messages[0].content
+
+    @pytest.mark.asyncio
+    async def test_injection_for_duplicate_session_when_flag_disabled(self) -> None:
+        """Case 2: inject_random_session_id=False, duplicate session -> YES injection (automatic)."""
+        registry = EventOutputRegistry()
+        tracker = WorkerSessionTracker()
+
+        messages = [
+            {"role": "user", "content": "What is the capital of France?"},
+        ]
+
+        segments = [
+            InputSegment(type="unique", message_count=1, token_count=10, source_event_id=None),
+        ]
+
+        api_data = OTelChatCompletionAPIData(
+            messages=[ChatMessage(role="user", content="What is the capital of France?")],
+            max_tokens=50,
+            event_id="session_001_dup1:event_0",  # Duplicate session (has _dup1)
+            registry=registry,
+            worker_tracker=tracker,
+            completion_queue=None,
+            total_events_in_session=1,
+            inject_random_session_id=False,  # Flag disabled
+            session_random_string="test123456789abc",  # Random string available
+            input_segments=segments,
+            original_messages=messages,
+        )
+
+        # Trigger substitution
+        await api_data.wait_for_predecessors_and_substitute()
+
+        # Message SHOULD have random string injected (duplicate session, automatic)
+        assert "[SESS:test123456789abc]" in api_data.messages[0].content
+        assert "What is the capital of France?" in api_data.messages[0].content
+
+    @pytest.mark.asyncio
+    async def test_injection_for_original_session_when_flag_enabled(self) -> None:
+        """Case 3: inject_random_session_id=True, original session -> YES injection (flag enabled)."""
+        registry = EventOutputRegistry()
+        tracker = WorkerSessionTracker()
+
+        messages = [
+            {"role": "user", "content": "What is the capital of France?"},
+        ]
+
+        segments = [
+            InputSegment(type="unique", message_count=1, token_count=10, source_event_id=None),
+        ]
+
+        api_data = OTelChatCompletionAPIData(
+            messages=[ChatMessage(role="user", content="What is the capital of France?")],
+            max_tokens=50,
[…]
