### 2026-04-18 #10314 @neerazz
Thanks @achandrasekar for the updates. I went back through the paper and repo again, the gRPC claim is gone, price/perf is framed correctly now, and the contrib docs are in better shape. So that's good.

The one thing I'm still stuck on is the streaming metrics. Looking at the current parser, it still timestamps every SSE message before it checks whether there's actually output text in it. The tests seem to expect that too. Like the one that asserts 3 timestamps for 2 content chunks + [DONE]. Since TTFT / TPOT / ITL all get computed from that timestamp list, and those metrics are basically the point of the paper, I don't think I can just let that go.

This is my first JOSS review so maybe I'm overthinking it, but for a benchmarking tool paper this feels like it matters. Everything else looks solid to me. Happy to clear the rest once the streaming side is sorted out.
