### 2026-04-20 5932ab5
chore: reuse Distribution config in conversation_replay
Addresses PR review: ConversationReplayDistribution was a near-duplicate
of the existing Distribution config. Removed the duplicate type and
switched ConversationReplayConfig fields to use Distribution directly.

- Added FIXED to DistributionType enum (previously only in the
  conversation-replay subtype) and handled it in sample_from_distribution
- ConversationReplayDataGenerator now samples via sample_from_distribution
  instead of re-implementing the dispatch
- Adapted to upstream UserSessionCompletionAPIData, which now exposes
  user_
[…]
