### 2026-04-22 f4bd34d
add multimodal benchmarking
