+++ inference_perf/reportgen/base.py
@@ -25,10 +25,10 @@ import numpy as np
 from pydantic import BaseModel
 
 from inference_perf.apis import RequestLifecycleMetric, SessionLifecycleMetric, StreamedInferenceResponseInfo
-from inference_perf.client.metricsclient import MetricsClient, PerfRuntimeParameters
-from inference_perf.client.metricsclient.base import ModelServerMetrics
-from inference_perf.client.metricsclient.prometheus_client import PrometheusMetricsClient
-from inference_perf.client.requestdatacollector import RequestDataCollector
+from inference_perf.client.server_metrics import ServerMetricsClient, PerfRuntimeParameters
+from inference_perf.client.server_metrics.base import ModelServerMetrics
+from inference_perf.client.server_metrics.prometheus_client import PrometheusMetricsClient
+from inference_perf.metrics.request_collector import RequestMetricCollector
 from inference_perf.config import (
     Config,
     PrometheusMetricsReportConfig,
@@ -484,7 +484,23 @@ def summarize_requests(
     valid_tpot = [v for v in tpot_values if v is not None]
     valid_ttft = [v for v in ttft_values if v is not None]
 
-    successes_dict = {
+    request_sizes = [len(x.request_data.encode("utf-8")) for x in all_successful]
+    all_images = []
+    all_videos = []
+    all_audios = []
+    for success in all_successful:
+        if success.info.payload.image:
+            all_images.extend(success.info.payload.image.instances)
+        if success.info.payload.video:
+            all_videos.extend(success.info.payload.video.instances)
+        if success.info.payload.audio:
+            all_audios.extend(success.info.payload.audio.instances)
+
+    image_counts = [safe_float(s.info.payload.image.count if s.info.payload.image else 0) for s in all_successful]
+    video_counts = [safe_float(s.info.payload.video.count if s.info.payload.video else 0) for s in all_successful]
+    audio_counts = [safe_float(s.info.payload.audio.count if s.info.payload.audio else 0) for s in all_successful]
+
+    successes_dict: dict[str, Any] = {
         "count": len(all_successful),
         "latency": {
             "request_latency": summarize(request_latency_values, percentiles),
@@ -495,7 +511,9 @@ def summarize_requests(
         },
         "throughput": {
             "input_tokens_per_sec": (
-                sum(safe_float(x.info.input_tokens) for x in all_successful) / total_time if total_time > 0 else 0.0
+                sum(safe_float(x.info.payload.text.input_tokens) for x in all_successful) / total_time
+                if total_time > 0
+                else 0.0
             ),
             "output_tokens_per_sec": (
                 sum(safe_float(x.info.response_info.output_tokens) if x.info.response_info else 0.0 for x in all_successful)
@@ -514,8 +532,32 @@ def summarize_requests(
                 else 0.0
             ),
             "requests_per_sec": (len(all_successful) / total_time if total_time > 0 else 0.0),
+            "images_per_sec": (sum(image_counts) / total_time if total_time > 0 else 0.0),
+            "videos_per_sec": (sum(video_counts) / total_time if total_time > 0 else 0.0),
+            "audios_per_sec": (sum(audio_counts) / total_time if total_time > 0 else 0.0),
         },
-        "prompt_len": summarize([safe_float(success.info.input_tokens) for success in all_successful], percentiles),
+        "request_size_bytes": summarize([float(x) for x in request_sizes], percentiles),
+        "prompt_len": summarize(
+            [safe_float(success.info.payload.text.input_tokens) for success in all_successful], percentiles
+        ),
+        "image": {
+            "count": summarize(image_counts, percentiles),
+            "pixels": summarize([safe_float(inst.pixels) for inst in all_images], percentiles),
[…]
