+++ inference_perf/payloads/__init__.py
@@ -11,21 +11,24 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from .base import MetricsClient, PerfRuntimeParameters, ModelServerMetrics
-from typing import Optional
+from .payload import (
+    Text,
+    Image,
+    Images,
+    Video,
+    Videos,
+    Audio,
+    Audios,
+    Payload,
+)
 
-
-class MockMetricsClient(MetricsClient):
-    def __init__(self) -> None:
-        pass
-
-    def collect_metrics_summary(self, runtime_parameters: PerfRuntimeParameters) -> Optional[ModelServerMetrics]:
-        return None
-
-    def collect_metrics_for_stage(
-        self, runtime_parameters: PerfRuntimeParameters, stage_id: int
-    ) -> Optional[ModelServerMetrics]:
-        return None
-
-    def wait(self) -> None:
-        pass
+__all__ = [
+    "Text",
+    "Image",
+    "Images",
+    "Video",
+    "Videos",
+    "Audio",
+    "Audios",
+    "Payload",
+]
