### 2026-04-22 4a43163
add multimodal benchmarking
