+++ inference_perf/datagen/otel_trace_replay_datagen.py
@@ -64,6 +64,7 @@ from multiprocessing.managers import SyncManager
 from pathlib import Path
 from typing import List, Optional
 
+from datasets import load_dataset
 from inference_perf.config import APIConfig, DataConfig
 from inference_perf.datagen.replay_graph_session_datagen import (
     ReplaySession,
@@ -122,6 +123,105 @@ def resolve_trace_files(trace_files: List[str]) -> List[Path]:
 
     # Convert to sorted list for consistent output
     return [Path(f) for f in all_files]
+
+
+def _download_hf_dataset(dataset_path: str) -> tuple[Path, List[Path]]:
+    """Download HuggingFace dataset and extract trace JSON files.
+
+    Args:
+        dataset_path: HuggingFace dataset identifier (e.g., 'lenadan/otel-test-snippet')
+
+    Returns:
+        Tuple of (temp_dir, list of trace file paths)
+
+    Raises:
+        ValueError: If dataset cannot be downloaded or doesn't contain trace data
+    """
+    try:
+        logger.info(f"Loading HuggingFace dataset: {dataset_path}")
+        # Load dataset - this downloads it to HuggingFace cache if not already present
+        dataset = load_dataset(dataset_path)
+
+        # Create a temporary directory to store extracted trace files
+        import tempfile
+
+        temp_dir = Path(tempfile.mkdtemp(prefix="hf_otel_traces_"))
+        logger.info(f"Extracting traces to temporary directory: {temp_dir}")
+
+        trace_files = []
+        row_count = 0
+
+        # Iterate through all splits in the dataset
+        for split_name, split_data in dataset.items():
+            logger.info(f"Processing split '{split_name}' with {len(split_data)} rows")
+
+            # Each row should contain trace data
+            for idx, row in enumerate(split_data):
+                # Try to find trace data in the row
+                # Common column names: 'trace', 'json', 'data', 'content'
+                trace_content = None
+                for col_name in ["trace", "json", "data", "content", "text", "spans"]:
+                    if col_name in row:
+                        trace_content = row[col_name]
+                        break
+
+                if trace_content is None:
+                    # If no standard column found, try the first column
+                    if len(row) > 0:
+                        trace_content = list(row.values())[0]
+
+                if trace_content:
+                    # Write trace to a JSON file
+                    trace_file = temp_dir / f"{split_name}_{idx}.json"
+
+                    # Parse trace_content and wrap in "spans" key if needed
+                    if isinstance(trace_content, str):
+                        try:
+                            trace_data = json.loads(trace_content)
+                        except json.JSONDecodeError:
+                            # If not valid JSON, skip this row
+                            logger.warning(f"Skipping row {idx} in split '{split_name}': invalid JSON")
+                            continue
+                    else:
+                        # Assume it's already a dict/object
+                        trace_data = trace_content
+
+                    # Wrap trace_data in "spans" key if needed
+                    if isinstance(trace_data, list):
+                        # trace_data is a list of spans, wrap it
+                        trace_data = {"spans": trace_data}
+                    elif isinstance(trace_data, dict) and "spans" not in trace_data:
+                        # trace_data is a dict but doesn't have "spans" key, wrap it
+                        trace_data = {"spans": trace_data}
+                    # else: trace_data is already a dict with "spans" key, use as-is
+
+                    trace_file.write_text(json.dumps(trace_data, indent=2))
+                    trace_files.append(trace_file)
+                    row_count += 1
+
+        if not trace_files:
+            raise ValueError(
[…]
