+++ inference_perf/datagen/otel_trace_replay_datagen.py
@@ -33,7 +33,7 @@ event_id → actual output text. All registry data is stored in per-worker plain
 worker affinity ensures that the event writing an output and the events reading it always run on
 the same worker process.
 
-The actual output is captured in `OTelChatCompletionAPIData.process_response`, which overrides
+The actual output is captured in `SessionChatCompletionAPIData.process_response`, which overrides
 the base `ChatCompletionAPIData.process_response` to register the output after each LLM call.
 
 Cross-process coordination
@@ -53,32 +53,22 @@ Two orthogonal coordination problems exist in this system:
     - session_completion_queue: Workers push completion data when last event finishes
     - Main process consumes from queue in check_session_completed()
 
-OTelChatCompletionAPIData holds refs to WorkerSessionTracker and completion_queue.
+SessionChatCompletionAPIData holds refs to WorkerSessionTracker and completion_queue.
 """
 
-import asyncio
 import glob
 import json
 import logging
 import random
-import time
-from dataclasses import dataclass, field
 from multiprocessing.managers import SyncManager
 from pathlib import Path
-from typing import Any, Dict, List, Optional, Set
+from typing import List, Optional
 
-from aiohttp import ClientResponse
-
-from inference_perf.apis.chat import ChatCompletionAPIData, ChatMessage
-from inference_perf.apis.streaming_parser import parse_sse_stream
-from inference_perf.apis.user_session import InferenceInfo
 from inference_perf.config import APIConfig, DataConfig
 from inference_perf.datagen.replay_graph_session_datagen import (
-    EventFailedError,
     ReplaySession,
     ReplayGraphSessionGeneratorBase,
 )
-from inference_perf.datagen.replay_graph_types import InputSegment, ReplayGraph
 from inference_perf.datagen.otel_trace_to_replay_graph import (
     build_raw_calls,
     build_graph,
@@ -86,682 +76,6 @@ from inference_perf.datagen.otel_trace_to_replay_graph import (
 from inference_perf.utils.custom_tokenizer import CustomTokenizer
 
 logger = logging.getLogger(__name__)
-
-
-# ---------------------------------------------------------------------------
-# SessionGraphState — tracks graph traversal state for one session
-# ---------------------------------------------------------------------------
-
-
-@dataclass
-class SessionGraphState:
-    """Tracks graph traversal state for one session.
-
-    Manages which events are ready to dispatch, which have been dispatched,
-    and which have completed. Used for graph-based traversal.
-    """
-
-    session_id: str
-    graph: ReplayGraph
-    ready_events: Set[str]  # event_ids ready to dispatch (predecessors done)
-    dispatched_events: Set[str]  # event_ids already dispatched
-    completed_events: Set[str]  # event_ids that have completed
-    event_completion_times: Dict[str, float]  # event_id → completion timestamp
-    is_active: bool = False  # Is this session currently active?
-    is_complete: bool = False  # Has this session finished all events?
-    failed: bool = False  # Did this session fail?
-    cancelled_events: int = 0  # Number of events cancelled due to failure
-
-
-# ---------------------------------------------------------------------------
-# OTelInferenceInfo — InferenceInfo subclass that carries the output text
-# ---------------------------------------------------------------------------
-
-
-class OTelInferenceInfo(InferenceInfo):
-    """InferenceInfo subclass that also carries the raw output text.
-
-    Inherits all fields from InferenceInfo (input_tokens, output_tokens, etc.)
-    and adds output_text so the OTel replay path can register the actual
-    generated text in the EventOutputRegistry without touching the base class.
-    """
-
-    output_text: Optional[str] = None
-    output_tokens: int = 0
-    output_token_times: list[float] = field(default_factory=list)
-
-
[…]
