### 2026-04-23 kubernetes-sigs/inference-perf#392 @jjk-g
Todo: quantify the impact of this on these metrics

With the fix, document our behavior for the impact in the cases.
