### 2026-04-24 b9993c0
test: Set integer seed in create_data_config mock
Mocked shared_prefix.seed was a Mock, causing
np.random.default_rng(...) to raise TypeError in SharedPrefixDataGenerator
tests. Pin it to an int so the RNG initializes.

Signed-off-by: Hyeonki Hong <hhk7734@gmail.com>
