### 2026-04-25 515962d
fix(security): archive extraction vulnerable to path traversal (t
The `extract_tarball` helper invokes `tar -xzvf` directly on a provided archive path without validating archive member paths. A crafted tarball can contain entries with `../` segments, absolute paths, or symlinks that write outside the intended destination directory, leading to arbitrary file overwrite.

Signed-off-by: tomaioo <203048277+tomaioo@users.noreply.github.com>
