### 2026-04-27 c45f804
feat: tool-call capture, structured substitution, and session replay fixes
  Capture structured tool-call outputs from live model responses and substitute
  them correctly into downstream requests during OTel trace replay. The live
  tool_calls message is stored and injected into successor events, IDs are
  rewritten to match the live model's IDs, and tool_choice is injected to guide
  the model toward tool-call output.

  Also fixes several bugs that caused session replay to hang: missing completion
  queue notifications on pre-request skip paths, tool_definitions stored as a raw
  JSON strin
[…]
