+++ inference_perf/datagen/replay_graph_session_datagen.py
@@ -28,7 +28,7 @@ import logging
 import time
 from dataclasses import dataclass, field, replace as dc_replace
 from multiprocessing.managers import SyncManager
-from typing import Any, Dict, List, Optional, Set
+from typing import Any, Dict, List, Optional, Set, Tuple
 
 from aiohttp import ClientResponse
 
@@ -63,6 +63,7 @@ class SessionInferenceInfo(InferenceInfo):
     """InferenceInfo subclass that also carries the raw output text."""
 
     output_text: Optional[str] = None
+    output_message: Optional[Dict[str, Any]] = None
 
 
 class WorkerSessionTracker:
@@ -101,11 +102,18 @@ class EventOutputRegistry:
 
     def __init__(self) -> None:
         self._event_output_text: Dict[str, str] = {}
+        self._event_output_message: Dict[str, Dict[str, Any]] = {}
         self._event_input_messages: Dict[str, Any] = {}
         self._event_signals: Dict[str, asyncio.Event] = {}
         self._failed_event_ids: Set[str] = set()
 
-    def record(self, event_id: str, output_text: str, messages: List[Any]) -> None:
+    def record(
+        self,
+        event_id: str,
+        output_text: str,
+        messages: List[Any],
+        output_message: Optional[Dict[str, Any]] = None,
+    ) -> None:
         if event_id in self._event_output_text:
             raise ValueError(
                 f"Event {event_id} has already been recorded. "
@@ -114,6 +122,8 @@ class EventOutputRegistry:
 
         self._event_output_text[event_id] = output_text
         self._event_input_messages[event_id] = list(messages) if messages else []
+        if output_message is not None:
+            self._event_output_message[event_id] = output_message
 
         if event_id in self._event_signals:
             self._event_signals[event_id].set()
@@ -121,6 +131,9 @@ class EventOutputRegistry:
 
     def get_output_by_event_id(self, event_id: str) -> Optional[str]:
         return self._event_output_text.get(event_id)
+
+    def get_message_by_event_id(self, event_id: str) -> Optional[Dict[str, Any]]:
+        return self._event_output_message.get(event_id)
 
     def get_messages_by_event_id(self, event_id: str) -> Optional[List[Any]]:
         return self._event_input_messages.get(event_id)
@@ -193,17 +206,80 @@ class SessionChatCompletionAPIData(ChatCompletionAPIData):
     original_messages: List[Dict[str, Any]] = field(default_factory=list)
     expected_output_content: Optional[str] = None
     skip_request: bool = False
+    expected_output_is_tool_call: bool = False
+    expected_output_tool_names: Optional[List[str]] = None
+
+    async def to_payload(
+        self, effective_model_name: str, max_tokens: int, ignore_eos: bool, streaming: bool
+    ) -> Dict[str, Any]:
+        payload = await super().to_payload(effective_model_name, max_tokens, ignore_eos, streaming)
+
+        if self.expected_output_is_tool_call and self.tool_definitions:
+            if "tool_choice" in payload:
+                logger.warning(
+                    f"Event {self.event_id}: payload already has tool_choice={payload['tool_choice']!r}; "
+                    f"overwriting with replay-enforced value."
+                )
+            names = self.expected_output_tool_names or []
+            # Build available set from self.tool_definitions (the raw list). The name
+            # field is set explicitly in to_payload and is NOT passed through
+            # _clean_parameters, so it survives schema cleaning unchanged.
+            available = {t["name"] for t in self.tool_definitions if "name" in t}
+            if len(names) == 1 and names[0] in available:
+                # Force the exact function the original trace recorded. This maximises
+                # faithfulness to the trace and ensures the successor's role:tool messages
+                # (which reference this function by name/index) remain coherent.
[…]
