+++ tests/test_tool_call_capture.py
@@ -1,0 +1,587 @@
+# Copyright 2026 The Kubernetes Authors.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""Tests for tool-call output capture and structured substitution."""
+
+from typing import Any, Dict, List, Optional
+
+import pytest
+from inference_perf.datagen.replay_graph_session_datagen import (
+    EventOutputRegistry,
+    SessionChatCompletionAPIData,
+    SessionInferenceInfo,
+)
+from inference_perf.datagen.replay_graph_types import InputSegment
+
+
+_TOOL_CALLS = [
+    {
+        "id": "call_abc",
+        "type": "function",
+        "function": {"name": "get_weather", "arguments": '{"location":"Paris"}'},
+    }
+]
+
+
+class TestEventOutputRegistryStructured:
+    def test_record_with_tool_call_message_stored(self) -> None:
+        registry = EventOutputRegistry()
+        msg = {"role": "assistant", "tool_calls": _TOOL_CALLS}
+        registry.record("evt1", "", [], output_message=msg)
+        assert registry.get_message_by_event_id("evt1") == msg
+
+    def test_record_text_response_no_message_entry(self) -> None:
+        registry = EventOutputRegistry()
+        registry.record("evt1", "hello", [])
+        assert registry.get_message_by_event_id("evt1") is None
+        assert registry.get_output_by_event_id("evt1") == "hello"
+
+    def test_record_text_response_with_explicit_message(self) -> None:
+        registry = EventOutputRegistry()
+        msg = {"role": "assistant", "content": "hello"}
+        registry.record("evt1", "hello", [], output_message=msg)
+        assert registry.get_message_by_event_id("evt1") == msg
+        assert registry.get_output_by_event_id("evt1") == "hello"
+
+    def test_cleanup_removes_output_message(self) -> None:
+        registry = EventOutputRegistry()
+        msg = {"role": "assistant", "tool_calls": _TOOL_CALLS}
+        registry.record("sess:evt1", "", [], output_message=msg)
+        registry._event_output_message.pop("sess:evt1", None)
+        assert registry.get_message_by_event_id("sess:evt1") is None
+
+    def test_get_message_absent_returns_none(self) -> None:
+        registry = EventOutputRegistry()
+        assert registry.get_message_by_event_id("nonexistent") is None
+
+
+class TestSubstitutionWithToolCalls:
+    """Tests for _build_messages_with_substitution when predecessor emitted tool calls."""
+
+    def _make_api_data(
+        self,
+        registry: EventOutputRegistry,
+        original_messages: List[Dict[str, Any]],
+        input_segments: List[InputSegment],
+    ) -> SessionChatCompletionAPIData:
+        from inference_perf.datagen.replay_graph_session_datagen import WorkerSessionTracker
+        from inference_perf.apis.chat import ChatMessage
+
+        chat_messages = [ChatMessage(role=m["role"], content=m.get("content")) for m in original_messages]
+        return SessionChatCompletionAPIData(
+            messages=chat_messages,
+            max_tokens=100,
+            event_id="sess:evt2",
+            registry=registry,
+            worker_tracker=WorkerSessionTracker(),
+            completion_queue=None,
+            total_events_in_session=2,
[…]

[…]
