### 2026-04-27 44ab5ea
fix: correct response_info attribute access in OTEL metrics and update tests
