### 2026-04-27 99ad6c0
fix: replace load_dataset with huggingface_hub for HF trace loading
  datasets.load_dataset() fails on this dataset because datasets>=3.3.2
  doesn't support the 'Json' feature type that the Hub auto-infers for
  nested objects in JSONL files.

  Switch to huggingface_hub.list_repo_files + hf_hub_download to download
  raw JSONL files directly, bypassing all schema inference.
