### 2026-04-27 4c465fa
chore: allow virtual address style for s3 storage
