### 2026-04-28 #10314 @achandrasekar
The non-content token [issue](https://github.com/kubernetes-sigs/inference-perf/issues/392) is now addressed as well. The effect of the issue was essentially 1 or 2 tokens getting counted more per request which usually has 100s to 1000s of tokens. While it does not affect real-world tests in a meaningful way, fixing it still makes the reporting more accurate. So, it's now been fixed. Please take another look. Thank you and appreciate the review!
