### 2026-04-28 87e1f15
fix: add missing cancelled_events field and apply formatter fixes
Add cancelled_events: int = 0 to ReplaySessionState dataclass (required
by worker session tracking lifecycle) and apply formatter changes from
pdm run validate.
