+++ tests/test_otel_replay_datagen.py
@@ -45,6 +45,7 @@ import pytest
 # Ensure project root is on path
 sys.path.insert(0, str(Path(__file__).parent.parent))
 
+from inference_perf.datagen.base import SessionGenerator
 from inference_perf.datagen.otel_trace_replay_datagen import OTelTraceReplayDataGenerator
 from inference_perf.datagen.replay_graph_session_datagen import (
     EventFailedError,
@@ -1310,29 +1311,16 @@ class TestRandomSessionIDInjection:
 
     def test_is_duplicate_session_detection(self) -> None:
         """Test the _is_duplicate_session() method with various session ID patterns."""
-        registry = EventOutputRegistry()
-        tracker = WorkerSessionTracker()
-
-        # Create a test instance to access the method
-        api_data = SessionChatCompletionAPIData(
-            messages=[ChatMessage(role="user", content="test")],
-            max_tokens=50,
-            event_id="test:event_0",
-            registry=registry,
-            worker_tracker=tracker,
-            completion_queue=None,
-            total_events_in_session=1,
-        )
 
         # Test cases that SHOULD be detected as duplicates
-        assert api_data._is_duplicate_session("session_001_dup1") is True
-        assert api_data._is_duplicate_session("my_session_dup123") is True
-        assert api_data._is_duplicate_session("trace_abc_dup5") is True
-        assert api_data._is_duplicate_session("a_dup999") is True
+        assert SessionGenerator.is_duplicate_session("session_001_dup1") is True
+        assert SessionGenerator.is_duplicate_session("my_session_dup123") is True
+        assert SessionGenerator.is_duplicate_session("trace_abc_dup5") is True
+        assert SessionGenerator.is_duplicate_session("a_dup999") is True
 
         # Test cases that should NOT be detected as duplicates
-        assert api_data._is_duplicate_session("session_001") is False
-        assert api_data._is_duplicate_session("my_dup_session") is False  # _dup not at end
-        assert api_data._is_duplicate_session("session_dup") is False  # No number after _dup
-        assert api_data._is_duplicate_session("duplicate_session") is False  # Contains "dup" but wrong pattern
-        assert api_data._is_duplicate_session("session_001_duplicate") is False  # Wrong suffix
+        assert SessionGenerator.is_duplicate_session("session_001") is False
+        assert SessionGenerator.is_duplicate_session("my_dup_session") is False  # _dup not at end
+        assert SessionGenerator.is_duplicate_session("session_dup") is False  # No number after _dup
+        assert SessionGenerator.is_duplicate_session("duplicate_session") is False  # Contains "dup" but wrong pattern
+        assert SessionGenerator.is_duplicate_session("session_001_duplicate") is False  # Wrong suffix
