### 2026-04-28 7bae372
add test build replay schedule with duplicate session and random session id injection
