+++ tests/test_otel_replay_datagen.py
@@ -50,6 +50,7 @@ from inference_perf.datagen.replay_graph_session_datagen import (
     EventFailedError,
     EventOutputRegistry,
     ReplayGraphSessionGeneratorBase,
+    ReplaySession,
     ReplaySessionEvent,
     ReplaySessionState,
     SessionChatCompletionAPIData,
@@ -61,9 +62,9 @@ from inference_perf.datagen.otel_trace_to_replay_graph import (
     build_graph,
     build_raw_calls,
 )
-from inference_perf.datagen.replay_graph_types import InputSegment
+from inference_perf.datagen.replay_graph_types import GraphCall, GraphEvent, InputSegment, ReplayGraph
 from inference_perf.apis.chat import ChatMessage
-from inference_perf.config import APIConfig, APIType
+from inference_perf.config import APIConfig, APIType, SessionReplayConfig
 
 
 # ---------------------------------------------------------------------------
@@ -1326,3 +1327,178 @@ class TestRandomSessionIDInjection:
             ReplayGraphSessionGeneratorBase.is_duplicate_session("duplicate_session") is False
         )  # Contains "dup" but wrong pattern
         assert ReplayGraphSessionGeneratorBase.is_duplicate_session("session_001_duplicate") is False  # Wrong suffix
+
+
+# ---------------------------------------------------------------------------
+# _build_replay_schedule: random session ID assignment for duplicates
+# ---------------------------------------------------------------------------
+
+
+def _make_simple_graph(event_count: int = 1) -> ReplayGraph:
+    """Build a minimal ReplayGraph with `event_count` independent root events."""
+    events: dict[str, GraphEvent] = {}
+    for i in range(event_count):
+        eid = f"event_{i}"
+        events[eid] = GraphEvent(
+            event_id=eid,
+            call=GraphCall(
+                call_id=f"call_{i}",
+                model="test-model",
+                messages=[{"role": "user", "content": f"msg {i}"}],
+                expected_output="ok",
+                input_segments=[InputSegment(type="unique", message_count=1, token_count=5)],
+                total_input_tokens=5,
+                expected_output_tokens=2,
+                temperature=None,
+                max_tokens_recorded=None,
+            ),
+            predecessor_event_ids=[],
+            predecessor_dependency_types={},
+            wait_ms=0,
+            t_start_ms=i * 1000,
+            t_end_ms=i * 1000 + 500,
+        )
+    return ReplayGraph(
+        events=events,
+        root_event_ids=[f"event_{i}" for i in range(event_count)],
+        source_file="test",
+    )
+
+
+def _make_generator(
+    replay_config: Optional[SessionReplayConfig] = None,
+) -> ReplayGraphSessionGeneratorBase:
+    """Create a ReplayGraphSessionGeneratorBase bypassing __init__."""
+    gen = object.__new__(ReplayGraphSessionGeneratorBase)
+    gen.replay_config = replay_config
+    gen.output_registry = EventOutputRegistry()
+    gen.worker_tracker = WorkerSessionTracker()
+    gen.session_completion_queue = None
+    gen.sessions = []
+    gen.session_graph_state = {}
+    gen.all_events = []
+    return gen
+
+
+class TestBuildReplayScheduleRandomSessionID:
+    """Verify _build_replay_schedule assigns random_string correctly for duplicate sessions."""
+
+    def test_duplicate_sessions_get_random_string_without_flag(self) -> None:
+        """Duplicate sessions (suffix _dup{N}) get a random_string even when inject_random_session_id=False."""
+        graph = _make_simple_graph()
+        gen = _make_generator(replay_config=SessionReplayConfig(inject_random_session_id=False))
+        gen.sessions = [
+            ReplaySession(session_id="sess_001_dup1", source_id="src", session_index=0, graph=graph),
[…]
