### 2026-04-28 76e491f
fix: harden tool schema cleaning for vLLM/xgrammar compatibility
Remove numeric constraints (minimum, maximum) from the allowed JSON
Schema keys, coerce non-boolean additionalProperties to false, and
inject empty properties for type:object schemas missing them.

OTel traces from real workloads include these JSON Schema features in
tool definitions, but vLLM's xgrammar backend rejects them. Since we
only need the server to accept the schema and generate tokens, stripping
them has no impact on replay fidelity.
