+++ inference_perf/apis/chat.py
@@ -41,6 +41,11 @@ from inference_perf.utils.custom_tokenizer import CustomTokenizer
 
 # Fields allowed by the OpenAI/vLLM JSON Schema validator for tool parameters.
 # Extra fields generated by Pydantic (e.g. 'default', 'title') are stripped.
+# Numeric constraints (minimum, maximum, exclusiveMinimum, exclusiveMaximum) and
+# array constraints (minItems, maxItems) are intentionally excluded: vLLM's
+# xgrammar backend and the llm-d-inference-sim both reject them as unsupported
+# additionalProperties. They are irrelevant for load testing — we only need the
+# server to accept the tool schema and generate tokens.
 _ALLOWED_JSON_SCHEMA_KEYS = {
     "type",
     "description",
@@ -54,8 +59,6 @@ _ALLOWED_JSON_SCHEMA_KEYS = {
     "$ref",
     "const",
     "additionalProperties",
-    "minimum",
-    "maximum",
 }
 
 
@@ -76,6 +79,11 @@ def _clean_schema(obj: Any, is_properties_map: bool = False) -> Any:
         for k, v in obj.items():
             if k not in _ALLOWED_JSON_SCHEMA_KEYS:
                 continue
+            # vLLM/simulator require additionalProperties to be a boolean.
+            # Object-valued schemas (e.g. {"type": "string"}) are valid JSON
+            # Schema but rejected by the server — coerce to false.
+            if k == "additionalProperties" and not isinstance(v, bool):
+                v = False
             cleaned[k] = _clean_schema(v, is_properties_map=(k == "properties"))
         # vLLM requires a plain 'type' on every property schema and does not
         # support anyOf/oneOf. Collapse to 'type': 'string' as a safe fallback.
@@ -86,6 +94,11 @@ def _clean_schema(obj: Any, is_properties_map: bool = False) -> Any:
         elif "anyOf" in cleaned or "oneOf" in cleaned:
             cleaned.pop("anyOf", None)
             cleaned.pop("oneOf", None)
+        # vLLM/simulator require "type": "object" schemas to include
+        # "properties". Map-like schemas (using only additionalProperties /
+        # propertyNames) lack it — inject an empty one.
+        if cleaned.get("type") == "object" and "properties" not in cleaned:
+            cleaned["properties"] = {}
         return cleaned
     if isinstance(obj, list):
         return [_clean_schema(v) for v in obj]
