### 2026-04-29 8d26860
add lllm-d-benchmark unified benchmark report generation
