+++ inference_perf/reportgen/br/v0_2/collectors/k8s/per_pod_prom_scraper.py
@@ -1,0 +1,182 @@
+# Copyright 2026 The Kubernetes Authors.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+"""Per-pod Prometheus scraping.
+
+Each pod's /metrics endpoint is scraped on an interval. Samples are bucketed
+per-pod, aggregated at collect() time into the BR0.2
+observability.components[] resource metrics + the ad-hoc `vllm_*` / `epp_*`
+fields seen in the upstream example. No PromQL server required.
+"""
+
+import asyncio
+import logging
+from collections import defaultdict
+from typing import Any, Dict, List, NamedTuple, Optional, Tuple
+
+import aiohttp
+import numpy as np
+
+logger = logging.getLogger(__name__)
+
+
+class PodScrapeTarget(NamedTuple):
+    pod_name: str
+    component_label: str
+    role: str
+    url: str
+
+
+# Metrics we know how to project into BR0.2. Each entry: (prom_name, bucket_key, units).
+_TRACKED_METRICS: Tuple[Tuple[str, str, str], ...] = (
+    ("vllm:gpu_cache_usage_perc", "vllm_kv_cache_usage_perc", "percent"),
+    ("vllm:num_requests_running", "vllm_num_requests_running", "count"),
+    ("vllm:num_requests_waiting", "vllm_num_requests_waiting", "count"),
+    ("vllm:num_preemptions_total", "vllm_num_preemptions_total", "count"),
+    ("vllm:prefix_cache_hits_total", "vllm_prefix_cache_hits_total", "tokens"),
+    ("vllm:prefix_cache_queries_total", "vllm_prefix_cache_queries_total", "tokens"),
+    (
+        "inference_extension:inference_pool:average_kv_cache_utilization",
+        "epp_pool_avg_kv_cache_utilization",
+        "percent",
+    ),
+    ("inference_extension:inference_pool:average_queue_size", "epp_pool_avg_queue_size", "count"),
+    ("inference_extension:inference_pool:ready_pods", "epp_pool_ready_pods", "count"),
+)
+# Additional categories of the schema that a future, richer mapping should fill:
+#   - GPU utilization / memory / power (require DCGM-exporter)
+#   - vllm:nixl_xfer_time_seconds_{sum,count} (histograms)
+
+
+class PerPodPromScraper:
+    """Direct-scrape Prometheus text from each pod's /metrics endpoint.
+
+    Targets are supplied by the composing collector (it knows which pods are
+    in the stack). The scraper is safe to start with an empty target list.
+    """
+
+    def __init__(
+        self,
+        targets: Optional[List[PodScrapeTarget]] = None,
+        interval_seconds: float = 15.0,
+        request_timeout_seconds: float = 5.0,
+    ) -> None:
+        self.targets: List[PodScrapeTarget] = list(targets or [])
+        self.interval_seconds = max(1.0, interval_seconds)
+        self.request_timeout_seconds = max(1.0, request_timeout_seconds)
+        self._task: Optional[asyncio.Task[None]] = None
+        # samples[bucket_key][pod_name] = list[float] (one entry per scrape tick)
+        self._samples: Dict[str, Dict[str, List[float]]] = defaultdict(lambda: defaultdict(list))
[…]
