### 2026-04-30 67c7bab
fix: count output tokens from server usage, including tool_calls
Previously output_len only counted tokens from message.content, reporting
0 for tool-call responses. Now prefers server-reported usage.completion_tokens
and falls back to counting content + serialized tool_calls.
