+++ inference_perf/datagen/replay_graph_session_datagen.py
@@ -602,7 +602,14 @@ class SessionChatCompletionAPIData(ChatCompletionAPIData):
 
             prompt_text = "".join([_get_text(msg.content) for msg in self.messages if msg.content])
             prompt_len = tokenizer.count_tokens(prompt_text)
-            output_len = tokenizer.count_tokens(output_text)
+            server_completion_tokens = server_usage.get("completion_tokens") if server_usage else None
+            if server_completion_tokens is not None:
+                output_len = int(server_completion_tokens)
+            else:
+                tc_text = ""
+                if tool_call_chunks:
+                    tc_text = json.dumps([tool_call_chunks[i] for i in sorted(tool_call_chunks)], ensure_ascii=False)
+                output_len = tokenizer.count_tokens(output_text + tc_text)
             info = SessionInferenceInfo(
                 request_metrics=RequestMetrics(text=Text(input_tokens=prompt_len)),
                 response_metrics=StreamedResponseMetrics(
@@ -622,6 +629,7 @@ class SessionChatCompletionAPIData(ChatCompletionAPIData):
             prompt_len = tokenizer.count_tokens("".join([_get_text(m.content) for m in self.messages]))
             choices = data.get("choices", [])
             output_message: Optional[Dict[str, Any]] = None
+            tool_calls = None
             if choices:
                 msg_dict = choices[0].get("message", {})
                 output_text = msg_dict.get("content", "") or ""
@@ -632,7 +640,15 @@ class SessionChatCompletionAPIData(ChatCompletionAPIData):
                         output_message["content"] = output_text
                 else:
                     output_message = {"role": "assistant", "content": output_text}
-            output_len = tokenizer.count_tokens(output_text)
+            usage = data.get("usage") or {}
+            server_completion_tokens = usage.get("completion_tokens")
+            if server_completion_tokens is not None:
+                output_len = int(server_completion_tokens)
+            else:
+                tc_text = ""
+                if tool_calls:
+                    tc_text = json.dumps(tool_calls, ensure_ascii=False)
+                output_len = tokenizer.count_tokens(output_text + tc_text)
             info = SessionInferenceInfo(
                 request_metrics=RequestMetrics(text=Text(input_tokens=prompt_len)),
                 response_metrics=UnaryResponseMetrics(output_tokens=output_len),
