### 2026-04-30 848e28f
feat: add max_wait_ms config to cap inter-event delays in trace replay
Limits the wait time between predecessor completion and event dispatch,
preventing unusually long tool/agent execution times from the original
trace from inflating replay duration. Defaults to 15000ms.
