### 2026-04-30 071f9b5
align tool call/use formatting
