### 2026-05-01 bc0b50a
Update code generation config in the catalog
