+++ workload-catalog/code-generation/inference-perf.yaml
@@ -26,17 +26,17 @@ data:
     num_conversations: 20
     shared_system_prompt_len: 3000
     dynamic_system_prompt_len:
-      type: uniform
-      min: 15000
-      max: 120000
-      mean: 55000
-      std_dev: 25000
+      type: lognormal
+      min: 10000
+      max: 990000
+      mean: 160000
+      std_dev: 233600
     turns_per_conversation:
-      type: normal
-      min: 2
-      max: 20
-      mean: 6
-      std_dev: 3.5
+      type: lognormal
+      min: 1
+      max: 3000
+      mean: 540
+      std_dev: 48600
     input_tokens_per_turn:
       type: lognormal
       min: 100
@@ -46,12 +46,12 @@ data:
     output_tokens_per_turn:
       type: lognormal
       min: 50
-      max: 4000
-      mean: 800
-      std_dev: 600
+      max: 10000
+      mean: 425
+      std_dev: 825
     tool_call_latency_sec:
       type: lognormal
-      min: 15
-      max: 120
-      mean: 45
-      std_dev: 25
+      min: 1
+      max: 1000
+      mean: 15
+      std_dev: 55
