### 2026-05-03 1143ffd
Update and improve otel examples (content and config). set cache dir for traces from hf
