+++ examples/otel/configs/per_case_config/exgentic_tau2_caude_code.yml
@@ -1,43 +1,0 @@
-# Config for code_review_workflow.json - Code review workflow test case
-
-api:
-  type: chat
-  streaming: false
-
-data:
-  type: otel_trace_replay
-  otel_trace_replay:
-    trace_files:
-      - "examples/otel/test_traces/exgentic/tau2_retail_claude_code_315397.json"
-
-    use_static_model: true
-    static_model_name: "HuggingFaceTB/SmolLM2-135M-Instruct"
-
-    default_max_tokens: 150
-
-    include_errors: false
-    skip_invalid_files: true
-
-load:
-  type: trace_session_replay
-  stages:
-    - concurrent_sessions: 1  # Sequential execution
-  num_workers: 2
-  worker_max_concurrency: 10
-
-server:
-  type: vllm
-  base_url: "http://localhost:8000"
-  model_name: "HuggingFaceTB/SmolLM2-135M-Instruct"
-
-storage:
-  local_storage:
-    path: "reports-exgentic-tau2-claude-code"
-
-report:
-  request_lifecycle:
-    summary: true
-    per_stage: false
-    per_request: false
-
-
