### 2026-05-03 5506a2c
Add logic for max-model-len handling
