### 2026-05-04 5a9e8e8
minor changes in extract messages and added cli flag
