### 2026-05-04 bdd667e
Fix non-deterministic event counts after session shuffle
Reassign session_index to match list position after shuffling/duplication
to ensure consistent event counting in multi-stage replay.
