### 2026-05-05 94ecdab
Fix tool call replay: disable ignore_eos and raise max_tokens for tool call requests
