### 2026-05-05 70db6dc
Add regression test for conversation_replay session re-prime
Covers PR #466's fix: after LoadGenerator clears LocalUserSession._instances
between stages, load_lazy_data must re-register the touched slot with its
original system_prompt. Without the fix, the assertion that conv_1 is back
in the registry fails.

Also adds an autouse fixture to isolate _instances across tests, and a
companion test that guards against the re-prime path clobbering a live
session mid-stage.
