### 2026-05-05 ae96700
Test system_prompt stays identical across repeated stage clears
Runs 4 simulated stage transitions and asserts each re-primed session's
context matches the blueprint's original system_prompt every time, not
just on the first clear.
