### 2026-05-05 3c7c4a3
Re-prime conversation_replay sessions after clear_instances() (#466)
When LoadGenerator clears LocalUserSession._instances between load
stages (PR #459), conversation_replay's user_sessions[].user_session_id
no longer resolves in the registry. The next get_instance() lazily
creates a fresh session with empty context, dropping the
per-conversation system_prompt and producing dramatically smaller
prompts in stages 1+.

Detect the cleared registry at the top of load_lazy_data and re-register
the session with its original system_prompt before dispatching.
