### 2026-05-06 a2c1b33
Fix session replay hang when server is down
Always cancel+drain queue after session stage completes, not only on
stage failure. When a session fails, the completion notification reaches
the main process before the worker picks up all items from the queue,
leaving stranded items without task_done() and blocking join() forever.
