### 2026-05-06 f9f0b11
Fix tool_call_id missing from role:tool messages in replay
