### 2026-05-06 0d2829e
bug fix in shared prefix substitution - In a segment with x messages, take only first x messages from parted for the substitution.
