### 2026-05-06 5b006cd
add support also for output messages are not passed as strings that represent jsons.
