### 2026-05-06 55db6f5
Fix queue.qsize() NotImplementedError on macOS
