### 2026-05-07 3e22ac0
Feat: update hf support + example config file with exgentic dataset.
