### 2026-05-09 #10314 @neerazz
Re-reviewed against the regenerated PDF and current main. All nine issues I filed are addressed. Recommending acceptance from my side. Deferring to @vipulg13 for their independent review.
