### 2026-05-10 ecfa306
Fix format/list errors + update cli_flags.md
