+++ inference_perf/datagen/otel_trace_replay_datagen.py
@@ -126,11 +126,11 @@ def resolve_trace_files(trace_files: List[str]) -> List[Path]:
 
 def _validate_hf_dataset_schema(dataset: Any, dataset_path: str) -> None:
     """Validate that the HuggingFace dataset has the expected schema for OTel traces.
-    
+
     Args:
         dataset: The loaded HuggingFace dataset
         dataset_path: Dataset identifier for error messages
-        
+
     Raises:
         ValueError: If schema validation fails
     """
@@ -138,21 +138,21 @@ def _validate_hf_dataset_schema(dataset: Any, dataset_path: str) -> None:
     # Official dataset: Exgentic/agent-llm-traces
     # Custom datasets must follow this schema to be compatible
     EXPECTED_SCHEMA = {
-        'harness': str,
-        'benchmark': str,
-        'models': list,
-        'session_id': str,
-        'spans': list,  # Critical field - must be present
+        "harness": str,
+        "benchmark": str,
+        "models": list,
+        "session_id": str,
+        "spans": list,  # Critical field - must be present
     }
-    
+
     if len(dataset) == 0:
         raise ValueError(f"No data found in HuggingFace dataset '{dataset_path}'")
-    
+
     # Validate schema by checking the first record
     first_record = dataset[0]
     missing_fields = []
     type_mismatches = []
-    
+
     for field_name, expected_type in EXPECTED_SCHEMA.items():
         if field_name not in first_record:
             missing_fields.append(field_name)
@@ -163,34 +163,33 @@ def _validate_hf_dataset_schema(dataset: Any, dataset_path: str) -> None:
                 type_mismatches.append(
                     f"  - '{field_name}': expected {expected_type.__name__}, got {type(actual_value).__name__}"
                 )
-    
+
     # Raise error if schema validation fails
     if missing_fields or type_mismatches:
         error_parts = [
             f"Dataset schema validation failed for '{dataset_path}'.",
-            f"Official supported dataset: 'Exgentic/agent-llm-traces'",
-            f"Custom datasets must follow the same schema to be compatible.",
+            "Official supported dataset: 'Exgentic/agent-llm-traces'",
+            "Custom datasets must follow the same schema to be compatible.",
         ]
-        
+
         if missing_fields:
             error_parts.append(f"\nMissing required fields: {', '.join(missing_fields)}")
-        
+
         if type_mismatches:
             error_parts.append("\nType mismatches:")
             error_parts.extend(type_mismatches)
-        
+
         error_parts.append(f"\nExpected schema: {list(EXPECTED_SCHEMA.keys())}")
         error_parts.append(f"Actual fields: {list(first_record.keys())}")
-        
+
         raise ValueError("\n".join(error_parts))
-    
+
     # Critical validation: ensure 'spans' field exists and is not empty in at least one record
-    if 'spans' not in first_record:
+    if "spans" not in first_record:
         raise ValueError(
-            f"Dataset '{dataset_path}' is missing the critical 'spans' field. "
-            f"This field is required for OTel trace replay."
+            f"Dataset '{dataset_path}' is missing the critical 'spans' field. This field is required for OTel trace replay."
         )
-    
+
     logger.info(f"Schema validation passed for dataset '{dataset_path}'")
 
 
@@ -214,30 +213,32 @@ def _download_hf_dataset(dataset_config: Union[str, Dict[str, Any]]) -> List[Dic
         dataset_path = dataset_config
         load_kwargs = {}  # No additional kwargs for string format
     elif isinstance(dataset_config, dict):
-        dataset_path = dataset_config.get("path")
-        if not dataset_path:
+        dataset_path_value = dataset_config.get("path")
[…]
