### 2026-05-11 b938f01
Limit number of traces in exgenetic example; add comments explaining fields
