### 2026-05-11 f23cb85
Add Multimodal Benchmarking (#450)
## Multi-Modal Benchmarking

Inference-perf now supports image, video, and audio payloads for
chat-style
benchmarks. Datagen samples typed multimodal specs; the API
implementation
materializes wire bytes at request-build time so heavy media never lives
on
per-request lifecycle metrics.

### What's new

- **New `data.multimodal` config block** with `image`, `video`, and
`audio`
  sub-configs (count, insertion point, resolution, profile, duration
  distributions). Works under `data.type: synthetic` and under
  `shared_prefix` (both prefix-side and payload-side)
[…]
