+++ inference_perf/client/modelserver/mock_client.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from inference_perf.client.requestdatacollector import RequestDataCollector
+from inference_perf.metrics.request_collector import RequestMetricCollector
 from typing import List, Optional
 from inference_perf.config import APIConfig, APIType
 from inference_perf.apis import (
@@ -20,8 +20,9 @@ from inference_perf.apis import (
     InferenceInfo,
     RequestLifecycleMetric,
     ErrorResponseInfo,
-    UnaryInferenceResponseInfo,
+    UnaryResponseMetrics,
 )
+from inference_perf.payloads import RequestMetrics, Text
 from .base import ModelServerClient, ModelServerPrometheusMetric, PrometheusMetricMetadata
 import asyncio
 import time
@@ -33,7 +34,7 @@ logger = logging.getLogger(__name__)
 class MockModelServerClient(ModelServerClient):
     def __init__(
         self,
-        metrics_collector: RequestDataCollector,
+        metrics_collector: RequestMetricCollector,
         api_config: APIConfig,
         timeout: Optional[float] = None,
         mock_latency: float = 1,
@@ -58,18 +59,17 @@ class MockModelServerClient(ModelServerClient):
                     await asyncio.sleep(self.mock_latency)
 
                 info = InferenceInfo(
-                    input_tokens=0,
-                    output_tokens=0,
+                    request_metrics=RequestMetrics(text=Text(input_tokens=0)),
                     lora_adapter=lora_adapter,
                 )
                 data.on_completion(info)
                 self.metrics_collector.record_metric(
                     RequestLifecycleMetric(
                         stage_id=stage_id,
-                        request_data=str(await data.to_payload(effective_model_name, 3, False, False)),
+                        request_data=str(await data.to_request_body(effective_model_name, 3, False, False)),
                         info=InferenceInfo(
-                            input_tokens=0,
-                            response_info=UnaryInferenceResponseInfo(output_tokens=0),
+                            request_metrics=RequestMetrics(text=Text(input_tokens=0)),
+                            response_metrics=UnaryResponseMetrics(output_tokens=0),
                             lora_adapter=lora_adapter,
                         ),
                         error=None,
@@ -83,10 +83,10 @@ class MockModelServerClient(ModelServerClient):
             self.metrics_collector.record_metric(
                 RequestLifecycleMetric(
                     stage_id=stage_id,
-                    request_data=str(data.to_payload(effective_model_name, 3, False, False)),
+                    request_data=str(data.to_request_body(effective_model_name, 3, False, False)),
                     info=InferenceInfo(
-                        input_tokens=0,
-                        response_info=UnaryInferenceResponseInfo(output_tokens=0),
+                        request_metrics=RequestMetrics(text=Text(input_tokens=0)),
+                        response_metrics=UnaryResponseMetrics(output_tokens=0),
                         lora_adapter=lora_adapter,
                     ),
                     error=ErrorResponseInfo(
