### 2026-05-11 80e7a19
Fix test_main.py patch target after collector rename
The multimodal PR (#450) renamed MultiprocessRequestDataCollector to
MultiprocessRequestMetricCollector. Update the patch targets in
tests/test_main.py to match so the unit tests pass.

Co-Authored-By: Claude Opus 4.7 (1M context) <noreply@anthropic.com>
