### 2026-05-11 d8c3408
autogenerate config docs:
