+++ scripts/sync_config_reference_doc.py
@@ -1,0 +1,257 @@
+"""Generate docs/config-reference.md from the Pydantic models in inference_perf.config.
+
+Discovers every BaseModel reachable from `Config`, emits one section per model with a
+table of fields (type, default, description), and enforces that every field has a
+non-empty description. Run via:
+
+    pdm run update:config-docs   # regenerate the doc
+    pdm run check:config-docs    # fail if doc is stale or descriptions are missing
+"""
+
+from __future__ import annotations
+
+import argparse
+import difflib
+import sys
+import typing
+from enum import Enum
+from pathlib import Path
+from typing import Any
+
+from pydantic import BaseModel
+from pydantic.fields import FieldInfo
+
+from inference_perf import config as config_module
+from inference_perf.config import Config
+
+DOC_PATH = Path("docs/config-reference.md")
+
+HEADER = """# Configuration Reference
+
+<!--
+  This file is auto-generated from the Pydantic models in inference_perf/config.py.
+  Do not edit by hand. Run `pdm run update:config-docs` after changing the schema.
+-->
+
+This reference enumerates every configuration option, generated directly from the
+schema. For tutorial-style examples and prose, see [config.md](config.md).
+
+Each section corresponds to one model. Nested models are linked.
+
+"""
+
+
+def _model_anchor(model_cls: type[BaseModel]) -> str:
+    return model_cls.__name__.lower()
+
+
+def _is_basemodel(tp: Any) -> bool:
+    return isinstance(tp, type) and issubclass(tp, BaseModel)
+
+
+def _is_enum(tp: Any) -> bool:
+    return isinstance(tp, type) and issubclass(tp, Enum)
+
+
+def _collect_models(root: type[BaseModel], seen: dict[str, type[BaseModel]]) -> None:
+    if root.__name__ in seen:
+        return
+    seen[root.__name__] = root
+    for field in root.model_fields.values():
+        for tp in _iter_annotated_types(field.annotation):
+            if _is_basemodel(tp):
+                _collect_models(tp, seen)
+
+
+def _iter_annotated_types(annotation: Any) -> list[Any]:
+    """Walk a type annotation and yield every concrete leaf type it references."""
+    out: list[Any] = []
+    stack: list[Any] = [annotation]
+    while stack:
+        tp = stack.pop()
+        if tp is None or tp is type(None):
+            continue
+        origin = typing.get_origin(tp)
+        if origin is not None:
+            stack.extend(typing.get_args(tp))
+            continue
+        out.append(tp)
+    return out
+
+
+def _render_type(annotation: Any) -> str:
+    """Render a type annotation as readable markdown."""
+    if annotation is None or annotation is type(None):
+        return "None"
+
+    origin = typing.get_origin(annotation)
+    args = typing.get_args(annotation)
+
+    if origin is typing.Union:
+        non_none = [a for a in args if a is not type(None)]
+        rendered = " \\| ".join(_render_type(a) for a in non_none)
+        if len(non_none) < len(args):
+            rendered = f"{rendered} (optional)"
+        return rendered
+
+    if origin in (list, typing.List):
+        inner = _render_type(args[0]) if args else "Any"
+        return f"List[{inner}]"
+
+    if origin in (dict, typing.Dict):
+        if len(args) == 2:
+            return f"Dict[{_render_type(args[0])}, {_render_type(args[1])}]"
+        return "Dict"
+
+    if origin in (tuple, typing.Tuple):
+        return "Tuple[" + ", ".join(_render_type(a) for a in args) + "]"
+
+    if typing.get_origin(annotation) is typing.Literal or origin is typing.Literal:
[…]
