### 2026-05-12 05bc546
Surface prompt cache token metrics
