+++ inference_perf/payloads/__init__.py
@@ -11,49 +11,133 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from typing import Any
+"""Modality data contracts: pre-flight specs and post-flight metric records.
 
-from .multimodal_spec import (
-    AudioInstanceSpec,
-    ImageRepresentation,
-    ImageInstanceSpec,
-    MultimodalSpec,
-    VideoInstanceSpec,
-    VideoRepresentation,
-)
-from .payload import (
+Modality is the primary organizing axis — each per-modality subpackage
+co-locates the spec types (the *plan*: sampled by datagen, consumed by the
+materializer) and the metric types (the *measurement*: recorded by the
+materializer, consumed by reportgen) for that modality. Inside each modality
+package:
+
+- ``spec/`` — pre-flight request specs, factored modality-primary /
+  provenance-secondary (synthetic / pre_encoded / remote / local). The video
+  variants are unioned into :data:`VideoSpecUnion`.
+- ``metrics.py`` — post-flight measurement records for that modality.
+
+The text modality has no spec types — text-side request fields live on
+:class:`InferenceAPIData` subclasses — and only contributes :class:`Text` to
+the metric aggregator.
+
+This module holds the two cross-modality aggregators that mirror each other
+field-for-field:
+
+- :class:`MultimodalSpec` — request-scoped container of per-modality specs.
+- :class:`RequestMetrics` — request-scoped container of per-modality metrics
+  plus :class:`Text`.
+
+:class:`VideoRepresentation` lives in ``video/spec/base.py`` as the
+user-facing wire-format enum; Spec classes themselves don't carry it.
+"""
+
+from typing import Any, List, Optional
+
+from pydantic import BaseModel
+
+from .audio import (
     Audio,
     Audios,
-    Image,
-    Images,
-    RequestMetrics,
-    Text,
-    Video,
-    Videos,
+    AudioSpec,
+    AudioSpecUnion,
+    LocalFileAudioSpec,
+    PreEncodedAudioSpec,
+    RemoteAudioSpec,
+    SyntheticAudioSpec,
 )
+from .image import (
+    Image,
+    ImageRepresentation,
+    Images,
+    ImageSpec,
+    ImageSpecUnion,
+    LocalFileImageSpec,
+    PreEncodedImageSpec,
+    RemoteImageSpec,
+    SyntheticImageSpec,
+)
+from .text import Text
+from .video import (
+    LocalFileVideoSpec,
+    PreEncodedFramesVideoSpec,
+    RemoteVideoSpec,
+    SyntheticFramesVideoSpec,
+    SyntheticMp4VideoSpec,
+    Video,
+    VideoRepresentation,
+    Videos,
+    VideoSpec,
+    VideoSpecUnion,
+)
+
+
+class MultimodalSpec(BaseModel):
+    """Aggregate container of every modality's pre-flight specs on one request."""
+
+    images: List[ImageSpecUnion] = []
+    videos: List[VideoSpecUnion] = []
+    audios: List[AudioSpecUnion] = []
+
+
+class RequestMetrics(BaseModel):
+    """Aggregate container of every modality's post-flight metrics on one request."""
+
+    text: Text
+    image: Optional[Images] = None
+    video: Optional[Videos] = None
+    audio: Optional[Audios] = None
+
 
 # In-flight wire body returned by ``InferenceAPIData.to_request_body`` and
[…]
