+++ inference_perf/config.py
@@ -23,7 +23,7 @@ import yaml
 from pydantic import AliasChoices, BaseModel, ConfigDict, Field, HttpUrl, model_validator
 
 from inference_perf.circuit_breaker import CircuitBreakerConfig
-from inference_perf.payloads.multimodal_spec import ImageRepresentation, VideoRepresentation
+from inference_perf.payloads import ImageRepresentation, VideoRepresentation
 
 
 class APIType(Enum):
@@ -90,6 +90,7 @@ class DataGenType(Enum):
     BillsumConversations = "billsum_conversations"
     OTelTraceReplay = "otel_trace_replay"
     ConversationReplay = "conversation_replay"
+    ShareGPT4Video = "sharegpt4video"
 
 
 class DistributionType(str, Enum):
@@ -395,6 +396,95 @@ class OTelTraceReplayConfig(SessionReplayConfig):
         return self
 
 
+class GatedHFDatasetConfig(BaseModel):
+    """Shared config base for loaders backed by gated HuggingFace datasets."""
+
+    token: Optional[str] = Field(
+        default=None,
+        description=(
+            "HuggingFace access token used to download the gated dataset. "
+            "Falls back to ``HF_TOKEN`` / ``HUGGING_FACE_HUB_TOKEN`` env vars when omitted."
+        ),
+    )
+
+
+class ShareGPT4VideoConfig(GatedHFDatasetConfig):
+    """Configuration for the ShareGPT4Video dataset loader.
+
+    The loader streams captions + keyframe indices from the gated HuggingFace
+    dataset and fetches the source MP4 zips (hosted on the same dataset repo
+    under ``zip_folder/<source>/``) into a local cache directory via a
+    background thread. The hot request path never blocks on the network — each
+    request samples from videos that are already on disk, and the on-disk pool
+    grows over time as background downloads complete. Init blocks only until
+    the first zip has been extracted so there is always at least one usable
+    video when load generation begins.
+
+    Each request decodes the configured keyframes via PyAV, resizes to
+    ``target_resolution``, re-encodes as PNG or JPEG, and emits the frames in
+    the Frames wire format (one ``image_url`` block per frame at a single
+    insertion point) — compatible with VLMs that don't accept ``video_url``
+    natively.
+
+    .. warning::
+       The HF repo is ~1.5 TB total; a single source zip is 15-21 GB. The
+       background downloader will keep pulling zips as the benchmark runs, so
+       monitor disk usage. Pre-populate ``cache_dir`` to skip downloads
+       entirely. Set ``HF_HUB_ENABLE_HF_TRANSFER=1`` for faster parallel
+       downloads if you want more variety sooner.
+    """
+
+    cache_dir: Optional[str] = Field(
+        default=None,
+        description=(
+            "Directory used for both the HuggingFace datasets cache and the extracted "
+            "video files. Defaults to ``./sharegpt4video_cache`` under the process's "
+            "current working directory. The loader logs the resolved absolute path at "
+            "startup so users can pre-populate it."
+        ),
+    )
+    representation: VideoRepresentation = Field(
+        default=VideoRepresentation.JPEG_FRAMES,
+        description=(
+            "Frame wire encoding: ``png_frames`` (lossless) or ``jpeg_frames`` (smaller payload). "
+            "``mp4`` is not supported here — this loader is Frames-only for v1."
+        ),
+    )
+    target_resolution: AnyResolution = Field(
+        default=ResolutionPreset.P720,
+        description="Frames are resized to this resolution before encoding.",
+    )
+    max_frames_per_request: int = Field(
[…]
