### 2026-05-12 d20069e
add MMMU dataset loader
Real-dataset loader for the gated HuggingFace MMMU corpus
(https://huggingface.co/datasets/MMMU/MMMU), a college-level VLM evaluation
benchmark spanning 30 subjects across 6 disciplines. Streams examples,
extracts embedded PIL images, re-encodes them as PNG or JPEG, and emits
chat-completion requests with question + options text plus 1-7 image_url
blocks per request. Document- and diagram-heavy workload shape that the
synthetic path doesn't exercise well.

Wires PreEncodedImageSpec into the chat.py materializer (image loop now
isinstance-dispatches on the discriminated 
[…]
