+++ inference_perf/datagen/mmmu_datagen.py
@@ -1,0 +1,238 @@
+# Copyright 2026 The Kubernetes Authors.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+"""MMMU real-dataset loader for multimodal benchmarking.
+
+MMMU (Massive Multi-discipline Multimodal Understanding,
+https://mmmu-benchmark.github.io/) is a college-level VLM evaluation
+benchmark spanning 30 subjects across 6 disciplines. The loader streams
+examples from the gated HuggingFace mirror at ``MMMU/MMMU``, extracts the
+embedded PIL images, re-encodes them, and emits chat-completion requests
+with the question + options as text and the images attached as
+:class:`PreEncodedImageSpec` blocks.
+
+Unlike :class:`ShareGPT4VideoDataGenerator`, MMMU is small enough to fit in
+memory — init loads all configured subjects synchronously and
+``load_lazy_data`` does an O(1) lookup. No background download thread.
+"""
+
+from __future__ import annotations
+
+import ast
+import io
+import logging
+from typing import Any, Generator, List, Optional, Tuple
+
+from PIL import Image as PILImage
+
+from inference_perf.apis.base import InferenceAPIData, LazyLoadInferenceAPIData
+from inference_perf.apis.chat import ChatCompletionAPIData, ChatMessage
+from inference_perf.config import APIConfig, APIType, DataConfig, MMMUConfig
+from inference_perf.datagen.multimodal_sampling import resolution_to_wh
+from inference_perf.payloads import (
+    ImageRepresentation,
+    MultimodalSpec,
+    PreEncodedImageSpec,
+)
+from inference_perf.utils.custom_tokenizer import CustomTokenizer
+
+from .base import DataGenerator, LazyLoadDataMixin
+from .gated_hf_dataset import GatedHFDataset
+
+logger = logging.getLogger(__name__)
+
+
+# MMMU canonical subject configs. The dataset is partitioned per-subject on
+# HuggingFace, so loading "all" means iterating these and concatenating. List
+# sourced from the dataset card at https://huggingface.co/datasets/MMMU/MMMU;
+# override via ``MMMUConfig.subjects`` if upstream changes.
+_MMMU_SUBJECTS: Tuple[str, ...] = (
+    "Accounting",
+    "Agriculture",
+    "Architecture_and_Engineering",
+    "Art",
+    "Art_Theory",
+    "Basic_Medical_Science",
+    "Biology",
+    "Chemistry",
+    "Clinical_Medicine",
+    "Computer_Science",
+    "Design",
+    "Diagnostics_and_Laboratory_Medicine",
+    "Economics",
+    "Electronics",
+    "Energy_and_Power",
+    "Finance",
+    "Geography",
+    "History",
+    "Literature",
+    "Manufacturing",
+    "Marketing",
+    "Materials",
+    "Math",
+    "Mechanical_Engineering",
+    "Music",
+    "Pharmacy",
+    "Physics",
+    "Psychology",
+    "Public_Health",
+    "Sociology",
+)
+
+_MAX_IMAGES_PER_EXAMPLE = 7
+
+
+class MMMUDataGenerator(DataGenerator, LazyLoadDataMixin, GatedHFDataset):
[…]
