+++ inference_perf/main.py
@@ -44,6 +44,7 @@ from inference_perf.datagen import (
     OTelTraceReplayDataGenerator,
     ConversationReplayDataGenerator,
     ShareGPT4VideoDataGenerator,
+    MMMUDataGenerator,
 )
 from inference_perf.client.modelserver import (
     ModelServerClient,
@@ -289,6 +290,7 @@ def main_cli() -> None:
                 DataGenType.OTelTraceReplay,
                 DataGenType.ConversationReplay,
                 DataGenType.ShareGPT4Video,
+                DataGenType.MMMU,
             }
         ):
             if tokenizer is None:
@@ -336,6 +338,9 @@ def main_cli() -> None:
         if config.data.type == DataGenType.ShareGPT4Video and config.data.sharegpt4video is None:
             raise Exception(f"{config.data.type.value} data generator requires 'sharegpt4video' to be configured")
 
+        if config.data.type == DataGenType.MMMU and config.data.mmmu is None:
+            raise Exception(f"{config.data.type.value} data generator requires 'mmmu' to be configured")
+
         if config.data.type == DataGenType.ShareGPT:
             datagen = HFShareGPTDataGenerator(config.api, config.data, tokenizer)
         elif config.data.type == DataGenType.CNNDailyMail:
@@ -361,6 +366,8 @@ def main_cli() -> None:
             )
         elif config.data.type == DataGenType.ShareGPT4Video:
             datagen = ShareGPT4VideoDataGenerator(config.api, config.data, tokenizer)
+        elif config.data.type == DataGenType.MMMU:
+            datagen = MMMUDataGenerator(config.api, config.data, tokenizer)
         else:
             datagen = MockDataGenerator(config.api, config.data, tokenizer)
     else:
