+++ tests/required/datagen/test_mmmu_datagen.py
@@ -1,0 +1,286 @@
+# Copyright 2026 The Kubernetes Authors.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+"""Tests for :class:`MMMUDataGenerator`.
+
+HuggingFace network access is mocked out — ``_load_dataset`` is monkey-patched
+to return synthetic in-memory example lists with PIL images, so these tests
+run offline and don't require an HF token.
+"""
+
+import base64
+import io
+from typing import Any, List
+from unittest.mock import MagicMock
+
+import pytest
+from PIL import Image as PILImage
+
+from inference_perf.apis.base import LazyLoadInferenceAPIData
+from inference_perf.apis.chat import ChatCompletionAPIData, ChatMessage
+from inference_perf.config import (
+    APIConfig,
+    APIType,
+    DataConfig,
+    DataGenType,
+    MMMUConfig,
+    Resolution,
+)
+from inference_perf.datagen.mmmu_datagen import MMMUDataGenerator
+from inference_perf.payloads import ImageRepresentation, PreEncodedImageSpec
+
+
+def _make_mock_tokenizer() -> MagicMock:
+    m = MagicMock()
+    m.count_tokens.return_value = 10
+    return m
+
+
+def _fake_example(
+    qid: str = "test/1",
+    n_images: int = 1,
+    options: str = "['Alpha', 'Beta', 'Gamma']",
+    question: str = "What is shown in <image 1>?",
+    image_size: tuple[int, int] = (32, 32),
+) -> dict[str, Any]:
+    ex: dict[str, Any] = {
+        "id": qid,
+        "question": question,
+        "options": options,
+        "answer": "A",
+    }
+    for i in range(1, 8):
+        ex[f"image_{i}"] = (
+            PILImage.new("RGB", image_size, color=(10 * i, 20 * i, 30 * i)) if i <= n_images else None
+        )
+    return ex
+
+
+def _load_first(gen: MMMUDataGenerator, data_index: int = 0) -> ChatCompletionAPIData:
+    lazy = LazyLoadInferenceAPIData(data_index=data_index)
+    item = gen.load_lazy_data(lazy)
+    assert isinstance(item, ChatCompletionAPIData)
+    return item
+
+
+def _build_generator(
+    monkeypatch: pytest.MonkeyPatch,
+    examples: List[dict[str, Any]],
+    **cfg_overrides: Any,
+) -> MMMUDataGenerator:
+    monkeypatch.setenv("HF_TOKEN", "test-token")
+    # Stub HF: return the same example list for every (subject, split) request.
+    monkeypatch.setattr(
+        MMMUDataGenerator,
+        "_load_dataset",
+        lambda self, path, **kwargs: list(examples),
+    )
+    cfg = MMMUConfig(subjects=["Math"], **cfg_overrides)
+    data_config = DataConfig(type=DataGenType.MMMU, mmmu=cfg)
+    return MMMUDataGenerator(APIConfig(type=APIType.Chat), data_config, _make_mock_tokenizer())
+
+
+def test_loader_requires_token(monkeypatch: pytest.MonkeyPatch) -> None:
+    monkeypatch.delenv("HF_TOKEN", raising=False)
+    monkeypatch.delenv("HUGGING_FACE_HUB_TOKEN", raising=False)
[…]
