### 2026-05-12 29612f9
add MMMU optional tests
Mirrors the sharegpt4video pattern from the parent commit:

- ``tests/optional/manual/mmmu/cases/{png,jpeg}/`` — Kubernetes manual e2e
  cases. Each case deploys a Qwen3-VL vLLM server and runs an inference-perf
  Job that hits MMMU through the gated HuggingFace dataset. ``hf-secret`` is
  expected in the ``default`` namespace; the orchestrator copies it into the
  per-case namespace before applying the workload.
- ``tests/optional/manual/mmmu/e2e_test.sh`` — orchestrator that iterates
  every case folder, runs each in its own namespace, extracts the lifecycle
  reports
[…]
