### 2026-05-14 cb777a4
regenerate system prompts per stage in conversation replay
