+++ inference_perf/datagen/conversation_replay_datagen.py
@@ -126,6 +126,7 @@ class ConversationBlueprint:
     conversation_id: int
     num_turns: int
     system_prompt: str
+    system_prompt_tokens: int
     turn_prompts: List[str] = field(default_factory=list)
     turn_output_lens: List[int] = field(default_factory=list)
     turn_tool_call_latencies: List[float] = field(default_factory=list)
@@ -217,11 +218,13 @@ class ConversationReplayDataGenerator(DataGenerator, LazyLoadDataMixin):
         # system_prompt the conversation was built around.
         expected_session_id = self.user_sessions[conv_idx].user_session_id
         if expected_session_id not in LocalUserSession._instances:
+            # Regenerate the entire system prompt for this stage
+            bp.system_prompt = self._generate_random_token_text(bp.system_prompt_tokens)
             self.user_sessions[conv_idx] = self._new_session(
                 user_session_id=expected_session_id,
                 context=bp.system_prompt,
             )
-            logger.debug("Slot %d: re-primed cleared session %s", conv_idx, expected_session_id)
+            logger.debug("Slot %d: refreshed system prompt and re-primed session %s", conv_idx, expected_session_id)
 
         # Closed-loop replenishment: when a conversation finishes all its turns,
         # reset the session so the slot immediately starts a fresh conversation.
@@ -366,6 +369,7 @@ class ConversationReplayDataGenerator(DataGenerator, LazyLoadDataMixin):
                 conversation_id=conv_id,
                 num_turns=num_turns,
                 system_prompt=system_prompt,
+                system_prompt_tokens=cfg.shared_system_prompt_len + dynamic_lens[conv_id],
                 turn_prompts=turn_prompts,
                 turn_output_lens=turn_output_lens_list,
                 turn_tool_call_latencies=turn_tool_latencies,
