+++ inference_perf/logger.py
@@ -14,6 +14,17 @@
 
 import logging
 import sys
+from typing import Optional
+
+from rich.console import Console
+from rich.logging import RichHandler
+
+_console: Optional[Console] = None
+
+
+def get_console() -> Optional[Console]:
+    """Shared Console so RichHandler logs don't tear the Progress bar. None in non-TTY."""
+    return _console
 
 
 def setup_logging(level: str) -> None:
@@ -23,19 +34,33 @@ def setup_logging(level: str) -> None:
     Args:
         level: Logging level (DEBUG, INFO, WARNING, ERROR, CRITICAL)
     """
+    global _console
+
     numeric_level = getattr(logging, level.upper(), None)
     if not isinstance(numeric_level, int):
         raise ValueError(f"Invalid log level: {level}")
 
-    # Use detailed format for DEBUG level, original format for others
-    if numeric_level == logging.DEBUG:
-        log_format = "%(asctime)s - %(levelname)s - %(filename)s:%(lineno)d - %(funcName)s - %(message)s"
+    handler: logging.Handler
+    if sys.stdout.isatty():
+        _console = Console()
+        handler = RichHandler(
+            console=_console,
+            show_path=numeric_level == logging.DEBUG,
+            rich_tracebacks=True,
+            markup=False,
+        )
+        log_format = "%(name)s - %(message)s"
     else:
-        log_format = "%(asctime)s - %(name)s - %(levelname)s - %(message)s"
+        _console = None
+        handler = logging.StreamHandler(sys.stdout)
+        if numeric_level == logging.DEBUG:
+            log_format = "%(asctime)s - %(levelname)s - %(filename)s:%(lineno)d - %(funcName)s - %(message)s"
+        else:
+            log_format = "%(asctime)s - %(name)s - %(levelname)s - %(message)s"
 
     logging.basicConfig(
         level=numeric_level,
         format=log_format,
-        handlers=[logging.StreamHandler(sys.stdout)],
+        handlers=[handler],
         force=True,
     )
