### 2026-05-14 bb7108c
test: assert prefix image sets partition by group
Replaces the loose "at least 2 groups diverge" check with a strong
partition assertion: with 3 groups × 2 prompts/group, the 6 prompts
must form exactly 3 groups of 2, each with a within-group-stable prefix
image set that does not match any other group's set.
