### 2026-05-14 114bd16
fix multimodal shared prefix groups
