### 2026-05-14 7904308
feat: tool-call replay, HuggingFace trace loading, and session replay hardening (#468)
- **Tool-call replay**: Capture structured `tool_calls` from live model
responses and substitute them into successor events. IDs are rewritten
positionally to keep `role: "tool"` messages valid, and `tool_choice` is
injected to guide the model toward the expected tool call.
- **HuggingFace trace loading**: Load OTel traces directly from
HuggingFace Hub datasets via `huggingface_hub`.
- **Duplicate session isolation**: Inject random strings into unique
messages of duplicated sessions to prevent KV-cache hits 
[…]
