+++ tests/test_otel_trace_to_replay_graph.py
@@ -577,52 +577,69 @@ def test_growing_prefix() -> None:
     assert graph.events["event_002_span_T3"].wait_ms == 1000
     assert graph.events["event_003_span_T4"].wait_ms == 1000
 
-    # span_T1: all unique (no predecessors)
-    segs_T1 = graph.events["event_000_span_T1"].call.input_segments
-    assert all(s.type == "unique" for s in segs_T1)
-    assert sum(s.message_count for s in segs_T1) == 2
 
-    # span_T2: SHARED(2) + OUTPUT(1) + UNIQUE(1)
-    segs_T2 = graph.events["event_001_span_T2"].call.input_segments
-    assert segs_T2[0].type == "shared"
-    assert segs_T2[0].source_event_id == "event_000_span_T1"
-    shared_msgs_T2 = segs_T2[0].message_count
-    print(f"\n  span_T2 shared prefix: {shared_msgs_T2} messages (source: {segs_T2[0].source_event_id})")
-    print(f"  span_T2 segments: {[(s.type, s.message_count) for s in segs_T2]}")
-    assert shared_msgs_T2 == 2, f"T2 should share 2 messages with T1, got {shared_msgs_T2}"
-    assert sum(s.message_count for s in segs_T2) == 4
+# ---------------------------------------------------------------------------
+# Test: tool_definitions flow
+# ---------------------------------------------------------------------------
 
-    # span_T3: SHARED(4) + OUTPUT(1) + UNIQUE(1)
-    segs_T3 = graph.events["event_002_span_T3"].call.input_segments
-    assert segs_T3[0].type == "shared"
-    assert segs_T3[0].source_event_id == "event_001_span_T2"
-    shared_msgs_T3 = segs_T3[0].message_count
-    print(f"  span_T3 shared prefix: {shared_msgs_T3} messages (source: {segs_T3[0].source_event_id})")
-    print(f"  span_T3 segments: {[(s.type, s.message_count) for s in segs_T3]}")
-    assert shared_msgs_T3 == 4, f"T3 should share 4 messages with T2, got {shared_msgs_T3}"
-    assert sum(s.message_count for s in segs_T3) == 6
 
-    # span_T4: SHARED(6) + OUTPUT(1) + UNIQUE(1)
-    segs_T4 = graph.events["event_003_span_T4"].call.input_segments
-    assert segs_T4[0].type == "shared"
-    assert segs_T4[0].source_event_id == "event_002_span_T3"
-    shared_msgs_T4 = segs_T4[0].message_count
-    print(f"  span_T4 shared prefix: {shared_msgs_T4} messages (source: {segs_T4[0].source_event_id})")
-    print(f"  span_T4 segments: {[(s.type, s.message_count) for s in segs_T4]}")
-    assert shared_msgs_T4 == 6, f"T4 should share 6 messages with T3, got {shared_msgs_T4}"
-    assert sum(s.message_count for s in segs_T4) == 8
+def test_tool_definitions_flow() -> None:
+    """Tool definitions in gen_ai.tool.definitions flow through to GraphCall.tool_definitions."""
+    tool_defs = [
+        {
+            "type": "function",
+            "name": "get_weather",
+            "description": "Get the current weather",
+            "parameters": {
+                "type": "object",
+                "properties": {"location": {"type": "string"}},
+                "required": ["location"],
+            },
+        }
+    ]
+    span = {
+        "trace_id": "trace_test",
+        "span_id": "span_tools",
+        "name": "chat gpt-4",
+        "start_time": "2024-01-01T00:00:00Z",
+        "end_time": "2024-01-01T00:00:01Z",
+        "attributes": {
+            "gen_ai.request.model": "gpt-4",
[…]

[…]
