### 2026-05-14 60151a2
Add native llm-d-benchmark v0.2.0 benchmark report generation
Addresses: #448. Adds native BR0.2 (llm-d-benchmark v0.2 benchmark report)
emission to inference-perf, retiring llm-d-benchmark's hand-written
translator and the duplicate metadata derivation that goes with it.

inference-perf is strictly responsible for the BR0.2 results section
(performance measurements derived from the run). Everything else (stack
configuration, run/scenario metadata) is supplied by the user as a
partial report: a YAML or JSON file in BR0.2 shape with results omitted.
inference-perf loads the partial, validates i
[…]
