+++ tests/reportgen/br/v0_2/test_partial_report.py
@@ -1,0 +1,168 @@
+# Copyright 2026 The Kubernetes Authors.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+"""Tests for inference-perf's BR0.2 partial-report emission.
+
+Inference-perf emits its slice of a BR0.2 report (``version``, ``run`` with
+the generated uid and stage time window, ``results`` from request metrics).
+A downstream composer merges in stack/scenario/observability/etc to produce
+a full BR0.2 document. These tests pin down the shape and mergeability of
+the emitted partial.
+"""
+
+import time
+from typing import Any, Dict
+
+from inference_perf.apis import (
+    InferenceInfo,
+    RequestLifecycleMetric,
+    StreamedResponseMetrics,
+)
+from inference_perf.payloads import RequestMetrics, Text
+from inference_perf.reportgen.br.v0_2 import build_partial_report, generate_run_uid
+from inference_perf.reportgen.br.v0_2.schema_v0_2 import VERSION
+
+import yaml
+
+
+def _metric(start: float, end: float) -> RequestLifecycleMetric:
+    return RequestLifecycleMetric(
+        stage_id=0,
+        scheduled_time=start - 0.001,
+        start_time=start,
+        end_time=end,
+        request_data="{}",
+        response_data="ok",
+        info=InferenceInfo(
+            request_metrics=RequestMetrics(text=Text(input_tokens=128)),
+            response_metrics=StreamedResponseMetrics(
+                response_chunks=[],
+                chunk_times=[start + 0.05, end - 0.005],
+                output_tokens=2,
+                output_token_times=[start + 0.05, end - 0.005],
+            ),
+        ),
+        error=None,
+    )
+
+
+def test_generate_run_uid_is_unique_and_stage_tagged() -> None:
+    a = generate_run_uid(0)
+    b = generate_run_uid(0)
+    assert a != b
+    assert a.startswith("inference-perf-stage-0-")
+    assert generate_run_uid(7).startswith("inference-perf-stage-7-")
+
+
+def test_partial_report_top_level_shape() -> None:
+    now = time.time()
+    metrics = [_metric(now + i * 0.1, now + i * 0.1 + 0.08) for i in range(5)]
+
+    partial = build_partial_report(metrics, tokenizer=None, run_uid="test-uid-abc")
+
+    assert set(partial.keys()) == {"version", "run", "results"}
+    assert partial["version"] == VERSION
+    assert partial["run"]["uid"] == "test-uid-abc"
+    assert "time" in partial["run"]
+    assert "results" in partial
+    assert "request_performance" in partial["results"]
+
+
+def test_partial_report_run_time_is_iso8601() -> None:
+    start = 1_700_000_000.0
+    end = start + 4.5
+    metrics = [_metric(start, end)]
+
+    partial = build_partial_report(metrics, tokenizer=None, run_uid="uid")
+
+    run_time = partial["run"]["time"]
+    assert run_time["start"].startswith("2023-")  # 1.7e9 → 2023
+    assert run_time["start"].endswith("+00:00") or run_time["start"].endswith("Z")
+    assert run_time["duration"].startswith("PT") and run_time["duration"].endswith("S")
[…]
