### 2026-05-15 eb304b5
pass stage in to the generation
