### 2026-05-15 1da791e
Address lint / typecheck issues
