+++ tests/required/apis/test_user_session.py
@@ -260,7 +260,7 @@ class TestUserSessionTruncation:
         hf = tok.get_tokenizer.return_value
         hf.encode = MagicMock(side_effect=lambda text: [1] * tok.count_tokens(text))
 
-        session = LocalUserSession(user_session_id="sess_1", system_prompt="tok_5", tokenizer=tok, max_model_len=70)
+        session = LocalUserSession(user_session_id="sess_1", system_prompt="tok_5", tokenizer=tok, max_model_len=220)
         LocalUserSession._instances["sess_1"] = session
 
         session.history = ["tok_5", "tok_5"]
@@ -268,7 +268,7 @@ class TestUserSessionTruncation:
 
         data = UserSessionCompletionAPIData(user_session_id="sess_1", target_round=1, prompt="tok_10", max_tokens=0)
 
-        payload = await data.to_payload("model", 0, False, False)
+        payload = await data.to_request_body("model", 0, False, False)
 
         assert session.history == ["tok_5"]
         assert payload["prompt"] == "tok_5 tok_5 tok_10"
@@ -279,7 +279,7 @@ class TestUserSessionTruncation:
         hf = tok.get_tokenizer.return_value
         hf.encode = MagicMock(side_effect=lambda text: [1] * tok.count_tokens(text))
 
-        session = LocalUserSession(user_session_id="sess_2", system_prompt="tok_15", tokenizer=tok, max_model_len=70)
+        session = LocalUserSession(user_session_id="sess_2", system_prompt="tok_15", tokenizer=tok, max_model_len=220)
         LocalUserSession._instances["sess_2"] = session
 
         session.history = []
@@ -287,7 +287,7 @@ class TestUserSessionTruncation:
 
         data = UserSessionCompletionAPIData(user_session_id="sess_2", target_round=0, prompt="tok_10", max_tokens=0)
 
-        payload = await data.to_payload("model", 0, False, False)
+        payload = await data.to_request_body("model", 0, False, False)
 
         assert session.system_prompt == "tok_10"
         assert payload["prompt"] == "tok_10 tok_10"
