### 2026-05-15 e6a5ead
fix: per-group cache_key for shared_prefix multimodal bytes (#485)
Addresses: #484
