+++ inference_perf/apis/chat.py
@@ -192,6 +192,14 @@ class ChatCompletionAPIData(InferenceAPIData):
     # group — required for prefix-cache benchmarks.
     prefix_text: Optional[str] = None
     prefix_multimodal_spec: Optional[MultimodalSpec] = None
+    # Identifies which prefix-cache group this request belongs to. Two requests
+    # with the same prefix_multimodal_spec AND the same prefix_cache_key render
+    # byte-identical prefix media (so the server's prefix cache hits); requests
+    # with the same spec but different prefix_cache_key render different bytes
+    # (so distinct logical prefixes don't collide in the cache). Left None when
+    # there's only one prefix group, which preserves the old single-stream
+    # behavior. The shared_prefix datagen sets this to the group id.
+    prefix_cache_key: Optional[int] = None
 
     # Realized per-instance metrics populated by ``to_request_body`` and read
     # by ``process_response`` to build ``RequestMetrics``. Aggregates across
@@ -205,13 +213,15 @@ class ChatCompletionAPIData(InferenceAPIData):
         text_str: str,
         spec: MultimodalSpec,
         deterministic: bool,
+        cache_key: Optional[int] = None,
     ) -> Tuple[list[dict[str, Any]], list[Image], list[Video], list[Audio]]:
         """Materialize bytes from ``spec`` and weave into ``text_str``.
 
         Returns ``(content_list, images, videos, audios)``. Pure (no mutation
         of class state). ``deterministic=True`` seeds RNGs per-instance from
-        spec content, yielding identical bytes for identical specs across
-        calls — required for prefix-cache benchmarks.
+        spec content + ``cache_key``, yielding identical bytes for identical
+        (spec, cache_key) pairs across calls — required for prefix-cache
+        benchmarks. ``cache_key`` is ignored when ``deterministic=False``.
         """
         media_items: list[Tuple[dict[str, Any], float]] = []
         image_instances: list[Image] = []
@@ -222,7 +232,7 @@ class ChatCompletionAPIData(InferenceAPIData):
         def _rng_for(*key_parts: object) -> np.random.Generator:
             if not deterministic:
                 return fresh_rng
-            return np.random.default_rng(hash(key_parts) & 0xFFFFFFFF)
+            return np.random.default_rng(hash((cache_key,) + key_parts) & 0xFFFFFFFF)
 
         for i, img in enumerate(spec.images):
             raw_bytes, data_url = _encode_image(
@@ -308,7 +318,7 @@ class ChatCompletionAPIData(InferenceAPIData):
             prefix_text = self.prefix_text or ""
             if self.prefix_multimodal_spec is not None:
                 prefix_content, p_imgs, p_vids, p_auds = self._materialize_multimodal_content(
-                    prefix_text, self.prefix_multimodal_spec, deterministic=True
+                    prefix_text, self.prefix_multimodal_spec, deterministic=True, cache_key=self.prefix_cache_key
                 )
                 all_images.extend(p_imgs)
                 all_videos.extend(p_vids)
