### 2026-05-15 cc620f4
Make per-stage progress visible in non-TTY logs (#487)
Addresses: #486

From local testing:
```
[05/15/26 17:09:33] INFO     inference_perf.loadgen.load_generator - Stage 0 - run started                                                                                       
Overall Progress ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━   0% 0/3   -:--:-- 0:00:07
Stage 0 Requests ━━━━━━━━━━━━━━━━━━━━━╺━━━━━━━━━━━━━━━━━━  52% 21/40 0:00:05 0:00:07
```

```
[05/15/26 17:09:33] INFO     inference_perf.loadgen.load_generator - Stage 0 - run started                                                         
[…]
