### 2026-05-15 f722d77
WIP: fix shared_prefix off-by-one by composing at token level
Fixes #490. shared_prefix_datagen built each prompt by string concat
(prefix_text + " " + suffix_text) and then re-tokenized the whole thing,
so the BPE merge at the prefix/suffix boundary could shift the count by
+1. When the sampled question length was 1, the convergence loop could
not shed the extra token and raised ValueError after 20 attempts.

Compose at the token level instead: keep the prefix's token IDs from the
generation step and produce the prompt as decode(prefix_ids + suffix_ids).
The boundary disappears entirely, no c
[…]
