### 2026-05-15 d7d54fe
Pin suffix boundary to word-start tokens for prefix-cache stability
The previous commit fixed the off-by-one count bug (#490) by composing
prompts at the token level. But the BPE-merge-at-boundary effect that
caused the count drift also caused the prefix's server-side tokenization
to drift across requests in a group — breaking prefix-cache hits when
the suffix's first token happened to merge across the boundary.

Both invariants share one root cause: an unconstrained boundary token.
Fix both by pinning the first suffix token to a "word-start" token (one
whose decoded form begins with whitespac
[…]
