### 2026-05-15 58eaafe
trim comments in datagen helpers and tests
