### 2026-05-15 b5296ea
Converge on suffix count to handle random-token roundtrip drift
The previous commit pinned the first suffix token to a word-start, which
fixed boundary-merge effects but left the suffix vulnerable to internal
BPE merges: random subsequent tokens can collapse on decode→encode,
causing the full prompt count to drift by a few tokens.

Run the same convergence loop on the suffix (preserving the word-start
first token), so count_tokens(decode(suffix_ids)) == q_len exactly. With
the word-start boundary on top of that, the full prompt always tokenizes
to count_tokens(prefix_text) + q_len.

Fixes the 
[…]
