### 2026-05-18 a313a19
fix: OTEL metrics correctly parse SSE streaming responses
Previously _record_otel_metrics assumed response_content was always a
single JSON object, causing json.loads to fail on SSE streaming
responses. This fix detects SSE format, parses all chunks, and
concatenates content deltas into the full output text following GenAI
semantic conventions. Also adds an early-return guard when OTEL is
disabled or span is None, and uses getattr for safer attribute access.
