### 2026-05-18 dbb8ff7
refactor multimodal payloads into per-modality trees
Splits the flat ``inference_perf/payloads/multimodal_spec.py`` into
``payloads/{text,image,video,audio}/`` subpackages, each with its own
``metrics.py`` and ``spec/`` discriminated union (synthetic / local /
remote / pre_encoded variants). ``payloads/payload.py`` moves to
``payloads/video/metrics.py``; the package ``__init__`` re-exports the
public names so existing call sites only need their import paths
updated.

``VideoSpec`` becomes polymorphic. ``chat.py`` switches from enum-on-
``representation`` dispatch to ``isinstance`` dispatch ove
[…]
