+++ inference_perf/datagen/sharegpt4video_datagen.py
@@ -1,0 +1,390 @@
+# Copyright 2026 The Kubernetes Authors.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+"""ShareGPT4Video real-dataset loader for multimodal benchmarking.
+
+The hot request path NEVER blocks on the network. A background thread pulls
+source MP4 zips from the gated HuggingFace dataset repo into the local cache;
+``load_lazy_data`` only ever samples from videos that are already extracted
+on disk. Init blocks only until the first zip has been extracted so there's
+always a non-empty pool when load generation starts.
+
+Emits requests in the Frames wire format from PR #450 (N ``image_url`` blocks
+at one insertion point), compatible with VLMs that don't accept ``video_url``
+natively.
+"""
+
+from __future__ import annotations
+
+import io
+import logging
+import os
+import threading
+import zipfile
+from typing import Any, Generator, List, Optional
+
+from filelock import FileLock
+from PIL import Image as PILImage
+
+from inference_perf.apis.base import InferenceAPIData, LazyLoadInferenceAPIData
+from inference_perf.apis.chat import ChatCompletionAPIData, ChatMessage
+from inference_perf.config import APIConfig, APIType, DataConfig, ShareGPT4VideoConfig
+from inference_perf.datagen.multimodal_sampling import resolution_to_wh
+from inference_perf.payloads import (
+    ImageRepresentation,
+    MultimodalSpec,
+    PreEncodedFramesVideoSpec,
+    VideoRepresentation,
+)
+from inference_perf.utils.custom_tokenizer import CustomTokenizer
+
+from .base import DataGenerator, LazyLoadDataMixin
+from .gated_hf_dataset import GatedHFDataset
+
+logger = logging.getLogger(__name__)
+
+DEFAULT_CACHE_DIRNAME = "sharegpt4video_cache"
+ZIP_FOLDER_PREFIX = "zip_folder/"
+
+
+class ShareGPT4VideoDataGenerator(DataGenerator, LazyLoadDataMixin, GatedHFDataset):
+    """Emits ShareGPT4Video requests using the Frames wire format.
+
+    Source zips download asynchronously in the background; the on-disk video
+    pool grows over time. ``load_lazy_data`` samples from whatever is currently
+    available and never blocks on the network.
+    """
+
+    def __init__(self, api_config: APIConfig, config: DataConfig, tokenizer: Optional[CustomTokenizer]) -> None:
+        DataGenerator.__init__(self, api_config, config, tokenizer)
+
+        if config.sharegpt4video is None:
+            raise ValueError("sharegpt4video config is required for ShareGPT4VideoDataGenerator")
+        self.sg4v_config: ShareGPT4VideoConfig = config.sharegpt4video
+
+        GatedHFDataset.__init__(self, token=self.sg4v_config.token)
+
+        self.cache_dir = os.path.abspath(self.sg4v_config.cache_dir or os.path.join(os.getcwd(), DEFAULT_CACHE_DIRNAME))
+        self.videos_dir = os.path.join(self.cache_dir, "videos")
+        self.hf_cache_dir = os.path.join(self.cache_dir, "hf")
+        self._extracted_marker_dir = os.path.join(self.cache_dir, ".extracted")
[…]
