### 2026-05-18 1a84f92
add sharegpt4video dataset loader
ShareGPT4Video is a gated HuggingFace video-captioning corpus
(https://huggingface.co/datasets/ShareGPT4Video/ShareGPT4Video,
~1.5 TB total, 15-21 GB per source zip). The loader streams
captions + keyframe indices and pulls source MP4 zips into a local
cache directory via a background thread, so the hot request path
never blocks on the network. Init blocks only until the first zip
has been extracted so there is always at least one usable video
when load generation begins; the on-disk pool grows over time as
background downloads complete.

Each request decodes 
[…]
