+++ tests/required/datagen/test_sharegpt4video_datagen.py
@@ -1,0 +1,353 @@
+# Copyright 2026 The Kubernetes Authors.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+"""Unit tests for the ShareGPT4Video loader and the pre-encoded-frames path."""
+
+from __future__ import annotations
+
+import io
+import logging
+import os
+from typing import Any, List
+from unittest.mock import MagicMock
+
+import pytest
+from PIL import Image as PILImage
+
+from inference_perf.apis.base import LazyLoadInferenceAPIData
+from inference_perf.apis.chat import ChatCompletionAPIData, ChatMessage
+from inference_perf.config import (
+    APIConfig,
+    APIType,
+    DataConfig,
+    DataGenType,
+    ShareGPT4VideoConfig,
+)
+from inference_perf.datagen import sharegpt4video_datagen
+from inference_perf.datagen.sharegpt4video_datagen import ShareGPT4VideoDataGenerator
+from inference_perf.payloads import (
+    ImageRepresentation,
+    MultimodalSpec,
+    PreEncodedFramesVideoSpec,
+    SyntheticMp4VideoSpec,
+    VideoRepresentation,
+)
+
+
+def _make_mock_tokenizer() -> MagicMock:
+    mock_tokenizer = MagicMock()
+    mock_tokenizer.count_tokens.return_value = 10
+    hf_tokenizer = MagicMock()
+    hf_tokenizer.vocab_size = 1000
+    hf_tokenizer.decode.return_value = "decoded prompt text"
+    mock_tokenizer.get_tokenizer.return_value = hf_tokenizer
+    return mock_tokenizer
+
+
+def _load_first(gen: ShareGPT4VideoDataGenerator, data_index: int = 0) -> ChatCompletionAPIData:
+    """Materialize one request from the generator with isinstance asserts for mypy."""
+    lazy_item = next(gen.get_data())
+    assert isinstance(lazy_item, LazyLoadInferenceAPIData)
+    lazy_item.data_index = data_index
+    item = gen.load_lazy_data(lazy_item)
+    assert isinstance(item, ChatCompletionAPIData)
+    return item
+
+
+def _fake_row(video_path: str = "pixabay/abc.mp4", n_keyframes: int = 4) -> dict[str, Any]:
+    return {
+        "video_id": os.path.basename(video_path).split(".")[0],
+        "video_path": video_path,
+        "keyframe": list(range(n_keyframes)),
+        "captions": [
+            {"idx": 0, "content": "A scene begins."},
+            {"idx": 1, "content": "The action continues."},
+        ],
+    }
+
+
+def _build_generator(
+    monkeypatch: pytest.MonkeyPatch,
+    tmp_path: Any,
+    rows: List[dict[str, Any]],
+    *,
+    representation: VideoRepresentation = VideoRepresentation.JPEG_FRAMES,
+    max_frames: int = 16,
+    available_zips: List[str] | None = None,
+    create_videos: bool = True,
+) -> ShareGPT4VideoDataGenerator:
+    monkeypatch.setenv("HF_TOKEN", "test-token")
+
+    monkeypatch.setattr("inference_perf.datagen.gated_hf_dataset.load_dataset", lambda *a, **kw: list(rows))
+
[…]
