+++ inference_perf/config/__init__.py
@@ -36,12 +36,14 @@ from inference_perf.config.datagen import (
     ConversationReplayConfig,
     DataConfig,
     DataGenType,
+    GatedHFDatasetConfig,
     ImageDatagenConfig,
     MediaDatagenConfig,
     OTelTraceReplayConfig,
     Resolution,
     ResolutionPreset,
     SessionReplayConfig,
+    ShareGPT4VideoConfig,
     SharedPrefix,
     SyntheticMultimodalDatagenConfig,
     TraceConfig,
@@ -89,6 +91,7 @@ __all__ = [
     "DataGenType",
     "Distribution",
     "DistributionType",
+    "GatedHFDatasetConfig",
     "GoodputConfig",
     "GoogleCloudStorageConfig",
     "ImageDatagenConfig",
@@ -112,6 +115,7 @@ __all__ = [
     "ResponseFormatType",
     "SessionLifecycleReportConfig",
     "SessionReplayConfig",
+    "ShareGPT4VideoConfig",
     "SharedPrefix",
     "SimpleStorageServiceConfig",
     "StageGenType",
