### 2026-05-18 276376c
add sharegpt4video optional tests
- ``tests/optional/manual/sharegpt4video/cases/{png,jpeg}_frames/``
  — Kubernetes manual e2e cases. Each case deploys a Qwen3-VL vLLM
  server and runs an inference-perf Job that hits ShareGPT4Video
  through the gated HuggingFace dataset. ``hf-secret`` is expected
  in the ``default`` namespace; the orchestrator copies it into the
  per-case namespace before applying the workload.
- ``tests/optional/manual/sharegpt4video/e2e_test.sh`` — orchestrator
  that iterates every case folder, runs each in its own namespace,
  extracts the lifecycle reports, and verif
[…]
