### 2026-05-19 57a484a
Log synthetic datagen materialization progress
SyntheticDataGenerator materializes prompts lazily inside worker
processes. On large runs with slow tokenizers there is no other signal
of forward progress between the run start and the first response,
making the benchmark look hung. Emit a per-worker INFO heartbeat from
load_lazy_data at most every 10s so logs show prompts are still being
produced.
