### 2026-05-19 9c22c13
Use logger.warning so heartbeat surfaces in worker processes
Python 3.14's default multiprocessing start method is forkserver, which
spawns workers from a fresh interpreter that does not inherit the
parent's basicConfig. INFO-level logger calls from inside a worker are
dropped because no handler is installed at that level. WARNING+ goes
through Python's lastResort handler to stderr without any setup, so the
heartbeat reaches users in both single-process and multi-worker runs.
