### 2026-05-19 fc212f5
Configure logging inside Worker.run() so worker INFO logs surface
The heartbeat is semantically INFO, not WARNING. Python 3.14's default
forkserver start method gives workers a fresh interpreter without the
parent's basicConfig, so without explicit setup any logger.info() call
made inside a worker (datagen heartbeat and pre-existing worker logs
alike) is dropped by the lastResort handler.

Capture the parent's effective root level in Worker.__init__ (runs in
the parent after setup_logging) and install a matching basicConfig at
the top of Worker.run() (runs in the child). Heartbeat reverts to
l
[…]
