### 2026-05-19 bf37b0c
Patch time.monotonic via string path for mypy strict
patch.object(synthetic_datagen.time, ...) tripped mypy --strict because
time is an implicitly re-exported module attribute. String-based patch
bypasses that without changing behavior.
