### 2026-05-19 57e47ec
Log synthetic datagen materialization progress (#496)
SyntheticDataGenerator materializes prompts lazily inside worker
processes. On large runs with slow tokenizers there is no other signal
of forward progress between the run start and the first response, making
the benchmark look hung. Emit a per-worker INFO heartbeat from
load_lazy_data at most every 10s so logs show prompts are still being
produced.


Logs snippet:
```
2026-05-19 17:22:47,282 - inference_perf.loadgen.load_generator - INFO - Stage 0 - run started
2026-05-19 17:22:49,313 - inference_perf.apis.chat - INFO - Multimodal datagen 
[…]
