### 2026-05-20 803b44a
Surface prompt cache token metrics in lifecycle summary (#473)
Fixes #467.

## Summary
- add `successes.prompt_tokens` with total/cached/uncached counts in
request lifecycle summaries
- read vLLM/OpenAI-compatible
`usage.prompt_tokens_details.cached_tokens` from streamed `server_usage`
- cover the summary output with a focused unit test

## Validation
- `uv run --group lint ruff format --check inference_perf tests scripts`
- `uv run --group lint ruff check`
- `uv run --group lint --group test --group types mypy --strict
./inference_perf ./tests`
- `uv run --extra analysis --group test pytest t
[…]
