### 2026-05-20 d811f88
[conversation_replay] force min_tokens == max_tokens for deterministic output length
Conversation replay benchmarks compare model throughput by sampling a fixed
output-length distribution per turn and asserting that each turn emits exactly
that many tokens. In practice the sampled length is not honored: different model
families terminate at different chat-template stop tokens, so the same workload
yields different total output token counts on Qwen vs Llama vs Nemotron, which
invalidates tok/sec and convos/hr comparisons.

`ignore_eos=True` does not fix this. vLLM's `ignore_eos` only suppresses
[…]
