### 2026-05-21 f7ba153
make sure in the same stage the system prompt is still different for different convs
