+++ inference_perf/datagen/conversation_replay_datagen.py
@@ -128,6 +128,7 @@ class ConversationBlueprint:
     num_turns: int
     system_prompt: str
     system_prompt_tokens: int
+    dynamic_system_prompt: str = ""
     turn_prompts: List[str] = field(default_factory=list)
     turn_output_lens: List[int] = field(default_factory=list)
     turn_tool_call_latencies: List[float] = field(default_factory=list)
@@ -179,6 +180,10 @@ class ConversationReplayDataGenerator(DataGenerator, LazyLoadDataMixin):
         # Seeded RNG for deterministic generation
         self.rng = np.random.default_rng(self.cr_config.seed)
 
+        # Cache for the currently active stage's shared system prompt
+        self._current_stage_id: Optional[int] = None
+        self._current_shared_prompt: Optional[str] = None
+
         # Build conversation blueprints
         self.blueprints: List[ConversationBlueprint] = []
         self.user_sessions: List[LocalUserSession] = []
@@ -219,13 +224,10 @@ class ConversationReplayDataGenerator(DataGenerator, LazyLoadDataMixin):
         # system_prompt the conversation was built around.
         expected_session_id = self.user_sessions[conv_idx].user_session_id
         if expected_session_id not in LocalUserSession._instances:
-            # Regenerate the entire system prompt for this stage using a derived stable seed.
-            seed_str = f"{self.cr_config.seed}_stage_{data.stage_idx}"
-            hash_digest = hashlib.sha256(seed_str.encode("utf-8")).digest()
-            derived_seed = int.from_bytes(hash_digest[:4], byteorder="little")
-            local_rng = np.random.default_rng(derived_seed)
+            # Get or generate the shared system prompt for this stage.
+            shared_prompt = self._get_or_generate_shared_prompt(data.stage_id)
+            bp.system_prompt = self._build_system_prompt(shared_prompt, bp.dynamic_system_prompt)
 
-            bp.system_prompt = self._generate_random_token_text(bp.system_prompt_tokens, rng=local_rng)
             self.user_sessions[conv_idx] = self._new_session(
                 user_session_id=expected_session_id,
                 context=bp.system_prompt,
@@ -290,6 +292,34 @@ class ConversationReplayDataGenerator(DataGenerator, LazyLoadDataMixin):
         LocalUserSession._instances[user_session_id] = session
         return session
 
+    def _get_or_generate_shared_prompt(self, stage_id: int) -> str:
+        """Get or generate the shared system prompt prefix for a specific stage using a derived stable seed."""
+        if self._current_shared_prompt is None or self._current_stage_id != stage_id:
+            if stage_id == 0:
+                # BACKWARDS COMPATIBILITY: Stage 0 must use the shared global self.rng
+                # to maintain the exact sequence of calls at startup, ensuring existing
+                # benchmark seeds generate identical conversation blueprints.
+                self._current_shared_prompt = self._generate_random_token_text(
+                    self.cr_config.shared_system_prompt_len
+                )
+            else:
+                # Later stages derive their seed stably from the base seed and stage ID.
+                # This happens dynamically at runtime (post-setup) and uses local isolated RNGs.
+                seed_str = f"{self.cr_config.seed}_stage_{stage_id}"
+                hash_digest = hashlib.sha256(seed_str.encode("utf-8")).digest()
+                derived_seed = int.from_bytes(hash_digest[:4], byteorder="little")
+                local_rng = np.random.default_rng(derived_seed)
+
+                self._current_shared_prompt = self._generate_random_token_text(
+                    self.cr_config.shared_system_prompt_len, rng=local_rng
+                )
[…]
