+++ tests/test_conversation_replay_datagen.py
@@ -13,7 +13,7 @@
 # limitations under the License.
 """Tests for ConversationReplayDataGenerator."""
 
-from typing import Generator
+from typing import Any, Generator
 
 import pytest
 from unittest.mock import MagicMock
@@ -49,8 +49,8 @@ def _make_mock_tokenizer(vocab_size: int = 32000) -> MagicMock:
     mock_tokenizer = MagicMock()
     hf_tok = MagicMock()
     hf_tok.vocab_size = vocab_size
-    hf_tok.decode.return_value = "mock decoded text"
-    hf_tok.batch_decode.return_value = ["mock decoded text"]
+    hf_tok.decode.side_effect = lambda ids, **kwargs: f"decoded_{ids}"
+    hf_tok.batch_decode.side_effect = lambda list_ids, **kwargs: [f"decoded_{ids}" for ids in list_ids]
     mock_tokenizer.get_tokenizer.return_value = hf_tok
     return mock_tokenizer
 
@@ -225,7 +225,7 @@ class TestConversationReplayDataGenerator:
         assert LocalUserSession._instances == {}
 
         # First request after the clear should re-prime the slot it touches AND regenerate prompt.
-        gen.load_lazy_data(LazyLoadInferenceAPIData(data_index=1, preferred_worker_id=1))
+        gen.load_lazy_data(LazyLoadInferenceAPIData(data_index=1, preferred_worker_id=1, stage_id=1))
 
         assert "conv_1" in LocalUserSession._instances
         # It should be different from original
@@ -243,10 +243,10 @@ class TestConversationReplayDataGenerator:
 
         last_contexts = {i: "" for i in range(3)}
 
-        for _ in range(3):
+        for stage_idx in range(3):
             LocalUserSession.clear_instances()
             for conv_idx in range(3):
-                gen.load_lazy_data(LazyLoadInferenceAPIData(data_index=conv_idx, preferred_worker_id=conv_idx))
+                gen.load_lazy_data(LazyLoadInferenceAPIData(data_index=conv_idx, preferred_worker_id=conv_idx, stage_id=stage_idx))
                 current_context = LocalUserSession._instances[f"conv_{conv_idx}"].context
                 assert current_context != last_contexts[conv_idx]
                 last_contexts[conv_idx] = current_context
@@ -312,13 +312,13 @@ class TestConversationReplayDataGenerator:
         # Run 1
         gen1 = ConversationReplayDataGenerator(api_config, data_config, make_deterministic_mock_tok())
         # Stage 0
-        gen1.load_lazy_data(LazyLoadInferenceAPIData(data_index=0, preferred_worker_id=0, stage_idx=0))
+        gen1.load_lazy_data(LazyLoadInferenceAPIData(data_index=0, preferred_worker_id=0, stage_id=0))
         context1_stage0 = LocalUserSession._instances["conv_0"].context
 
         LocalUserSession.clear_instances()
 
         # Stage 1
-        gen1.load_lazy_data(LazyLoadInferenceAPIData(data_index=2, preferred_worker_id=0, stage_idx=1))
+        gen1.load_lazy_data(LazyLoadInferenceAPIData(data_index=2, preferred_worker_id=0, stage_id=1))
         context1_stage1 = LocalUserSession._instances["conv_0"].context
 
         # Isolate Run 2
@@ -327,13 +327,13 @@ class TestConversationReplayDataGenerator:
         # Run 2
         gen2 = ConversationReplayDataGenerator(api_config, data_config, make_deterministic_mock_tok())
         # Stage 0
[…]
