### 2026-05-21 e2009a1
WIP: seed Prometheus exposition surface for runtime observability
Adds a minimal HTTP /metrics endpoint exposing a single gauge,
inference_perf_run_elapsed_seconds, as the first step toward #489.
Default port 9464 (OTel Collector convention), fresh CollectorRegistry
per instance, no wiring into run lifecycle yet (intentional, to keep
this reviewable).

Metric name and naming conventions are placeholder pending the design
sync; this PR is structural only.
