### 2026-05-21 b2748da
Add per-modality distinct-media pool for multimodal payload datagen
Introduces MediaPoolConfig{size} on the per-modality config so benchmarks
can express a bounded corpus of N distinct synthetic images / videos /
audios reused across requests. Pools are eagerly materialized at datagen
init and sampled uniformly.

Architecture follows the modality-primary scheme from #477: each pool
lives in its own payloads/<modality>/pool.py alongside the modality's
spec and metrics, with a per-process singleton + set/get/reset helpers.
The pool_index field lives only on the Synthetic*Spec variants (the
prove
[…]
