+++ inference_perf/apis/chat.py
@@ -37,6 +37,9 @@ from inference_perf.apis.streaming_parser import parse_sse_stream
 from inference_perf.config import APIConfig, APIType
 from inference_perf.mediagen.pool import get_video_pool
 from inference_perf.mediagen.synthesis import generate_jpeg_bytes, generate_mp4_bytes, generate_png_bytes, generate_wav_bytes
+from inference_perf.payloads.audio import pool as audio_pool_mod
+from inference_perf.payloads.image import pool as image_pool_mod
+from inference_perf.payloads.video import pool as video_pool_mod
 from inference_perf.payloads import (
     Audio,
     Audios,
@@ -292,16 +295,26 @@ class ChatCompletionAPIData(InferenceAPIData):
                 return fresh_rng
             return np.random.default_rng(hash((cache_key,) + key_parts) & 0xFFFFFFFF)
 
+        # Per-modality bounded pools are only consulted on the payload-side
+        # path (``deterministic=False``). Prefix-side specs always re-encode
+        # with a seeded RNG so prefix bytes stay byte-identical across
+        # requests in the same group.
+        image_pool = image_pool_mod.get_pool() if not deterministic else None
         for i, img in enumerate(spec.images):
             # Dispatch on the polymorphic image spec subtype for wire emission.
             # ``img.get_metrics(wire_bytes)`` already encodes the per-provenance
             # measurement rule, so each branch just computes ``wire_bytes`` and
             # calls it. Provenance variants still missing a branch raise below.
             if isinstance(img, SyntheticImageSpec):
-                raw_bytes, data_url = _encode_image(
-                    img.width, img.height, img.representation, _rng_for("img", i, img.width, img.height)
-                )
-                wire_bytes = len(raw_bytes)
+                if image_pool is not None and img.pool_index is not None:
+                    img_entry = image_pool.get(img.pool_index)
+                    data_url = img_entry.data_url
+                    wire_bytes = len(img_entry.raw_bytes)
+                else:
+                    raw_bytes, data_url = _encode_image(
+                        img.width, img.height, img.representation, _rng_for("img", i, img.width, img.height)
+                    )
+                    wire_bytes = len(raw_bytes)
             elif isinstance(img, PreEncodedImageSpec):
                 # Bytes are already encoded by the dataset loader; just base64-wrap
                 # them, picking the data-URL mime type from the declared representation.
@@ -312,6 +325,7 @@ class ChatCompletionAPIData(InferenceAPIData):
             media_items.append(({"type": "image_url", "image_url": {"url": data_url}}, img.insertion_point))
             image_instances.append(img.get_metrics(wire_bytes))
 
+        video_pool = video_pool_mod.get_pool() if not deterministic else None
         for vi, vid in enumerate(spec.videos):
             # Dispatch on the polymorphic spec subtype for wire emission.
             # Metric construction is delegated to ``vid.get_metrics(total_bytes)``
@@ -319,27 +333,44 @@ class ChatCompletionAPIData(InferenceAPIData):
             # pre-encoded report real pixel/byte/frame counts; remote/local
             # stubs would return zero-or-declared-dims per their own contract).
             if isinstance(vid, SyntheticMp4VideoSpec):
-                if deterministic:
-                    # Deterministic mode bypasses the pool — pool sampling is
-                    # non-deterministic, and this path is rare (prefix-side
-                    # videos in shared-prefix benchmarks) so per-init encode
-                    # cost is acceptable.
-                    mp4_bytes = generate_mp4_bytes(
-                        vid.width, vid.height, vid.frames, _rng_for("vid_mp4", vi, vid.width, vid.height, vid.frames)
+                if video_pool is not None and vid.pool_index is not None:
[…]

[…]
