+++ inference_perf/datagen/shared_prefix_datagen.py
@@ -21,23 +21,8 @@ from inference_perf.apis.completion import CompletionAPIData
 from inference_perf.apis.chat import ChatCompletionAPIData, ChatMessage
 from inference_perf.apis.user_session import LocalUserSession, UserSessionCompletionAPIData
 from inference_perf.config import APIConfig, APIType, DataConfig, Distribution, SyntheticMultimodalDatagenConfig
-from inference_perf.datagen.multimodal_sampling import (
-    resolution_to_wh,
-    sample_audio_duration,
-    sample_image_resolution,
-    sample_insertion_point,
-    sample_video_profile,
-)
-from inference_perf.payloads import (
-    ImageRepresentation,
-    MultimodalSpec,
-    SyntheticAudioSpec,
-    SyntheticFramesVideoSpec,
-    SyntheticImageSpec,
-    SyntheticMp4VideoSpec,
-    VideoRepresentation,
-    VideoSpecUnion,
-)
+from inference_perf.datagen.multimodal_sampling import configure_payload_pools, sample_spec_from_config
+from inference_perf.payloads import MultimodalSpec
 from inference_perf.utils.custom_tokenizer import CustomTokenizer
 from inference_perf.utils.distribution import sample_from_distribution
 
@@ -84,6 +69,12 @@ class SharedPrefixDataGenerator(DataGenerator, LazyLoadDataMixin):
 
         self.prefix_multimodal: Optional[SyntheticMultimodalDatagenConfig] = self.shared_prefix.multimodal
         self.payload_multimodal: Optional[SyntheticMultimodalDatagenConfig] = config.multimodal
+
+        # Eagerly materialize per-modality payload pools when configured.
+        # Only payload-side: prefix-side bytes flow through chat.py's seeded
+        # deterministic path so prefix-cache hits per group are preserved.
+        if self.payload_multimodal is not None:
+            configure_payload_pools(self.payload_multimodal, self.rng)
 
         # Resolve all parameters to Distribution
         system_prompt_dist = self._resolve_distribution(self.shared_prefix.system_prompt_len)
@@ -141,7 +132,7 @@ class SharedPrefixDataGenerator(DataGenerator, LazyLoadDataMixin):
         """Sample a fresh payload-side multimodal spec for one request."""
         if not self.payload_multimodal:
             return None
-        return _sample_spec(self.payload_multimodal, self.rng)
+        return sample_spec_from_config(self.payload_multimodal, self.rng, use_pool=True)
 
     def load_lazy_data(self, data: LazyLoadInferenceAPIData) -> InferenceAPIData:
         i = data.data_index % len(self.prompts)
@@ -218,7 +209,9 @@ class SharedPrefixDataGenerator(DataGenerator, LazyLoadDataMixin):
                 # ``ChatCompletionAPIData.prefix_cache_key``), so requests in
                 # this group produce identical wire bytes (server prefix-cache
                 # hits) while across-group bytes diverge.
-                self.prefix_specs_by_group[group_id] = _sample_spec(self.prefix_multimodal, self.rng)
+                self.prefix_specs_by_group[group_id] = sample_spec_from_config(
+                    self.prefix_multimodal, self.rng, use_pool=False
+                )
 
             for prompt_id in range(self.num_prompts_per_group):
                 q_len = self.question_len_list_per_group[group_id][prompt_id]
@@ -259,61 +252,3 @@ class SharedPrefixDataGenerator(DataGenerator, LazyLoadDataMixin):
             self.user_sessions = [self.user_sessions[i] for i in indices]
 
 
-def _sample_spec(cfg: SyntheticMultimodalDatagenConfig, rng: np.random.Generator) -> MultimodalSpec:
[…]
