### 2026-05-21 e3a6c42
Fix mypy errors in prometheus test, apply unrelated ruff format fix
- Add Iterator[PrometheusMetricsServer] return type to the fixture
- Annotate _scrape's read() locally to satisfy --strict
- Apply ruff format to tests/required/apis/test_chat.py (pre-existing
  drift from #496 that was blocking CI on this PR)
