### 2026-05-21 b8ffb80
Payload specs and measurement rules for each (media, provenance) pair (#477)
Partially addresses: #482, #476, and #475

Establishes a typed (media × provenance) grid for how request media is
described pre-flight and measured post-flight, covering the variants
datasets actually need.

## Why

A single `MultimodalSpec` should be able to carry mixed-provenance
media, and each provenance has different measurement semantics:

| | Pixels / duration | Wire bytes |
|---|---|---|
| **Synthetic** | exact | exact |
| **Pre-encoded** | trusted declared dims | exact |
| **Local file** | trusted declared di
[…]
