+++ tests/required/apis/test_chat.py
@@ -11,6 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+import base64
 import logging
 from typing import Any, Iterator
 from unittest.mock import MagicMock, patch
@@ -20,11 +21,13 @@ import pytest
 from inference_perf.apis import chat as chat_module
 from inference_perf.apis.chat import ChatCompletionAPIData, ChatMessage
 from inference_perf.config import APIType
-from inference_perf.payloads.multimodal_spec import (
-    AudioInstanceSpec,
-    ImageInstanceSpec,
+from inference_perf.payloads import (
+    ImageRepresentation,
     MultimodalSpec,
-    VideoInstanceSpec,
+    PreEncodedFramesVideoSpec,
+    SyntheticAudioSpec,
+    SyntheticFramesVideoSpec,
+    SyntheticImageSpec,
 )
 
 
@@ -78,9 +81,9 @@ def _reset_multimodal_progress_state() -> None:
 
 def _make_multimodal_request(images: int = 1, audios: int = 0, videos: int = 0) -> ChatCompletionAPIData:
     spec = MultimodalSpec(
-        images=[ImageInstanceSpec(width=8, height=8, insertion_point=0.0) for _ in range(images)],
-        videos=[VideoInstanceSpec(width=8, height=8, frames=2, insertion_point=0.0) for _ in range(videos)],
-        audios=[AudioInstanceSpec(duration=0.1, insertion_point=0.0) for _ in range(audios)],
+        images=[SyntheticImageSpec(width=8, height=8, insertion_point=0.0) for _ in range(images)],
+        videos=[SyntheticFramesVideoSpec(width=8, height=8, frames=2, insertion_point=0.0) for _ in range(videos)],
+        audios=[SyntheticAudioSpec(duration=0.1, insertion_point=0.0) for _ in range(audios)],
     )
     return ChatCompletionAPIData(
         messages=[ChatMessage(role="user", content="hi")],
@@ -129,3 +132,63 @@ async def test_multimodal_heartbeat_skips_within_interval() -> None:
     assert mock_logger.info.call_count == 1
     assert chat_module._multimodal_materialized_requests == 4
     assert chat_module._multimodal_materialized_images == 4
+
+
+@pytest.mark.asyncio
+async def test_materialize_pre_encoded_frames_video() -> None:
+    """``PreEncodedFramesVideoSpec`` is the only materializer branch reached
+    by dataset-loader provenance (frame bytes supplied, not synthesized).
+    Verifies the loader contract: one ``image_url`` block per frame, bytes
+    emitted verbatim (base64-wrapped, no re-encoding), mime-typed by
+    ``frame_representation``, and the realized ``Video`` metric reports the
+    summed input bytes."""
+    frame_bytes_list = [b"PNG_FRAME_ONE_BYTES", b"PNG_FRAME_TWO_BYTES", b"PNG_FRAME_THREE_BYTES"]
+    video_spec = PreEncodedFramesVideoSpec(
+        width=128,
+        height=64,
+        frames=len(frame_bytes_list),
+        insertion_point=0.0,
+        frame_representation=ImageRepresentation.PNG,
+        frames_bytes=frame_bytes_list,
+    )
+    data = ChatCompletionAPIData(
+        messages=[ChatMessage(role="user", content="describe this video")],
+        multimodal_spec=MultimodalSpec(videos=[video_spec]),
+    )
+
+    payload = await data.to_request_body(effective_model_name="gpt-vlm", max_tokens=100, ignore_eos=False, streaming=False)
+    content = payload["messages"][0]["content"]
[…]
