### 2026-05-22 4e64892
Add session-level metrics to CLI summary tables (#493)
- Add print_session_summary_tables() function to display session
lifecycle metrics
- Display session summary, duration, events, and token statistics
- Include tests for session summary table generation
