+++ inference_perf/utils/cli_summary.py
@@ -28,6 +28,14 @@ def extract_stage_id(report_name: str) -> Optional[int]:
     return None
 
 
+def extract_session_stage_id(report_name: str) -> Optional[int]:
+    """Extract stage ID from session report name (e.g. 'stage_0_session_lifecycle_metrics')."""
+    match = re.match(r"stage_(\d+)_session_lifecycle_metrics", report_name)
+    if match:
+        return int(match.group(1))
+    return None
+
+
 def print_summary_table(reports: List[ReportFile]) -> None:
     """Print a summary table of all stages to stdout using rich."""
     stage_reports: Dict[int, Dict[str, Any]] = {}
@@ -270,3 +278,150 @@ def print_summary_table(reports: List[ReportFile]) -> None:
     console.print(latency_table)
     console.print(speed_table)
     console.print(token_table)
+
+    # Print session-level metrics if available
+    print_session_summary_tables(reports)
+
+
+def print_session_summary_tables(reports: List[ReportFile]) -> None:
+    """Print session-level summary tables for session-based data generators."""
+    session_reports: Dict[int, Dict[str, Any]] = {}
+
+    # Extract session lifecycle metrics reports
+    for report in reports:
+        stage_id = extract_session_stage_id(report.name)
+        if stage_id is not None:
+            session_reports[stage_id] = report.contents
+
+    if not session_reports:
+        # No session metrics found, skip
+        return
+
+    # Sort stages by ID
+    sorted_stages = sorted(session_reports.keys())
+
+    console = Console()
+
+    # Table 1: Session Summary
+    session_summary_table = Table(
+        title="[bold magenta]Session Summary[/bold magenta]", show_header=True, header_style="bold cyan"
+    )
+    session_summary_table.add_column("Stage", justify="right")
+    session_summary_table.add_column("Sessions/s", justify="right")
+    session_summary_table.add_column("Total Sessions", justify="right")
+    session_summary_table.add_column("Succeeded", justify="right")
+    session_summary_table.add_column("Failed", justify="right")
+    session_summary_table.add_column("Error %", justify="right")
+    session_summary_table.add_column("Total Events", justify="right")
+    session_summary_table.add_column("Events Completed", justify="right")
+    session_summary_table.add_column("Events Cancelled", justify="right")
+
+    # Table 2: Session Duration & Events
+    session_duration_table = Table(
+        title="[bold magenta]Session Duration & Events[/bold magenta]", show_header=True, header_style="bold cyan"
+    )
+    session_duration_table.add_column("Stage", justify="right")
+    session_duration_table.add_column("Duration Mean (s)", justify="right")
+    session_duration_table.add_column("Duration Med (s)", justify="right")
+    session_duration_table.add_column("Duration P90 (s)", justify="right")
+    session_duration_table.add_column("Events Mean", justify="right")
+    session_duration_table.add_column("Events Med", justify="right")
+    session_duration_table.add_column("Events P90", justify="right")
+
+    # Table 3: Session Token Totals
+    session_tokens_table = Table(
+        title="[bold magenta]Session Token Totals (per session)[/bold magenta]", show_header=True, header_style="bold cyan"
+    )
+    session_tokens_table.add_column("Stage", justify="right")
+    session_tokens_table.add_column("In Tok/Sess Mean", justify="right")
+    session_tokens_table.add_column("In Tok/Sess Med", justify="right")
+    session_tokens_table.add_column("In Tok/Sess P90", justify="right")
+    session_tokens_table.add_column("Out Tok/Sess Mean", justify="right")
+    session_tokens_table.add_column("Out Tok/Sess Med", justify="right")
[…]
