+++ inference_perf/config/client/__init__.py
@@ -1,0 +1,34 @@
+# Copyright 2026 The Kubernetes Authors.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+from inference_perf.config.client.filestorage import (
+    GoogleCloudStorageConfig,
+    SimpleStorageServiceConfig,
+    StorageConfig,
+    StorageConfigBase,
+)
+from inference_perf.config.client.modelserver import (
+    ModelServerClientConfig,
+    ModelServerType,
+)
+from inference_perf.config.client.server_metrics import PrometheusClientConfig
+
+__all__ = [
+    "GoogleCloudStorageConfig",
+    "ModelServerClientConfig",
+    "ModelServerType",
+    "PrometheusClientConfig",
+    "SimpleStorageServiceConfig",
+    "StorageConfig",
+    "StorageConfigBase",
+]

[…]
