### 2026-05-26 204b011
fix: shared_prefix off-by-one by composing prompts at token level
shared_prefix built prompts by string concatenation (prefix + " " + suffix)
and re-tokenized, so the boundary did not tokenize as len(prefix_ids) +
len(suffix_ids). The explicit separator cost an extra token the additive
length target never budgeted for: exact-length convergence crashed at
question_len == 1 (the single suffix token could not be trimmed to absorb
it), and the prefix region re-tokenized differently across a group, causing
prefix-cache misses.

Compose at the token level instead, building decode(prefix_ids + suffix
[…]
