+++ inference_perf/datagen/datagen_utils.py
@@ -62,48 +62,43 @@ def random_token_ids(rng: np.random.Generator, valid_token_ids: np.ndarray, leng
     return rng.choice(valid_token_ids, size=length).tolist()  # type: ignore[no-any-return]
 
 
+def build_word_start_token_ids(tokenizer: CustomTokenizer, valid_token_ids: np.ndarray) -> np.ndarray:
+    """Token IDs whose decoded form starts with whitespace.
+
+    Used to pin the first token of a suffix when appending it to a prefix:
+    a whitespace-prefixed token prevents BPE merges across the boundary,
+    so ``len(decode(prefix_ids + suffix_ids))`` retokenizes exactly and the
+    prefix's server-side tokens stay stable. Falls back to the full vocab
+    when no such tokens exist.
+    """
+    hf_tokenizer = tokenizer.get_tokenizer()
+    word_starts: List[int] = []
+    for tid in valid_token_ids:
+        decoded = hf_tokenizer.decode([int(tid)], skip_special_tokens=True)
+        if decoded and decoded[0].isspace():
+            word_starts.append(int(tid))
+    if not word_starts:
+        return valid_token_ids
+    return np.array(word_starts, dtype=np.int64)
+
+
 def converge_to_exact_length_text(
     tokenizer: CustomTokenizer,
     target_len: int,
-    prefix_text: str,
     initial_tokens: List[int],
     adjust_tokens_fn: Callable[[List[int], int, int], List[int]],
-) -> str:
-    """Generates a string that tokenizes to exactly target_len, optionally prefixed.
+) -> Tuple[str, List[int]]:
+    """Generate text tokenizing to exactly target_len; return (text, ids).
 
-    Args:
-        tokenizer: The custom tokenizer.
-        target_len: The target token length.
-        prefix_text: Optional prefix to include in the length calculation.
-        initial_tokens: The initial list of token IDs to start with.
-        adjust_tokens_fn: A callback function to adjust the token list when the
-          length doesn't match. It takes (current_tokens, current_len,
-          target_len) and returns new_tokens.
-
-    Raises:
-        ValueError: If we cannot land on exactly `target_len` within the
-          iteration budget. Most often this happens when the tokenizer's BPE
-          merges around the prefix/suffix boundary keep flipping the count by
-          more than one token per adjustment, or when the requested length is
-          near the tokenizer's vocabulary edge cases (e.g. very small
-          target_len with a tokenizer that always prepends a BOS). Mitigation:
-          try a slightly different `target_len`, verify the tokenizer config
-          matches the model server's, or pre-tokenize the prefix once and
-          adjust around it instead of regenerating each iteration.
+    ``adjust_tokens_fn`` takes ``(current_tokens, current_len, target_len)``
+    and returns new tokens. The returned ids let callers compose with another
+    chunk at the token level instead of by string concat (which would
+    re-tokenize across the boundary).
     """
     if target_len <= 0:
-        return ""
+        return "", []
 
     hf_tokenizer = tokenizer.get_tokenizer()
-
-    prefix_len = 0
-    if prefix_text:
-        prefix_len = tokenizer.count_tokens(prefix_text)
-        assert prefix_len < target_len, (
-            f"target_len ({target_len}) must be > prefix_len ({prefix_len}). "
-            f"This helper generates suffix tokens such that prefix_len + suffix_len == target_len; "
-            f"the caller is responsible for choosing target_len with room for a non-empty suffix."
-        )
 
     current_tokens = initial_tokens
 
@@ -112,23 +107,18 @@ def converge_to_exact_length_text(
     for _ in range(max_iterations):
         text = hf_tokenizer.decode(current_tokens, skip_special_tokens=True)
 
[…]
