+++ inference_perf/datagen/shared_prefix_datagen.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 import logging
-from typing import Dict, Generator, List, Optional, Union
+from typing import Dict, Generator, List, Optional, Tuple, Union
 
 import numpy as np
 
@@ -38,7 +38,13 @@ from inference_perf.utils.custom_tokenizer import CustomTokenizer
 from inference_perf.utils.distribution import sample_from_distribution
 
 from .base import DataGenerator, LazyLoadDataMixin
-from .datagen_utils import generate_random_exact_length_text, init_vocab_sampling, random_token_ids
+from .datagen_utils import (
+    build_word_start_token_ids,
+    converge_to_exact_length_text,
+    generate_random_exact_length_text,
+    init_vocab_sampling,
+    random_token_ids,
+)
 
 logger = logging.getLogger(__name__)
 
@@ -67,6 +73,7 @@ class SharedPrefixDataGenerator(DataGenerator, LazyLoadDataMixin):
             raise ValueError("Tokenizer is required for SharedPrefixDataGenerator but was not initialized.")
 
         self.vocab_size, self.special_token_ids, self.valid_token_ids = init_vocab_sampling(self.tokenizer)
+        self.word_start_token_ids = build_word_start_token_ids(self.tokenizer, self.valid_token_ids)
 
         if self.shared_prefix is None:
             raise ValueError("Shared Prefix config is required for SharedPrefixDataGenerator")
@@ -184,15 +191,45 @@ class SharedPrefixDataGenerator(DataGenerator, LazyLoadDataMixin):
         """Generates a list of random token IDs of a specified length."""
         return random_token_ids(self.rng, self.valid_token_ids, length)
 
-    def _generate_exact_length_text(self, target_len: int, prefix_text: str = "") -> str:
-        """Generates a string that tokenizes to exactly target_len, optionally prefixed.
+    def _sample_suffix_ids(self, length: int) -> List[int]:
+        """Sample suffix IDs decoding to exactly ``length`` tokens, first token word-start.
 
-        If prefix_text is provided, the TOTAL length including prefix will be target_len.
-        Returns the full combined text if prefix_text is provided, else just the generated text.
+        The first token is word-start so the suffix decodes with a leading whitespace
+        character, preventing BPE merges across the prefix/suffix boundary. The
+        convergence loop then settles the rest of the suffix to the exact count
+        regardless of random-token roundtrip drift.
         """
+        if length <= 0:
+            return []
+        if self.tokenizer is None:
+            raise ValueError("Tokenizer is required for sampling suffix IDs.")
+        initial = [int(self.rng.choice(self.word_start_token_ids))]
+        if length > 1:
+            initial += random_token_ids(self.rng, self.valid_token_ids, length - 1)
+
+        def adjust(current: List[int], current_len: int, target_len: int) -> List[int]:
+            if current_len < target_len:
+                current.extend(random_token_ids(self.rng, self.valid_token_ids, target_len - current_len))
+                return current
+            diff = current_len - target_len
+            # Preserve the word-start first token so the boundary stays stable.
+            if diff < len(current) - 1:
+                return current[:-diff]
+            return current[:1]
+
+        _, ids = converge_to_exact_length_text(
+            tokenizer=self.tokenizer,
+            target_len=length,
+            initial_tokens=initial,
+            adjust_tokens_fn=adjust,
+        )
+        return ids
+
+    def _generate_exact_length_text(self, target_len: int) -> Tuple[str, List[int]]:
+        """Generates a string + its underlying token IDs, tokenizing to exactly target_len."""
         if self.tokenizer is None:
[…]
