### 2026-05-27 40d6daf
Address max-model-len truncation CPU usage
