+++ inference_perf/apis/user_session.py
@@ -31,6 +31,7 @@ class LocalUserSession:
     system_prompt: str
     max_model_len: Optional[int]
     history: list[str]
+    history_tokens: list[int]
 
     _instances: dict[str, "LocalUserSession"] = {}
 
@@ -48,6 +49,7 @@ class LocalUserSession:
         self.tokenizer = tokenizer
         self.max_model_len = max_model_len
         self.history = []
+        self.history_tokens = []
         self._current_round = 0
         self._in_flight: Optional[asyncio.Lock] = None
         self._waiting_rounds: Optional[asyncio.Queue[asyncio.Future[bool]]] = None
@@ -81,7 +83,7 @@ class LocalUserSession:
         self._current_round += 1
         return self.context
 
-    def update_context(self, response: str) -> None:
+    def update_context(self, response: str, response_len: Optional[int] = None) -> None:
         if self.system_prompt and self.tokenizer and self.max_model_len:
             history_context = " ".join(self.history) if self.history else ""
             base_len = len(self.system_prompt)
@@ -90,18 +92,19 @@ class LocalUserSession:
             turn_content = response[base_len:].strip()
             if turn_content:
                 self.history.append(turn_content)
+                if response_len is not None:
+                    self.history_tokens.append(response_len)
+                else:
+                    self.history_tokens.append(self.tokenizer.count_tokens(turn_content))
 
-            history_str = " ".join(self.history)
             system_tokens = self.tokenizer.count_tokens(self.system_prompt)
-            history_tokens = self.tokenizer.count_tokens(history_str)
+            total_tokens = system_tokens + sum(self.history_tokens)
 
-            if system_tokens + history_tokens > self.max_model_len:
-                while self.history:
-                    history_str = " ".join(self.history)
-                    history_tokens = self.tokenizer.count_tokens(history_str)
-                    if system_tokens + history_tokens <= self.max_model_len:
-                        break
+            if total_tokens > self.max_model_len:
+                while self.history and total_tokens > self.max_model_len:
                     self.history.pop(0)
+                    popped_len = self.history_tokens.pop(0)
+                    total_tokens -= popped_len
 
             self.context = (
                 self.system_prompt + " " + " ".join(self.history)
@@ -133,6 +136,7 @@ class UserSessionCompletionAPIData(CompletionAPIData):
     model_config = ConfigDict(arbitrary_types_allowed=True)
     user_session_id: str = Field(exclude=True)
     target_round: int
+    prompt_len: Optional[int] = None
 
     @property
     def user_session(self) -> LocalUserSession:
@@ -150,32 +154,32 @@ class UserSessionCompletionAPIData(CompletionAPIData):
 
             system_prompt = self.user_session.system_prompt
             history = list(self.user_session.history)
+            history_tokens = list(self.user_session.history_tokens)
+            if len(history_tokens) != len(history):
+                history_tokens = [self.user_session.tokenizer.count_tokens(h) for h in history]
+                self.user_session.history_tokens = list(history_tokens)
             current_prompt = self.prompt
 
-            def get_text(sys: str, hist: list[str], curr: str) -> str:
-                parts = []
-                if sys:
-                    parts.append(sys)
-                if hist:
-                    parts.append(" ".join(hist))
-                if curr:
-                    parts.append(curr)
-                return " ".join(parts)
+            system_tokens = self.user_session.tokenizer.count_tokens(system_prompt)
+            
+            if self.prompt_len is not None:
+                current_prompt_tokens = self.prompt_len
+            else:
+                current_prompt_tokens = self.user_session.tokenizer.count_tokens(current_prompt)
 
-            combined_text = get_text(system_prompt, history, current_prompt)
[…]
