### 2026-05-27 ad65fe0
fix: shared_prefix off-by-one by composing at token level (#491)
## Summary

Fixes #490 and a related prefix-cache stability hazard. Both share one
root cause: prompts were composed by string concatenation (`prefix_text
+ " " + suffix_text`) and then re-tokenized, so the prefix/suffix
boundary did not tokenize as `len(prefix_ids) + len(suffix_ids)`.

**Symptom 1 (loud, #490)**: `shared_prefix` raised `ValueError: Could
not generate a prompt of exactly N tokens after 20 attempts (got N+1)`
whenever the sampled question length was 1. The convergence target is
additive (`sys_prompt_len + question
[…]
