### 2026-05-28 9dc473f
Updating paper charts and description
Signed-off-by: Sachin Varghese <sachin.mathew31@gmail.com>
