### 2026-05-28 497a8b2
Updating chart font size
Signed-off-by: Sachin Varghese <sachin.mathew31@gmail.com>
