+++ inference_perf/analysis/analyze.py
@@ -40,7 +40,11 @@ def _extract_throughput_metric(throughput_data: Dict[str, Any], metric_name: str
 
 
 def _generate_multi_plot(
-    chartset_to_generate: List[List[Dict[str, Any]]], num_charts: int, names: List[str], suptitle: str, output_path: Path
+    chartset_to_generate: List[List[Dict[str, Any]]],
+    num_charts: int,
+    names: List[str],
+    suptitle: str,
+    output_path: Path,
 ) -> None:
     """Generates and saves a plot with multiple subplots."""
     import matplotlib.pyplot as plt
@@ -54,7 +58,7 @@ def _generate_multi_plot(
         return
 
     fig, axes = plt.subplots(1, num_charts, figsize=(7 * num_charts, 6), squeeze=False)
-    fig.suptitle(suptitle, fontsize=16)
+    fig.suptitle(suptitle, fontsize=18)
 
     for charts_to_generate in chartset_to_generate:
         if not charts_to_generate:
@@ -67,9 +71,9 @@ def _generate_multi_plot(
             y_values = [x[1] for x in data]
 
             ax.plot(qps_values, y_values, marker="o", linestyle="-")
-            ax.set_title(chart_info["title"])
-            ax.set_xlabel(chart_info.get("xlabel", "QPS (requested rate)"))
-            ax.set_ylabel(chart_info["ylabel"])
+            ax.set_title(chart_info["title"], fontsize=16)
+            ax.set_xlabel(chart_info.get("xlabel", "QPS (requested rate)"), fontsize=14)
+            ax.set_ylabel(chart_info["ylabel"], fontsize=14)
             if not names:
                 ax.legend([f"Report {idx + 1}" for idx in range(len(chartset_to_generate))])
             else:
@@ -92,7 +96,7 @@ def _generate_plot(charts_to_generate: List[Dict[str, Any]], suptitle: str, outp
 
     num_charts = len(charts_to_generate)
     fig, axes = plt.subplots(1, num_charts, figsize=(7 * num_charts, 6), squeeze=False)
-    fig.suptitle(suptitle, fontsize=16)
+    fig.suptitle(suptitle, fontsize=18)
 
     for i, chart_info in enumerate(charts_to_generate):
         ax = axes[0, i]
@@ -101,9 +105,9 @@ def _generate_plot(charts_to_generate: List[Dict[str, Any]], suptitle: str, outp
         y_values = [x[1] for x in data]
 
         ax.plot(qps_values, y_values, marker="o", linestyle="-")
-        ax.set_title(chart_info["title"])
-        ax.set_xlabel(chart_info.get("xlabel", "QPS (requested rate)"))
-        ax.set_ylabel(chart_info["ylabel"])
+        ax.set_title(chart_info["title"], fontsize=16)
+        ax.set_xlabel(chart_info.get("xlabel", "QPS (requested rate)"), fontsize=14)
+        ax.set_ylabel(chart_info["ylabel"], fontsize=14)
         ax.grid(True)
 
     fig.tight_layout(rect=(0, 0.03, 1, 0.95))
@@ -195,7 +199,11 @@ def analyze_reports(report_dirs: List[str], analysis_dir: Optional[str] = None) 
                         else:
                             qps_vs_ttft.append((qps, ttft))
 
-                    ntpot = _extract_latency_metric(latency_data, "normalized_time_per_output_token", convert_to_ms=True)
+                    ntpot = _extract_latency_metric(
+                        latency_data,
+                        "normalized_time_per_output_token",
+                        convert_to_ms=True,
+                    )
                     if ntpot is not None:
                         if concurrency:
                             concurrency_vs_ntpot.append((concurrency, ntpot))
