### 2026-05-28 ea85724
paper verbiage update
Signed-off-by: Sachin Varghese <sachin.mathew31@gmail.com>
