+++ paper/paper.md
@@ -110,13 +110,15 @@ Price-performance metrics below are not directly reported by inference-perf but 
 - Price per million input tokens  
 - Throughput per dollar
 
-The above metrics can also be plotted into charts using the analyze command in the tool at various request rates (QPS) to understand how the latency and throughput scales with the load. The charts below demonstrates an analysis presented at KubeCon NA 2025 using `inference-perf` and details of the benchmark conditions can be found in the source repository [@inference-perf-demos].
+The above metrics can also be plotted into charts using the analyze command in the tool at various request rates (QPS) to understand how the latency and throughput scales with the load. The charts below demonstrates an analysis presented at KubeCon NA 2025 using `inference-perf`.
 
 ![Throughput vs QPS](assets/throughput_vs_qps.png)
 
 ![Latency vs QPS](assets/latency_vs_qps.png)
 
 ![Throughput vs Latency](assets/throughput_vs_latency.png)
+
+Further details of the above metrics analysis in terms of the inference server utilized, hardware specification and other benchmark conditions can be found in the demos section of the source repository [@inference-perf-demos].
 
 # Research Impact Statement
 
