### 2026-05-28 9e92d8b
test: add unit tests for the split config package
Adds 56 unit tests covering the config package introduced in #505:
top-level config read and deep-merge, common distributions, and
per-package validation for apis, loadgen, client (storage and
server_metrics), the datagen data block, and reportgen.
