+++ tests/required/config/datagen/test_data_config.py
@@ -1,0 +1,225 @@
+# Copyright 2026 The Kubernetes Authors.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+"""Validity rules for the ``DataConfig`` / ``SharedPrefix`` surface.
+
+These exercise the ``data:`` block of the config (``Config.data``).
+"""
+
+import os
+import tempfile
+
+import pytest
+import yaml
+
+from inference_perf.config import (
+    Config,
+    DataGenType,
+    Distribution,
+    DistributionType,
+    VideoProfile,
+    read_config,
+)
+
+
+# --- SharedPrefix aliases / distribution syntax --------------------------
+
+
+def test_shared_prefix_short_names() -> None:
+    config = Config.model_validate(
+        {
+            "data": {
+                "type": DataGenType.SharedPrefix,
+                "shared_prefix": {"num_groups": 5, "num_prompts_per_group": 20},
+            }
+        }
+    )
+    assert config.data.shared_prefix is not None
+    assert config.data.shared_prefix.num_groups == 5
+    assert config.data.shared_prefix.num_prompts_per_group == 20
+
+
+def test_shared_prefix_alias_names() -> None:
+    config = Config.model_validate(
+        {
+            "data": {
+                "type": DataGenType.SharedPrefix,
+                "shared_prefix": {"num_unique_system_prompts": 7, "num_users_per_system_prompt": 15},
+            }
+        }
+    )
+    assert config.data.shared_prefix is not None
+    assert config.data.shared_prefix.num_groups == 7
+    assert config.data.shared_prefix.num_prompts_per_group == 15
+
+
+def test_shared_prefix_serializes_with_aliases() -> None:
+    config = Config.model_validate(
+        {
+            "data": {
+                "type": DataGenType.SharedPrefix,
+                "shared_prefix": {"num_unique_system_prompts": 7, "num_users_per_system_prompt": 15},
+            }
+        }
+    )
+    dumped = config.model_dump(mode="json", by_alias=True)
+    shared_prefix_dump = dumped["data"]["shared_prefix"]
+    assert shared_prefix_dump["num_unique_system_prompts"] == 7
+    assert shared_prefix_dump["num_users_per_system_prompt"] == 15
+
+
+def test_shared_prefix_inline_distribution() -> None:
+    config = Config.model_validate(
+        {
+            "data": {
+                "type": DataGenType.SharedPrefix,
+                "shared_prefix": {
+                    "num_groups": 5,
+                    "question_len": {
+                        "type": "skew_normal",
+                        "mean": 200,
+                        "std_dev": 80,
+                        "skew": 2.5,
+                        "min": 10,
+                        "max": 2000,
+                    },
+                },
+            }
+        }
+    )
+    sp = config.data.shared_prefix
+    assert sp is not None
+    assert isinstance(sp.question_len, Distribution)
+    assert sp.question_len.type == DistributionType.SKEW_NORMAL
+    assert sp.question_len.mean == 200
+    assert sp.question_len.skew == 2.5
+
+
+def test_shared_prefix_fixed_int_unchanged() -> None:
+    config = Config.model_validate(
+        {
+            "data": {
+                "type": DataGenType.SharedPrefix,
[…]
