### 2026-05-29 8f1c22e
Paper update for charts
Signed-off-by: Sachin Varghese <sachin.mathew31@gmail.com>
