### 2026-05-29 87a61cc
minor chart text size updates
Signed-off-by: Sachin Varghese <sachin.mathew31@gmail.com>
