### 2026-05-29 75176e8
Analyze charts font size increase and paper update (#513)
This PR increases the analyze charts font sizes and updates the paper to
utilize charts from the kubecon demo.

---------

Signed-off-by: Sachin Varghese <sachin.mathew31@gmail.com>
