+++ paper/paper.bib
@@ -164,3 +164,13 @@
   pages = {446--459},
   doi = {10.1109/ISCA45697.2020.00045}
 }
+
+@misc{inference-perf-demos,
+  author = {{Inference Perf Contributors}},
+  title = {Inference Perf Contributors},
+  year = {2025},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  url = {https://github.com/kubernetes-sigs/inference-perf/tree/main/demos/kubecon-na-2025},
+  note = {Accessed: 2026-05-26}
+}
