### 2026-05-29 3546c2a
Add VisionArena-Chat dataset support
Adds a data generator for the public lmarena-ai/VisionArena-Chat image
dataset. The loader streams a bounded pool of rows, uses the first user turn
of each conversation as the prompt, and attaches the row's image(s) as
pre-encoded blocks. The dataset stores images undecoded, so encoded bytes are
forwarded verbatim (with optional downscaling and normalization of non
JPEG/PNG formats to PNG); no HuggingFace token is required.

Wires PreEncodedImageSpec into the chat materializer (previously a stub that
raised for image dataset-loader provenance), mirroring th
[…]
