+++ inference_perf/config/__init__.py
@@ -48,6 +48,7 @@ from inference_perf.config.datagen import (
     TraceFormat,
     VideoDatagenConfig,
     VideoProfile,
+    VisionArenaConfig,
     WeightedDuration,
     WeightedResolution,
     WeightedVideoProfile,
@@ -125,6 +126,7 @@ __all__ = [
     "TraceSessionReplayLoadStage",
     "VideoDatagenConfig",
     "VideoProfile",
+    "VisionArenaConfig",
     "WeightedDuration",
     "WeightedResolution",
     "WeightedVideoProfile",
