### 2026-05-31 cdd244d
Add K8s Slack invitation link (#510)
Fixed #483

Signed-off-by: ChengHao Yang <17496418+tico88612@users.noreply.github.com>
