### 2026-06-01 19e42bd
Support multi-tenant headers and OTel mapping
Introduce core capabilities for multi-tenant load generation in the
OTel Trace Replay engine. Enables multiplexing requests across tenant
identities and priority classes dynamically.

Why:
Currently, inference-perf does not support passing per-request
headers or labels from traces, preventing validation of llm-d-router
Flow Control multi-tenancy features under shared, saturated loads.

How:
- Extend InferenceAPIData and subclasses to store request-specific
  headers and labels.
- Dynamically merge custom headers in openai_client.py case-
  insensit
[…]
