+++ tests/test_multitenant_loadgen.py
@@ -1,0 +1,447 @@
+# Copyright 2026 The Kubernetes Authors.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+import json
+import http.server
+import threading
+import socket
+import sys
+import logging
+from pathlib import Path
+from typing import Any, ClassVar, Dict, Generator, List
+from unittest.mock import AsyncMock, MagicMock, patch
+
+import pytest
+
+# Ensure project root is on path
+sys.path.insert(0, str(Path(__file__).parent.parent))
+
+from inference_perf.apis import (
+    LazyLoadInferenceAPIData,
+    RequestLifecycleMetric,
+    InferenceInfo,
+    StreamedResponseMetrics,
+)
+from inference_perf.config import APIConfig, APIType, DataConfig, OTelTraceReplayConfig
+from inference_perf.datagen.otel_trace_replay_datagen import OTelTraceReplayDataGenerator
+from inference_perf.client.modelserver.openai_client import openAIModelServerClient, openAIModelServerClientSession
+from inference_perf.payloads import RequestMetrics, Text
+
+# Bypassed Hub Connection globally
+import inference_perf.client.modelserver.openai_client
+
+inference_perf.client.modelserver.openai_client.CustomTokenizer = MagicMock()  # type: ignore[attr-defined]
+
+
+class ConcreteOpenAIClient(openAIModelServerClient):
+    def get_supported_apis(self) -> List[APIType]:
+        return [APIType.Chat, APIType.Completion]
+
+    def get_prometheus_metric_metadata(self) -> Any:
+        return MagicMock()
+
+
+# ---------------------------------------------------------------------------
+# Mock HTTP Server Fixture
+# ---------------------------------------------------------------------------
+
+
+class MockHandler(http.server.BaseHTTPRequestHandler):
+    received_headers: ClassVar[Dict[str, str]] = {}
+    received_body: ClassVar[bytes | None] = None
+
+    def do_POST(self) -> None:
+        content_length = int(self.headers.get("Content-Length", 0))
+        body = self.rfile.read(content_length)
+        MockHandler.received_body = body
+
+        # Capture headers case-sensitively
+        MockHandler.received_headers = {k: v for k, v in self.headers.items()}
+
+        self.send_response(200)
+        self.send_header("Content-Type", "application/json")
+        self.end_headers()
+
+        # Return a mock OpenAI-like response
+        response = {
+            "choices": [{"message": {"role": "assistant", "content": "Mock response content"}}],
+            "usage": {"prompt_tokens": 5, "completion_tokens": 5},
+        }
+        self.wfile.write(json.dumps(response).encode("utf-8"))
+
+    def log_message(self, format: str, *args: Any) -> None:
+        pass  # Suppress logging
+
+
+@pytest.fixture
+def mock_server() -> Generator[str, None, None]:
+    # Find a free port
+    s = socket.socket()
+    s.bind(("", 0))
+    port = s.getsockname()[1]
+    s.close()
+
+    server = http.server.HTTPServer(("127.0.0.1", port), MockHandler)
+    thread = threading.Thread(target=server.serve_forever)
+    thread.daemon = True
+    thread.start()
+
[…]

[…]
