### 2026-06-01 c943f41
Split config file (#505)
NOTE:
* No functional changes
* No changes to tests

Splits the 792-line inference_perf/config.py into per-subsystem modules
under inference_perf/config/ (datagen, loadgen, reportgen,
client/{modelserver,filestorage,server_metrics}, metrics, apis, utils),
each as a config.py inside its own directory so it can grow into
siblings without restructuring later. This mirrors the runtime tree so
the config surface for a subsystem lives next to its other concerns.
Keeps the public import path stable via
inference_perf/config/__init__.py re-exports so no call sites had to
chang
[…]
