### 2026-06-02 f723342
enable filtering of hf dataset
# Conflicts:
#	inference_perf/config.py
