+++ inference_perf/main.py
@@ -12,12 +12,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 import json
-import multiprocessing as mp
-import sys
 from argparse import ArgumentParser
 
 import yaml
 from inference_perf.analysis.analyze import analyze_reports
+from inference_perf.utils.mp_context import MP_CONTEXT
 from typing import List, Optional
 from inference_perf.client.modelserver.tgi_client import TGImodelServerClient
 from inference_perf.datagen.base import BaseGenerator
@@ -115,14 +114,9 @@ class InferencePerfRunner:
 
 
 def main_cli() -> None:
-    # Set multiprocessing start method to 'fork' on macOS to avoid pickle issues
-    # This must be done before any multiprocessing operations
-    if sys.platform == "darwin":  # macOS
-        try:
-            mp.set_start_method("fork", force=True)
-        except RuntimeError:
-            # Start method already set, ignore
-            pass
+    # The load generator pins its own ``forkserver`` multiprocessing context
+    # (see inference_perf.utils.mp_context), so the process-global default start
+    # method is no longer relied upon and is intentionally left untouched here.
 
     # Parse command line arguments
     parser = ArgumentParser()
@@ -276,14 +270,15 @@ def main_cli() -> None:
         raise Exception("Load stages must be configured, or sweep must be configured")
 
     # Create multiprocessing manager for session replay datagens if needed.
-    # Must be created before workers are forked.
+    # Must be created before workers are started, and from the same context as
+    # the workers so its proxies are usable across the worker boundary.
     mp_manager = None
     if (
         config.data
         and config.data.type in (DataGenType.OTelTraceReplay, DataGenType.WekaTraceReplay, DataGenType.SyntheticAgentic)
         and config.load.num_workers > 0
     ):
-        mp_manager = mp.Manager()
+        mp_manager = MP_CONTEXT.Manager()
 
     datagen: BaseGenerator
     if config.data:
