### 2026-06-03 7052215
Start load generator workers with forkserver instead of fork
Python 3.14 deprecates fork() from a multi-threaded process (the main
process runs an asyncio event loop and executor threads); the child can
deadlock on locks held by threads that no longer exist after the fork. The
load generator forked workers from this multi-threaded process, which
emitted a DeprecationWarning on every run.

Switch workers to a single shared forkserver context
(inference_perf/utils/mp_context.py). The forkserver server process is
single-threaded, so the forks that create workers are safe. Unlike fork,
forkserver 
[…]
