### 2026-06-03 ee5bab5
Add absolute coverage metrics, artifacts, badge, and multiprocessing collection
Augments the coverage gate (scripts/check_coverage_regression.py) beyond the
delta-vs-main ratchet:

- Enable multiprocessing-aware coverage so subprocess code (the Worker request
  collector in inference_perf/loadgen/load_generator.py) reports real numbers
  instead of reading as uncovered (loadgen 41.6% -> 51.3%, total 63.3% -> 64.2%).
- Print absolute total coverage and enforce a single repo-wide floor
  (MIN_TOTAL_COVERAGE) in addition to the existing delta ratchet; both hard-fail.
- Emit coverage.json/coverage
[…]
