+++ scripts/check_coverage_regression.py
@@ -22,6 +22,18 @@ import argparse
 # Ensure PDM doesn't reuse the current virtualenv when we are trying to use another one (e.g. for baseline)
 os.environ["PDM_IGNORE_ACTIVE_VENV"] = "1"
 
+# Absolute repo-wide coverage floor. Set just below the current total so the
+# build fails if coverage regresses meaningfully in absolute terms, independent
+# of the delta-vs-main ratchet. Raise this as coverage improves.
+MIN_TOTAL_COVERAGE = 63.0
+
+# Coverage config (pyproject.toml) used for both the current and baseline runs.
+# Exporting COVERAGE_PROCESS_START lets coverage start measuring inside the
+# multiprocessing Workers (see [tool.coverage.run] concurrency=multiprocessing),
+# so subprocess code reports real numbers instead of reading as uncovered.
+COVERAGE_CONFIG = Path("pyproject.toml").absolute()
+os.environ["COVERAGE_PROCESS_START"] = str(COVERAGE_CONFIG)
+
 
 def get_coverage_data(report_path):
     if not Path(report_path).exists():
@@ -101,11 +113,46 @@ def generate_baseline(output_path: Path):
 
 
 def generate_current(output_path: Path):
-    """Generates coverage for the current branch/environment."""
+    """Generates coverage for the current branch/environment.
+
+    Also emits coverage.xml (Cobertura) as a CI artifact. Uses the same
+    --cov-config as the baseline run so multiprocessing-aware settings apply
+    consistently and the two reports are comparable.
+    """
     print("--- Generating current coverage ---")
-    cmd = f"pdm run pytest --cov=inference_perf --cov-report=json:{output_path.absolute()} tests/"
+    cmd = (
+        f"pdm run pytest --cov=inference_perf --cov-config={COVERAGE_CONFIG} "
+        f"--cov-report=json:{output_path.absolute()} --cov-report=xml:coverage.xml tests/"
+    )
     run_command(cmd)
-    print(f"✅ Current report generated: {output_path.name}")
+    print(f"✅ Current report generated: {output_path.name} (+ coverage.xml)")
+
+
+def _badge_color(coverage: float) -> str:
+    """Maps a coverage percentage to a shields.io badge color."""
+    for threshold, color in ((90, "brightgreen"), (80, "green"), (70, "yellowgreen"), (60, "yellow"), (50, "orange")):
+        if coverage >= threshold:
+            return color
+    return "red"
+
+
+def write_badge_endpoint(coverage: float, output_path: Path = Path("coverage-endpoint.json")):
+    """Writes a shields.io endpoint-badge JSON describing total coverage.
+
+    The README references this file on the 'badges' branch via
+    https://img.shields.io/endpoint?url=...coverage-endpoint.json so that
+    shields.io renders the badge (proxied by GitHub's camo) rather than linking a
+    raw SVG directly, which renders inconsistently.
+    """
+    payload = {
+        "schemaVersion": 1,
+        "label": "coverage",
+        "message": f"{coverage:.2f}%",
+        "color": _badge_color(coverage),
+    }
+    with open(output_path, "w") as f:
+        json.dump(payload, f)
+    print(f"✅ Badge endpoint written: {output_path.name} ({payload['message']}, {payload['color']})")
 
 
 def print_detailed_report(current_data, baseline_data):
@@ -186,20 +233,41 @@ def main():
     current_val = current_data["totals"]["percent_covered"]
     baseline_val = baseline_data["totals"]["percent_covered"]
 
+    # Emit the badge endpoint JSON for the current coverage (published to the
[…]
