+++ tests/required/client/modelserver/test_result_types.py
@@ -1,0 +1,19 @@
+from inference_perf.client.modelserver.base import GaugeResult, HistogramResult
+
+
+def test_gauge_result_as_summary() -> None:
+    """as_summary projects avg -> mean and keeps the four percentile keys."""
+    summary = GaugeResult(avg=1.0, median=2.0, p90=3.0, p99=4.0).as_summary()
+
+    assert summary == {"mean": 1.0, "median": 2.0, "p90": 3.0, "p99": 4.0}
+
+
+def test_histogram_result_as_summary_drops_per_second() -> None:
+    """HistogramResult inherits as_summary and is narrowed to the gauge keys.
+
+    The extra per_second field is not part of the per-metric report summary.
+    """
+    summary = HistogramResult(avg=1.0, median=2.0, p90=3.0, p99=4.0, per_second=5.0).as_summary()
+
+    assert summary == {"mean": 1.0, "median": 2.0, "p90": 3.0, "p99": 4.0}
+    assert "per_second" not in summary
