### 2026-06-03 0a57518
Move report projection onto GaugeResult.as_summary()
Replace the free _hist() helper in reportgen with a GaugeResult.as_summary()
method so the avg->mean projection lives next to the typed result instead of
being hand-unpacked at each call site. HistogramResult inherits it and stays
narrowed to the four gauge keys (per_second is not part of the per-metric
summary). Add unit tests for both.
