### 2026-06-03 6475c8d
Assemble ModelServerMetrics via typed iteration instead of reflection
Address review feedback to drop getattr/setattr in favor of strict typing:

- Add Metric.collect() so each metric owns its query execution + parse.
- Make BaseMetrics iterable via an explicit _iter_metrics() hook; OpenAIMetrics
  yields its named fields rather than walking __dict__.
- Build ModelServerMetrics by collecting {target_field: result} and validating
  through Pydantic, replacing the reflective setattr and the hand-rolled
  field-name guard. This enforces each field's declared result type at
  construction (the pri
[…]
