+++ tests/required/client/modelserver/test_result_types.py
@@ -1,4 +1,6 @@
-from inference_perf.client.modelserver.base import GaugeResult, HistogramResult
+from typing import List
+
+from inference_perf.client.modelserver.base import GaugeMetric, GaugeResult, HistogramResult
 
 
 def test_gauge_result_as_summary() -> None:
@@ -17,3 +19,19 @@ def test_histogram_result_as_summary_drops_per_second() -> None:
 
     assert summary == {"mean": 1.0, "median": 2.0, "p90": 3.0, "p99": 4.0}
     assert "per_second" not in summary
+
+
+def test_metric_collect_runs_queries_and_parses() -> None:
+    """collect() executes each of the metric's queries in order and parses the results."""
+    metric = GaugeMetric(target_field="kv_cache_usage", metric_name="vllm:kv_cache_usage_perc", filters=[])
+    seen_queries: List[str] = []
+
+    def execute(query: str) -> float:
+        seen_queries.append(query)
+        return float(len(seen_queries))  # 1.0, 2.0, 3.0, 4.0 in query order
+
+    result = metric.collect(execute, duration=30)
+
+    assert seen_queries == metric.get_queries(30)
+    assert isinstance(result, GaugeResult)
+    assert (result.avg, result.median, result.p90, result.p99) == (1.0, 2.0, 3.0, 4.0)
