+++ tests/required/client/metricsclient/test_prometheus_client.py
@@ -1,9 +1,11 @@
-from typing import Any, List
+from typing import Any, Iterator, List
+import pytest
+from pydantic import ValidationError
 from unittest.mock import patch
 from inference_perf.client.server_metrics.prometheus_client.base import PrometheusMetricsClient
 from inference_perf.client.server_metrics.base import ModelServerMetrics
 from inference_perf.config import PrometheusClientConfig
-from inference_perf.client.modelserver.base import BaseMetrics, HistogramMetric, Metric
+from inference_perf.client.modelserver.base import BaseMetrics, CounterResult, HistogramMetric, Metric
 
 
 def test_get_model_server_metrics_base_metrics() -> None:
@@ -12,11 +14,9 @@ def test_get_model_server_metrics_base_metrics() -> None:
     client = PrometheusMetricsClient(config)
 
     class FakeBaseMetrics(BaseMetrics):
-        def get_all_metrics(self) -> List[Metric[Any]]:
-            return [
-                HistogramMetric("inter_token_latency", "fake:itl", []),
-                HistogramMetric("time_per_output_token", "fake:tpot", []),
-            ]
+        def _iter_metrics(self) -> Iterator[Metric[Any]]:
+            yield HistogramMetric("inter_token_latency", "fake:itl", [])
+            yield HistogramMetric("time_per_output_token", "fake:tpot", [])
 
     def mock_execute(query: str, eval_time: str) -> float:
         if "fake:itl" in query and "_sum" in query and "/" in query:
@@ -45,3 +45,29 @@ def test_get_model_server_metrics_uses_custom_metrics_by_default() -> None:
 
     assert isinstance(result, ModelServerMetrics)
     assert result.inter_token_latency.avg == 1.23
+
+
+def test_get_model_server_metrics_rejects_wrong_result_type() -> None:
+    """A metric whose result type does not match its target field fails at construction.
+
+    This is the strictness the typed assembly buys over the old setattr: the result
+    is validated against the field's declared type instead of being written blindly.
+    """
+    config = PrometheusClientConfig(url="http://localhost:9090")
+    client = PrometheusMetricsClient(config)
+
+    class WrongTypeMetric(Metric[CounterResult]):
+        metric_name = "fake:wrong"
+        target_field = "request_latency"  # declared as HistogramResult on ModelServerMetrics
+
+        def get_queries(self, duration: float) -> List[str]:
+            return ["q"]
+
+        def parse(self, results: List[float]) -> CounterResult:
+            return CounterResult(value=1.0)
+
+    metadata = BaseMetrics(custom_metrics=[WrongTypeMetric()])
+
+    with patch.object(PrometheusMetricsClient, "execute_query", return_value=1.0):
+        with pytest.raises(ValidationError):
+            client.get_model_server_metrics(metadata, query_duration=30, query_eval_time=100)
