### 2026-06-03 5cbfb21
Make Reasoning YAML output dist consistent with config.json and README.md
Signed-off-by: Nick Masluk <nick@randombytes.net>
