### 2026-06-03 4c45cfe
Restrict builtins in filter expression eval()
