### 2026-06-03 369bf46
Split modelserver metrics into a dedicated metrics package
Move the metric implementations out of modelserver/base.py into a new
modelserver/metrics package with one implementation per metric type
(gauge, histogram, counter, requests, custom), each under its own
subpackage base.py. base.py now holds only the ModelServerClient and
ModelServerClientSession abstractions and imports BaseMetrics for typing.

Pure move plus import rewrite; no behavior change.
