+++ inference_perf/client/modelserver/base.py
@@ -12,198 +12,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 from abc import ABC, abstractmethod
-from typing import Any, Callable, Dict, Generic, Iterator, List, Optional, Tuple, TypeVar
-from pydantic import BaseModel
+from typing import List, Optional, Tuple
 from inference_perf.config import APIConfig, APIType
 from inference_perf.apis import InferenceAPIData
-
-
-class GaugeResult(BaseModel):
-    avg: float = 0.0
-    median: float = 0.0
-    p90: float = 0.0
-    p99: float = 0.0
-
-    def as_summary(self) -> Dict[str, float]:
-        """Project to the report's mean/median/p90/p99 shape.
-
-        HistogramResult inherits this and is intentionally narrowed to the same
-        four keys (its per_second field is not part of the per-metric summary).
-        """
-        return {"mean": self.avg, "median": self.median, "p90": self.p90, "p99": self.p99}
-
-
-class HistogramResult(GaugeResult):
-    per_second: float = 0.0
-
-
-class RequestsResult(BaseModel):
-    total: float = 0.0
-    per_second: float = 0.0
-
-
-class CounterResult(BaseModel):
-    value: float = 0.0
-
-
-R = TypeVar("R", bound=BaseModel)
-
-
-class Metric(ABC, Generic[R]):
-    metric_name: str
-    target_field: str
-
-    @abstractmethod
-    def get_queries(self, duration: float) -> List[str]:
-        """Returns the ordered list of PromQL queries this metric requires."""
-        ...
-
-    @abstractmethod
-    def parse(self, results: List[float]) -> R:
-        """Convert the ordered query results into a typed result object."""
-        ...
-
-    def collect(self, execute: Callable[[str], float], duration: float) -> R:
-        """Run this metric's queries via execute and parse them into its typed result.
-
-        Keeps query execution and parsing together on the metric so callers never
-        need to know the query/result shape of a particular metric type.
-        """
-        return self.parse([execute(query) for query in self.get_queries(duration)])
-
-
-class GaugeMetric(Metric[GaugeResult]):
-    def __init__(self, target_field: str, metric_name: str, filters: List[str]) -> None:
-        self.target_field = target_field
-        self.metric_name = metric_name
-        self.filters = ",".join(filters)
-
-    def get_queries(self, duration: float) -> List[str]:
-        f, m = self.filters, self.metric_name
-        return [
-            f"avg_over_time({m}{{{f}}}[{duration:.0f}s])",
-            f"quantile_over_time(0.5, {m}{{{f}}}[{duration:.0f}s])",
-            f"quantile_over_time(0.9, {m}{{{f}}}[{duration:.0f}s])",
-            f"quantile_over_time(0.99, {m}{{{f}}}[{duration:.0f}s])",
-        ]
-
-    def parse(self, results: List[float]) -> GaugeResult:
-        return GaugeResult(avg=results[0], median=results[1], p90=results[2], p99=results[3])
-
-
-class CounterMetric(Metric[HistogramResult]):
-    """Returns HistogramResult (not CounterResult) because target fields like prompt_tokens may
-    be populated by either CounterMetric (vllm/sglang) or HistogramMetric (tgi). Counter fills
-    only avg+per_second; percentile fields default to 0."""
-
-    def __init__(self, target_field: str, metric_name: str, filters: List[str]) -> None:
-        self.target_field = target_field
-        self.metric_name = metric_name
-        self.filters = ",".join(filters)
-
-    def get_queries(self, duration: float) -> List[str]:
-        f, m = self.filters, self.metric_name
[…]
