+++ inference_perf/client/modelserver/metrics/__init__.py
@@ -1,0 +1,33 @@
+# Copyright 2026 The Kubernetes Authors.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+from .base import Metric, BaseMetrics
+from .gauge.base import GaugeResult, GaugeMetric
+from .histogram.base import HistogramResult, HistogramMetric
+from .counter.base import CounterResult, CounterMetric
+from .requests.base import RequestsResult, RequestsMetric
+from .custom.base import CustomMetric
+
+__all__ = [
+    "Metric",
+    "BaseMetrics",
+    "GaugeResult",
+    "GaugeMetric",
+    "HistogramResult",
+    "HistogramMetric",
+    "CounterResult",
+    "CounterMetric",
+    "RequestsResult",
+    "RequestsMetric",
+    "CustomMetric",
+]

[…]
