+++ inference_perf/client/modelserver/openai_client.py
@@ -29,7 +29,7 @@ from inference_perf.utils import CustomTokenizer
 from .base import ModelServerClient, ModelServerClientSession
 from .metrics import Metric, BaseMetrics
 from .otel_instrumentation import get_otel_instrumentation
-from typing import Iterator, List, Optional, Any, Dict
+from typing import Iterator, List, Optional, Any, Dict, Tuple
 import aiohttp
 import asyncio
 import json
@@ -45,15 +45,16 @@ logger = logging.getLogger(__name__)
 class OpenAIMetrics(BaseMetrics):
     def __init__(
         self,
+        filters: List[str],
         prompt_tokens: Metric[Any],
         output_tokens: Metric[Any],
         requests: Metric[Any],
         request_latency: Metric[Any],
         queue_length: Metric[Any],
         time_per_output_token: Metric[Any],
-        custom_metrics: Optional[List[Metric[Any]]] = None,
+        custom_metrics: Optional[Dict[str, Metric[Any]]] = None,
     ) -> None:
-        super().__init__(custom_metrics)
+        super().__init__(filters, custom_metrics)
         self.prompt_tokens = prompt_tokens
         self.output_tokens = output_tokens
         self.requests = requests
@@ -61,13 +62,13 @@ class OpenAIMetrics(BaseMetrics):
         self.queue_length = queue_length
         self.time_per_output_token = time_per_output_token
 
-    def _iter_metrics(self) -> Iterator[Metric[Any]]:
-        yield self.prompt_tokens
-        yield self.output_tokens
-        yield self.requests
-        yield self.request_latency
-        yield self.queue_length
-        yield self.time_per_output_token
+    def _iter_metrics(self) -> Iterator[Tuple[str, Metric[Any]]]:
+        yield "prompt_tokens", self.prompt_tokens
+        yield "output_tokens", self.output_tokens
+        yield "requests", self.requests
+        yield "request_latency", self.request_latency
+        yield "queue_length", self.queue_length
+        yield "time_per_output_token", self.time_per_output_token
         yield from super()._iter_metrics()
 
 
