### 2026-06-03 2c4ae4d
Drop redundant PYTHONPATH injection in e2e benchmark runner
The e2e tests that run inference_perf/main.py directly already set
PYTHONPATH=repo_root via extra_env (which overwrites this block), and
the tests using the installed inference-perf console script resolve the
package from site-packages and never needed it. Revert benchmark.py to
match main.
