+++ inference_perf/reportgen/base.py
@@ -395,9 +395,9 @@ def summarize_prometheus_metrics(metrics: ModelServerMetrics) -> ResponsesSummar
             "time_to_first_token": metrics.time_to_first_token.as_summary(),
             "time_per_output_token": metrics.time_per_output_token.as_summary(),
             "kv_cache_usage_percentage": metrics.kv_cache_usage.as_summary(),
-            "num_requests_swapped": {"mean": metrics.num_requests_swapped.value},
-            "num_preemptions_total": {"mean": metrics.num_preemptions_total.value},
-            "prefix_cache_hit_percent": {"mean": _ratio(metrics.prefix_cache_hits.value, metrics.prefix_cache_queries.value)},
+            "num_requests_swapped": {"mean": metrics.num_requests_swapped.total},
+            "num_preemptions_total": {"mean": metrics.num_preemptions_total.total},
+            "prefix_cache_hit_percent": {"mean": _ratio(metrics.prefix_cache_hits.total, metrics.prefix_cache_queries.total)},
             "inter_token_latency": metrics.inter_token_latency.as_summary(),
             "num_requests_running": {"mean": metrics.num_requests_running.avg},
             "request_queue_time": metrics.request_queue_time.as_summary(),
@@ -409,15 +409,15 @@ def summarize_prometheus_metrics(metrics: ModelServerMetrics) -> ResponsesSummar
             "request_max_num_generation_tokens": metrics.request_max_num_generation_tokens.as_summary(),
             "request_params_n": metrics.request_params_n.as_summary(),
             "request_params_max_tokens": metrics.request_params_max_tokens.as_summary(),
-            "request_success_count": metrics.request_success_count.value,
+            "request_success_count": metrics.request_success_count.total,
             "iteration_tokens": metrics.iteration_tokens.as_summary(),
-            "prompt_tokens_cached": metrics.prompt_tokens_cached.value,
-            "prompt_tokens_recomputed": metrics.prompt_tokens_recomputed.value,
+            "prompt_tokens_cached": metrics.prompt_tokens_cached.total,
+            "prompt_tokens_recomputed": metrics.prompt_tokens_recomputed.total,
             "external_prefix_cache_hit_percent": {
-                "mean": _ratio(metrics.external_prefix_cache_hits.value, metrics.external_prefix_cache_queries.value)
+                "mean": _ratio(metrics.external_prefix_cache_hits.total, metrics.external_prefix_cache_queries.total)
             },
-            "mm_cache_hit_percent": {"mean": _ratio(metrics.mm_cache_hits.value, metrics.mm_cache_queries.value)},
-            "corrupted_requests": metrics.corrupted_requests.value,
+            "mm_cache_hit_percent": {"mean": _ratio(metrics.mm_cache_hits.total, metrics.mm_cache_queries.total)},
+            "corrupted_requests": metrics.corrupted_requests.total,
             "request_prefill_kv_computed_tokens": metrics.request_prefill_kv_computed_tokens.as_summary(),
             "kv_block_idle_before_evict": metrics.kv_block_idle_before_evict.as_summary(),
             "kv_block_lifetime": metrics.kv_block_lifetime.as_summary(),
