### 2026-06-03 a59e46c
test(optional): add live-backend multimodal test tier with cluster matching
Formalizes the optional (live-backend) tests per #472. A `live` pytest
marker gates tests that need a real model server; default CI excludes them
via `-m "not live"`.

Hardware requirements are inferred from each case's manifest (its
nodeSelector is the single source of truth) and matched against clusters
passed via --kubeconfigs, live nodes only. A case skips when no cluster has
a matching node, and contending cases serialize onto the available nodes via
a per-(cluster, nodeSelector) slot semaphore, reusing hardware s
[…]
