+++ pyproject.toml
@@ -41,6 +41,7 @@ dev = [
     "hypothesis>=6.146.0",
     "pytest-asyncio>=1.3.0",
     "pytest-picked",
+    "types-PyYAML>=6.0.0",
 ]
 otel = [
     "opentelemetry-api>=1.30.0",
