+++ tests/optional/conftest.py
@@ -21,8 +21,12 @@ Flow per live test:
     -> hand the test a Cluster handle (kubeconfig + fresh namespace)
     -> on teardown, delete the namespace (frees the GPU) and release the slot.
 """
+
 from __future__ import annotations
 
+import json
+import os
+import subprocess
 import sys
 import uuid
 from collections.abc import Iterator
@@ -35,7 +39,11 @@ import pytest
 sys.path.insert(0, str(Path(__file__).parent))
 
 from harness import requirements, slots  # noqa: E402
-from harness.runner import Cluster  # noqa: E402
+from harness.runner import DEFAULT_IMAGE, Cluster  # noqa: E402
+
+# Namespace naming, shared by per-case creation and the orphan sweep.
+NAMESPACE_PREFIX = "inference-perf-e2e"
+_IMAGE_ENV = "INFERENCE_PERF_IMAGE"
 
 
 def pytest_addoption(parser: pytest.Parser) -> None:
@@ -45,13 +53,24 @@ def pytest_addoption(parser: pytest.Parser) -> None:
         help="Comma-separated kubeconfig paths; each is one candidate cluster. "
         "Empty means use the ambient kubeconfig as the single cluster.",
     )
+    parser.addoption(
+        "--image",
+        default="",
+        help=f"inference-perf image for the job (default: ${_IMAGE_ENV} or {DEFAULT_IMAGE}).",
+    )
+    parser.addoption(
+        "--sweep-orphan-namespaces",
+        action="store_true",
+        default=False,
+        help=f"Before running, delete leftover {NAMESPACE_PREFIX}-* namespaces from "
+        "killed prior runs. Off by default; unsafe alongside concurrent runs.",
+    )
 
 
 def pytest_configure(config: pytest.Config) -> None:
     config.addinivalue_line(
         "markers",
-        "live: requires a real model-server backend on a matching cluster "
-        "(excluded from default CI via -m 'not live').",
+        "live: requires a real model-server backend on a matching cluster (excluded from default CI via -m 'not live').",
     )
 
 
@@ -61,10 +80,32 @@ def kubeconfigs(request: pytest.FixtureRequest) -> list[str | None]:
     return [p.strip() for p in raw.split(",") if p.strip()] or [None]
 
 
+@pytest.fixture(scope="session")
+def image(request: pytest.FixtureRequest) -> str:
+    return str(request.config.getoption("--image")) or os.environ.get(_IMAGE_ENV, "") or DEFAULT_IMAGE
+
+
+@pytest.fixture(scope="session", autouse=True)
+def _sweep_orphan_namespaces(request: pytest.FixtureRequest, kubeconfigs: list[str | None]) -> None:
+    """Reclaim namespaces leaked by killed prior runs (opt-in via flag).
+
+    Per-case teardown already deletes namespaces on a clean exit; this covers
+    hard kills (SIGKILL, OOM, CI timeout) where teardown never ran.
+    """
+    if not request.config.getoption("--sweep-orphan-namespaces"):
+        return
+    for kubeconfig in kubeconfigs:
+        raw = _kubectl(kubeconfig, "get", "namespaces", "-o", "json", capture=True, check=False)
+        if not raw.strip():
+            continue
+        names = [ns.get("metadata", {}).get("name", "") for ns in json.loads(raw).get("items", [])]
+        orphans = [n for n in names if n.startswith(f"{NAMESPACE_PREFIX}-")]
+        for name in orphans:
+            _kubectl(kubeconfig, "delete", "namespace", name, "--wait=false", check=False)
+
+
 @pytest.fixture
-def cluster_for_case(
-    request: pytest.FixtureRequest, kubeconfigs: list[str | None]
[…]
