### 2026-06-03 d2cbdfe
test(optional): persist reports, sweep orphans, image override; drop e2e_test.sh
Removes manual/multimodal/e2e_test.sh now that test_multimodal.py covers the
same cases, and folds in the script's remaining capabilities:

- run_case writes per-case reports to cases/<case>/output/ (summary and
  stage-0 lifecycle metrics), matching the script's output layout
- --sweep-orphan-namespaces reclaims inference-perf-e2e-* namespaces left by
  killed runs (opt-in; per-case teardown already covers clean exits)
- --image / $INFERENCE_PERF_IMAGE overrides the job image

Also flattens the misnamed manual/ l
[…]
