+++ tests/optional/harness/runner.py
@@ -13,11 +13,11 @@
 # limitations under the License.
 """Run one multimodal case end-to-end against a cluster.
 
-Ports steps 1-4 of tests/optional/manual/multimodal/e2e_test.sh into Python so the
-pytest tier owns deploy, run, and verification. The namespace lifecycle and
-concurrency control live in the fixture (conftest.py); this module assumes the
-namespace already exists and does not delete it.
+Owns deploy, run, verification, and artifact extraction for a single case. The
+namespace lifecycle and concurrency control live in the fixture (conftest.py);
+this module assumes the namespace already exists and does not delete it.
 """
+
 from __future__ import annotations
 
 import json
@@ -35,7 +35,7 @@ class Cluster:
     manifest_path: Path
 
 
-_IMAGE = "quay.io/inference-perf/inference-perf:latest"
+DEFAULT_IMAGE = "quay.io/inference-perf/inference-perf:latest"
 _TEMPLATE_IMAGE = "quay.io/inference-perf/inference-perf:latest"
 _VLLM_DEPLOYMENT = "qwen3-32b-vllm-deployment"
 _CONFIG_MAP = "inference-perf-config"
@@ -60,17 +60,22 @@ def _kubectl(
         cmd[1:1] = ["-n", namespace]
     if kubeconfig:
         cmd[1:1] = ["--kubeconfig", kubeconfig]
-    result = subprocess.run(
-        cmd, input=stdin, capture_output=capture, text=True, check=check
-    )
+    result = subprocess.run(cmd, input=stdin, capture_output=capture, text=True, check=check)
     return result.stdout if capture else ""
 
 
 def _copy_hf_secret(kubeconfig: str | None, namespace: str) -> None:
     """Copy hf-secret from the default namespace into the case namespace."""
     raw = _kubectl(
-        kubeconfig, "default", "get", "secret", "hf-secret", "-o", "json",
-        capture=True, check=False,
+        kubeconfig,
+        "default",
+        "get",
+        "secret",
+        "hf-secret",
+        "-o",
+        "json",
+        capture=True,
+        check=False,
     )
     if not raw.strip():
         # Mirrors the bash warning: gated models will fail to start without it.
@@ -84,7 +89,7 @@ def _copy_hf_secret(kubeconfig: str | None, namespace: str) -> None:
     _kubectl(kubeconfig, namespace, "apply", "-f", "-", stdin=json.dumps(secret))
 
 
-def run_case(kubeconfig: str | None, namespace: str, case_dir: Path, image: str = _IMAGE) -> None:
+def run_case(kubeconfig: str | None, namespace: str, case_dir: Path, image: str = DEFAULT_IMAGE) -> None:
     """Deploy the model server, run inference-perf, and assert all requests succeeded.
 
     Raises AssertionError on zero successes or any failures; lets subprocess errors
@@ -99,14 +104,25 @@ def run_case(kubeconfig: str | None, namespace: str, case_dir: Path, image: str 
     if vllm.is_file():
         _kubectl(kubeconfig, namespace, "apply", "-f", str(vllm))
     _kubectl(
-        kubeconfig, namespace, "rollout", "status",
-        f"deployment/{_VLLM_DEPLOYMENT}", f"--timeout={_ROLLOUT_TIMEOUT}",
+        kubeconfig,
+        namespace,
+        "rollout",
+        "status",
+        f"deployment/{_VLLM_DEPLOYMENT}",
+        f"--timeout={_ROLLOUT_TIMEOUT}",
     )
 
     # Step 2: config.
     cm = _kubectl(
[…]
