### 2026-06-03 b22ae9d
Make streaming datagens forkserver-safe (picklable)
forkserver/spawn load-generator workers pickle the Worker process and its
datagen at start, unlike fork which clones memory without serializing. The
shareGPT/CNN-DailyMail/Billsum/InfinityInstruct datagens each hold a live HF
streaming iterator (iter/itertools.cycle over load_dataset(streaming=True)),
which is generator-backed and unpicklable, so .start() raised
'TypeError: cannot pickle generator object' (caught by the shareGPT e2e cases;
the other three were latent).

Add a StreamingDatasetMixin that drops the iterator in __getstate__ and r
[…]
