### 2026-06-04 4006896
test(optional): make live harness suite-agnostic, add text suite, fix default pytest run
Decouple the live e2e harness from multimodal so any suite can reuse it:

- runner.run_case no longer hardcodes the qwen3-32b Deployment; it waits on the
  Deployment(s) declared in the case manifest (new requirements.deployment_names),
  with an optional override. This is what lets a suite bring its own server.
- Add tests/optional/text/, a text-only worked-example suite (differently named
  Deployment, discovered from the manifest). A single discovery-based test_live.py
  globs <suite>/cases/<case>/vllm.
[…]
