+++ tests/test_requirements.py
@@ -20,12 +20,20 @@ is the whole point of splitting inference (testable anywhere) from matching
 
 from __future__ import annotations
 
+import sys
 import textwrap
 from pathlib import Path
 
 import pytest
 
-from harness import requirements
+# The live-tier harness package lives under tests/optional/ (it is e2e
+# infrastructure). Its requirement-inference logic is pure and unit-tested here,
+# so put that dir on the path to import `harness` by absolute name. Done in-file
+# rather than via a tests/conftest.py to avoid a second `conftest` module, which
+# `mypy --strict ./tests` rejects as a duplicate.
+sys.path.insert(0, str(Path(__file__).parent / "optional"))
+
+from harness import requirements  # noqa: E402
 
 
 def _write(tmp_path: Path, body: str) -> Path:
@@ -100,3 +108,43 @@ def test_conflicting_selectors_raise(tmp_path: Path) -> None:
 )
 def test_node_fits(selector: dict[str, str], labels: dict[str, str], fits: bool) -> None:
     assert requirements.node_fits(selector, labels) is fits
+
+
+def test_deployment_names_read_from_manifest_in_order(tmp_path: Path) -> None:
+    manifest = _write(
+        tmp_path,
+        """
+        apiVersion: apps/v1
+        kind: Deployment
+        metadata:
+          name: qwen3-text-vllm-deployment
+        ---
+        apiVersion: v1
+        kind: Service
+        metadata:
+          name: qwen3-text-vllm-service
+        ---
+        apiVersion: apps/v1
+        kind: Deployment
+        metadata:
+          name: sidecar-deployment
+        """,
+    )
+    # Only Deployments, in document order; the Service is ignored. This is what
+    # lets a suite name its server anything without the harness hardcoding it.
+    assert requirements.deployment_names(manifest) == ["qwen3-text-vllm-deployment", "sidecar-deployment"]
+
+
+def test_deployment_names_empty_when_none(tmp_path: Path) -> None:
+    manifest = _write(
+        tmp_path,
+        """
+        apiVersion: v1
+        kind: Pod
+        metadata:
+          name: standalone
+        spec:
+          containers: [{name: c, image: busybox}]
+        """,
+    )
+    assert requirements.deployment_names(manifest) == []
