### 2026-06-04 1493620
fix: preserve partial response body when a streaming request fails
When a streaming response broke partway through on an HTTP 200 (truncated
SSE stream, dropped connection, or a proxy that returns 200 then sends an
error page), parse_sse_stream discarded the bytes it had already received.
The OpenAI client only assigned response_content after process_response
returned, so on a parse failure the per-request report recorded an empty
response body, leaving 200-but-failed requests undiagnosable.

parse_sse_stream now re-raises failures as StreamInterruptedError, which
carries the raw bytes receive
[…]
