### 2026-06-05 dbe26ca
regenerate system prompts per stage in conversation replay (#480)
Fixes: #479

This PR updates the ConversationReplayDataGenerator to automatically
regenerate system prompts at the start of every new benchmark stage.
