### 2026-06-05 eca15d4
feat: add reasoning output support for OTEL trace replay (#499)
- Combine reasoning_content with output text in content field
- Store separate reasoning_content and output_content fields
- Support both streaming and non-streaming responses
- Fix streaming: use info.output_text instead of parsing empty
response_content
- Consistent behavior across replay and OTEL instrumentation
