+++ inference_perf/client/modelserver/openai_client.py
@@ -220,74 +220,71 @@ class openAIModelServerClientSession(ModelServerClientSession):
             # Extract input and output following GenAI semantic conventions
             try:
                 # Extract input based on request type
-                messages = getattr(data, "messages", None)
-                if messages is not None:
+                if hasattr(data, "messages") and data.messages:
                     # Serialize each message, preserving tool_calls when present.
-                    input_messages = [msg.to_dict() for msg in messages]
+                    input_messages = [msg.to_dict() for msg in data.messages]
                     otel_response_info["input_messages"] = json.dumps(input_messages)
 
                     # Record tool definitions so they appear in Jaeger alongside the request.
-                    tool_definitions = getattr(data, "tool_definitions", None)
-                    if tool_definitions:
-                        otel_response_info["tool_definitions"] = json.dumps(tool_definitions)
-                else:
-                    prompt = getattr(data, "prompt", None)
-                    if prompt is not None:
-                        # Text completion - store as prompt string (gen_ai.prompt)
-                        otel_response_info["input_prompt"] = prompt
+                    if hasattr(data, "tool_definitions") and data.tool_definitions:
+                        otel_response_info["tool_definitions"] = json.dumps(data.tool_definitions)
+                elif hasattr(data, "prompt"):
+                    # Text completion - store as prompt string (gen_ai.prompt)
+                    otel_response_info["input_prompt"] = data.prompt
 
                 # Extract output text (gen_ai.output.text)
-                if response and response.status == 200 and response_content:
-                    response_json = None
+                if self.client.api_config.streaming and response_content:
                     stripped_response = response_content.strip()
-
                     if stripped_response.startswith("data:"):
-                        sse_payloads = []
+                        output_parts = []
+                        reasoning_parts = []
+                        tool_call_parts = []
                         for line in stripped_response.splitlines():
                             line = line.strip()
                             if not line or not line.startswith("data:"):
                                 continue
-
                             payload = line[len("data:") :].strip()
                             if not payload or payload == "[DONE]":
                                 continue
-
                             try:
-                                sse_payloads.append(json.loads(payload))
+                                chunk = json.loads(payload)
                             except json.JSONDecodeError:
                                 continue
+                            choices = chunk.get("choices", [])
+                            if not choices:
+                                continue
+                            delta = choices[0].get("delta", {})
+                            if delta.get("content"):
+                                output_parts.append(delta["content"])
+                            if delta.get("reasoning_content"):
+                                reasoning_parts.append(delta["reasoning_content"])
+                            if delta.get("tool_calls"):
+                                tool_call_parts.append(delta)
 
-                        if sse_payloads:
-                            response_json = sse_payloads[-1]
-                            # Concatenate all content deltas for full output text
-                            full_text = "".join(
-                                p.get("choices", [{}])[0].get("delta", {}).get("content", "") for p in sse_payloads
-                            )
-                            if full_text:
-                                otel_response_info["output_text"] = full_text
-                            # Check last chunk for tool_calls
-                            last_delta = sse_payloads[-1].get("choices", [{}])[0].get("delta", {})
-                            if last_delta.get("tool_calls"):
-                                otel_response_info["output_message"] = json.dumps(last_delta)
-                    else:
-                        response_json = json.loads(response_content)
+                        if output_parts:
+                            otel_response_info["output_text"] = "".join(output_parts)
+                        if reasoning_parts:
+                            otel_response_info["reasoning_text"] = "".join(reasoning_parts)
+                        if tool_call_parts:
+                            otel_response_info["output_message"] = json.dumps(tool_call_parts)
+                elif response and response.status == 200 and response_content:
+                    response_json = json.loads(response_content)
+                    choices = response_json.get("choices", [])
+                    if choices:
[…]
