+++ tests/test_reasoning_output_support.py
@@ -1,0 +1,342 @@
+#!/usr/bin/env python3
+
+# Copyright 2025 The Kubernetes Authors.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""
+Tests for reasoning output support in OTEL trace replay.
+
+Exercises the actual production code in SessionChatCompletionAPIData.process_response
+to verify correct handling of reasoning_content from reasoning models.
+"""
+
+from __future__ import annotations
+
+import json
+from typing import Any, AsyncGenerator, Dict, List, Optional, cast
+from unittest.mock import AsyncMock, MagicMock
+
+import pytest
+from aiohttp import ClientResponse
+
+from inference_perf.apis.chat import ChatMessage
+from inference_perf.datagen.replay_graph_session_datagen import (
+    EventOutputRegistry,
+    SessionChatCompletionAPIData,
+    SessionInferenceInfo,
+    WorkerSessionTracker,
+)
+
+
+# ---------------------------------------------------------------------------
+# Helpers
+# ---------------------------------------------------------------------------
+
+
+def _make_tokenizer() -> MagicMock:
+    tok = MagicMock()
+    tok.count_tokens = lambda text: max(1, len((text or "").split()))
+    return tok
+
+
+def _make_config(streaming: bool = False) -> MagicMock:
+    cfg = MagicMock()
+    cfg.streaming = streaming
+    return cfg
+
+
+def _make_non_streaming_response(
+    content: str = "",
+    reasoning_content: Optional[str] = None,
+    tool_calls: Optional[List[Dict[str, Any]]] = None,
+) -> MagicMock:
+    message: Dict[str, Any] = {"role": "assistant", "content": content}
+    if reasoning_content is not None:
+        message["reasoning_content"] = reasoning_content
+    if tool_calls is not None:
+        message["tool_calls"] = tool_calls
+    response = MagicMock()
+    response.json = AsyncMock(return_value={"choices": [{"message": message}]})
+    return response
+
+
+class FakeStreamingResponse:
+    """Minimal aiohttp ClientResponse stand-in that yields preset SSE bytes."""
+
+    def __init__(self, chunks: List[bytes]) -> None:
+        self.status = 200
+        self.headers = {"content-type": "text/event-stream"}
+        self._chunks = chunks
+        self.content = self._make_content()
+
+    def _make_content(self) -> MagicMock:
+        content = MagicMock()
+
+        async def iter_any() -> AsyncGenerator[bytes, None]:
+            for chunk in self._chunks:
+                yield chunk
+
+        content.iter_any = iter_any
+        return content
+
+
+def _build_sse_stream(deltas: List[Dict[str, Any]], completion_tokens: Optional[int] = None) -> bytes:
+    """Build an OpenAI-style SSE byte stream from a list of delta dicts."""
+    parts: List[bytes] = []
+    for delta in deltas:
+        chunk = json.dumps({"choices": [{"delta": delta}]})
+        parts.append(f"data: {chunk}\n\n".encode())
+    if completion_tokens is not None:
+        usage_chunk = json.dumps({"choices": [], "usage": {"completion_tokens": completion_tokens}})
[…]
