### 2026-06-05 0984e29
Address JOSS paper reviewer feedback: clarify extensions sentence, correct online references, add Prometheus citation
