+++ paper/paper.bib
@@ -46,15 +46,15 @@
 
 @misc{k6,
   author = {{Grafana Labs}},
-  title = {{k6}: Open-source load testing tool},
-  year = {2021},
+  title = {{Grafana} {k6}: Load testing for engineering teams},
+  year = {2017},
   url = {https://k6.io/},
   note = {Accessed: 2026-01-08}
 }
 
 @misc{locust,
   author = {Jonatan Heyman and Carl Byström and Joakim Hamrén and Hugo Heyman and Lars Holmberg},
-  title = {Locust: An Open Source Load Testing Tool},
+  title = {{Locust}: A modern load testing framework},
   year = {2011},
   url = {https://locust.io/},
   note = {Accessed: 2026-01-08}
@@ -69,7 +69,7 @@
 
 @misc{tgi-benchmark,
   author = {{Hugging Face}},
-  title = {Text Generation Inference Benchmark Tool},
+  title = {{Text Generation Inference} Benchmarking Tool},
   year = {2023},
   url = {https://github.com/huggingface/text-generation-inference/tree/main/benchmark}
 }
@@ -84,7 +84,7 @@
 
 @misc{llm-d,
   author = {{llm-d Team}},
-  title = {{llm-d}: A {Kubernetes}-native High-performance Distributed {LLM} Inference Framework},
+  title = {{llm-d}: A high-performance and scalable distributed {LLM} inference framework},
   year = {2025},
   url = {https://llm-d.ai/},
   note = {Accessed: 2026-01-08}
@@ -94,38 +94,38 @@
   author = {{Linux Foundation}},
   title = {Generative {AI} Policy},
   year = {2023},
-  url = {https://linuxfoundation.org/legal/generative-ai-policy},
+  url = {https://www.linuxfoundation.org/legal/generative-ai},
   note = {Accessed: 2026-01-08}
 }
 
 @misc{dynamo,
   author = {{NVIDIA}},
-  title = {{NVIDIA} {Dynamo}},
-  year = {2024},
+  title = {{Dynamo}: A Datacenter Scale Distributed Inference Serving Framework},
+  year = {2025},
   url = {https://github.com/ai-dynamo/dynamo},
   note = {Accessed: 2026-01-08}
 }
 
 @misc{llm-d-kv-cache,
   author = {{llm-d Team}},
-  title = {{llm-d} {KV} Cache Aware Routing},
+  title = {{Inference-perf} Benchmark Report for {llm-d} well-lit scheduling strategies},
   year = {2025},
-  url = {https://github.com/llm-d/llm-d-kv-cache/blob/main/benchmarking/37-capacity/README.md},
+  url = {https://github.com/llm-d/llm-d-kv-cache/blob/9f03606d62dd4b4ff8dca9bcc323b6e8935a94ac/benchmarking/37-capacity/README.md},
   note = {Accessed: 2026-01-08}
 }
 
 @misc{inference-perf-contributors,
   author = {{Inference Perf Contributors}},
-  title = {{Inference} {Perf} Contributors},
-  year = {2024},
+  title = {Contributors to {kubernetes-sigs/inference-perf}},
+  year = {2025},
   url = {https://github.com/kubernetes-sigs/inference-perf/graphs/contributors},
   note = {Accessed: 2026-01-08}
 }
 
 @misc{inference-perf-metrics,
   author = {{Inference Perf Contributors}},
-  title = {{Inference} {Perf} Metrics},
-  year = {2024},
+  title = {{Inference Performance Metrics Definition}},
+  year = {2025},
   url = {https://github.com/kubernetes-sigs/inference-perf/blob/release-v0.3.0/docs/metrics.md},
[…]
