### 2026-06-05 6caf16b
Copy-editing of bibliography (#533)
* Remove definite article from institutional author names
* Force capitalization of proper names
* Remove GitHub from publisher and journal fields
* Use en dash for page ranges
