+++ paper/paper.bib
@@ -1,23 +1,21 @@
 @misc{kubernetes,
   title = {Kubernetes},
-  author = {{The Kubernetes Authors}},
+  author = {{Kubernetes Authors}},
   year = {2014},
   publisher = {Cloud Native Computing Foundation},
   url = {https://kubernetes.io/}
 }
 
 @misc{wg-serving,
-  author = {{The Kubernetes Authors}},
+  author = {{Kubernetes Authors}},
   title = {Kubernetes Serving Working Group},
   year = {2024},
-  publisher = {GitHub},
-  journal = {GitHub repository},
   url = {https://github.com/kubernetes/community/tree/master/archive/wg-serving}
 }
 
 @inproceedings{vllm,
   author = {Kwon, Woosuk and Li, Zhuohan and Zhuang, Siyuan and Sheng, Ying and Zheng, Lianmin and Yu, Cody Hao and Rostaing, Joseph E. and Zhang, Hao and Hendry, Gilbert and Stoica, Ion},
-  title = {vLLM: Easy, Fast, and Cheap LLM Serving with PagedAttention},
+  title = {{vLLM}: Easy, Fast, and Cheap {LLM} Serving with {PagedAttention}},
   year = {2023},
   isbn = {9798400702233},
   publisher = {Association for Computing Machinery},
@@ -25,11 +23,11 @@
   url = {https://doi.org/10.1145/3600006.3613165},
   doi = {10.1145/3600006.3613165},
   booktitle = {Proceedings of the 29th ACM Symposium on Operating Systems Principles},
-  pages = {611–626}
+  pages = {611--626}
 }
 
 @inproceedings{sglang,
-  title = {SGLang: Efficient Execution of Structured Language Model Programs},
+  title = {{SGLang}: Efficient Execution of Structured Language Model Programs},
   author = {Zheng, Lianmin and Yin, Liangsheng and Xie, Zhiqiang and Sun, Chuyue and Huang, Jeff and Yu, Cody Hao and Cao, Shiyi and Kozyrakis, Christos and Stoica, Ion and Gonzalez, Joseph E. and Barrett, Clark and Sheng, Ying},
   booktitle = {Advances in Neural Information Processing Systems 37},
   pages = {62557--62583},
@@ -43,14 +41,12 @@
   author = {{Hugging Face}},
   title = {Text Generation Inference},
   year = {2023},
-  publisher = {GitHub},
-  journal = {GitHub repository},
   url = {https://github.com/huggingface/text-generation-inference}
 }
 
 @misc{k6,
   author = {{Grafana Labs}},
-  title = {k6: Open-source load testing tool},
+  title = {{k6}: Open-source load testing tool},
   year = {2021},
   url = {https://k6.io/},
   note = {Accessed: 2026-01-08}
@@ -58,7 +54,7 @@
 
 @misc{locust,
   author = {Jonatan Heyman and Carl Byström and Joakim Hamrén and Hugo Heyman and Lars Holmberg},
-  title = {Locust: An open source load testing tool},
+  title = {Locust: An Open Source Load Testing Tool},
   year = {2011},
   url = {https://locust.io/},
   note = {Accessed: 2026-01-08}
@@ -66,10 +62,8 @@
 
 @misc{vllm-benchmark,
   author = {{vLLM Team}},
-  title = {vLLM Benchmarks},
+  title = {{vLLM} Benchmarks},
   year = {2023},
-  publisher = {GitHub},
-  journal = {GitHub repository},
   url = {https://github.com/vllm-project/vllm/tree/main/benchmarks}
 }
 
@@ -77,30 +71,28 @@
   author = {{Hugging Face}},
   title = {Text Generation Inference Benchmark Tool},
[…]
