### 2026-06-08 2219aee
Inject session identity header for session replay requests  (#504)
- Added configurable session_id_header_key to APIConfig 
- Injected session_id as an HTTP header in openai_client when
session_id_header_key is set
- Updated example configs with session_id_header_key under api: section
- Added unit tests verifying header injection and skip behavior

Resolves #492
