### 2026-06-08 b8360e2
Add Support for VisionArena Dataset (#525)
Addresses: #225.

Adds a data generator for the public lmarena-ai/VisionArena-Chat image
dataset. The loader streams a bounded pool of rows, uses the first user
turn of each conversation as the prompt, and attaches the row's image(s)
as pre-encoded blocks. The dataset stores images undecoded, so encoded
JPEG, PNG, and WEBP bytes are forwarded verbatim at their native
resolution (preserving the real payload size); images in other formats
are skipped rather than silently re-encoded, and no HuggingFace token is
required.

Wires PreEncodedImageSpec into t
[…]
