+++ tests/required/datagen/test_visionarena_datagen.py
@@ -1,0 +1,236 @@
+# Copyright 2026 The Kubernetes Authors.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+"""Loader behavior for the VisionArena-Chat data generator.
+
+``load_dataset`` is monkeypatched with an in-memory iterable mirroring the
+VisionArena schema (``conversation`` turns + undecoded ``images``), so these
+tests never touch the network.
+"""
+
+import io
+from typing import Any, List
+
+import pytest
+from PIL import Image as PILImage
+
+from inference_perf.apis.base import LazyLoadInferenceAPIData
+from inference_perf.apis.chat import ChatCompletionAPIData
+from inference_perf.config import Config, DataGenType
+from inference_perf.datagen import visionarena_datagen
+from inference_perf.datagen.visionarena_datagen import VisionArenaDataGenerator
+from inference_perf.payloads import ImageRepresentation, PreEncodedImageSpec
+
+
+def _png_bytes(w: int = 16, h: int = 8) -> bytes:
+    buf = io.BytesIO()
+    PILImage.new("RGB", (w, h), (200, 30, 30)).save(buf, format="PNG")
+    return buf.getvalue()
+
+
+def _jpeg_bytes(w: int = 16, h: int = 8) -> bytes:
+    buf = io.BytesIO()
+    PILImage.new("RGB", (w, h), (30, 30, 200)).save(buf, format="JPEG")
+    return buf.getvalue()
+
+
+def _webp_bytes(w: int = 16, h: int = 8) -> bytes:
+    buf = io.BytesIO()
+    PILImage.new("RGB", (w, h), (30, 200, 30)).save(buf, format="WEBP")
+    return buf.getvalue()
+
+
+def _gif_bytes(w: int = 10, h: int = 10) -> bytes:
+    buf = io.BytesIO()
+    PILImage.new("RGB", (w, h), (1, 2, 3)).save(buf, format="GIF")
+    return buf.getvalue()
+
+
+def _image_field(raw: bytes) -> dict[str, Any]:
+    """Mirror the undecoded HuggingFace Image form (decode=false)."""
+    return {"bytes": raw, "path": None}
+
+
+def _row(prompt: str, image_blobs: List[bytes]) -> dict[str, Any]:
+    # VisionArena declares conversation as List[List[{role, content}]] (each
+    # turn wrapped in a single-element list); mirror that real shape here.
+    return {
+        "conversation": [[{"role": "user", "content": prompt}]],
+        "images": [_image_field(b) for b in image_blobs],
+    }
+
+
+def _make_generator(monkeypatch: pytest.MonkeyPatch, rows: List[Any], **va_overrides: Any) -> VisionArenaDataGenerator:
+    monkeypatch.setattr(visionarena_datagen, "load_dataset", lambda *a, **k: list(rows))
+    config = Config.model_validate(
+        {"api": {"type": "chat"}, "data": {"type": DataGenType.VisionArena, "visionarena": va_overrides}}
+    )
+    return VisionArenaDataGenerator(config.api, config.data, None)
+
+
+def _materialize(gen: VisionArenaDataGenerator, idx: int = 0) -> ChatCompletionAPIData:
[…]
