### 2026-06-09 bbc3120
fix: handle StopIteration in data generators for long-running benchmarks
When running benchmarks with high request rates over long durations,
the ShareGPT and InfinityInstruct dataset iterators can be exhausted,
causing StopIteration crashes.

Changes:
- hf_sharegpt_datagen.py: wrap initial next() in try/except to handle
  empty datasets gracefully; add StopIteration guards in get_completion_data
  and get_chat_data to prevent silent crashes
- infinity_instruct_datagen.py: replace iter() with itertools.cycle()
  for automatic iterator recycling, matching the pattern used by
  ShareGPT and CNN/
[…]
