### 2026-06-09 0a85afd
otel_trace_replay: add bad_tool_call_handling=use_recorded for malformed tool_call args (#535)
When a model server's tool-call parser emits malformed JSON in
tool_calls[i].function.arguments — e.g. vLLM's `qwen3_xml` parser leaks
closing XML markers (`</parameter></function>`) into the JSON string at
decode time — vLLM still returns 200 on the response. But on the *next*
turn, replaying that response verbatim in messages[] makes the chat
template's `json.loads(arguments)` raise JSONDecodeError, and vLLM
returns HTTP 400. The session can no longer make progress.

This patch introduces a new kno
[…]
