+++ inference_perf/datagen/replay_graph_session_datagen.py
@@ -47,11 +47,54 @@ from inference_perf.apis.chat import ChatMessage
 from inference_perf.payloads import RequestMetrics, Text
 from inference_perf.apis.streaming_parser import parse_sse_stream
 from inference_perf.config import APIConfig, APIType, DataConfig, SessionReplayConfig
+from inference_perf.config.datagen.replay import BadToolCallHandling
 from inference_perf.datagen.base import LazyLoadDataMixin, SessionGenerator
 from inference_perf.datagen.replay_graph_types import InputSegment, ReplayGraph
 from inference_perf.utils.custom_tokenizer import CustomTokenizer
 
 logger = logging.getLogger(__name__)
+
+# --- bad_tool_call_handling ------------------------------------------------
+# Server-side tool-call parsers can emit malformed JSON in
+# tool_calls[i].function.arguments — for example vLLM's `qwen3_xml` parser
+# leaks closing XML markers (`</parameter></function>`) into the JSON string
+# value at decode time. vLLM still returns 200 on the response, but on the
+# *next* turn the chat template's `json.loads(arguments)` raises and vLLM
+# returns HTTP 400. Replaying the bad bytes verbatim therefore halts the
+# session.
+#
+# This mitigation lives ENTIRELY in the substitution path
+# (`_build_messages_with_substitution`). The response path is byte-identical
+# to upstream main: it stores the raw tool_calls in the registry exactly as
+# the model emitted them. At substitution time, when a downstream event
+# pulls a predecessor's stored message, we run `_detect_bad_tool_calls` on
+# its `tool_calls` and, if `bad_tool_call_handling=use_recorded`, substitute
+# the recorded assistant message at this slot.
+#
+# Gated per-event by the `bad_tool_call_handling` field on the OTel replay
+# config. When the value is `none` (the default), the inline detection is
+# short-circuited and behavior is identical to upstream main.
+
+
+def _detect_bad_tool_calls(
+    tool_calls: Optional[List[Dict[str, Any]]],
+) -> List[Tuple[int, str, str]]:
+    """Return [(index, function_name, json_error)] for tool_calls whose
+    `arguments` field is a string that fails json.loads(). Empty list = ok."""
+    bad: List[Tuple[int, str, str]] = []
+    if not tool_calls:
+        return bad
+    for i, tc in enumerate(tool_calls):
+        fn = tc.get("function") or {}
+        args = fn.get("arguments", "")
+        if not isinstance(args, str):
+            continue
+        try:
+            json.loads(args)
+        except json.JSONDecodeError as e:
+            bad.append((i, fn.get("name", "?"), str(e)))
+    return bad
+# --- end bad_tool_call_handling --------------------------------------------
 
 
 class EventFailedError(Exception):
@@ -89,6 +132,13 @@ class WorkerSessionTracker:
     def __init__(self) -> None:
         self._event_completions: Dict[str, Dict[str, float]] = {}
         self._failed_sessions: Set[str] = set()
+        # Per-session set of predecessor event_ids whose live tool_call
+        # response was detected malformed at substitution time and replaced
+        # with the recorded assistant message. Empty when
+        # bad_tool_call_handling is `none`. Stored as a set so a
+        # predecessor with multiple downstream consumers (DAG fan-out) is
+        # counted once, not once per consumer.
+        self._recorded_substitution_event_ids: Dict[str, Set[str]] = {}
 
     def record_event_completed(self, session_id: str, event_id: str, completion_time: float) -> None:
         if session_id not in self._event_completions:
@@ -112,6 +162,15 @@ class WorkerSessionTracker:
 
     def get_session_completion_times(self, session_id: str) -> Dict[str, float]:
[…]
