+++ docs/cli_flags.md
@@ -128,6 +128,7 @@ These command line flags are automatically generated from the internal `Config` 
 | `--data.otel_trace_replay.default_max_tokens` | int | Default max_tokens if not specified in trace |
 | `--data.otel_trace_replay.override_tool_call_max_tokens` | boolean | Override tool call max_tokens to 4096 instead of using trace recorded length |
 | `--data.otel_trace_replay.inject_random_session_id` | boolean | Inject random string into unique segments to invalidate KV-cache between sessions |
+| `--data.otel_trace_replay.bad_tool_call_handling` | Enum (none, use_recorded) | Mitigation for tool_calls whose function.arguments is not valid JSON. `none` (default): no mitigation. `use_recorded`: substitute the recorded assistant message at the affected slot. See docs/otel_trace_replay.md |
 | `--data.otel_trace_replay.duplicate_sessions_target` | int | Target number of sessions to reach by duplicating existing sessions. If None, no duplication occurs. |
 | `--data.otel_trace_replay.max_wait_ms` | int | Maximum inter-event wait time in milliseconds. Caps the delay between predecessor completion and event dispatch to avoid reproducing unusually long tool/agent execution times from the original trace. |
 | `--data.otel_trace_replay.include_errors` | boolean | Include spans with error status |
