### 2026-06-10 7ce3a4e
fix: RNG state synchronization across multiworkers
