### 2026-06-10 3c99ab5
pdm run validate
